% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationcostprofiler_operations.R
\name{applicationcostprofiler_get_report_definition}
\alias{applicationcostprofiler_get_report_definition}
\title{Retrieves the definition of a report already configured in AWS
Application Cost Profiler}
\usage{
applicationcostprofiler_get_report_definition(reportId)
}
\arguments{
\item{reportId}{[required] ID of the report to retrieve.}
}
\description{
Retrieves the definition of a report already configured in AWS Application Cost Profiler.

See \url{https://paws-r.github.io/docs/applicationcostprofiler/get_report_definition.html} for full documentation.
}
\keyword{internal}
