% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationcostprofiler_operations.R
\name{applicationcostprofiler_import_application_usage}
\alias{applicationcostprofiler_import_application_usage}
\title{Ingests application usage data from Amazon Simple Storage Service
(Amazon S3)}
\usage{
applicationcostprofiler_import_application_usage(sourceS3Location)
}
\arguments{
\item{sourceS3Location}{[required] Amazon S3 location to import application usage data from.}
}
\description{
Ingests application usage data from Amazon Simple Storage Service (Amazon S3).

See \url{https://paws-r.github.io/docs/applicationcostprofiler/import_application_usage.html} for full documentation.
}
\keyword{internal}
