% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationcostprofiler_operations.R
\name{applicationcostprofiler_list_report_definitions}
\alias{applicationcostprofiler_list_report_definitions}
\title{Retrieves a list of all reports and their configurations for your AWS
account}
\usage{
applicationcostprofiler_list_report_definitions(
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{The token value from a previous call to access the next page of results.}

\item{maxResults}{The maximum number of results to return.}
}
\description{
Retrieves a list of all reports and their configurations for your AWS account.

See \url{https://www.paws-r-sdk.com/docs/applicationcostprofiler_list_report_definitions/} for full documentation.
}
\keyword{internal}
