% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationcostprofiler_operations.R
\name{applicationcostprofiler_put_report_definition}
\alias{applicationcostprofiler_put_report_definition}
\title{Creates the report definition for a report in Application Cost Profiler}
\usage{
applicationcostprofiler_put_report_definition(
  reportId,
  reportDescription,
  reportFrequency,
  format,
  destinationS3Location
)
}
\arguments{
\item{reportId}{[required] Required. ID of the report. You can choose any valid string matching the
pattern for the ID.}

\item{reportDescription}{[required] Required. Description of the report.}

\item{reportFrequency}{[required] Required. The cadence to generate the report.}

\item{format}{[required] Required. The format to use for the generated report.}

\item{destinationS3Location}{[required] Required. Amazon Simple Storage Service (Amazon S3) location where
Application Cost Profiler uploads the report.}
}
\description{
Creates the report definition for a report in Application Cost Profiler.

See \url{https://www.paws-r-sdk.com/docs/applicationcostprofiler_put_report_definition/} for full documentation.
}
\keyword{internal}
