% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_create_log_pattern}
\alias{applicationinsights_create_log_pattern}
\title{Adds an log pattern to a LogPatternSet}
\usage{
applicationinsights_create_log_pattern(ResourceGroupName,
  PatternSetName, PatternName, Pattern, Rank)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{PatternSetName}{[required] The name of the log pattern set.}

\item{PatternName}{[required] The name of the log pattern.}

\item{Pattern}{[required] The log pattern.}

\item{Rank}{[required] Rank of the log pattern.}
}
\description{
Adds an log pattern to a \code{LogPatternSet}.
}
\section{Request syntax}{
\preformatted{svc$create_log_pattern(
  ResourceGroupName = "string",
  PatternSetName = "string",
  PatternName = "string",
  Pattern = "string",
  Rank = 123
)
}
}

\keyword{internal}
