% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_delete_application}
\alias{applicationinsights_delete_application}
\title{Removes the specified application from monitoring}
\usage{
applicationinsights_delete_application(ResourceGroupName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}
}
\description{
Removes the specified application from monitoring. Does not delete the application.

See \url{https://paws-r.github.io/docs/applicationinsights/delete_application.html} for full documentation.
}
\keyword{internal}
