% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_delete_log_pattern}
\alias{applicationinsights_delete_log_pattern}
\title{Removes the specified log pattern from a LogPatternSet}
\usage{
applicationinsights_delete_log_pattern(ResourceGroupName,
  PatternSetName, PatternName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{PatternSetName}{[required] The name of the log pattern set.}

\item{PatternName}{[required] The name of the log pattern.}
}
\value{
An empty list.
}
\description{
Removes the specified log pattern from a \code{LogPatternSet}.
}
\section{Request syntax}{
\preformatted{svc$delete_log_pattern(
  ResourceGroupName = "string",
  PatternSetName = "string",
  PatternName = "string"
)
}
}

\keyword{internal}
