% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_workload}
\alias{applicationinsights_describe_workload}
\title{Describes a workload and its configuration}
\usage{
applicationinsights_describe_workload(
  ResourceGroupName,
  ComponentName,
  WorkloadId,
  AccountId = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{WorkloadId}{[required] The ID of the workload.}

\item{AccountId}{The AWS account ID for the workload owner.}
}
\description{
Describes a workload and its configuration.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_workload/} for full documentation.
}
\keyword{internal}
