% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_log_patterns}
\alias{applicationinsights_list_log_patterns}
\title{Lists the log patterns in the specific log LogPatternSet}
\usage{
applicationinsights_list_log_patterns(
  ResourceGroupName,
  PatternSetName = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  AccountId = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{PatternSetName}{The name of the log pattern set.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}

\item{AccountId}{The Amazon Web Services account ID for the resource group owner.}
}
\description{
Lists the log patterns in the specific log \code{LogPatternSet}.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_list_log_patterns/} for full documentation.
}
\keyword{internal}
