% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_workloads}
\alias{applicationinsights_list_workloads}
\title{Lists the workloads that are configured on a given component}
\usage{
applicationinsights_list_workloads(
  ResourceGroupName,
  ComponentName,
  MaxResults = NULL,
  NextToken = NULL,
  AccountId = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}

\item{AccountId}{The AWS account ID of the owner of the workload.}
}
\description{
Lists the workloads that are configured on a given component.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_list_workloads/} for full documentation.
}
\keyword{internal}
