% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_remove_workload}
\alias{applicationinsights_remove_workload}
\title{Remove workload from a component}
\usage{
applicationinsights_remove_workload(
  ResourceGroupName,
  ComponentName,
  WorkloadId
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{WorkloadId}{[required] The ID of the workload.}
}
\description{
Remove workload from a component.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_remove_workload/} for full documentation.
}
\keyword{internal}
