% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_untag_resource}
\alias{applicationinsights_untag_resource}
\title{Remove one or more tags (keys and values) from a specified application}
\usage{
applicationinsights_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the application that you want to
remove one or more tags from.}

\item{TagKeys}{[required] The tags (tag keys) that you want to remove from the resource. When you
specify a tag key, the action removes both that key and its associated
tag value.

To remove more than one tag from the application, append the \code{TagKeys}
parameter and argument for each additional tag to remove, separated by
an ampersand.}
}
\value{
An empty list.
}
\description{
Remove one or more tags (keys and values) from a specified application.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
