% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_update_workload}
\alias{applicationinsights_update_workload}
\title{Adds a workload to a component}
\usage{
applicationinsights_update_workload(
  ResourceGroupName,
  ComponentName,
  WorkloadId = NULL,
  WorkloadConfiguration
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{WorkloadId}{The ID of the workload.}

\item{WorkloadConfiguration}{[required] The configuration settings of the workload. The value is the escaped
JSON of the configuration.}
}
\description{
Adds a workload to a component. Each component can have at most five workloads.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_update_workload/} for full documentation.
}
\keyword{internal}
