% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_associate_attribute_group}
\alias{appregistry_associate_attribute_group}
\title{Associates an attribute group with an application to augment the
application's metadata with the group's attributes}
\usage{
appregistry_associate_attribute_group(application, attributeGroup)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}

\item{attributeGroup}{[required] The name, ID, or ARN of the attribute group that holds the attributes to
describe the application.}
}
\description{
Associates an attribute group with an application to augment the application's metadata with the group's attributes. This feature enables applications to be described with user-defined details that are machine-readable, such as third-party integrations.

See \url{https://www.paws-r-sdk.com/docs/appregistry_associate_attribute_group/} for full documentation.
}
\keyword{internal}
