% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_associate_resource}
\alias{appregistry_associate_resource}
\title{Associates a resource with an application}
\usage{
appregistry_associate_resource(application, resourceType, resource)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}

\item{resourceType}{[required] The type of resource of which the application will be associated.}

\item{resource}{[required] The name or ID of the resource of which the application will be
associated.}
}
\description{
Associates a resource with an application. The resource can be specified by its ARN or name. The application can be specified by ARN, ID, or name.

See \url{https://www.paws-r-sdk.com/docs/appregistry_associate_resource/} for full documentation.
}
\keyword{internal}
