% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_delete_application}
\alias{appregistry_delete_application}
\title{Deletes an application that is specified either by its application ID,
name, or ARN}
\usage{
appregistry_delete_application(application)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}
}
\description{
Deletes an application that is specified either by its application ID, name, or ARN. All associated attribute groups and resources must be disassociated from it before deleting an application.

See \url{https://www.paws-r-sdk.com/docs/appregistry_delete_application/} for full documentation.
}
\keyword{internal}
