% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_get_associated_resource}
\alias{appregistry_get_associated_resource}
\title{Gets the resource associated with the application}
\usage{
appregistry_get_associated_resource(
  application,
  resourceType,
  resource,
  nextToken = NULL,
  resourceTagStatus = NULL,
  maxResults = NULL
)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}

\item{resourceType}{[required] The type of resource associated with the application.}

\item{resource}{[required] The name or ID of the resource associated with the application.}

\item{nextToken}{A unique pagination token for each page of results. Make the call again
with the returned token to retrieve the next page of results.}

\item{resourceTagStatus}{States whether an application tag is applied, not applied, in the
process of being applied, or skipped.}

\item{maxResults}{The maximum number of results to return. If the parameter is omitted, it
defaults to 25. The value is optional.}
}
\description{
Gets the resource associated with the application.

See \url{https://www.paws-r-sdk.com/docs/appregistry_get_associated_resource/} for full documentation.
}
\keyword{internal}
