% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_associated_resources}
\alias{appregistry_list_associated_resources}
\title{Lists all of the resources that are associated with the specified
application}
\usage{
appregistry_list_associated_resources(
  application,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}

\item{nextToken}{The token to use to get the next page of results after a previous API
call.}

\item{maxResults}{The upper bound of the number of results to return (cannot exceed 25).
If this parameter is omitted, it defaults to 25. This value is optional.}
}
\description{
Lists all of the resources that are associated with the specified application. Results are paginated.

See \url{https://www.paws-r-sdk.com/docs/appregistry_list_associated_resources/} for full documentation.
}
\keyword{internal}
