% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_attribute_groups_for_application}
\alias{appregistry_list_attribute_groups_for_application}
\title{Lists the details of all attribute groups associated with a specific
application}
\usage{
appregistry_list_attribute_groups_for_application(
  application,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{application}{[required] The name or ID of the application.}

\item{nextToken}{This token retrieves the next page of results after a previous API call.}

\item{maxResults}{The upper bound of the number of results to return. The value cannot
exceed 25. If you omit this parameter, it defaults to 25. This value is
optional.}
}
\description{
Lists the details of all attribute groups associated with a specific application. The results display in pages.

See \url{https://paws-r.github.io/docs/appregistry/list_attribute_groups_for_application.html} for full documentation.
}
\keyword{internal}
