% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_tags_for_resource}
\alias{appregistry_list_tags_for_resource}
\title{Lists all of the tags on the resource}
\usage{
appregistry_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource name (ARN) that specifies the resource.}
}
\description{
Lists all of the tags on the resource.

See \url{https://paws-r.github.io/docs/appregistry/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
