% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_sync_resource}
\alias{appregistry_sync_resource}
\title{Syncs the resource with current AppRegistry records}
\usage{
appregistry_sync_resource(resourceType, resource)
}
\arguments{
\item{resourceType}{[required] The type of resource of which the application will be associated.}

\item{resource}{[required] An entity you can work with and specify with a name or ID. Examples
include an Amazon EC2 instance, an Amazon Web Services CloudFormation
stack, or an Amazon S3 bucket.}
}
\description{
Syncs the resource with current AppRegistry records.

See \url{https://paws-r.github.io/docs/appregistry/sync_resource.html} for full documentation.
}
\keyword{internal}
