% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_tag_resource}
\alias{appregistry_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified resource}
\usage{
appregistry_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource name (ARN) that specifies the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified resource.

See \url{https://paws-r.github.io/docs/appregistry/tag_resource.html} for full documentation.
}
\keyword{internal}
