% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_update_application}
\alias{appregistry_update_application}
\title{Updates an existing application with new attributes}
\usage{
appregistry_update_application(application, name = NULL, description = NULL)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application that will be updated.}

\item{name}{Deprecated: The new name of the application. The name must be unique in
the region in which you are updating the application. Please do not use
this field as we have stopped supporting name updates.}

\item{description}{The new description of the application.}
}
\description{
Updates an existing application with new attributes.

See \url{https://www.paws-r-sdk.com/docs/appregistry_update_application/} for full documentation.
}
\keyword{internal}
