% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_associate_assessment_report_evidence_folder}
\alias{auditmanager_associate_assessment_report_evidence_folder}
\title{Associates an evidence folder to an assessment report in an Audit
Manager assessment}
\usage{
auditmanager_associate_assessment_report_evidence_folder(
  assessmentId,
  evidenceFolderId
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{evidenceFolderId}{[required] The identifier for the folder that the evidence is stored in.}
}
\description{
Associates an evidence folder to an assessment report in an Audit Manager assessment.

See \url{https://paws-r.github.io/docs/auditmanager/associate_assessment_report_evidence_folder.html} for full documentation.
}
\keyword{internal}
