% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_associate_assessment_report_evidence}
\alias{auditmanager_batch_associate_assessment_report_evidence}
\title{Associates a list of evidence to an assessment report in an Audit
Manager assessment}
\usage{
auditmanager_batch_associate_assessment_report_evidence(
  assessmentId,
  evidenceFolderId,
  evidenceIds
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{evidenceFolderId}{[required] The identifier for the folder that the evidence is stored in.}

\item{evidenceIds}{[required] The list of evidence identifiers.}
}
\description{
Associates a list of evidence to an assessment report in an Audit Manager assessment.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_batch_associate_assessment_report_evidence/} for full documentation.
}
\keyword{internal}
