% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_delete_delegation_by_assessment}
\alias{auditmanager_batch_delete_delegation_by_assessment}
\title{Deletes a batch of delegations for an assessment in Audit Manager}
\usage{
auditmanager_batch_delete_delegation_by_assessment(delegationIds, assessmentId)
}
\arguments{
\item{delegationIds}{[required] The identifiers for the delegations.}

\item{assessmentId}{[required] The identifier for the assessment.}
}
\description{
Deletes a batch of delegations for an assessment in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/batch_delete_delegation_by_assessment.html} for full documentation.
}
\keyword{internal}
