% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_disassociate_assessment_report_evidence}
\alias{auditmanager_batch_disassociate_assessment_report_evidence}
\title{Disassociates a list of evidence from an assessment report in Audit
Manager}
\usage{
auditmanager_batch_disassociate_assessment_report_evidence(
  assessmentId,
  evidenceFolderId,
  evidenceIds
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{evidenceFolderId}{[required] The identifier for the folder that the evidence is stored in.}

\item{evidenceIds}{[required] The list of evidence identifiers.}
}
\description{
Disassociates a list of evidence from an assessment report in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_batch_disassociate_assessment_report_evidence/} for full documentation.
}
\keyword{internal}
