% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_import_evidence_to_assessment_control}
\alias{auditmanager_batch_import_evidence_to_assessment_control}
\title{Uploads one or more pieces of evidence to a control in an Audit Manager
assessment}
\usage{
auditmanager_batch_import_evidence_to_assessment_control(
  assessmentId,
  controlSetId,
  controlId,
  manualEvidence
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{controlSetId}{[required] The identifier for the control set.}

\item{controlId}{[required] The identifier for the control.}

\item{manualEvidence}{[required] The list of manual evidence objects.}
}
\description{
Uploads one or more pieces of evidence to a control in an Audit Manager assessment. You can upload manual evidence from any Amazon Simple Storage Service (Amazon S3) bucket by specifying the S3 URI of the evidence.

See \url{https://paws-r.github.io/docs/auditmanager/batch_import_evidence_to_assessment_control.html} for full documentation.
}
\keyword{internal}
