% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment}
\alias{auditmanager_create_assessment}
\title{Creates an assessment in Audit Manager}
\usage{
auditmanager_create_assessment(
  name,
  description = NULL,
  assessmentReportsDestination,
  scope,
  roles,
  frameworkId,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the assessment to be created.}

\item{description}{The optional description of the assessment to be created.}

\item{assessmentReportsDestination}{[required] The assessment report storage destination for the assessment that's being created.}

\item{scope}{[required] The wrapper that contains the Amazon Web Services accounts that are in scope for the assessment.

You no longer need to specify which Amazon Web Services services are in scope when you create or update an assessment. Audit Manager infers the services in scope by examining your assessment controls and their data sources, and then mapping this information to the relevant Amazon Web Services services.

If an underlying data source changes for your assessment, we automatically update the services scope as needed to reflect the correct Amazon Web Services services. This ensures that your assessment collects accurate and comprehensive evidence about all of the relevant services in your AWS environment.}

\item{roles}{[required] The list of roles for the assessment.}

\item{frameworkId}{[required] The identifier for the framework that the assessment will be created from.}

\item{tags}{The tags that are associated with the assessment.}
}
\description{
Creates an assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_create_assessment/} for full documentation.
}
\keyword{internal}
