% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment_framework}
\alias{auditmanager_create_assessment_framework}
\title{Creates a custom framework in Audit Manager}
\usage{
auditmanager_create_assessment_framework(
  name,
  description = NULL,
  complianceType = NULL,
  controlSets,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the new custom framework.}

\item{description}{An optional description for the new custom framework.}

\item{complianceType}{The compliance type that the new custom framework supports, such as CIS
or HIPAA.}

\item{controlSets}{[required] The control sets that are associated with the framework.

The \code{Controls} object returns a partial response when called through
Framework APIs. For a complete \code{Controls} object, use
\code{\link[=auditmanager_get_control]{get_control}}.}

\item{tags}{The tags that are associated with the framework.}
}
\description{
Creates a custom framework in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_create_assessment_framework/} for full documentation.
}
\keyword{internal}
