% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_control}
\alias{auditmanager_create_control}
\title{Creates a new custom control in Audit Manager}
\usage{
auditmanager_create_control(
  name,
  description = NULL,
  testingInformation = NULL,
  actionPlanTitle = NULL,
  actionPlanInstructions = NULL,
  controlMappingSources,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the control.}

\item{description}{The description of the control.}

\item{testingInformation}{The steps to follow to determine if the control is satisfied.}

\item{actionPlanTitle}{The title of the action plan for remediating the control.}

\item{actionPlanInstructions}{The recommended actions to carry out if the control isn't fulfilled.}

\item{controlMappingSources}{[required] The data mapping sources for the control.}

\item{tags}{The tags that are associated with the control.}
}
\description{
Creates a new custom control in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/create_control.html} for full documentation.
}
\keyword{internal}
