% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment}
\alias{auditmanager_delete_assessment}
\title{Deletes an assessment in Audit Manager}
\usage{
auditmanager_delete_assessment(assessmentId)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}
}
\description{
Deletes an assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_delete_assessment/} for full documentation.
}
\keyword{internal}
