% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment_framework}
\alias{auditmanager_delete_assessment_framework}
\title{Deletes a custom framework in Audit Manager}
\usage{
auditmanager_delete_assessment_framework(frameworkId)
}
\arguments{
\item{frameworkId}{[required] The identifier for the custom framework.}
}
\description{
Deletes a custom framework in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/delete_assessment_framework.html} for full documentation.
}
\keyword{internal}
