% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment_framework_share}
\alias{auditmanager_delete_assessment_framework_share}
\title{Deletes a share request for a custom framework in Audit Manager}
\usage{
auditmanager_delete_assessment_framework_share(requestId, requestType)
}
\arguments{
\item{requestId}{[required] The unique identifier for the share request to be deleted.}

\item{requestType}{[required] Specifies whether the share request is a sent request or a received
request.}
}
\description{
Deletes a share request for a custom framework in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_delete_assessment_framework_share/} for full documentation.
}
\keyword{internal}
