% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_disassociate_assessment_report_evidence_folder}
\alias{auditmanager_disassociate_assessment_report_evidence_folder}
\title{Disassociates an evidence folder from the specified assessment report in
Audit Manager}
\usage{
auditmanager_disassociate_assessment_report_evidence_folder(
  assessmentId,
  evidenceFolderId
)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{evidenceFolderId}{[required] The unique identifier for the folder that the evidence is stored in.}
}
\description{
Disassociates an evidence folder from the specified assessment report in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_disassociate_assessment_report_evidence_folder/} for full documentation.
}
\keyword{internal}
