% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_assessment}
\alias{auditmanager_get_assessment}
\title{Gets information about a specified assessment}
\usage{
auditmanager_get_assessment(assessmentId)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}
}
\description{
Gets information about a specified assessment.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_assessment/} for full documentation.
}
\keyword{internal}
