% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_assessment_framework}
\alias{auditmanager_get_assessment_framework}
\title{Gets information about a specified framework}
\usage{
auditmanager_get_assessment_framework(frameworkId)
}
\arguments{
\item{frameworkId}{[required] The identifier for the framework.}
}
\description{
Gets information about a specified framework.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_assessment_framework/} for full documentation.
}
\keyword{internal}
