% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_change_logs}
\alias{auditmanager_get_change_logs}
\title{Gets a list of changelogs from Audit Manager}
\usage{
auditmanager_get_change_logs(
  assessmentId,
  controlSetId = NULL,
  controlId = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{controlSetId}{The unique identifier for the control set.}

\item{controlId}{The unique identifier for the control.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Gets a list of changelogs from Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_change_logs/} for full documentation.
}
\keyword{internal}
