% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_control}
\alias{auditmanager_get_control}
\title{Gets information about a specified control}
\usage{
auditmanager_get_control(controlId)
}
\arguments{
\item{controlId}{[required] The identifier for the control.}
}
\description{
Gets information about a specified control.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_control/} for full documentation.
}
\keyword{internal}
