% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_delegations}
\alias{auditmanager_get_delegations}
\title{Returns a list of delegations from an audit owner to a delegate}
\usage{
auditmanager_get_delegations(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of delegations from an audit owner to a delegate.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_delegations/} for full documentation.
}
\keyword{internal}
