% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence}
\alias{auditmanager_get_evidence}
\title{Returns evidence from Audit Manager}
\usage{
auditmanager_get_evidence(
  assessmentId,
  controlSetId,
  evidenceFolderId,
  evidenceId
)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{controlSetId}{[required] The unique identifier for the control set.}

\item{evidenceFolderId}{[required] The unique identifier for the folder that the evidence is stored in.}

\item{evidenceId}{[required] The unique identifier for the evidence.}
}
\description{
Returns evidence from Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_evidence/} for full documentation.
}
\keyword{internal}
