% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_by_evidence_folder}
\alias{auditmanager_get_evidence_by_evidence_folder}
\title{Returns all evidence from a specified evidence folder in Audit Manager}
\usage{
auditmanager_get_evidence_by_evidence_folder(
  assessmentId,
  controlSetId,
  evidenceFolderId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{controlSetId}{[required] The identifier for the control set.}

\item{evidenceFolderId}{[required] The unique identifier for the folder that the evidence is stored in.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns all evidence from a specified evidence folder in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_evidence_by_evidence_folder/} for full documentation.
}
\keyword{internal}
