% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_folder}
\alias{auditmanager_get_evidence_folder}
\title{Gets an evidence folder from a specified assessment in Audit Manager}
\usage{
auditmanager_get_evidence_folder(assessmentId, controlSetId, evidenceFolderId)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{controlSetId}{[required] The unique identifier for the control set.}

\item{evidenceFolderId}{[required] The unique identifier for the folder that the evidence is stored in.}
}
\description{
Gets an evidence folder from a specified assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_evidence_folder/} for full documentation.
}
\keyword{internal}
