% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_folders_by_assessment_control}
\alias{auditmanager_get_evidence_folders_by_assessment_control}
\title{Gets a list of evidence folders that are associated with a specified
control in an Audit Manager assessment}
\usage{
auditmanager_get_evidence_folders_by_assessment_control(
  assessmentId,
  controlSetId,
  controlId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{controlSetId}{[required] The identifier for the control set.}

\item{controlId}{[required] The identifier for the control.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Gets a list of evidence folders that are associated with a specified control in an Audit Manager assessment.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_evidence_folders_by_assessment_control/} for full documentation.
}
\keyword{internal}
