% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_insights}
\alias{auditmanager_get_insights}
\title{Gets the latest analytics data for all your current active assessments}
\usage{
auditmanager_get_insights()
}
\description{
Gets the latest analytics data for all your current active assessments.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_insights/} for full documentation.
}
\keyword{internal}
