% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_insights_by_assessment}
\alias{auditmanager_get_insights_by_assessment}
\title{Gets the latest analytics data for a specific active assessment}
\usage{
auditmanager_get_insights_by_assessment(assessmentId)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}
}
\description{
Gets the latest analytics data for a specific active assessment.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_get_insights_by_assessment/} for full documentation.
}
\keyword{internal}
