% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_services_in_scope}
\alias{auditmanager_get_services_in_scope}
\title{Returns a list of the in-scope Amazon Web Services for the specified
assessment}
\usage{
auditmanager_get_services_in_scope()
}
\description{
Returns a list of the in-scope Amazon Web Services for the specified assessment.

See \url{https://paws-r.github.io/docs/auditmanager/get_services_in_scope.html} for full documentation.
}
\keyword{internal}
