% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_settings}
\alias{auditmanager_get_settings}
\title{Returns the settings for the specified Amazon Web Services account}
\usage{
auditmanager_get_settings(attribute)
}
\arguments{
\item{attribute}{[required] The list of setting attribute enum values.}
}
\description{
Returns the settings for the specified Amazon Web Services account.

See \url{https://paws-r.github.io/docs/auditmanager/get_settings.html} for full documentation.
}
\keyword{internal}
