% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessment_framework_share_requests}
\alias{auditmanager_list_assessment_framework_share_requests}
\title{Returns a list of sent or received share requests for custom frameworks
in Audit Manager}
\usage{
auditmanager_list_assessment_framework_share_requests(
  requestType,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{requestType}{[required] Specifies whether the share request is a sent request or a received
request.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of sent or received share requests for custom frameworks in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/list_assessment_framework_share_requests.html} for full documentation.
}
\keyword{internal}
