% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessment_frameworks}
\alias{auditmanager_list_assessment_frameworks}
\title{Returns a list of the frameworks that are available in the Audit Manager
framework library}
\usage{
auditmanager_list_assessment_frameworks(
  frameworkType,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{frameworkType}{[required] The type of framework, such as a standard framework or a custom
framework.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of the frameworks that are available in the Audit Manager framework library.

See \url{https://paws-r.github.io/docs/auditmanager/list_assessment_frameworks.html} for full documentation.
}
\keyword{internal}
