% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_control_domain_insights}
\alias{auditmanager_list_control_domain_insights}
\title{Lists the latest analytics data for control domains across all of your
active assessments}
\usage{
auditmanager_list_control_domain_insights(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Lists the latest analytics data for control domains across all of your active assessments.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_list_control_domain_insights/} for full documentation.
}
\keyword{internal}
