% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_keywords_for_data_source}
\alias{auditmanager_list_keywords_for_data_source}
\title{Returns a list of keywords that are pre-mapped to the specified control
data source}
\usage{
auditmanager_list_keywords_for_data_source(
  source,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{source}{[required] The control mapping data source that the keywords apply to.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of keywords that are pre-mapped to the specified control data source.

See \url{https://paws-r.github.io/docs/auditmanager/list_keywords_for_data_source.html} for full documentation.
}
\keyword{internal}
