% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_notifications}
\alias{auditmanager_list_notifications}
\title{Returns a list of all Audit Manager notifications}
\usage{
auditmanager_list_notifications(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of all Audit Manager notifications.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_list_notifications/} for full documentation.
}
\keyword{internal}
