% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_register_account}
\alias{auditmanager_register_account}
\title{Enables Audit Manager for the specified Amazon Web Services account}
\usage{
auditmanager_register_account(kmsKey = NULL, delegatedAdminAccount = NULL)
}
\arguments{
\item{kmsKey}{The KMS key details.}

\item{delegatedAdminAccount}{The delegated administrator account for Audit Manager.}
}
\description{
Enables Audit Manager for the specified Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_register_account/} for full documentation.
}
\keyword{internal}
