% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_register_organization_admin_account}
\alias{auditmanager_register_organization_admin_account}
\title{Enables an Amazon Web Services account within the organization as the
delegated administrator for Audit Manager}
\usage{
auditmanager_register_organization_admin_account(adminAccountId)
}
\arguments{
\item{adminAccountId}{[required] The identifier for the delegated administrator account.}
}
\description{
Enables an Amazon Web Services account within the organization as the delegated administrator for Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/register_organization_admin_account.html} for full documentation.
}
\keyword{internal}
