% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_untag_resource}
\alias{auditmanager_untag_resource}
\title{Removes a tag from a resource in Audit Manager}
\usage{
auditmanager_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the specified resource.}

\item{tagKeys}{[required] The name or key of the tag.}
}
\description{
Removes a tag from a resource in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/untag_resource.html} for full documentation.
}
\keyword{internal}
