% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_control}
\alias{auditmanager_update_assessment_control}
\title{Updates a control within an assessment in Audit Manager}
\usage{
auditmanager_update_assessment_control(
  assessmentId,
  controlSetId,
  controlId,
  controlStatus = NULL,
  commentBody = NULL
)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{controlSetId}{[required] The unique identifier for the control set.}

\item{controlId}{[required] The unique identifier for the control.}

\item{controlStatus}{The status of the control.}

\item{commentBody}{The comment body text for the control.}
}
\description{
Updates a control within an assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_update_assessment_control/} for full documentation.
}
\keyword{internal}
