% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_framework_share}
\alias{auditmanager_update_assessment_framework_share}
\title{Updates a share request for a custom framework in Audit Manager}
\usage{
auditmanager_update_assessment_framework_share(requestId, requestType, action)
}
\arguments{
\item{requestId}{[required] The unique identifier for the share request.}

\item{requestType}{[required] Specifies whether the share request is a sent request or a received
request.}

\item{action}{[required] Specifies the update action for the share request.}
}
\description{
Updates a share request for a custom framework in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_update_assessment_framework_share/} for full documentation.
}
\keyword{internal}
