% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_status}
\alias{auditmanager_update_assessment_status}
\title{Updates the status of an assessment in Audit Manager}
\usage{
auditmanager_update_assessment_status(assessmentId, status)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{status}{[required] The current status of the assessment.}
}
\description{
Updates the status of an assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_update_assessment_status/} for full documentation.
}
\keyword{internal}
