% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_settings}
\alias{auditmanager_update_settings}
\title{Updates Audit Manager settings for the current user account}
\usage{
auditmanager_update_settings(
  snsTopic = NULL,
  defaultAssessmentReportsDestination = NULL,
  defaultProcessOwners = NULL,
  kmsKey = NULL
)
}
\arguments{
\item{snsTopic}{The Amazon Simple Notification Service (Amazon SNS) topic that Audit
Manager sends notifications to.}

\item{defaultAssessmentReportsDestination}{The default storage destination for assessment reports.}

\item{defaultProcessOwners}{A list of the default audit owners.}

\item{kmsKey}{The KMS key details.}
}
\description{
Updates Audit Manager settings for the current user account.

See \url{https://paws-r.github.io/docs/auditmanager/update_settings.html} for full documentation.
}
\keyword{internal}
