% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_attach_load_balancers}
\alias{autoscaling_attach_load_balancers}
\title{To attach an Application Load Balancer, Network Load Balancer, or
Gateway Load Balancer, use the AttachLoadBalancerTargetGroups API
operation instead}
\usage{
autoscaling_attach_load_balancers(AutoScalingGroupName, LoadBalancerNames)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LoadBalancerNames}{[required] The names of the load balancers. You can specify up to 10 load
balancers.}
}
\description{
To attach an Application Load Balancer, Network Load Balancer, or Gateway Load Balancer, use the \code{\link[=autoscaling_attach_load_balancer_target_groups]{attach_load_balancer_target_groups}} API operation instead.

See \url{https://paws-r.github.io/docs/autoscaling/attach_load_balancers.html} for full documentation.
}
\keyword{internal}
