% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_batch_delete_scheduled_action}
\alias{autoscaling_batch_delete_scheduled_action}
\title{Deletes one or more scheduled actions for the specified Auto Scaling
group}
\usage{
autoscaling_batch_delete_scheduled_action(AutoScalingGroupName,
  ScheduledActionNames)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScheduledActionNames}{[required] The names of the scheduled actions to delete. The maximum number allowed
is 50.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FailedScheduledActions = list(
    list(
      ScheduledActionName = "string",
      ErrorCode = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes one or more scheduled actions for the specified Auto Scaling
group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_delete_scheduled_action(
  AutoScalingGroupName = "string",
  ScheduledActionNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
