% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_cancel_instance_refresh}
\alias{autoscaling_cancel_instance_refresh}
\title{Cancels an instance refresh or rollback that is in progress}
\usage{
autoscaling_cancel_instance_refresh(AutoScalingGroupName)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}
}
\description{
Cancels an instance refresh or rollback that is in progress. If an instance refresh or rollback is not in progress, an \code{ActiveInstanceRefreshNotFound} error occurs.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_cancel_instance_refresh/} for full documentation.
}
\keyword{internal}
