% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_launch_configuration}
\alias{autoscaling_delete_launch_configuration}
\title{Deletes the specified launch configuration}
\usage{
autoscaling_delete_launch_configuration(LaunchConfigurationName)
}
\arguments{
\item{LaunchConfigurationName}{[required] The name of the launch configuration.}
}
\description{
Deletes the specified launch configuration.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_delete_launch_configuration/} for full documentation.
}
\keyword{internal}
