% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_notification_configuration}
\alias{autoscaling_delete_notification_configuration}
\title{Deletes the specified notification}
\usage{
autoscaling_delete_notification_configuration(AutoScalingGroupName, TopicARN)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{TopicARN}{[required] The Amazon Resource Name (ARN) of the Amazon SNS topic.}
}
\description{
Deletes the specified notification.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_delete_notification_configuration/} for full documentation.
}
\keyword{internal}
