% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_policy}
\alias{autoscaling_delete_policy}
\title{Deletes the specified scaling policy}
\usage{
autoscaling_delete_policy(AutoScalingGroupName = NULL, PolicyName)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{PolicyName}{[required] The name or Amazon Resource Name (ARN) of the policy.}
}
\description{
Deletes the specified scaling policy.

See \url{https://paws-r.github.io/docs/autoscaling/delete_policy.html} for full documentation.
}
\keyword{internal}
