% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_warm_pool}
\alias{autoscaling_delete_warm_pool}
\title{Deletes the warm pool for the specified Auto Scaling group}
\usage{
autoscaling_delete_warm_pool(AutoScalingGroupName, ForceDelete)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ForceDelete}{Specifies that the warm pool is to be deleted along with all instances
associated with the warm pool, without waiting for all instances to be
terminated. This parameter also deletes any outstanding lifecycle
actions associated with the warm pool instances.}
}
\value{
An empty list.
}
\description{
Deletes the warm pool for the specified Auto Scaling group.
}
\section{Request syntax}{
\preformatted{svc$delete_warm_pool(
  AutoScalingGroupName = "string",
  ForceDelete = TRUE|FALSE
)
}
}

\keyword{internal}
