% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_auto_scaling_instances}
\alias{autoscaling_describe_auto_scaling_instances}
\title{Gets information about the Auto Scaling instances in the account and
Region}
\usage{
autoscaling_describe_auto_scaling_instances(
  InstanceIds = NULL,
  MaxRecords = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. If you omit this property, all Auto Scaling
instances are described. If you specify an ID that does not exist, it is
ignored with no error.

Array Members: Maximum number of 50 items.}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{50}.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Gets information about the Auto Scaling instances in the account and Region.

See \url{https://paws-r.github.io/docs/autoscaling/describe_auto_scaling_instances.html} for full documentation.
}
\keyword{internal}
