% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_auto_scaling_notification_types}
\alias{autoscaling_describe_auto_scaling_notification_types}
\title{Describes the notification types that are supported by Amazon EC2 Auto
Scaling}
\usage{
autoscaling_describe_auto_scaling_notification_types()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AutoScalingNotificationTypes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the notification types that are supported by Amazon EC2 Auto
Scaling.
}
\examples{
\dontrun{
# This example describes the available notification types.
svc$describe_auto_scaling_notification_types()
}

}
\keyword{internal}
