% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_metric_collection_types}
\alias{autoscaling_describe_metric_collection_types}
\title{Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling}
\usage{
autoscaling_describe_metric_collection_types()
}
\description{
Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_metric_collection_types/} for full documentation.
}
\keyword{internal}
