% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_notification_configurations}
\alias{autoscaling_describe_notification_configurations}
\title{Gets information about the Amazon SNS notifications that are configured
for one or more Auto Scaling groups}
\usage{
autoscaling_describe_notification_configurations(
  AutoScalingGroupNames = NULL,
  NextToken = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{AutoScalingGroupNames}{The name of the Auto Scaling group.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{100}.}
}
\description{
Gets information about the Amazon SNS notifications that are configured for one or more Auto Scaling groups.

See \url{https://paws-r.github.io/docs/autoscaling/describe_notification_configurations.html} for full documentation.
}
\keyword{internal}
