% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_policies}
\alias{autoscaling_describe_policies}
\title{Gets information about the scaling policies in the account and Region}
\usage{
autoscaling_describe_policies(
  AutoScalingGroupName = NULL,
  PolicyNames = NULL,
  PolicyTypes = NULL,
  NextToken = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{PolicyNames}{The names of one or more policies. If you omit this property, all
policies are described. If a group name is provided, the results are
limited to that group. If you specify an unknown policy name, it is
ignored with no error.

Array Members: Maximum number of 50 items.}

\item{PolicyTypes}{One or more policy types. The valid values are \code{SimpleScaling},
\code{StepScaling}, \code{TargetTrackingScaling}, and \code{PredictiveScaling}.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to be returned with each call. The default
value is \code{50} and the maximum value is \code{100}.}
}
\description{
Gets information about the scaling policies in the account and Region.

See \url{https://paws-r.github.io/docs/autoscaling/describe_policies.html} for full documentation.
}
\keyword{internal}
