% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_warm_pool}
\alias{autoscaling_describe_warm_pool}
\title{Describes a warm pool and its instances}
\usage{
autoscaling_describe_warm_pool(AutoScalingGroupName, MaxRecords,
  NextToken)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{MaxRecords}{The maximum number of instances to return with this call. The maximum
value is \code{50}.}

\item{NextToken}{The token for the next set of instances to return. (You received this
token from a previous call.)}
}
\value{
A list with the following syntax:\preformatted{list(
  WarmPoolConfiguration = list(
    MaxGroupPreparedCapacity = 123,
    MinSize = 123,
    PoolState = "Stopped"|"Running",
    Status = "PendingDelete"
  ),
  Instances = list(
    list(
      InstanceId = "string",
      InstanceType = "string",
      AvailabilityZone = "string",
      LifecycleState = "Pending"|"Pending:Wait"|"Pending:Proceed"|"Quarantined"|"InService"|"Terminating"|"Terminating:Wait"|"Terminating:Proceed"|"Terminated"|"Detaching"|"Detached"|"EnteringStandby"|"Standby"|"Warmed:Pending"|"Warmed:Pending:Wait"|"Warmed:Pending:Proceed"|"Warmed:Terminating"|"Warmed:Terminating:Wait"|"Warmed:Terminating:Proceed"|"Warmed:Terminated"|"Warmed:Stopped"|"Warmed:Running",
      HealthStatus = "string",
      LaunchConfigurationName = "string",
      LaunchTemplate = list(
        LaunchTemplateId = "string",
        LaunchTemplateName = "string",
        Version = "string"
      ),
      ProtectedFromScaleIn = TRUE|FALSE,
      WeightedCapacity = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Describes a warm pool and its instances.
}
\section{Request syntax}{
\preformatted{svc$describe_warm_pool(
  AutoScalingGroupName = "string",
  MaxRecords = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
