% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_detach_instances}
\alias{autoscaling_detach_instances}
\title{Removes one or more instances from the specified Auto Scaling group}
\usage{
autoscaling_detach_instances(
  InstanceIds = NULL,
  AutoScalingGroupName,
  ShouldDecrementDesiredCapacity
)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ShouldDecrementDesiredCapacity}{[required] Indicates whether the Auto Scaling group decrements the desired capacity
value by the number of instances detached.}
}
\description{
Removes one or more instances from the specified Auto Scaling group.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_detach_instances/} for full documentation.
}
\keyword{internal}
