% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_detach_load_balancers}
\alias{autoscaling_detach_load_balancers}
\title{Detaches one or more Classic Load Balancers from the specified Auto
Scaling group}
\usage{
autoscaling_detach_load_balancers(AutoScalingGroupName, LoadBalancerNames)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LoadBalancerNames}{[required] The names of the load balancers. You can specify up to 10 load
balancers.}
}
\description{
Detaches one or more Classic Load Balancers from the specified Auto Scaling group.

See \url{https://paws-r.github.io/docs/autoscaling/detach_load_balancers.html} for full documentation.
}
\keyword{internal}
