% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_detach_traffic_sources}
\alias{autoscaling_detach_traffic_sources}
\title{Detaches one or more traffic sources from the specified Auto Scaling
group}
\usage{
autoscaling_detach_traffic_sources(AutoScalingGroupName, TrafficSources)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{TrafficSources}{[required] The unique identifiers of one or more traffic sources. You can specify
up to 10 traffic sources.}
}
\description{
Detaches one or more traffic sources from the specified Auto Scaling group.

See \url{https://paws-r.github.io/docs/autoscaling/detach_traffic_sources.html} for full documentation.
}
\keyword{internal}
