% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_execute_policy}
\alias{autoscaling_execute_policy}
\title{Executes the specified policy}
\usage{
autoscaling_execute_policy(
  AutoScalingGroupName = NULL,
  PolicyName,
  HonorCooldown = NULL,
  MetricValue = NULL,
  BreachThreshold = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{PolicyName}{[required] The name or ARN of the policy.}

\item{HonorCooldown}{Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period
to complete before executing the policy.

Valid only if the policy type is \code{SimpleScaling}. For more information,
see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/ec2-auto-scaling-scaling-cooldowns.html}{Scaling cooldowns for Amazon EC2 Auto Scaling}
in the \emph{Amazon EC2 Auto Scaling User Guide}.}

\item{MetricValue}{The metric value to compare to \code{BreachThreshold}. This enables you to
execute a policy of type \code{StepScaling} and determine which step
adjustment to use. For example, if the breach threshold is 50 and you
want to use a step adjustment with a lower bound of 0 and an upper bound
of 10, you can set the metric value to 59.

If you specify a metric value that doesn't correspond to a step
adjustment for the policy, the call returns an error.

Required if the policy type is \code{StepScaling} and not supported
otherwise.}

\item{BreachThreshold}{The breach threshold for the alarm.

Required if the policy type is \code{StepScaling} and not supported
otherwise.}
}
\description{
Executes the specified policy. This can be useful for testing the design of your scaling policy.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_execute_policy/} for full documentation.
}
\keyword{internal}
