% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_exit_standby}
\alias{autoscaling_exit_standby}
\title{Moves the specified instances out of the standby state}
\usage{
autoscaling_exit_standby(InstanceIds, AutoScalingGroupName)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Activities = list(
    list(
      ActivityId = "string",
      AutoScalingGroupName = "string",
      Description = "string",
      Cause = "string",
      StartTime = as.POSIXct(
        "2015-01-01"
      ),
      EndTime = as.POSIXct(
        "2015-01-01"
      ),
      StatusCode = "PendingSpotBidPlacement"|"WaitingForSpotInstanceRequestId"|"WaitingForSpotInstanceId"|"WaitingForInstanceId"|"PreInService"|"InProgress"|"WaitingForELBConnectionDraining"|"MidLifecycleAction"|"WaitingForInstanceWarmup"|"Successful"|"Failed"|"Cancelled",
      StatusMessage = "string",
      Progress = 123,
      Details = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Moves the specified instances out of the standby state.

After you put the instances back in service, the desired capacity is
incremented.

For more information, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-enter-exit-standby.html}{Temporarily removing instances from your Auto Scaling group}
in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$exit_standby(
  InstanceIds = list(
    "string"
  ),
  AutoScalingGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This example moves the specified instance out of standby mode.
svc$exit_standby(
  AutoScalingGroupName = "my-auto-scaling-group",
  InstanceIds = list(
    "i-93633f9b"
  )
)
}

}
\keyword{internal}
