% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_launch_instances}
\alias{autoscaling_launch_instances}
\title{Launches a specified number of instances in an Auto Scaling group}
\usage{
autoscaling_launch_instances(
  AutoScalingGroupName,
  RequestedCapacity,
  ClientToken,
  AvailabilityZones = NULL,
  AvailabilityZoneIds = NULL,
  SubnetIds = NULL,
  RetryStrategy = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group to launch instances into.}

\item{RequestedCapacity}{[required] The number of instances to launch. Although this value can exceed 100
for instance weights, the actual instance count is limited to 100
instances per launch.}

\item{ClientToken}{[required] A unique, case-sensitive identifier to ensure idempotency of the
request.}

\item{AvailabilityZones}{The Availability Zones for the instance launch. Must match or be
included in the Auto Scaling group's Availability Zone configuration.
Either \code{AvailabilityZones} or \code{SubnetIds} must be specified for groups
with multiple Availability Zone configurations.}

\item{AvailabilityZoneIds}{A list of Availability Zone IDs where instances should be launched. Must
match or be included in the group's AZ configuration. You cannot specify
both AvailabilityZones and AvailabilityZoneIds. Required for multi-AZ
groups, optional for single-AZ groups.}

\item{SubnetIds}{The subnet IDs for the instance launch. Either \code{AvailabilityZones} or
\code{SubnetIds} must be specified. If both are specified, the subnets must
reside in the specified Availability Zones.}

\item{RetryStrategy}{Specifies whether to retry asynchronously if the synchronous launch
fails. Valid values are NONE (default, no async retry) and
RETRY_WITH_GROUP_CONFIGURATION (increase desired capacity and retry with
group configuration).}
}
\description{
Launches a specified number of instances in an Auto Scaling group. Returns instance IDs and other details if launch is successful or error details if launch is unsuccessful.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_launch_instances/} for full documentation.
}
\keyword{internal}
