% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_put_warm_pool}
\alias{autoscaling_put_warm_pool}
\title{Adds a warm pool to the specified Auto Scaling group}
\usage{
autoscaling_put_warm_pool(AutoScalingGroupName,
  MaxGroupPreparedCapacity, MinSize, PoolState)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{MaxGroupPreparedCapacity}{Specifies the total maximum number of instances that are allowed to be
in the warm pool or in any state except \code{Terminated} for the Auto
Scaling group. This is an optional property. Specify it only if the warm
pool size should not be determined by the difference between the group's
maximum capacity and its desired capacity.

Amazon EC2 Auto Scaling will launch and maintain either the difference
between the group's maximum capacity and its desired capacity, if a
value for \code{MaxGroupPreparedCapacity} is not specified, or the difference
between the \code{MaxGroupPreparedCapacity} and the desired capacity, if a
value for \code{MaxGroupPreparedCapacity} is specified.

The size of the warm pool is dynamic. Only when
\code{MaxGroupPreparedCapacity} and \code{MinSize} are set to the same value does
the warm pool have an absolute size.

If the desired capacity of the Auto Scaling group is higher than the
\code{MaxGroupPreparedCapacity}, the capacity of the warm pool is 0. To
remove a value that you previously set, include the property but specify
-1 for the value.}

\item{MinSize}{Specifies the minimum number of instances to maintain in the warm pool.
This helps you to ensure that there is always a certain number of warmed
instances available to handle traffic spikes. Defaults to 0 if not
specified.}

\item{PoolState}{Sets the instance state to transition to after the lifecycle hooks
finish. Valid values are: \code{Stopped} (default) or \code{Running}.}
}
\value{
An empty list.
}
\description{
Adds a warm pool to the specified Auto Scaling group. A warm pool is a
pool of pre-initialized EC2 instances that sits alongside the Auto
Scaling group. Whenever your application needs to scale out, the Auto
Scaling group can draw on the warm pool to meet its new desired
capacity. For more information, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/ec2-auto-scaling-warm-pools.html}{Warm pools for Amazon EC2 Auto Scaling}
in the \emph{Amazon EC2 Auto Scaling User Guide}.

This operation must be called from the Region in which the Auto Scaling
group was created. This operation cannot be called on an Auto Scaling
group that has a mixed instances policy or a launch template or launch
configuration that requests Spot Instances.

You can view the instances in the warm pool using the
\code{\link[=autoscaling_describe_warm_pool]{describe_warm_pool}} API call. If you
are no longer using a warm pool, you can delete it by calling the
\code{\link[=autoscaling_delete_warm_pool]{delete_warm_pool}} API.
}
\section{Request syntax}{
\preformatted{svc$put_warm_pool(
  AutoScalingGroupName = "string",
  MaxGroupPreparedCapacity = 123,
  MinSize = 123,
  PoolState = "Stopped"|"Running"
)
}
}

\examples{
\dontrun{
# This example adds a warm pool to the specified Auto Scaling group.
svc$put_warm_pool(
  AutoScalingGroupName = "my-auto-scaling-group",
  MinSize = 30L,
  PoolState = "Stopped"
)
}

}
\keyword{internal}
