% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_resume_processes}
\alias{autoscaling_resume_processes}
\title{Resumes the specified suspended auto scaling processes, or all suspended
process, for the specified Auto Scaling group}
\usage{
autoscaling_resume_processes(AutoScalingGroupName, ScalingProcesses = NULL)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScalingProcesses}{One or more of the following processes:
\itemize{
\item \code{Launch}
\item \code{Terminate}
\item \code{AddToLoadBalancer}
\item \code{AlarmNotification}
\item \code{AZRebalance}
\item \code{HealthCheck}
\item \code{InstanceRefresh}
\item \code{ReplaceUnhealthy}
\item \code{ScheduledActions}
}

If you omit this property, all processes are specified.}
}
\description{
Resumes the specified suspended auto scaling processes, or all suspended process, for the specified Auto Scaling group.

See \url{https://paws-r.github.io/docs/autoscaling/resume_processes.html} for full documentation.
}
\keyword{internal}
