% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_rollback_instance_refresh}
\alias{autoscaling_rollback_instance_refresh}
\title{Cancels an instance refresh that is in progress and rolls back any
changes that it made}
\usage{
autoscaling_rollback_instance_refresh(AutoScalingGroupName = NULL)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}
}
\description{
Cancels an instance refresh that is in progress and rolls back any changes that it made. Amazon EC2 Auto Scaling replaces any instances that were replaced during the instance refresh. This restores your Auto Scaling group to the configuration that it was using before the start of the instance refresh.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_rollback_instance_refresh/} for full documentation.
}
\keyword{internal}
