% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_set_instance_protection}
\alias{autoscaling_set_instance_protection}
\title{Updates the instance protection settings of the specified instances}
\usage{
autoscaling_set_instance_protection(
  InstanceIds,
  AutoScalingGroupName,
  ProtectedFromScaleIn
)
}
\arguments{
\item{InstanceIds}{[required] One or more instance IDs. You can specify up to 50 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ProtectedFromScaleIn}{[required] Indicates whether the instance is protected from termination by Amazon
EC2 Auto Scaling when scaling in.}
}
\description{
Updates the instance protection settings of the specified instances. This operation cannot be called on instances in a warm pool.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_set_instance_protection/} for full documentation.
}
\keyword{internal}
