% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_terminate_instance_in_auto_scaling_group}
\alias{autoscaling_terminate_instance_in_auto_scaling_group}
\title{Terminates the specified instance and optionally adjusts the desired
group size}
\usage{
autoscaling_terminate_instance_in_auto_scaling_group(
  InstanceId = NULL,
  InstanceIds = NULL,
  AutoScalingGroupName = NULL,
  ShouldDecrementDesiredCapacity
)
}
\arguments{
\item{InstanceId}{The ID of the instance.}

\item{InstanceIds}{The IDs of the instances. You can specify up to 100 instances.

This parameter requires that you also specify \code{AutoScalingGroupName}.}

\item{AutoScalingGroupName}{The name of the Auto Scaling group. Required when using \code{InstanceIds}.}

\item{ShouldDecrementDesiredCapacity}{[required] Indicates whether terminating the instance also decrements the size of the Auto Scaling group.}
}
\description{
Terminates the specified instance and optionally adjusts the desired group size. This operation cannot be called on instances in a warm pool.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_terminate_instance_in_auto_scaling_group/} for full documentation.
}
\keyword{internal}
