% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscalingplans_operations.R
\name{autoscalingplans_create_scaling_plan}
\alias{autoscalingplans_create_scaling_plan}
\title{Creates a scaling plan}
\usage{
autoscalingplans_create_scaling_plan(
  ScalingPlanName,
  ApplicationSource,
  ScalingInstructions
)
}
\arguments{
\item{ScalingPlanName}{[required] The name of the scaling plan. Names cannot contain vertical bars,
colons, or forward slashes.}

\item{ApplicationSource}{[required] A CloudFormation stack or set of tags. You can create one scaling plan
per application source.

For more information, see
\href{https://docs.aws.amazon.com/autoscaling/plans/APIReference/API_ApplicationSource.html}{ApplicationSource}
in the \emph{AWS Auto Scaling API Reference}.}

\item{ScalingInstructions}{[required] The scaling instructions.

For more information, see
\href{https://docs.aws.amazon.com/autoscaling/plans/APIReference/API_ScalingInstruction.html}{ScalingInstruction}
in the \emph{AWS Auto Scaling API Reference}.}
}
\description{
Creates a scaling plan.

See \url{https://www.paws-r-sdk.com/docs/autoscalingplans_create_scaling_plan/} for full documentation.
}
\keyword{internal}
