% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_activate_organizations_access}
\alias{cloudformation_activate_organizations_access}
\title{Activate trusted access with Organizations}
\usage{
cloudformation_activate_organizations_access()
}
\description{
Activate trusted access with Organizations. With trusted access between StackSets and Organizations activated, the management account has permissions to create and manage StackSets for your organization.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_activate_organizations_access/} for full documentation.
}
\keyword{internal}
