% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_batch_describe_type_configurations}
\alias{cloudformation_batch_describe_type_configurations}
\title{Returns configuration data for the specified CloudFormation extensions,
from the CloudFormation registry for the account and region}
\usage{
cloudformation_batch_describe_type_configurations(TypeConfigurationIdentifiers)
}
\arguments{
\item{TypeConfigurationIdentifiers}{[required] The list of identifiers for the desired extension configurations.}
}
\description{
Returns configuration data for the specified CloudFormation extensions, from the CloudFormation registry for the account and region.

See \url{https://paws-r.github.io/docs/cloudformation/batch_describe_type_configurations.html} for full documentation.
}
\keyword{internal}
