% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_delete_change_set}
\alias{cloudformation_delete_change_set}
\title{Deletes the specified change set}
\usage{
cloudformation_delete_change_set(ChangeSetName, StackName = NULL)
}
\arguments{
\item{ChangeSetName}{[required] The name or Amazon Resource Name (ARN) of the change set that you want
to delete.}

\item{StackName}{If you specified the name of a change set to delete, specify the stack
name or Amazon Resource Name (ARN) that's associated with it.}
}
\description{
Deletes the specified change set. Deleting change sets ensures that no one executes the wrong change set.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_delete_change_set/} for full documentation.
}
\keyword{internal}
