% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_delete_generated_template}
\alias{cloudformation_delete_generated_template}
\title{Deleted a generated template}
\usage{
cloudformation_delete_generated_template(GeneratedTemplateName)
}
\arguments{
\item{GeneratedTemplateName}{[required] The name or Amazon Resource Name (ARN) of a generated template.}
}
\description{
Deleted a generated template.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_delete_generated_template/} for full documentation.
}
\keyword{internal}
