% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_delete_stack}
\alias{cloudformation_delete_stack}
\title{Deletes a specified stack}
\usage{
cloudformation_delete_stack(
  StackName,
  RetainResources = NULL,
  RoleARN = NULL,
  ClientRequestToken = NULL,
  DeletionMode = NULL,
  DeploymentConfig = NULL
)
}
\arguments{
\item{StackName}{[required] The name or the unique stack ID that's associated with the stack.}

\item{RetainResources}{For stacks in the \code{DELETE_FAILED} state, a list of resource logical IDs that are associated with the resources you want to retain. During deletion, CloudFormation deletes the stack but doesn't delete the retained resources.

Retaining resources is useful when you can't delete a resource, such as a non-empty S3 bucket, but you want to delete the stack.}

\item{RoleARN}{The Amazon Resource Name (ARN) of an IAM role that CloudFormation assumes to delete the stack. CloudFormation uses the role's credentials to make calls on your behalf.

If you don't specify a value, CloudFormation uses the role that was previously associated with the stack. If no role is available, CloudFormation uses a temporary session that's generated from your user credentials.}

\item{ClientRequestToken}{A unique identifier for this \code{\link[=cloudformation_delete_stack]{delete_stack}} request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to delete a stack with the same name. You might retry \code{\link[=cloudformation_delete_stack]{delete_stack}} requests to ensure that CloudFormation successfully received them.

All events initiated by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a \code{\link[=cloudformation_create_stack]{create_stack}} operation with the token \code{token1}, then all the \code{StackEvents} generated by that operation will have \code{ClientRequestToken} set as \code{token1}.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format \emph{Console-StackOperation-ID}, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: \verb{Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002}.}

\item{DeletionMode}{Specifies the deletion mode for the stack. Possible values are:
\itemize{
\item \code{STANDARD} - Use the standard behavior. Specifying this value is the same as not specifying this parameter.
\item \code{FORCE_DELETE_STACK} - Delete the stack if it's stuck in a \code{DELETE_FAILED} state due to resource deletion failure.
}}

\item{DeploymentConfig}{The deployment configuration for this stack operation, including the deployment mode.}
}
\description{
Deletes a specified stack. Once the call completes successfully, stack deletion starts. Deleted stacks don't show up in the \code{\link[=cloudformation_describe_stacks]{describe_stacks}} operation if the deletion has been completed successfully.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_delete_stack/} for full documentation.
}
\keyword{internal}
