% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_deregister_type}
\alias{cloudformation_deregister_type}
\title{Marks an extension or extension version as DEPRECATED in the
CloudFormation registry, removing it from active use}
\usage{
cloudformation_deregister_type(Arn, Type, TypeName, VersionId)
}
\arguments{
\item{Arn}{The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{Type}{The kind of extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{TypeName}{The name of the extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{VersionId}{The ID of a specific version of the extension. The version ID is the
value at the end of the Amazon Resource Name (ARN) assigned to the
extension version when it is registered.}
}
\value{
An empty list.
}
\description{
Marks an extension or extension version as \code{DEPRECATED} in the
CloudFormation registry, removing it from active use. Deprecated
extensions or extension versions cannot be used in CloudFormation
operations.

To deregister an entire extension, you must individually deregister all
active versions of that extension. If an extension has only a single
active version, deregistering that version results in the extension
itself being deregistered and marked as deprecated in the registry.

You cannot deregister the default version of an extension if there are
other active version of that extension. If you do deregister the default
version of an extension, the textensionype itself is deregistered as
well and marked as deprecated.

To view the deprecation status of an extension or extension version, use
\code{\link[=cloudformation_describe_type]{describe_type}}.
}
\section{Request syntax}{
\preformatted{svc$deregister_type(
  Arn = "string",
  Type = "RESOURCE"|"MODULE",
  TypeName = "string",
  VersionId = "string"
)
}
}

\keyword{internal}
