% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_generated_template}
\alias{cloudformation_describe_generated_template}
\title{Describes a generated template}
\usage{
cloudformation_describe_generated_template(GeneratedTemplateName)
}
\arguments{
\item{GeneratedTemplateName}{[required] The name or Amazon Resource Name (ARN) of a generated template.}
}
\description{
Describes a generated template. The output includes details about the progress of the creation of a generated template started by a \code{\link[=cloudformation_create_generated_template]{create_generated_template}} API action or the update of a generated template started with an \code{\link[=cloudformation_update_generated_template]{update_generated_template}} API action.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_generated_template/} for full documentation.
}
\keyword{internal}
