% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_publisher}
\alias{cloudformation_describe_publisher}
\title{Returns information about a CloudFormation extension publisher}
\usage{
cloudformation_describe_publisher(PublisherId = NULL)
}
\arguments{
\item{PublisherId}{The ID of the extension publisher.

If you don't supply a \code{PublisherId}, and you have registered as an
extension publisher,
\code{\link[=cloudformation_describe_publisher]{describe_publisher}} returns
information about your own publisher account.}
}
\description{
Returns information about a CloudFormation extension publisher.

See \url{https://paws-r.github.io/docs/cloudformation/describe_publisher.html} for full documentation.
}
\keyword{internal}
