% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_resource_scan}
\alias{cloudformation_describe_resource_scan}
\title{Describes details of a resource scan}
\usage{
cloudformation_describe_resource_scan(ResourceScanId)
}
\arguments{
\item{ResourceScanId}{[required] The Amazon Resource Name (ARN) of the resource scan.}
}
\description{
Describes details of a resource scan.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_resource_scan/} for full documentation.
}
\keyword{internal}
