% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_events}
\alias{cloudformation_describe_stack_events}
\title{Returns all stack related events for a specified stack in reverse
chronological order}
\usage{
cloudformation_describe_stack_events(StackName = NULL, NextToken = NULL)
}
\arguments{
\item{StackName}{The name or the unique stack ID that's associated with the stack, which
aren't always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack's name or its
unique stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}

Default: There is no default value.}

\item{NextToken}{A string that identifies the next page of events that you want to
retrieve.}
}
\description{
Returns all stack related events for a specified stack in reverse chronological order. For more information about a stack's event history, go to \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cfn-whatis-concepts.html}{Stacks} in the CloudFormation User Guide.

See \url{https://paws-r.github.io/docs/cloudformation/describe_stack_events.html} for full documentation.
}
\keyword{internal}
