% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_refactor}
\alias{cloudformation_describe_stack_refactor}
\title{Describes the stack refactor status}
\usage{
cloudformation_describe_stack_refactor(StackRefactorId)
}
\arguments{
\item{StackRefactorId}{[required] The ID associated with the stack refactor created from the
\code{\link[=cloudformation_create_stack_refactor]{create_stack_refactor}} action.}
}
\description{
Describes the stack refactor status.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_stack_refactor/} for full documentation.
}
\keyword{internal}
