% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_get_hook_result}
\alias{cloudformation_get_hook_result}
\title{Retrieves detailed information and remediation guidance for a Hook
invocation result}
\usage{
cloudformation_get_hook_result(HookResultId = NULL)
}
\arguments{
\item{HookResultId}{The unique identifier (ID) of the Hook invocation result that you want
details about. You can get the ID from the
\code{\link[=cloudformation_list_hook_results]{list_hook_results}} operation.}
}
\description{
Retrieves detailed information and remediation guidance for a Hook invocation result.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_get_hook_result/} for full documentation.
}
\keyword{internal}
