% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_get_stack_policy}
\alias{cloudformation_get_stack_policy}
\title{Returns the stack policy for a specified stack}
\usage{
cloudformation_get_stack_policy(StackName)
}
\arguments{
\item{StackName}{[required] The name or unique stack ID that is associated with the stack whose
policy you want to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StackPolicyBody = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the stack policy for a specified stack. If a stack doesn't have
a policy, a null value is returned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_stack_policy(
  StackName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
