% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_generated_templates}
\alias{cloudformation_list_generated_templates}
\title{Lists your generated templates in this Region}
\usage{
cloudformation_list_generated_templates(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{If the number of available results exceeds this maximum, the response
includes a \code{NextToken} value that you can use for the \code{NextToken}
parameter to get the next set of results. By default the
\code{\link[=cloudformation_list_generated_templates]{list_generated_templates}}
API action will return at most 50 results in each response. The maximum
value is 100.}
}
\description{
Lists your generated templates in this Region.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_generated_templates/} for full documentation.
}
\keyword{internal}
