% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_resource_scans}
\alias{cloudformation_list_resource_scans}
\title{List the resource scans from newest to oldest}
\usage{
cloudformation_list_resource_scans(
  NextToken = NULL,
  MaxResults = NULL,
  ScanTypeFilter = NULL
)
}
\arguments{
\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{If the number of available results exceeds this maximum, the response
includes a \code{NextToken} value that you can use for the \code{NextToken}
parameter to get the next set of results. The default value is 10. The
maximum value is 100.}

\item{ScanTypeFilter}{The scan type that you want to get summary information about. The
default is \code{FULL}.}
}
\description{
List the resource scans from newest to oldest. By default it will return up to 10 resource scans.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_resource_scans/} for full documentation.
}
\keyword{internal}
