% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_stack_refactor_actions}
\alias{cloudformation_list_stack_refactor_actions}
\title{Lists the stack refactor actions that will be taken after calling the
ExecuteStackRefactor action}
\usage{
cloudformation_list_stack_refactor_actions(
  StackRefactorId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{StackRefactorId}{[required] The ID associated with the stack refactor created from the
\code{\link[=cloudformation_create_stack_refactor]{create_stack_refactor}} action.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of results to be returned with a single call. If the
number of available results exceeds this maximum, the response includes
a \code{NextToken} value that you can assign to the \code{NextToken} request
parameter to get the next set of results.}
}
\description{
Lists the stack refactor actions that will be taken after calling the \code{\link[=cloudformation_execute_stack_refactor]{execute_stack_refactor}} action.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_stack_refactor_actions/} for full documentation.
}
\keyword{internal}
