% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_stack_refactors}
\alias{cloudformation_list_stack_refactors}
\title{Lists all account stack refactor operations and their statuses}
\usage{
cloudformation_list_stack_refactors(
  ExecutionStatusFilter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ExecutionStatusFilter}{Execution status to use as a filter. Specify one or more execution
status codes to list only stack refactors with the specified execution
status codes.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of results to be returned with a single call. If the
number of available results exceeds this maximum, the response includes
a \code{NextToken} value that you can assign to the \code{NextToken} request
parameter to get the next set of results.}
}
\description{
Lists all account stack refactor operations and their statuses.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_stack_refactors/} for full documentation.
}
\keyword{internal}
