% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_stack_set_operations}
\alias{cloudformation_list_stack_set_operations}
\title{Returns summary information about operations performed on a stack set}
\usage{
cloudformation_list_stack_set_operations(StackSetName, NextToken,
  MaxResults)
}
\arguments{
\item{StackSetName}{[required] The name or unique ID of the stack set that you want to get operation
summaries for.}

\item{NextToken}{If the previous paginated request didn't return all of the remaining
results, the response object's \code{NextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{\link[=cloudformation_list_stack_set_operations]{list_stack_set_operations}}
again and assign that token to the request object's \code{NextToken}
parameter. If there are no remaining results, the previous response
object's \code{NextToken} parameter is set to \code{null}.}

\item{MaxResults}{The maximum number of results to be returned with a single call. If the
number of available results exceeds this maximum, the response includes
a \code{NextToken} value that you can assign to the \code{NextToken} request
parameter to get the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Summaries = list(
    list(
      OperationId = "string",
      Action = "CREATE"|"UPDATE"|"DELETE"|"DETECT_DRIFT",
      Status = "RUNNING"|"SUCCEEDED"|"FAILED"|"STOPPING"|"STOPPED"|"QUEUED",
      CreationTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      EndTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns summary information about operations performed on a stack set.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_stack_set_operations(
  StackSetName = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
