% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_stop_stack_set_operation}
\alias{cloudformation_stop_stack_set_operation}
\title{Stops an in-progress operation on a stack set and its associated stack
instances}
\usage{
cloudformation_stop_stack_set_operation(StackSetName, OperationId,
  CallAs)
}
\arguments{
\item{StackSetName}{[required] The name or unique ID of the stack set that you want to stop the
operation for.}

\item{OperationId}{[required] The ID of the stack operation.}

\item{CallAs}{[Service-managed permissions] Specifies whether you are acting as an
account administrator in the organization's management account or as a
delegated administrator in a member account.

By default, \code{SELF} is specified. Use \code{SELF} for stack sets with
self-managed permissions.
\itemize{
\item If you are signed in to the management account, specify \code{SELF}.
\item If you are signed in to a delegated administrator account, specify
\code{DELEGATED_ADMIN}.

Your AWS account must be registered as a delegated administrator in
the management account. For more information, see \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/stacksets-orgs-delegated-admin.html}{Register a delegated administrator}
in the \emph{AWS CloudFormation User Guide}.
}}
}
\value{
An empty list.
}
\description{
Stops an in-progress operation on a stack set and its associated stack
instances.
}
\section{Request syntax}{
\preformatted{svc$stop_stack_set_operation(
  StackSetName = "string",
  OperationId = "string",
  CallAs = "SELF"|"DELEGATED_ADMIN"
)
}
}

\keyword{internal}
