% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_update_termination_protection}
\alias{cloudformation_update_termination_protection}
\title{Updates termination protection for the specified stack}
\usage{
cloudformation_update_termination_protection(
  EnableTerminationProtection, StackName)
}
\arguments{
\item{EnableTerminationProtection}{[required] Whether to enable termination protection on the specified stack.}

\item{StackName}{[required] The name or unique ID of the stack for which you want to set termination
protection.}
}
\value{
A list with the following syntax:\preformatted{list(
  StackId = "string"
)
}
}
\description{
Updates termination protection for the specified stack. If a user
attempts to delete a stack with termination protection enabled, the
operation fails and the stack remains unchanged. For more information,
see \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-protect-stacks.html}{Protecting a Stack From Being Deleted}
in the \emph{AWS CloudFormation User Guide}.

For \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-nested-stacks.html}{nested stacks},
termination protection is set on the root stack and cannot be changed
directly on the nested stack.
}
\section{Request syntax}{
\preformatted{svc$update_termination_protection(
  EnableTerminationProtection = TRUE|FALSE,
  StackName = "string"
)
}
}

\keyword{internal}
