% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_cancel_query}
\alias{cloudtrail_cancel_query}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_cancel_query(
  EventDataStore = NULL,
  QueryId,
  EventDataStoreOwnerAccountId = NULL
)
}
\arguments{
\item{EventDataStore}{The ARN (or the ID suffix of the ARN) of an event data store on which the specified query is running.}

\item{QueryId}{[required] The ID of the query that you want to cancel. The \code{QueryId} comes from the response of a \code{\link[=cloudtrail_start_query]{start_query}} operation.}

\item{EventDataStoreOwnerAccountId}{The account ID of the event data store owner.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_cancel_query/} for full documentation.
}
\keyword{internal}
