% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_create_channel}
\alias{cloudtrail_create_channel}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_create_channel(Name, Source, Destinations, Tags = NULL)
}
\arguments{
\item{Name}{[required] The name of the channel.}

\item{Source}{[required] The name of the partner or external event source. You cannot change this name after you create the channel. A maximum of one channel is allowed per source.

A source can be either \code{Custom} for all valid non-Amazon Web Services events, or the name of a partner event source. For information about the source names for available partners, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/query-event-data-store-integration.html#cloudtrail-lake-partner-information}{Additional information about integration partners} in the CloudTrail User Guide.}

\item{Destinations}{[required] One or more event data stores to which events arriving through a channel will be logged.}

\item{Tags}{A list of tags.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_create_channel/} for full documentation.
}
\keyword{internal}
