% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_deregister_organization_delegated_admin}
\alias{cloudtrail_deregister_organization_delegated_admin}
\title{Removes CloudTrail delegated administrator permissions from a member
account in an organization}
\usage{
cloudtrail_deregister_organization_delegated_admin(DelegatedAdminAccountId)
}
\arguments{
\item{DelegatedAdminAccountId}{[required] A delegated administrator account ID. This is a member account in an
organization that is currently designated as a delegated administrator.}
}
\description{
Removes CloudTrail delegated administrator permissions from a member account in an organization.

See \url{https://paws-r.github.io/docs/cloudtrail/deregister_organization_delegated_admin.html} for full documentation.
}
\keyword{internal}
