% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_describe_query}
\alias{cloudtrail_describe_query}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_describe_query(
  EventDataStore = NULL,
  QueryId = NULL,
  QueryAlias = NULL,
  RefreshId = NULL,
  EventDataStoreOwnerAccountId = NULL
)
}
\arguments{
\item{EventDataStore}{The ARN (or the ID suffix of the ARN) of an event data store on which the specified query was run.}

\item{QueryId}{The query ID.}

\item{QueryAlias}{The alias that identifies a query template.}

\item{RefreshId}{The ID of the dashboard refresh.}

\item{EventDataStoreOwnerAccountId}{The account ID of the event data store owner.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_describe_query/} for full documentation.
}
\keyword{internal}
