% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_enable_federation}
\alias{cloudtrail_enable_federation}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_enable_federation(EventDataStore, FederationRoleArn)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or ID suffix of the ARN) of the event data store for which you want to enable Lake query federation.}

\item{FederationRoleArn}{[required] The ARN of the federation role to use for the event data store. Amazon Web Services services like Lake Formation use this federation role to access data for the federated event data store. The federation role must exist in your account and provide the \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/query-federation.html#query-federation-permissions-role}{required minimum permissions}.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_enable_federation/} for full documentation.
}
\keyword{internal}
