% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_query_results}
\alias{cloudtrail_get_query_results}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_get_query_results(
  EventDataStore = NULL,
  QueryId,
  NextToken = NULL,
  MaxQueryResults = NULL,
  EventDataStoreOwnerAccountId = NULL
)
}
\arguments{
\item{EventDataStore}{The ARN (or ID suffix of the ARN) of the event data store against which the query was run.}

\item{QueryId}{[required] The ID of the query for which you want to get results.}

\item{NextToken}{A token you can use to get the next page of query results.}

\item{MaxQueryResults}{The maximum number of query results to display on a single page.}

\item{EventDataStoreOwnerAccountId}{The account ID of the event data store owner.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_get_query_results/} for full documentation.
}
\keyword{internal}
