% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_channels}
\alias{cloudtrail_list_channels}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_list_channels(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of CloudTrail channels to display on a single page.}

\item{NextToken}{The token to use to get the next page of results after a previous API call. This token must be passed in with the same parameters that were specified in the original call. For example, if the original call specified an AttributeKey of 'Username' with a value of 'root', the call with NextToken should include those same parameters.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_channels/} for full documentation.
}
\keyword{internal}
