% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_insights_data}
\alias{cloudtrail_list_insights_data}
\title{Returns Insights events generated on a trail that logs data events}
\usage{
cloudtrail_list_insights_data(
  InsightSource,
  DataType,
  Dimensions = NULL,
  StartTime = NULL,
  EndTime = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InsightSource}{[required] The Amazon Resource Name(ARN) of the trail for which you want to
retrieve Insights events.}

\item{DataType}{[required] Specifies the category of events returned. To fetch Insights events,
specify \code{InsightsEvents} as the value of \code{DataType}}

\item{Dimensions}{Contains a map of dimensions. Currently the map can contain only one
item.}

\item{StartTime}{Specifies that only events that occur after or at the specified time are
returned. If the specified start time is after the specified end time,
an error is returned.}

\item{EndTime}{Specifies that only events that occur before or at the specified time
are returned. If the specified end time is before the specified start
time, an error is returned.}

\item{MaxResults}{The number of events to return. Possible values are 1 through 50. The
default is 50.}

\item{NextToken}{The token to use to get the next page of results after a previous API
call. This token must be passed in with the same parameters that were
specified in the original call. For example, if the original call
specified a EventName as a dimension with \code{PutObject} as a value, the
call with NextToken should include those same parameters.}
}
\description{
Returns Insights events generated on a trail that logs data events. You can list Insights events that occurred in a Region within the last 90 days.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_insights_data/} for full documentation.
}
\keyword{internal}
