% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_queries}
\alias{cloudtrail_list_queries}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_list_queries(
  EventDataStore,
  NextToken = NULL,
  MaxResults = NULL,
  StartTime = NULL,
  EndTime = NULL,
  QueryStatus = NULL
)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or the ID suffix of the ARN) of an event data store on which queries were run.}

\item{NextToken}{A token you can use to get the next page of results.}

\item{MaxResults}{The maximum number of queries to show on a page.}

\item{StartTime}{Use with \code{EndTime} to bound a \code{\link[=cloudtrail_list_queries]{list_queries}} request, and limit its results to only those queries run within a specified time period.}

\item{EndTime}{Use with \code{StartTime} to bound a \code{\link[=cloudtrail_list_queries]{list_queries}} request, and limit its results to only those queries run within a specified time period.}

\item{QueryStatus}{The status of queries that you want to return in results. Valid values for \code{QueryStatus} include \code{QUEUED}, \code{RUNNING}, \code{FINISHED}, \code{FAILED}, \code{TIMED_OUT}, or \code{CANCELLED}.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_queries/} for full documentation.
}
\keyword{internal}
