% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_put_insight_selectors}
\alias{cloudtrail_put_insight_selectors}
\title{Lets you enable Insights event logging by specifying the Insights
selectors that you want to enable on an existing trail}
\usage{
cloudtrail_put_insight_selectors(TrailName, InsightSelectors)
}
\arguments{
\item{TrailName}{[required] The name of the CloudTrail trail for which you want to change or add
Insights selectors.}

\item{InsightSelectors}{[required] A JSON string that contains the insight types you want to log on a
trail. \code{ApiCallRateInsight} and \code{ApiErrorRateInsight} are valid Insight
types.

The \code{ApiCallRateInsight} Insights type analyzes write-only management
API calls that are aggregated per minute against a baseline API call
volume.

The \code{ApiErrorRateInsight} Insights type analyzes management API calls
that result in error codes. The error is shown if the API call is
unsuccessful.}
}
\description{
Lets you enable Insights event logging by specifying the Insights selectors that you want to enable on an existing trail. You also use \code{\link[=cloudtrail_put_insight_selectors]{put_insight_selectors}} to turn off Insights event logging, by passing an empty list of insight types. The valid Insights event types in this release are \code{ApiErrorRateInsight} and \code{ApiCallRateInsight}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_put_insight_selectors/} for full documentation.
}
\keyword{internal}
