% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_register_organization_delegated_admin}
\alias{cloudtrail_register_organization_delegated_admin}
\title{Registers an organization’s member account as the CloudTrail delegated
administrator}
\usage{
cloudtrail_register_organization_delegated_admin(MemberAccountId)
}
\arguments{
\item{MemberAccountId}{[required] An organization member account ID that you want to designate as a
delegated administrator.}
}
\description{
Registers an organization’s member account as the CloudTrail delegated administrator.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_register_organization_delegated_admin/} for full documentation.
}
\keyword{internal}
