% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_search_sample_queries}
\alias{cloudtrail_search_sample_queries}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_search_sample_queries(
  SearchPhrase,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{SearchPhrase}{[required] The natural language phrase to use for the semantic search. The phrase must be in English. The length constraint is in characters, not words.}

\item{MaxResults}{The maximum number of results to return on a single page. The default value is 10.}

\item{NextToken}{A token you can use to get the next page of results. The length constraint is in characters, not words.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_search_sample_queries/} for full documentation.
}
\keyword{internal}
