% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_start_import}
\alias{cloudtrail_start_import}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_start_import(
  Destinations = NULL,
  ImportSource = NULL,
  StartEventTime = NULL,
  EndEventTime = NULL,
  ImportId = NULL
)
}
\arguments{
\item{Destinations}{The ARN of the destination event data store. Use this parameter for a new import.}

\item{ImportSource}{The source S3 bucket for the import. Use this parameter for a new import.}

\item{StartEventTime}{Use with \code{EndEventTime} to bound a \code{\link[=cloudtrail_start_import]{start_import}} request, and limit imported trail events to only those events logged within a specified time period. When you specify a time range, CloudTrail checks the prefix and log file names to verify the names contain a date between the specified \code{StartEventTime} and \code{EndEventTime} before attempting to import events.}

\item{EndEventTime}{Use with \code{StartEventTime} to bound a \code{\link[=cloudtrail_start_import]{start_import}} request, and limit imported trail events to only those events logged within a specified time period. When you specify a time range, CloudTrail checks the prefix and log file names to verify the names contain a date between the specified \code{StartEventTime} and \code{EndEventTime} before attempting to import events.}

\item{ImportId}{The ID of the import. Use this parameter when you are retrying an import.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_start_import/} for full documentation.
}
\keyword{internal}
