% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_start_logging}
\alias{cloudtrail_start_logging}
\title{Starts the recording of Amazon Web Services API calls and log file
delivery for a trail}
\usage{
cloudtrail_start_logging(Name)
}
\arguments{
\item{Name}{[required] Specifies the name or the CloudTrail ARN of the trail for which
CloudTrail logs Amazon Web Services API calls. The following is the
format of a trail ARN.

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}
}
\description{
Starts the recording of Amazon Web Services API calls and log file delivery for a trail. For a trail that is enabled in all Regions, this operation must be called from the Region in which the trail was created. This operation cannot be called on the shadow trails (replicated trails in other Regions) of a trail that is enabled in all Regions.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_start_logging/} for full documentation.
}
\keyword{internal}
