% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_start_query}
\alias{cloudtrail_start_query}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_start_query(
  QueryStatement = NULL,
  DeliveryS3Uri = NULL,
  QueryAlias = NULL,
  QueryParameters = NULL,
  EventDataStoreOwnerAccountId = NULL
)
}
\arguments{
\item{QueryStatement}{The SQL code of your query.}

\item{DeliveryS3Uri}{The URI for the S3 bucket where CloudTrail delivers the query results.}

\item{QueryAlias}{The alias that identifies a query template.}

\item{QueryParameters}{The query parameters for the specified \code{QueryAlias}.}

\item{EventDataStoreOwnerAccountId}{The account ID of the event data store owner.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_start_query/} for full documentation.
}
\keyword{internal}
