% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_update_channel}
\alias{cloudtrail_update_channel}
\title{CloudTrail Lake will no longer be open to new customers starting May 31,
2026}
\usage{
cloudtrail_update_channel(Channel, Destinations = NULL, Name = NULL)
}
\arguments{
\item{Channel}{[required] The ARN or ID (the ARN suffix) of the channel that you want to update.}

\item{Destinations}{The ARNs of event data stores that you want to log events arriving through the channel.}

\item{Name}{Changes the name of the channel.}
}
\description{
CloudTrail Lake will no longer be open to new customers starting May 31, 2026. If you would like to use CloudTrail Lake, sign up prior to that date. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-lake-service-availability-change.html}{CloudTrail Lake availability change}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_update_channel/} for full documentation.
}
\keyword{internal}
