% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_update_trail}
\alias{cloudtrail_update_trail}
\title{Updates trail settings that control what events you are logging, and how
to handle log files}
\usage{
cloudtrail_update_trail(
  Name,
  S3BucketName = NULL,
  S3KeyPrefix = NULL,
  SnsTopicName = NULL,
  IncludeGlobalServiceEvents = NULL,
  IsMultiRegionTrail = NULL,
  EnableLogFileValidation = NULL,
  CloudWatchLogsLogGroupArn = NULL,
  CloudWatchLogsRoleArn = NULL,
  KmsKeyId = NULL,
  IsOrganizationTrail = NULL,
  RecursiveLogging = NULL
)
}
\arguments{
\item{Name}{[required] Specifies the name of the trail or trail ARN. If \code{Name} is a trail name, the string must meet the following requirements:
\itemize{
\item Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)
\item Start with a letter or number, and end with a letter or number
\item Be between 3 and 128 characters
\item Have no adjacent periods, underscores or dashes. Names like \verb{my-_namespace} and \code{my--namespace} are not valid.
\item Not be in IP address format (for example, 192.168.5.4)
}

If \code{Name} is a trail ARN, it must be in the following format.

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}

\item{S3BucketName}{Specifies the name of the Amazon S3 bucket designated for publishing log files. See \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/bucketnamingrules.html}{Amazon S3 Bucket naming rules}.}

\item{S3KeyPrefix}{Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/get-and-view-cloudtrail-log-files.html#cloudtrail-find-log-files}{Finding Your CloudTrail Log Files}. The maximum length is 200 characters.}

\item{SnsTopicName}{Specifies the name or ARN of the Amazon SNS topic defined for notification of log file delivery. The maximum length is 256 characters.}

\item{IncludeGlobalServiceEvents}{Specifies whether the trail is publishing events from global services such as IAM to the log files. Setting this value to \code{true} only delivers global service events to the trail if the trail is multi-Region or if the trail's home Region is the partition leader Region (for example, us-east-1).}

\item{IsMultiRegionTrail}{Specifies whether the trail applies only to the current Region or to all Regions. The default is false. If the trail exists only in the current Region and this value is set to true, shadow trails (replications of the trail) will be created in the other Regions. If the trail exists in all Regions and this value is set to false, the trail will remain in the Region where it was created, and its shadow trails in other Regions will be deleted. As a best practice, consider using trails that log events in all Regions.}

\item{EnableLogFileValidation}{Specifies whether log file validation is enabled. The default is false.

When you disable log file integrity validation, the chain of digest files is broken after one hour. CloudTrail does not create digest files for log files that were delivered during a period in which log file integrity validation was disabled. For example, if you enable log file integrity validation at noon on January 1, disable it at noon on January 2, and re-enable it at noon on January 10, digest files will not be created for the log files delivered from noon on January 2 to noon on January 10. The same applies whenever you stop CloudTrail logging or delete a trail.}

\item{CloudWatchLogsLogGroupArn}{Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs are delivered. You must use a log group that exists in your account.

Not required unless you specify \code{CloudWatchLogsRoleArn}.}

\item{CloudWatchLogsRoleArn}{Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group. You must use a role that exists in your account.}

\item{KmsKeyId}{Specifies the KMS key ID to use to encrypt the logs and digest files delivered by CloudTrail. The value can be an alias name prefixed by "alias/", a fully specified ARN to an alias, a fully specified ARN to a key, or a globally unique identifier.

CloudTrail also supports KMS multi-Region keys. For more information about multi-Region keys, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/multi-region-keys-overview.html}{Using multi-Region keys} in the \emph{Key Management Service Developer Guide}.

Examples:
\itemize{
\item alias/MyAliasName
\item arn:aws:kms:us-east-2:123456789012:alias/MyAliasName
\item arn:aws:kms:us-east-2:123456789012:key/12345678-1234-1234-1234-123456789012
\item 12345678-1234-1234-1234-123456789012
}}

\item{IsOrganizationTrail}{Specifies whether the trail is applied to all accounts in an organization in Organizations, or only for the current Amazon Web Services account. The default is false, and cannot be true unless the call is made on behalf of an Amazon Web Services account that is the management account for an organization in Organizations. If the trail is not an organization trail and this is set to \code{true}, the trail will be created in all Amazon Web Services accounts that belong to the organization. If the trail is an organization trail and this is set to \code{false}, the trail will remain in the current Amazon Web Services account but be deleted from all member accounts in the organization.

Only the management account for the organization can convert an organization trail to a non-organization trail, or convert a non-organization trail to an organization trail.}

\item{RecursiveLogging}{Specifies whether recursive logging is enabled for the trail. If you set \code{RecursiveLogging} to \code{false}, CloudTrail suppresses events generated by CloudTrail when it delivers log files to your trail's destinations, including Amazon S3 and CloudWatch Logs. The default value is \code{true}.}
}
\description{
Updates trail settings that control what events you are logging, and how to handle log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket. \code{\link[=cloudtrail_update_trail]{update_trail}} must be called from the Region in which the trail was created; otherwise, an \code{InvalidHomeRegionException} is thrown.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_update_trail/} for full documentation.
}
\keyword{internal}
