% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtraildataservice_operations.R
\name{cloudtraildataservice_put_audit_events}
\alias{cloudtraildataservice_put_audit_events}
\title{Ingests your application events into CloudTrail Lake}
\usage{
cloudtraildataservice_put_audit_events(
  auditEvents,
  channelArn,
  externalId = NULL
)
}
\arguments{
\item{auditEvents}{[required] The JSON payload of events that you want to ingest. You can also point
to the JSON event payload in a file.}

\item{channelArn}{[required] The ARN or ID (the ARN suffix) of a channel.}

\item{externalId}{A unique identifier that is conditionally required when the channel's
resource policy includes an external ID. This value can be any string,
such as a passphrase or account number.}
}
\description{
Ingests your application events into CloudTrail Lake. A required parameter, \code{auditEvents}, accepts the JSON records (also called \emph{payload}) of events that you want CloudTrail to ingest. You can add up to 100 of these events (or up to 1 MB) per \code{\link[=cloudtraildataservice_put_audit_events]{put_audit_events}} request.

See \url{https://www.paws-r-sdk.com/docs/cloudtraildataservice_put_audit_events/} for full documentation.
}
\keyword{internal}
