% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_associate_dataset_kms_key}
\alias{cloudwatch_associate_dataset_kms_key}
\title{Associates an Amazon Web Services Key Management Service (Amazon Web
Services KMS) customer managed key with the specified dataset}
\usage{
cloudwatch_associate_dataset_kms_key(DatasetIdentifier, KmsKeyArn)
}
\arguments{
\item{DatasetIdentifier}{[required] Specifies the identifier of the dataset that you want to associate the KMS key with. For the \code{default} dataset, you can specify either \code{default} or the full dataset Amazon Resource Name (ARN) in the format \code{arn:aws:cloudwatch:Region:account-id:dataset/default}.}

\item{KmsKeyArn}{[required] Specifies the Amazon Resource Name (ARN) of the customer managed KMS key to associate with the dataset. The key must be a symmetric encryption KMS key (\code{SYMMETRIC_DEFAULT}) in the same Amazon Web Services Region as the dataset.

The ARN must be in the format \code{arn:aws:kms:Region:account-id:key/key-id }. Key IDs, aliases, and alias ARNs are not accepted.

For more information about KMS key ARNs, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html#key-id-key-ARN}{Key ARN} in the \emph{Amazon Web Services Key Management Service Developer Guide}.}
}
\description{
Associates an Amazon Web Services Key Management Service (Amazon Web Services KMS) customer managed key with the specified dataset. After this operation completes, all data published to the dataset is encrypted at rest using the specified KMS key. Callers must have \code{kms:Decrypt} permission on the key to read the encrypted data.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_associate_dataset_kms_key/} for full documentation.
}
\keyword{internal}
