% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_create_resource_metrics_configuration}
\alias{cloudwatch_create_resource_metrics_configuration}
\title{Creates a resource metrics configuration for an Amazon Web Services
resource}
\usage{
cloudwatch_create_resource_metrics_configuration(
  ResourceArn,
  MetricSelections = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to enable detailed monitoring for.}

\item{MetricSelections}{Specifies which metrics Amazon CloudWatch collects for the resource. If you omit this parameter, Amazon CloudWatch collects all available detailed metrics for the resource.}
}
\description{
Creates a resource metrics configuration for an Amazon Web Services resource. After you create a configuration, Amazon CloudWatch collects detailed metrics for that resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_create_resource_metrics_configuration/} for full documentation.
}
\keyword{internal}
