% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_alarm_mute_rule}
\alias{cloudwatch_delete_alarm_mute_rule}
\title{Deletes a specific alarm mute rule}
\usage{
cloudwatch_delete_alarm_mute_rule(AlarmMuteRuleName)
}
\arguments{
\item{AlarmMuteRuleName}{[required] The name of the alarm mute rule to delete.}
}
\description{
Deletes a specific alarm mute rule.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_delete_alarm_mute_rule/} for full documentation.
}
\keyword{internal}
