% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_alarms}
\alias{cloudwatch_delete_alarms}
\title{Deletes the specified alarms}
\usage{
cloudwatch_delete_alarms(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The alarms to be deleted. Do not enclose the alarm names in quote marks.}
}
\description{
Deletes the specified alarms. You can delete up to 100 alarms in one operation. However, this total can include no more than one composite alarm. For example, you could delete 99 metric alarms and one composite alarms with one operation, but you can't delete two composite alarms with one operation. Log alarms cannot be batch deleted.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_delete_alarms/} for full documentation.
}
\keyword{internal}
