% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_insight_rules}
\alias{cloudwatch_delete_insight_rules}
\title{Permanently deletes the specified Contributor Insights rules}
\usage{
cloudwatch_delete_insight_rules(RuleNames)
}
\arguments{
\item{RuleNames}{[required] An array of the rule names to delete. If you need to find out the names
of your rules, use
\code{\link[=cloudwatch_describe_insight_rules]{describe_insight_rules}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Failures = list(
    list(
      FailureResource = "string",
      ExceptionType = "string",
      FailureCode = "string",
      FailureDescription = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Permanently deletes the specified Contributor Insights rules.

If you create a rule, delete it, and then re-create it with the same
name, historical data from the first time the rule was created might not
be available.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_insight_rules(
  RuleNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
