% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_resource_metrics_configuration}
\alias{cloudwatch_delete_resource_metrics_configuration}
\title{Deletes the resource metrics configuration for an Amazon Web Services
resource}
\usage{
cloudwatch_delete_resource_metrics_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to delete the resource metrics configuration for.}
}
\description{
Deletes the resource metrics configuration for an Amazon Web Services resource. After you delete the configuration, Amazon CloudWatch stops collecting detailed metrics for the resource. Metric data that Amazon CloudWatch already collected for the resource is not deleted.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_delete_resource_metrics_configuration/} for full documentation.
}
\keyword{internal}
