% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_alarm_contributors}
\alias{cloudwatch_describe_alarm_contributors}
\title{Returns the information of the current alarm contributors that are in
ALARM state}
\usage{
cloudwatch_describe_alarm_contributors(AlarmName, NextToken = NULL)
}
\arguments{
\item{AlarmName}{[required] The name of the alarm for which to retrieve contributor information.}

\item{NextToken}{The token returned by a previous call to indicate that there is more
data available.}
}
\description{
Returns the information of the current alarm contributors that are in \code{ALARM} state. This operation returns details about the individual time series that contribute to the alarm's state.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_describe_alarm_contributors/} for full documentation.
}
\keyword{internal}
