% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_insight_rules}
\alias{cloudwatch_describe_insight_rules}
\title{Returns a list of all the Contributor Insights rules in your account}
\usage{
cloudwatch_describe_insight_rules(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of rules.}

\item{MaxResults}{The maximum number of results to return in one operation. If you omit
this parameter, the default of 500 is used.}
}
\description{
Returns a list of all the Contributor Insights rules in your account.

For more information about Contributor Insights, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/ContributorInsights.html}{Using Contributor Insights to Analyze High-Cardinality Data}.
}
\section{Request syntax}{
\preformatted{svc$describe_insight_rules(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
