% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_disable_alarm_actions}
\alias{cloudwatch_disable_alarm_actions}
\title{Disables the actions for the specified alarms}
\usage{
cloudwatch_disable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\description{
Disables the actions for the specified alarms. When an alarm's actions are disabled, the alarm actions do not execute when the alarm state changes.

See \url{https://paws-r.github.io/docs/cloudwatch/disable_alarm_actions.html} for full documentation.
}
\keyword{internal}
