% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_disable_insight_rules}
\alias{cloudwatch_disable_insight_rules}
\title{Disables the specified Contributor Insights rules}
\usage{
cloudwatch_disable_insight_rules(RuleNames)
}
\arguments{
\item{RuleNames}{[required] An array of the rule names to disable. If you need to find out the names
of your rules, use
\code{\link[=cloudwatch_describe_insight_rules]{describe_insight_rules}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Failures = list(
    list(
      FailureResource = "string",
      ExceptionType = "string",
      FailureCode = "string",
      FailureDescription = "string"
    )
  )
)
}
}
\description{
Disables the specified Contributor Insights rules. When rules are
disabled, they do not analyze log groups and do not incur costs.
}
\section{Request syntax}{
\preformatted{svc$disable_insight_rules(
  RuleNames = list(
    "string"
  )
)
}
}

\keyword{internal}
