% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_alarm_mute_rule}
\alias{cloudwatch_get_alarm_mute_rule}
\title{Retrieves details for a specific alarm mute rule}
\usage{
cloudwatch_get_alarm_mute_rule(AlarmMuteRuleName)
}
\arguments{
\item{AlarmMuteRuleName}{[required] The name of the alarm mute rule to retrieve.}
}
\description{
Retrieves details for a specific alarm mute rule.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_get_alarm_mute_rule/} for full documentation.
}
\keyword{internal}
