% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_dashboard}
\alias{cloudwatch_get_dashboard}
\title{Displays the details of the dashboard that you specify}
\usage{
cloudwatch_get_dashboard(DashboardName)
}
\arguments{
\item{DashboardName}{[required] The name of the dashboard to be described.}
}
\description{
Displays the details of the dashboard that you specify.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_get_dashboard/} for full documentation.
}
\keyword{internal}
