% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_dataset}
\alias{cloudwatch_get_dataset}
\title{Returns information about the specified dataset}
\usage{
cloudwatch_get_dataset(DatasetIdentifier)
}
\arguments{
\item{DatasetIdentifier}{[required] Specifies the identifier of the dataset to retrieve. For the \code{default} dataset, you can specify either \code{default} or the full dataset Amazon Resource Name (ARN) in the format \code{arn:aws:cloudwatch:Region:account-id:dataset/default}.}
}
\description{
Returns information about the specified dataset. This includes its identifier, Amazon Resource Name (ARN), and any customer managed Amazon Web Services Key Management Service (Amazon Web Services KMS) key that is currently associated with it.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_get_dataset/} for full documentation.
}
\keyword{internal}
