% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_metric_stream}
\alias{cloudwatch_get_metric_stream}
\title{Returns information about the metric stream that you specify}
\usage{
cloudwatch_get_metric_stream(Name)
}
\arguments{
\item{Name}{[required] The name of the metric stream to retrieve information about.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  Name = "string",
  IncludeFilters = list(
    list(
      Namespace = "string"
    )
  ),
  ExcludeFilters = list(
    list(
      Namespace = "string"
    )
  ),
  FirehoseArn = "string",
  RoleArn = "string",
  State = "string",
  CreationDate = as.POSIXct(
    "2015-01-01"
  ),
  LastUpdateDate = as.POSIXct(
    "2015-01-01"
  ),
  OutputFormat = "json"|"opentelemetry0.7"
)
}
}
\description{
Returns information about the metric stream that you specify.
}
\section{Request syntax}{
\preformatted{svc$get_metric_stream(
  Name = "string"
)
}
}

\keyword{internal}
