% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_o_tel_enrichment}
\alias{cloudwatch_get_o_tel_enrichment}
\title{Returns the current status of vended metric enrichment for the account,
including whether CloudWatch vended metrics are enriched with resource
ARN and resource tag labels and queryable using PromQL}
\usage{
cloudwatch_get_o_tel_enrichment()
}
\description{
Returns the current status of vended metric enrichment for the account, including whether CloudWatch vended metrics are enriched with resource ARN and resource tag labels and queryable using PromQL. For the list of supported resources, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/UsingResourceTagsForTelemetry.html}{Supported Amazon Web Services infrastructure metrics}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_get_o_tel_enrichment/} for full documentation.
}
\keyword{internal}
