% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_get_resource_metrics_configuration}
\alias{cloudwatch_get_resource_metrics_configuration}
\title{Retrieves the current resource metrics configuration for an Amazon Web
Services resource}
\usage{
cloudwatch_get_resource_metrics_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to retrieve the resource metrics configuration for.}
}
\description{
Retrieves the current resource metrics configuration for an Amazon Web Services resource. The response includes the resource ARN, any metric selections, and the times at which the configuration was created and last updated.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_get_resource_metrics_configuration/} for full documentation.
}
\keyword{internal}
