% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_list_alarm_mute_rules}
\alias{cloudwatch_list_alarm_mute_rules}
\title{Lists alarm mute rules in your Amazon Web Services account and region}
\usage{
cloudwatch_list_alarm_mute_rules(
  AlarmName = NULL,
  Statuses = NULL,
  MaxRecords = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AlarmName}{Filter results to show only mute rules that target the specified alarm
name.}

\item{Statuses}{Filter results to show only mute rules with the specified statuses.
Valid values are \code{SCHEDULED}, \code{ACTIVE}, or \code{EXPIRED}.}

\item{MaxRecords}{The maximum number of mute rules to return in one call. The default is
50.}

\item{NextToken}{The token returned from a previous call to indicate where to continue
retrieving results.}
}
\description{
Lists alarm mute rules in your Amazon Web Services account and region.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_list_alarm_mute_rules/} for full documentation.
}
\keyword{internal}
