% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_list_managed_insight_rules}
\alias{cloudwatch_list_managed_insight_rules}
\title{Returns a list that contains the number of managed Contributor Insights
rules in your account}
\usage{
cloudwatch_list_managed_insight_rules(
  ResourceARN,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ResourceARN}{[required] The ARN of an Amazon Web Services resource that has managed Contributor
Insights rules.}

\item{NextToken}{Include this value to get the next set of rules if the value was
returned by the previous operation.}

\item{MaxResults}{The maximum number of results to return in one operation. If you omit
this parameter, the default number is used. The default number is \code{100}.}
}
\description{
Returns a list that contains the number of managed Contributor Insights rules in your account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_list_managed_insight_rules/} for full documentation.
}
\keyword{internal}
