% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_list_metric_streams}
\alias{cloudwatch_list_metric_streams}
\title{Returns a list of metric streams in this account}
\usage{
cloudwatch_list_metric_streams(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Include this value, if it was returned by the previous call, to get the
next set of metric streams.}

\item{MaxResults}{The maximum number of results to return in one operation.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Entries = list(
    list(
      Arn = "string",
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdateDate = as.POSIXct(
        "2015-01-01"
      ),
      Name = "string",
      FirehoseArn = "string",
      State = "string",
      OutputFormat = "json"|"opentelemetry0.7"
    )
  )
)
}
}
\description{
Returns a list of metric streams in this account.
}
\section{Request syntax}{
\preformatted{svc$list_metric_streams(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
