% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_list_tags_for_resource}
\alias{cloudwatch_list_tags_for_resource}
\title{Displays the tags associated with a CloudWatch resource}
\usage{
cloudwatch_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the CloudWatch resource that you want to view tags for.

The ARN format of an alarm is \code{arn:aws:cloudwatch:Region:account-id:alarm:alarm-name }

The ARN format of a Contributor Insights rule is \code{arn:aws:cloudwatch:Region:account-id:insight-rule/insight-rule-name }

The ARN format of a dashboard is \code{arn:aws:cloudwatch::account-id:dashboard/dashboard-name }

The ARN format of a metric stream is \code{arn:aws:cloudwatch:Region:account-id:metric-stream/metric-stream-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/#amazoncloudwatch-resources-for-iam-policies}{Resource Types Defined by Amazon CloudWatch} in the \emph{Amazon Web Services General Reference}.}
}
\description{
Displays the tags associated with a CloudWatch resource. Currently, alarms, dashboards, metric streams and Contributor Insights rules support tagging.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
