% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_put_managed_insight_rules}
\alias{cloudwatch_put_managed_insight_rules}
\title{Creates a managed Contributor Insights rule for a specified Amazon Web
Services resource}
\usage{
cloudwatch_put_managed_insight_rules(ManagedRules)
}
\arguments{
\item{ManagedRules}{[required] A list of \code{ManagedRules} to enable.}
}
\description{
Creates a managed Contributor Insights rule for a specified Amazon Web Services resource. When you enable a managed rule, you create a Contributor Insights rule that collects data from Amazon Web Services services. You cannot edit these rules with \code{\link[=cloudwatch_put_insight_rule]{put_insight_rule}}. The rules can be enabled, disabled, and deleted using \code{\link[=cloudwatch_enable_insight_rules]{enable_insight_rules}}, \code{\link[=cloudwatch_disable_insight_rules]{disable_insight_rules}}, and \code{\link[=cloudwatch_delete_insight_rules]{delete_insight_rules}}. If a previously created managed rule is currently disabled, a subsequent call to this API will re-enable it. Use \code{\link[=cloudwatch_list_managed_insight_rules]{list_managed_insight_rules}} to describe all available rules.

See \url{https://paws-r.github.io/docs/cloudwatch/put_managed_insight_rules.html} for full documentation.
}
\keyword{internal}
