% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_stop_metric_streams}
\alias{cloudwatch_stop_metric_streams}
\title{Stops the streaming of metrics for one or more of your metric streams}
\usage{
cloudwatch_stop_metric_streams(Names)
}
\arguments{
\item{Names}{[required] The array of the names of metric streams to stop streaming.

This is an "all or nothing" operation. If you do not have permission to
access all of the metric streams that you list here, then none of the
streams that you list in the operation will stop streaming.}
}
\value{
An empty list.
}
\description{
Stops the streaming of metrics for one or more of your metric streams.
}
\section{Request syntax}{
\preformatted{svc$stop_metric_streams(
  Names = list(
    "string"
  )
)
}
}

\keyword{internal}
