% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_stop_o_tel_enrichment}
\alias{cloudwatch_stop_o_tel_enrichment}
\title{Disables enrichment and PromQL access for CloudWatch vended metrics for
supported Amazon Web Services resources in the account}
\usage{
cloudwatch_stop_o_tel_enrichment()
}
\description{
Disables enrichment and PromQL access for CloudWatch vended metrics for \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/UsingResourceTagsForTelemetry.html}{supported Amazon Web Services resources} in the account. After disabling, these metrics are no longer enriched with resource ARN and resource tag labels, and cannot be queried using PromQL.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_stop_o_tel_enrichment/} for full documentation.
}
\keyword{internal}
