% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_update_resource_metrics_configuration}
\alias{cloudwatch_update_resource_metrics_configuration}
\title{Updates the resource metrics configuration for an Amazon Web Services
resource}
\usage{
cloudwatch_update_resource_metrics_configuration(
  ResourceArn,
  MetricSelections = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to update the resource metrics configuration for.}

\item{MetricSelections}{Specifies which metrics Amazon CloudWatch collects for the resource. The selections that you provide completely replace any existing metric selections.

If you omit this parameter, Amazon CloudWatch removes any existing metric selection filter and collects all available detailed metrics for the resource.}
}
\description{
Updates the resource metrics configuration for an Amazon Web Services resource. The \code{MetricSelections} value that you provide replaces any existing metric selections for the resource; it is not merged with them.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_update_resource_metrics_configuration/} for full documentation.
}
\keyword{internal}
