% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_batch_get_service_level_objective_budget_report}
\alias{cloudwatchapplicationsignals_batch_get_service_level_objective_budget_report}
\title{Use this operation to retrieve one or more service level objective (SLO)
budget reports}
\usage{
cloudwatchapplicationsignals_batch_get_service_level_objective_budget_report(
  Timestamp,
  SloIds
)
}
\arguments{
\item{Timestamp}{[required] The date and time that you want the report to be for. It is expressed as
the number of milliseconds since Jan 1, 1970 00:00:00 UTC.}

\item{SloIds}{[required] An array containing the IDs of the service level objectives that you
want to include in the report.}
}
\description{
Use this operation to retrieve one or more \emph{service level objective (SLO) budget reports}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_batch_get_service_level_objective_budget_report/} for full documentation.
}
\keyword{internal}
