% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_batch_delete_instrumentation_configurations}
\alias{cloudwatchapplicationsignals_batch_delete_instrumentation_configurations}
\title{Deletes multiple instrumentation configurations in a single request}
\usage{
cloudwatchapplicationsignals_batch_delete_instrumentation_configurations(
  DeletionTarget
)
}
\arguments{
\item{DeletionTarget}{[required] The deletion target - either bulk by scope or targeted by ARN list.}
}
\description{
Deletes multiple instrumentation configurations in a single request. Supports two mutually exclusive selection methods:

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_batch_delete_instrumentation_configurations/} for full documentation.
}
\keyword{internal}
