% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_batch_update_exclusion_windows}
\alias{cloudwatchapplicationsignals_batch_update_exclusion_windows}
\title{Add or remove time window exclusions for one or more Service Level
Objectives (SLOs)}
\usage{
cloudwatchapplicationsignals_batch_update_exclusion_windows(
  SloIds,
  AddExclusionWindows = NULL,
  RemoveExclusionWindows = NULL
)
}
\arguments{
\item{SloIds}{[required] The list of SLO IDs to add or remove exclusion windows from.}

\item{AddExclusionWindows}{A list of exclusion windows to add to the specified SLOs. You can add up
to 10 exclusion windows per SLO.}

\item{RemoveExclusionWindows}{A list of exclusion windows to remove from the specified SLOs. The
window configuration must match an existing exclusion window.}
}
\description{
Add or remove time window exclusions for one or more Service Level Objectives (SLOs).

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_batch_update_exclusion_windows/} for full documentation.
}
\keyword{internal}
