% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_delete_instrumentation_configuration}
\alias{cloudwatchapplicationsignals_delete_instrumentation_configuration}
\title{Deletes the specified instrumentation configuration}
\usage{
cloudwatchapplicationsignals_delete_instrumentation_configuration(
  InstrumentationType,
  Service,
  Environment,
  SignalType,
  LocationIdentifier
)
}
\arguments{
\item{InstrumentationType}{[required] Type of instrumentation configuration (BREAKPOINT or PROBE). Required to identify the configuration to delete.}

\item{Service}{[required] Service name for the instrumentation configuration.}

\item{Environment}{[required] Environment name for the instrumentation configuration.}

\item{SignalType}{[required] Signal type for the instrumentation configuration.}

\item{LocationIdentifier}{[required] Location identifier - either full code location or a pre-computed hash.}
}
\description{
Deletes the specified instrumentation configuration. SDKs remove the instrumentation during their next sync after the configuration is deleted or expires.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_delete_instrumentation_configuration/} for full documentation.
}
\keyword{internal}
