% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_get_service_level_objective}
\alias{cloudwatchapplicationsignals_get_service_level_objective}
\title{Returns information about one SLO created in the account}
\usage{
cloudwatchapplicationsignals_get_service_level_objective(Id)
}
\arguments{
\item{Id}{[required] The ARN or name of the SLO that you want to retrieve information about.
You can find the ARNs of SLOs by using the
\code{\link[=cloudwatchapplicationsignals_list_service_level_objectives]{list_service_level_objectives}}
operation.}
}
\description{
Returns information about one SLO created in the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_get_service_level_objective/} for full documentation.
}
\keyword{internal}
