% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_service_level_objective_exclusion_windows}
\alias{cloudwatchapplicationsignals_list_service_level_objective_exclusion_windows}
\title{Retrieves all exclusion windows configured for a specific SLO}
\usage{
cloudwatchapplicationsignals_list_service_level_objective_exclusion_windows(
  Id,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Id}{[required] The ID of the SLO to list exclusion windows for.}

\item{MaxResults}{The maximum number of results to return in one operation. If you omit
this parameter, the default of 50 is used.}

\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of service level objectives.}
}
\description{
Retrieves all exclusion windows configured for a specific SLO.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_service_level_objective_exclusion_windows/} for full documentation.
}
\keyword{internal}
