% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_entity_events}
\alias{cloudwatchapplicationsignals_list_entity_events}
\title{Returns a list of change events for a specific entity, such as
deployments, configuration changes, or other state-changing activities}
\usage{
cloudwatchapplicationsignals_list_entity_events(
  Entity,
  StartTime,
  EndTime,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Entity}{[required] The entity for which to retrieve change events. This specifies the
service, resource, or other entity whose event history you want to
examine.

This is a string-to-string map. It can include the following fields.
\itemize{
\item \code{Type} designates the type of object this is.
\item \code{ResourceType} specifies the type of the resource. This field is
used only when the value of the \code{Type} field is \code{Resource} or
\code{AWS::Resource}.
\item \code{Name} specifies the name of the object. This is used only if the
value of the \code{Type} field is \code{Service}, \code{RemoteService}, or
\code{AWS::Service}.
\item \code{Identifier} identifies the resource objects of this resource. This
is used only if the value of the \code{Type} field is \code{Resource} or
\code{AWS::Resource}.
\item \code{Environment} specifies the location where this object is hosted, or
what it belongs to.
\item \code{AwsAccountId} specifies the account where this object is in.
}

Below is an example of a service.

\verb{\{ "Type": "Service", "Name": "visits-service", "Environment": "petclinic-test" \}}

Below is an example of a resource.

\verb{\{ "Type": "AWS::Resource", "ResourceType": "AWS::DynamoDB::Table", "Identifier": "Customers" \}}}

\item{StartTime}{[required] The start of the time period to retrieve change events for. When used in
a raw HTTP Query API, it is formatted as epoch time in seconds. For
example: \code{1698778057}}

\item{EndTime}{[required] The end of the time period to retrieve change events for. When used in a
raw HTTP Query API, it is formatted as epoch time in seconds. For
example: \code{1698778057}}

\item{MaxResults}{The maximum number of change events to return in one operation. If you
omit this parameter, the default of 50 is used.}

\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of change events.}
}
\description{
Returns a list of change events for a specific entity, such as deployments, configuration changes, or other state-changing activities. This operation helps track the history of changes that may have affected service performance.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_entity_events/} for full documentation.
}
\keyword{internal}
