% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_grouping_attribute_definitions}
\alias{cloudwatchapplicationsignals_list_grouping_attribute_definitions}
\title{Returns the current grouping configuration for this account, including
all custom grouping attribute definitions that have been configured}
\usage{
cloudwatchapplicationsignals_list_grouping_attribute_definitions(
  NextToken = NULL,
  AwsAccountId = NULL,
  IncludeLinkedAccounts = NULL
)
}
\arguments{
\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of grouping attribute definitions.}

\item{AwsAccountId}{The Amazon Web Services account ID to retrieve grouping attribute
definitions for. Use this when accessing grouping configurations from a
different account in cross-account monitoring scenarios.}

\item{IncludeLinkedAccounts}{If you are using this operation in a monitoring account, specify \code{true}
to include grouping attributes from source accounts in the returned
data.}
}
\description{
Returns the current grouping configuration for this account, including all custom grouping attribute definitions that have been configured. These definitions determine how services are logically grouped based on telemetry attributes, Amazon Web Services tags, or predefined mappings.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_grouping_attribute_definitions/} for full documentation.
}
\keyword{internal}
