% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_instrumentation_configurations}
\alias{cloudwatchapplicationsignals_list_instrumentation_configurations}
\title{Returns all active instrumentation configurations for a service and
environment}
\usage{
cloudwatchapplicationsignals_list_instrumentation_configurations(
  Service,
  Environment,
  InstrumentationType,
  SyncedAt = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Service}{[required] The name of the service to retrieve instrumentation configurations for.}

\item{Environment}{[required] The environment that the service is running in.}

\item{InstrumentationType}{[required] Type of instrumentation configuration (BREAKPOINT or PROBE). Required to determine which backing store to query.}

\item{SyncedAt}{The timestamp from the last successful sync. When provided, the response returns \code{Changed} as \code{false} if nothing is new since this time, or returns the latest configurations when changes exist.}

\item{MaxResults}{The maximum number of configurations to return in one call. The default is 50 and the maximum is 100.}

\item{NextToken}{Use the token returned by a previous call to retrieve the next page of configurations.}
}
\description{
Returns all active instrumentation configurations for a service and environment. SDKs use this operation to sync configurations and apply client-side filters locally.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_instrumentation_configurations/} for full documentation.
}
\keyword{internal}
