% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_service_states}
\alias{cloudwatchapplicationsignals_list_service_states}
\title{Returns information about the last deployment and other change states of
services}
\usage{
cloudwatchapplicationsignals_list_service_states(
  StartTime,
  EndTime,
  MaxResults = NULL,
  NextToken = NULL,
  IncludeLinkedAccounts = NULL,
  AwsAccountId = NULL,
  AttributeFilters = NULL
)
}
\arguments{
\item{StartTime}{[required] The start of the time period to retrieve service state information for.
When used in a raw HTTP Query API, it is formatted as epoch time in
seconds. For example, \code{1698778057}.}

\item{EndTime}{[required] The end of the time period to retrieve service state information for.
When used in a raw HTTP Query API, it is formatted as epoch time in
seconds. For example, \code{1698778057}.}

\item{MaxResults}{The maximum number of service states to return in one operation. If you
omit this parameter, the default of 20 is used.}

\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of service states.}

\item{IncludeLinkedAccounts}{If you are using this operation in a monitoring account, specify \code{true}
to include service states from source accounts in the returned data.}

\item{AwsAccountId}{The Amazon Web Services account ID to filter service states by. Use this
to limit results to services from a specific account.}

\item{AttributeFilters}{A list of attribute filters to narrow down the services. You can filter
by platform, environment, or other service attributes.}
}
\description{
Returns information about the last deployment and other change states of services. This API provides visibility into recent changes that may have affected service performance, helping with troubleshooting and change correlation.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_service_states/} for full documentation.
}
\keyword{internal}
