% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_list_tags_for_resource}
\alias{cloudwatchapplicationsignals_list_tags_for_resource}
\title{Displays the tags associated with a CloudWatch resource}
\usage{
cloudwatchapplicationsignals_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudWatch resource that you want to view tags for.

The ARN format of an Application Signals SLO is \code{arn:aws:cloudwatch:Region:account-id:slo:slo-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/#amazoncloudwatch-resources-for-iam-policies}{Resource Types Defined by Amazon CloudWatch} in the \emph{Amazon Web Services General Reference}.}
}
\description{
Displays the tags associated with a CloudWatch resource. Tags can be assigned to service level objectives.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
