% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_start_discovery}
\alias{cloudwatchapplicationsignals_start_discovery}
\title{Enables this Amazon Web Services account to be able to use CloudWatch
Application Signals by creating the
AWSServiceRoleForCloudWatchApplicationSignals service-linked role}
\usage{
cloudwatchapplicationsignals_start_discovery()
}
\description{
Enables this Amazon Web Services account to be able to use CloudWatch Application Signals by creating the \emph{AWSServiceRoleForCloudWatchApplicationSignals} service-linked role. This service- linked role has the following permissions:

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_start_discovery/} for full documentation.
}
\keyword{internal}
