% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_untag_resource}
\alias{cloudwatchapplicationsignals_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
cloudwatchapplicationsignals_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudWatch resource that you want to delete tags from.

The ARN format of an Application Signals SLO is \code{arn:aws:cloudwatch:Region:account-id:slo:slo-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/#amazoncloudwatch-resources-for-iam-policies}{Resource Types Defined by Amazon CloudWatch} in the \emph{Amazon Web Services General Reference}.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_untag_resource/} for full documentation.
}
\keyword{internal}
