% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_activate_event_source}
\alias{cloudwatchevents_activate_event_source}
\title{Activates a partner event source that has been deactivated}
\usage{
cloudwatchevents_activate_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to activate.}
}
\value{
An empty list.
}
\description{
Activates a partner event source that has been deactivated. Once
activated, your matching event bus will start receiving events from the
event source.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$activate_event_source(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
