% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_create_api_destination}
\alias{cloudwatchevents_create_api_destination}
\title{Creates an API destination, which is an HTTP invocation endpoint
configured as a target for events}
\usage{
cloudwatchevents_create_api_destination(Name, Description,
  ConnectionArn, InvocationEndpoint, HttpMethod,
  InvocationRateLimitPerSecond)
}
\arguments{
\item{Name}{[required] The name for the API destination to create.}

\item{Description}{A description for the API destination to create.}

\item{ConnectionArn}{[required] The ARN of the connection to use for the API destination. The
destination endpoint must support the authorization type specified for
the connection.}

\item{InvocationEndpoint}{[required] The URL to the HTTP invocation endpoint for the API destination.}

\item{HttpMethod}{[required] The method to use for the request to the HTTP invocation endpoint.}

\item{InvocationRateLimitPerSecond}{The maximum number of requests per second to send to the HTTP invocation
endpoint.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiDestinationArn = "string",
  ApiDestinationState = "ACTIVE"|"INACTIVE",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Creates an API destination, which is an HTTP invocation endpoint
configured as a target for events.
}
\section{Request syntax}{
\preformatted{svc$create_api_destination(
  Name = "string",
  Description = "string",
  ConnectionArn = "string",
  InvocationEndpoint = "string",
  HttpMethod = "POST"|"GET"|"HEAD"|"OPTIONS"|"PUT"|"PATCH"|"DELETE",
  InvocationRateLimitPerSecond = 123
)
}
}

\keyword{internal}
