% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_create_connection}
\alias{cloudwatchevents_create_connection}
\title{Creates a connection}
\usage{
cloudwatchevents_create_connection(
  Name,
  Description = NULL,
  AuthorizationType,
  AuthParameters
)
}
\arguments{
\item{Name}{[required] The name for the connection to create.}

\item{Description}{A description for the connection to create.}

\item{AuthorizationType}{[required] The type of authorization to use for the connection.}

\item{AuthParameters}{[required] A \code{CreateConnectionAuthRequestParameters} object that contains the
authorization parameters to use to authorize with the endpoint.}
}
\description{
Creates a connection. A connection defines the authorization type and credentials to use for authorization with an API destination HTTP endpoint.

See \url{https://paws-r.github.io/docs/cloudwatchevents/create_connection.html} for full documentation.
}
\keyword{internal}
