% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_api_destination}
\alias{cloudwatchevents_delete_api_destination}
\title{Deletes the specified API destination}
\usage{
cloudwatchevents_delete_api_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the destination to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified API destination.
}
\section{Request syntax}{
\preformatted{svc$delete_api_destination(
  Name = "string"
)
}
}

\keyword{internal}
