% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_archive}
\alias{cloudwatchevents_delete_archive}
\title{Deletes the specified archive}
\usage{
cloudwatchevents_delete_archive(ArchiveName)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to delete.}
}
\description{
Deletes the specified archive.

See \url{https://paws-r.github.io/docs/cloudwatchevents/delete_archive.html} for full documentation.
}
\keyword{internal}
