% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_connection}
\alias{cloudwatchevents_delete_connection}
\title{Deletes a connection}
\usage{
cloudwatchevents_delete_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to delete.}
}
\description{
Deletes a connection.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_delete_connection/} for full documentation.
}
\keyword{internal}
