% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_partner_event_source}
\alias{cloudwatchevents_delete_partner_event_source}
\title{This operation is used by SaaS partners to delete a partner event source}
\usage{
cloudwatchevents_delete_partner_event_source(Name, Account)
}
\arguments{
\item{Name}{[required] The name of the event source to delete.}

\item{Account}{[required] The Amazon Web Services account ID of the Amazon Web Services customer
that the event source was created for.}
}
\description{
This operation is used by SaaS partners to delete a partner event source. This operation is not used by Amazon Web Services customers.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_delete_partner_event_source/} for full documentation.
}
\keyword{internal}
