% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_event_bus}
\alias{cloudwatchevents_describe_event_bus}
\title{Displays details about an event bus in your account}
\usage{
cloudwatchevents_describe_event_bus(Name)
}
\arguments{
\item{Name}{The name or ARN of the event bus to show details for. If you omit this,
the default event bus is displayed.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Name = "string",
  Arn = "string",
  Policy = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Displays details about an event bus in your account. This can include
the external AWS accounts that are permitted to write events to your
default event bus, and the associated policy. For custom event buses and
partner event buses, it displays the name, ARN, policy, state, and
creation time.

To enable your account to receive events from other accounts on its
default event bus, use
\code{\link[=cloudwatchevents_put_permission]{put_permission}}.

For more information about partner event buses, see
\code{\link[=cloudwatchevents_create_event_bus]{create_event_bus}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_event_bus(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
