% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_list_api_destinations}
\alias{cloudwatchevents_list_api_destinations}
\title{Retrieves a list of API destination in the account in the current Region}
\usage{
cloudwatchevents_list_api_destinations(NamePrefix, ConnectionArn,
  NextToken, Limit)
}
\arguments{
\item{NamePrefix}{A name prefix to filter results returned. Only API destinations with a
name that starts with the prefix are returned.}

\item{ConnectionArn}{The ARN of the connection specified for the API destination.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of API destinations to include in the response.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiDestinations = list(
    list(
      ApiDestinationArn = "string",
      Name = "string",
      ApiDestinationState = "ACTIVE"|"INACTIVE",
      ConnectionArn = "string",
      InvocationEndpoint = "string",
      HttpMethod = "POST"|"GET"|"HEAD"|"OPTIONS"|"PUT"|"PATCH"|"DELETE",
      InvocationRateLimitPerSecond = 123,
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of API destination in the account in the current
Region.
}
\section{Request syntax}{
\preformatted{svc$list_api_destinations(
  NamePrefix = "string",
  ConnectionArn = "string",
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
