% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_list_connections}
\alias{cloudwatchevents_list_connections}
\title{Retrieves a list of connections from the account}
\usage{
cloudwatchevents_list_connections(NamePrefix, ConnectionState,
  NextToken, Limit)
}
\arguments{
\item{NamePrefix}{A name prefix to filter results returned. Only connections with a name
that starts with the prefix are returned.}

\item{ConnectionState}{The state of the connection.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of connections to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  Connections = list(
    list(
      ConnectionArn = "string",
      Name = "string",
      ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
      StateReason = "string",
      AuthorizationType = "BASIC"|"OAUTH_CLIENT_CREDENTIALS"|"API_KEY",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastAuthorizedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of connections from the account.
}
\section{Request syntax}{
\preformatted{svc$list_connections(
  NamePrefix = "string",
  ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
