% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_list_tags_for_resource}
\alias{cloudwatchevents_list_tags_for_resource}
\title{Displays the tags associated with an EventBridge resource}
\usage{
cloudwatchevents_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the EventBridge resource for which you want to view tags.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Displays the tags associated with an EventBridge resource. In
EventBridge, rules and event buses can be tagged.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
