% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_put_events}
\alias{cloudwatchevents_put_events}
\title{Sends custom events to Amazon EventBridge so that they can be matched to
rules}
\usage{
cloudwatchevents_put_events(Entries)
}
\arguments{
\item{Entries}{[required] The entry that defines an event in your system. You can specify several
parameters for the entry such as the source and type of the event,
resources associated with the event, and so on.}
}
\description{
Sends custom events to Amazon EventBridge so that they can be matched to rules.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_put_events/} for full documentation.
}
\keyword{internal}
