% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_put_partner_events}
\alias{cloudwatchevents_put_partner_events}
\title{This is used by SaaS partners to write events to a customer's partner
event bus}
\usage{
cloudwatchevents_put_partner_events(Entries)
}
\arguments{
\item{Entries}{[required] The list of events to write to the event bus.}
}
\description{
This is used by SaaS partners to write events to a customer's partner event bus. Amazon Web Services customers do not use this operation.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_put_partner_events/} for full documentation.
}
\keyword{internal}
