% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_put_targets}
\alias{cloudwatchevents_put_targets}
\title{Adds the specified targets to the specified rule, or updates the targets
if they are already associated with the rule}
\usage{
cloudwatchevents_put_targets(Rule, EventBusName = NULL, Targets)
}
\arguments{
\item{Rule}{[required] The name of the rule.}

\item{EventBusName}{The name or ARN of the event bus associated with the rule. If you omit
this, the default event bus is used.}

\item{Targets}{[required] The targets to update or add to the rule.}
}
\description{
Adds the specified targets to the specified rule, or updates the targets if they are already associated with the rule.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_put_targets/} for full documentation.
}
\keyword{internal}
