% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_update_api_destination}
\alias{cloudwatchevents_update_api_destination}
\title{Updates an API destination}
\usage{
cloudwatchevents_update_api_destination(Name, Description,
  ConnectionArn, InvocationEndpoint, HttpMethod,
  InvocationRateLimitPerSecond)
}
\arguments{
\item{Name}{[required] The name of the API destination to update.}

\item{Description}{The name of the API destination to update.}

\item{ConnectionArn}{The ARN of the connection to use for the API destination.}

\item{InvocationEndpoint}{The URL to the endpoint to use for the API destination.}

\item{HttpMethod}{The method to use for the API destination.}

\item{InvocationRateLimitPerSecond}{The maximum number of invocations per second to send to the API
destination.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiDestinationArn = "string",
  ApiDestinationState = "ACTIVE"|"INACTIVE",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Updates an API destination.
}
\section{Request syntax}{
\preformatted{svc$update_api_destination(
  Name = "string",
  Description = "string",
  ConnectionArn = "string",
  InvocationEndpoint = "string",
  HttpMethod = "POST"|"GET"|"HEAD"|"OPTIONS"|"PUT"|"PATCH"|"DELETE",
  InvocationRateLimitPerSecond = 123
)
}
}

\keyword{internal}
