% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_update_archive}
\alias{cloudwatchevents_update_archive}
\title{Updates the specified archive}
\usage{
cloudwatchevents_update_archive(ArchiveName, Description, EventPattern,
  RetentionDays)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to update.}

\item{Description}{The description for the archive.}

\item{EventPattern}{The event pattern to use to filter events sent to the archive.}

\item{RetentionDays}{The number of days to retain events in the archive.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ArchiveArn = "string",
  State = "ENABLED"|"DISABLED"|"CREATING"|"UPDATING"|"CREATE_FAILED"|"UPDATE_FAILED",
  StateReason = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified archive.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_archive(
  ArchiveName = "string",
  Description = "string",
  EventPattern = "string",
  RetentionDays = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
