% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_update_connection}
\alias{cloudwatchevents_update_connection}
\title{Updates settings for a connection}
\usage{
cloudwatchevents_update_connection(Name, Description, AuthorizationType,
  AuthParameters)
}
\arguments{
\item{Name}{[required] The name of the connection to update.}

\item{Description}{A description for the connection.}

\item{AuthorizationType}{The type of authorization to use for the connection.}

\item{AuthParameters}{The authorization parameters to use for the connection.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConnectionArn = "string",
  ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastAuthorizedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Updates settings for a connection.
}
\section{Request syntax}{
\preformatted{svc$update_connection(
  Name = "string",
  Description = "string",
  AuthorizationType = "BASIC"|"OAUTH_CLIENT_CREDENTIALS"|"API_KEY",
  AuthParameters = list(
    BasicAuthParameters = list(
      Username = "string",
      Password = "string"
    ),
    OAuthParameters = list(
      ClientParameters = list(
        ClientID = "string",
        ClientSecret = "string"
      ),
      AuthorizationEndpoint = "string",
      HttpMethod = "GET"|"POST"|"PUT",
      OAuthHttpParameters = list(
        HeaderParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        ),
        QueryStringParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        ),
        BodyParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        )
      )
    ),
    ApiKeyAuthParameters = list(
      ApiKeyName = "string",
      ApiKeyValue = "string"
    ),
    InvocationHttpParameters = list(
      HeaderParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      ),
      QueryStringParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      ),
      BodyParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      )
    )
  )
)
}
}

\keyword{internal}
