% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_batch_evaluate_feature}
\alias{cloudwatchevidently_batch_evaluate_feature}
\title{This operation assigns feature variation to user sessions}
\usage{
cloudwatchevidently_batch_evaluate_feature(project, requests)
}
\arguments{
\item{project}{[required] The name or ARN of the project that contains the feature being
evaluated.}

\item{requests}{[required] An array of structures, where each structure assigns a feature variation
to one user session.}
}
\description{
This operation assigns feature variation to user sessions. For each user session, you pass in an \code{entityID} that represents the user. Evidently then checks the evaluation rules and assigns the variation.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_batch_evaluate_feature/} for full documentation.
}
\keyword{internal}
