% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_delete_experiment}
\alias{cloudwatchevidently_delete_experiment}
\title{Deletes an Evidently experiment}
\usage{
cloudwatchevidently_delete_experiment(experiment, project)
}
\arguments{
\item{experiment}{[required] The name of the experiment to delete.}

\item{project}{[required] The name or ARN of the project that contains the experiment to delete.}
}
\description{
Deletes an Evidently experiment. The feature used for the experiment is not deleted.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_delete_experiment/} for full documentation.
}
\keyword{internal}
