% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_delete_launch}
\alias{cloudwatchevidently_delete_launch}
\title{Deletes an Evidently launch}
\usage{
cloudwatchevidently_delete_launch(launch, project)
}
\arguments{
\item{launch}{[required] The name of the launch to delete.}

\item{project}{[required] The name or ARN of the project that contains the launch to delete.}
}
\description{
Deletes an Evidently launch. The feature used for the launch is not deleted.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/delete_launch.html} for full documentation.
}
\keyword{internal}
