% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_delete_segment}
\alias{cloudwatchevidently_delete_segment}
\title{Deletes a segment}
\usage{
cloudwatchevidently_delete_segment(segment)
}
\arguments{
\item{segment}{[required] Specifies the segment to delete.}
}
\description{
Deletes a segment. You can't delete a segment that is being used in a launch or experiment, even if that launch or experiment is not currently running.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_delete_segment/} for full documentation.
}
\keyword{internal}
