% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_get_feature}
\alias{cloudwatchevidently_get_feature}
\title{Returns the details about one feature}
\usage{
cloudwatchevidently_get_feature(feature, project)
}
\arguments{
\item{feature}{[required] The name of the feature that you want to retrieve information for.}

\item{project}{[required] The name or ARN of the project that contains the feature.}
}
\description{
Returns the details about one feature. You must already know the feature name. To retrieve a list of features in your account, use \code{\link[=cloudwatchevidently_list_features]{list_features}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_get_feature/} for full documentation.
}
\keyword{internal}
