% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_get_launch}
\alias{cloudwatchevidently_get_launch}
\title{Returns the details about one launch}
\usage{
cloudwatchevidently_get_launch(launch, project)
}
\arguments{
\item{launch}{[required] The name of the launch that you want to see the details of.}

\item{project}{[required] The name or ARN of the project that contains the launch.}
}
\description{
Returns the details about one launch. You must already know the launch name. To retrieve a list of launches in your account, use \code{\link[=cloudwatchevidently_list_launches]{list_launches}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_get_launch/} for full documentation.
}
\keyword{internal}
