% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_get_project}
\alias{cloudwatchevidently_get_project}
\title{Returns the details about one launch}
\usage{
cloudwatchevidently_get_project(project)
}
\arguments{
\item{project}{[required] The name or ARN of the project that you want to see the details of.}
}
\description{
Returns the details about one launch. You must already know the project name. To retrieve a list of projects in your account, use \code{\link[=cloudwatchevidently_list_projects]{list_projects}}.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/get_project.html} for full documentation.
}
\keyword{internal}
