% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_get_segment}
\alias{cloudwatchevidently_get_segment}
\title{Returns information about the specified segment}
\usage{
cloudwatchevidently_get_segment(segment)
}
\arguments{
\item{segment}{[required] The ARN of the segment to return information for.}
}
\description{
Returns information about the specified segment. Specify the segment you want to view by specifying its ARN.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/get_segment.html} for full documentation.
}
\keyword{internal}
