% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_features}
\alias{cloudwatchevidently_list_features}
\title{Returns configuration details about all the features in the specified
project}
\usage{
cloudwatchevidently_list_features(maxResults = NULL, nextToken = NULL, project)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_features]{list_features}} operation.}

\item{project}{[required] The name or ARN of the project to return the feature list from.}
}
\description{
Returns configuration details about all the features in the specified project.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_features/} for full documentation.
}
\keyword{internal}
