% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_launches}
\alias{cloudwatchevidently_list_launches}
\title{Returns configuration details about all the launches in the specified
project}
\usage{
cloudwatchevidently_list_launches(
  maxResults = NULL,
  nextToken = NULL,
  project,
  status = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_launches]{list_launches}} operation.}

\item{project}{[required] The name or ARN of the project to return the launch list from.}

\item{status}{Use this optional parameter to limit the returned results to only the
launches with the status that you specify here.}
}
\description{
Returns configuration details about all the launches in the specified project.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_launches/} for full documentation.
}
\keyword{internal}
