% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_projects}
\alias{cloudwatchevidently_list_projects}
\title{Returns configuration details about all the projects in the current
Region in your account}
\usage{
cloudwatchevidently_list_projects(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_projects]{list_projects}} operation.}
}
\description{
Returns configuration details about all the projects in the current Region in your account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_projects/} for full documentation.
}
\keyword{internal}
