% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_segment_references}
\alias{cloudwatchevidently_list_segment_references}
\title{Use this operation to find which experiments or launches are using a
specified segment}
\usage{
cloudwatchevidently_list_segment_references(
  maxResults = NULL,
  nextToken = NULL,
  segment,
  type
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response. If you omit
this, the default of 50 is used.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_segment_references]{list_segment_references}}
operation.}

\item{segment}{[required] The ARN of the segment that you want to view information for.}

\item{type}{[required] Specifies whether to return information about launches or experiments
that use this segment.}
}
\description{
Use this operation to find which experiments or launches are using a specified segment.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_segment_references/} for full documentation.
}
\keyword{internal}
