% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_segments}
\alias{cloudwatchevidently_list_segments}
\title{Returns a list of audience segments that you have created in your
account in this Region}
\usage{
cloudwatchevidently_list_segments(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response. If you omit
this, the default of 50 is used.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_segments]{list_segments}} operation.}
}
\description{
Returns a list of audience segments that you have created in your account in this Region.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_segments/} for full documentation.
}
\keyword{internal}
