% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_put_project_events}
\alias{cloudwatchevidently_put_project_events}
\title{Sends performance events to Evidently}
\usage{
cloudwatchevidently_put_project_events(events, project)
}
\arguments{
\item{events}{[required] An array of event structures that contain the performance data that is
being sent to Evidently.}

\item{project}{[required] The name or ARN of the project to write the events to.}
}
\description{
Sends performance events to Evidently. These events can be used to evaluate a launch or an experiment.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_put_project_events/} for full documentation.
}
\keyword{internal}
