% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_stop_experiment}
\alias{cloudwatchevidently_stop_experiment}
\title{Stops an experiment that is currently running}
\usage{
cloudwatchevidently_stop_experiment(
  desiredState = NULL,
  experiment,
  project,
  reason = NULL
)
}
\arguments{
\item{desiredState}{Specify whether the experiment is to be considered \code{COMPLETED} or
\code{CANCELLED} after it stops.}

\item{experiment}{[required] The name of the experiment to stop.}

\item{project}{[required] The name or ARN of the project that contains the experiment to stop.}

\item{reason}{A string that describes why you are stopping the experiment.}
}
\description{
Stops an experiment that is currently running. If you stop an experiment, you can't resume it or restart it.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/stop_experiment.html} for full documentation.
}
\keyword{internal}
