% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_tag_resource}
\alias{cloudwatchevidently_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified CloudWatch
Evidently resource}
\usage{
cloudwatchevidently_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the CloudWatch Evidently resource that you're adding tags to.}

\item{tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified CloudWatch Evidently resource. Projects, features, launches, and experiments can be tagged.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_tag_resource/} for full documentation.
}
\keyword{internal}
