% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_test_segment_pattern}
\alias{cloudwatchevidently_test_segment_pattern}
\title{Use this operation to test a rules pattern that you plan to use to
create an audience segment}
\usage{
cloudwatchevidently_test_segment_pattern(pattern, payload)
}
\arguments{
\item{pattern}{[required] The pattern to test.}

\item{payload}{[required] A sample \code{evaluationContext} JSON block to test against the specified
pattern.}
}
\description{
Use this operation to test a rules pattern that you plan to use to create an audience segment. For more information about segments, see \code{\link[=cloudwatchevidently_create_segment]{create_segment}}.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/test_segment_pattern.html} for full documentation.
}
\keyword{internal}
