% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_untag_resource}
\alias{cloudwatchevidently_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
cloudwatchevidently_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the CloudWatch Evidently resource that you're removing tags
from.}

\item{tagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_untag_resource/} for full documentation.
}
\keyword{internal}
