% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_delete_monitor}
\alias{cloudwatchinternetmonitor_delete_monitor}
\title{Deletes a monitor in Amazon CloudWatch Internet Monitor}
\usage{
cloudwatchinternetmonitor_delete_monitor(MonitorName)
}
\arguments{
\item{MonitorName}{[required] The name of the monitor to delete.}
}
\description{
Deletes a monitor in Amazon CloudWatch Internet Monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_delete_monitor/} for full documentation.
}
\keyword{internal}
