% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_get_internet_event}
\alias{cloudwatchinternetmonitor_get_internet_event}
\title{Gets information that Amazon CloudWatch Internet Monitor has generated
about an internet event}
\usage{
cloudwatchinternetmonitor_get_internet_event(EventId)
}
\arguments{
\item{EventId}{[required] The \code{EventId} of the internet event to return information for.}
}
\description{
Gets information that Amazon CloudWatch Internet Monitor has generated about an internet event. Internet Monitor displays information about recent global health events, called internet events, on a global outages map that is available to all Amazon Web Services customers.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_get_internet_event/} for full documentation.
}
\keyword{internal}
