% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_list_tags_for_resource}
\alias{cloudwatchinternetmonitor_list_tags_for_resource}
\title{Lists the tags for a resource}
\usage{
cloudwatchinternetmonitor_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a resource.}
}
\description{
Lists the tags for a resource. Tags are supported only for monitors in Amazon CloudWatch Internet Monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
