% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_stop_query}
\alias{cloudwatchinternetmonitor_stop_query}
\title{Stop a query that is progress for a specific monitor}
\usage{
cloudwatchinternetmonitor_stop_query(MonitorName, QueryId)
}
\arguments{
\item{MonitorName}{[required] The name of the monitor.}

\item{QueryId}{[required] The ID of the query that you want to stop. A \code{QueryId} is an
internally-generated identifier for a specific query.}
}
\description{
Stop a query that is progress for a specific monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_stop_query/} for full documentation.
}
\keyword{internal}
