% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_tag_resource}
\alias{cloudwatchinternetmonitor_tag_resource}
\title{Adds a tag to a resource}
\usage{
cloudwatchinternetmonitor_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a tag that you add to a resource.
Tags are supported only for monitors in Amazon CloudWatch Internet
Monitor.}

\item{Tags}{[required] Tags that you add to a resource. You can add a maximum of 50 tags in
Internet Monitor.}
}
\description{
Adds a tag to a resource. Tags are supported only for monitors in Amazon CloudWatch Internet Monitor. You can add a maximum of 50 tags in Internet Monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_tag_resource/} for full documentation.
}
\keyword{internal}
