% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_untag_resource}
\alias{cloudwatchinternetmonitor_untag_resource}
\title{Removes a tag from a resource}
\usage{
cloudwatchinternetmonitor_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a tag you remove a resource from.}

\item{TagKeys}{[required] Tag keys that you remove from a resource.}
}
\description{
Removes a tag from a resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_untag_resource/} for full documentation.
}
\keyword{internal}
