% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_cancel_import_task}
\alias{cloudwatchlogs_cancel_import_task}
\title{Cancels an active import task and stops importing data from the
CloudTrail Lake Event Data Store}
\usage{
cloudwatchlogs_cancel_import_task(importId)
}
\arguments{
\item{importId}{[required] The ID of the import task to cancel.}
}
\description{
Cancels an active import task and stops importing data from the CloudTrail Lake Event Data Store.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_cancel_import_task/} for full documentation.
}
\keyword{internal}
