% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_create_delivery}
\alias{cloudwatchlogs_create_delivery}
\title{Creates a delivery}
\usage{
cloudwatchlogs_create_delivery(
  deliverySourceName,
  deliveryDestinationArn,
  tags = NULL
)
}
\arguments{
\item{deliverySourceName}{[required] The name of the delivery source to use for this delivery.}

\item{deliveryDestinationArn}{[required] The ARN of the delivery destination to use for this delivery.}

\item{tags}{An optional list of key-value pairs to associate with the resource.

For more information about tagging, see \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services resources}}
}
\description{
Creates a \emph{delivery}. A delivery is a connection between a logical \emph{delivery source} and a logical \emph{delivery destination} that you have already created.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_create_delivery/} for full documentation.
}
\keyword{internal}
