% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_create_import_task}
\alias{cloudwatchlogs_create_import_task}
\title{Starts an import from a data source to CloudWatch Log and creates a
managed log group as the destination for the imported data}
\usage{
cloudwatchlogs_create_import_task(
  importSourceArn,
  importRoleArn,
  importFilter = NULL
)
}
\arguments{
\item{importSourceArn}{[required] The ARN of the source to import from.}

\item{importRoleArn}{[required] The ARN of the IAM role that grants CloudWatch Logs permission to import
from the CloudTrail Lake Event Data Store.}

\item{importFilter}{Optional filters to constrain the import by CloudTrail event time. Times
are specified in Unix timestamp milliseconds. The range of data being
imported must be within the specified source's retention period.}
}
\description{
Starts an import from a data source to CloudWatch Log and creates a managed log group as the destination for the imported data. Currently, \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/query-event-data-store.html}{CloudTrail Event Data Store} is the only supported data source.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_create_import_task/} for full documentation.
}
\keyword{internal}
