% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_create_log_stream}
\alias{cloudwatchlogs_create_log_stream}
\title{Creates a log stream for the specified log group}
\usage{
cloudwatchlogs_create_log_stream(logGroupName, logStreamName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{logStreamName}{[required] The name of the log stream.}
}
\description{
Creates a log stream for the specified log group. A log stream is a sequence of log events that originate from a single source, such as an application instance or a resource that is being monitored.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_create_log_stream/} for full documentation.
}
\keyword{internal}
