% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_data_protection_policy}
\alias{cloudwatchlogs_delete_data_protection_policy}
\title{Deletes the data protection policy from the specified log group}
\usage{
cloudwatchlogs_delete_data_protection_policy(logGroupIdentifier)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group that you want to delete the data
protection policy for.}
}
\description{
Deletes the data protection policy from the specified log group.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_data_protection_policy/} for full documentation.
}
\keyword{internal}
