% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_delivery_destination}
\alias{cloudwatchlogs_delete_delivery_destination}
\title{Deletes a delivery destination}
\usage{
cloudwatchlogs_delete_delivery_destination(name)
}
\arguments{
\item{name}{[required] The name of the delivery destination that you want to delete. You can
find a list of delivery destination names by using the
\code{\link[=cloudwatchlogs_describe_delivery_destinations]{describe_delivery_destinations}}
operation.}
}
\description{
Deletes a \emph{delivery destination}. A delivery is a connection between a logical \emph{delivery source} and a logical \emph{delivery destination}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_delivery_destination/} for full documentation.
}
\keyword{internal}
