% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_delivery_destination_policy}
\alias{cloudwatchlogs_delete_delivery_destination_policy}
\title{Deletes a delivery destination policy}
\usage{
cloudwatchlogs_delete_delivery_destination_policy(deliveryDestinationName)
}
\arguments{
\item{deliveryDestinationName}{[required] The name of the delivery destination that you want to delete the policy
for.}
}
\description{
Deletes a delivery destination policy. For more information about these policies, see \code{\link[=cloudwatchlogs_put_delivery_destination_policy]{put_delivery_destination_policy}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_delivery_destination_policy/} for full documentation.
}
\keyword{internal}
