% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_delivery_source}
\alias{cloudwatchlogs_delete_delivery_source}
\title{Deletes a delivery source}
\usage{
cloudwatchlogs_delete_delivery_source(name)
}
\arguments{
\item{name}{[required] The name of the delivery source that you want to delete.}
}
\description{
Deletes a \emph{delivery source}. A delivery is a connection between a logical \emph{delivery source} and a logical \emph{delivery destination}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_delivery_source/} for full documentation.
}
\keyword{internal}
