% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_index_policy}
\alias{cloudwatchlogs_delete_index_policy}
\title{Deletes a log-group level field index policy that was applied to a
single log group}
\usage{
cloudwatchlogs_delete_index_policy(logGroupIdentifier)
}
\arguments{
\item{logGroupIdentifier}{[required] The log group to delete the index policy for. You can specify either the
name or the ARN of the log group.}
}
\description{
Deletes a log-group level field index policy that was applied to a single log group. The indexing of the log events that happened before you delete the policy will still be used for as many as 30 days to improve CloudWatch Logs Insights queries.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_index_policy/} for full documentation.
}
\keyword{internal}
