% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_log_anomaly_detector}
\alias{cloudwatchlogs_delete_log_anomaly_detector}
\title{Deletes the specified CloudWatch Logs anomaly detector}
\usage{
cloudwatchlogs_delete_log_anomaly_detector(anomalyDetectorArn)
}
\arguments{
\item{anomalyDetectorArn}{[required] The ARN of the anomaly detector to delete. You can find the ARNs of log
anomaly detectors in your account by using the
\code{\link[=cloudwatchlogs_list_log_anomaly_detectors]{list_log_anomaly_detectors}}
operation.}
}
\description{
Deletes the specified CloudWatch Logs anomaly detector.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_log_anomaly_detector/} for full documentation.
}
\keyword{internal}
