% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_log_stream}
\alias{cloudwatchlogs_delete_log_stream}
\title{Deletes the specified log stream and permanently deletes all the
archived log events associated with the log stream}
\usage{
cloudwatchlogs_delete_log_stream(logGroupName, logStreamName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{logStreamName}{[required] The name of the log stream.}
}
\value{
An empty list.
}
\description{
Deletes the specified log stream and permanently deletes all the
archived log events associated with the log stream.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_log_stream(
  logGroupName = "string",
  logStreamName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
