% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_lookup_table}
\alias{cloudwatchlogs_delete_lookup_table}
\title{Deletes a lookup table permanently}
\usage{
cloudwatchlogs_delete_lookup_table(lookupTableArn)
}
\arguments{
\item{lookupTableArn}{[required] The ARN of the lookup table to delete.}
}
\description{
Deletes a lookup table permanently. This operation cannot be undone.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_lookup_table/} for full documentation.
}
\keyword{internal}
