% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_metric_filter}
\alias{cloudwatchlogs_delete_metric_filter}
\title{Deletes the specified metric filter}
\usage{
cloudwatchlogs_delete_metric_filter(logGroupName, filterName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{filterName}{[required] The name of the metric filter.}
}
\description{
Deletes the specified metric filter.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_metric_filter/} for full documentation.
}
\keyword{internal}
