% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_resource_policy}
\alias{cloudwatchlogs_delete_resource_policy}
\title{Deletes a resource policy from this account}
\usage{
cloudwatchlogs_delete_resource_policy(policyName = NULL)
}
\arguments{
\item{policyName}{The name of the policy to be revoked. This parameter is required.}
}
\description{
Deletes a resource policy from this account. This revokes the access of the identities in that policy to put log events to this account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_resource_policy/} for full documentation.
}
\keyword{internal}
