% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_retention_policy}
\alias{cloudwatchlogs_delete_retention_policy}
\title{Deletes the specified retention policy}
\usage{
cloudwatchlogs_delete_retention_policy(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Deletes the specified retention policy.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/delete_retention_policy.html} for full documentation.
}
\keyword{internal}
