% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_scheduled_query}
\alias{cloudwatchlogs_delete_scheduled_query}
\title{Deletes a scheduled query and stops all future executions}
\usage{
cloudwatchlogs_delete_scheduled_query(identifier)
}
\arguments{
\item{identifier}{[required] The ARN or name of the scheduled query to delete.}
}
\description{
Deletes a scheduled query and stops all future executions. This operation also removes any configured actions and associated resources.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_scheduled_query/} for full documentation.
}
\keyword{internal}
