% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_subscription_filter}
\alias{cloudwatchlogs_delete_subscription_filter}
\title{Deletes the specified subscription filter}
\usage{
cloudwatchlogs_delete_subscription_filter(logGroupName, filterName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{filterName}{[required] The name of the subscription filter.}
}
\value{
An empty list.
}
\description{
Deletes the specified subscription filter.
}
\section{Request syntax}{
\preformatted{svc$delete_subscription_filter(
  logGroupName = "string",
  filterName = "string"
)
}
}

\keyword{internal}
