% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_syslog_configuration}
\alias{cloudwatchlogs_delete_syslog_configuration}
\title{Deletes a syslog configuration for a log group}
\usage{
cloudwatchlogs_delete_syslog_configuration(
  logGroupIdentifier,
  vpcEndpointId = NULL
)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group to remove the syslog configuration from.}

\item{vpcEndpointId}{The ID of the VPC endpoint associated with the syslog configuration to delete.}
}
\description{
Deletes a syslog configuration for a log group. After deletion, syslog data is no longer ingested through the specified VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_syslog_configuration/} for full documentation.
}
\keyword{internal}
