% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_transformer}
\alias{cloudwatchlogs_delete_transformer}
\title{Deletes the log transformer for the specified log group}
\usage{
cloudwatchlogs_delete_transformer(logGroupIdentifier)
}
\arguments{
\item{logGroupIdentifier}{[required] Specify either the name or ARN of the log group to delete the
transformer for. If the log group is in a source account and you are
using a monitoring account, you must use the log group ARN.}
}
\description{
Deletes the log transformer for the specified log group. As soon as you do this, the transformation of incoming log events according to that transformer stops. If this account has an account-level transformer that applies to this log group, the log group begins using that account-level transformer when this log-group level transformer is deleted.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_delete_transformer/} for full documentation.
}
\keyword{internal}
