% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_configuration_templates}
\alias{cloudwatchlogs_describe_configuration_templates}
\title{Use this operation to return the valid and default values that are used
when creating delivery sources, delivery destinations, and deliveries}
\usage{
cloudwatchlogs_describe_configuration_templates(
  service = NULL,
  logTypes = NULL,
  resourceTypes = NULL,
  deliveryDestinationTypes = NULL,
  nextToken = NULL,
  limit = NULL
)
}
\arguments{
\item{service}{Use this parameter to filter the response to include only the
configuration templates that apply to the Amazon Web Services service
that you specify here.}

\item{logTypes}{Use this parameter to filter the response to include only the
configuration templates that apply to the log types that you specify
here.}

\item{resourceTypes}{Use this parameter to filter the response to include only the
configuration templates that apply to the resource types that you
specify here.}

\item{deliveryDestinationTypes}{Use this parameter to filter the response to include only the
configuration templates that apply to the delivery destination types
that you specify here.}

\item{nextToken}{}

\item{limit}{Use this parameter to limit the number of configuration templates that
are returned in the response.}
}
\description{
Use this operation to return the valid and default values that are used when creating delivery sources, delivery destinations, and deliveries. For more information about deliveries, see \code{\link[=cloudwatchlogs_create_delivery]{create_delivery}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_configuration_templates/} for full documentation.
}
\keyword{internal}
