% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_deliveries}
\alias{cloudwatchlogs_describe_deliveries}
\title{Retrieves a list of the deliveries that have been created in the account}
\usage{
cloudwatchlogs_describe_deliveries(nextToken = NULL, limit = NULL)
}
\arguments{
\item{nextToken}{The token for the next set of items to return. The token expires after 24 hours.}

\item{limit}{Optionally specify the maximum number of deliveries to return in the response.}
}
\description{
Retrieves a list of the deliveries that have been created in the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_deliveries/} for full documentation.
}
\keyword{internal}
