% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_import_task_batches}
\alias{cloudwatchlogs_describe_import_task_batches}
\title{Gets detailed information about the individual batches within an import
task, including their status and any error messages}
\usage{
cloudwatchlogs_describe_import_task_batches(
  importId,
  batchImportStatus = NULL,
  limit = NULL,
  nextToken = NULL
)
}
\arguments{
\item{importId}{[required] The ID of the import task to get batch information for.}

\item{batchImportStatus}{Optional filter to list import batches by their status. Accepts multiple
status values: IN_PROGRESS, CANCELLED, COMPLETED and FAILED.}

\item{limit}{The maximum number of import batches to return in the response. Default:
10}

\item{nextToken}{The pagination token for the next set of results.}
}
\description{
Gets detailed information about the individual batches within an import task, including their status and any error messages. For CloudTrail Event Data Store sources, a batch refers to a subset of stored events grouped by their eventTime.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_import_task_batches/} for full documentation.
}
\keyword{internal}
