% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_import_tasks}
\alias{cloudwatchlogs_describe_import_tasks}
\title{Lists and describes import tasks, with optional filtering by import
status and source ARN}
\usage{
cloudwatchlogs_describe_import_tasks(
  importId = NULL,
  importStatus = NULL,
  importSourceArn = NULL,
  limit = NULL,
  nextToken = NULL
)
}
\arguments{
\item{importId}{Optional filter to describe a specific import task by its ID.}

\item{importStatus}{Optional filter to list imports by their status. Valid values are
IN_PROGRESS, CANCELLED, COMPLETED and FAILED.}

\item{importSourceArn}{Optional filter to list imports from a specific source}

\item{limit}{The maximum number of import tasks to return in the response. Default:
50}

\item{nextToken}{The pagination token for the next set of results.}
}
\description{
Lists and describes import tasks, with optional filtering by import status and source ARN.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_import_tasks/} for full documentation.
}
\keyword{internal}
