% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_log_groups}
\alias{cloudwatchlogs_describe_log_groups}
\title{Lists the specified log groups}
\usage{
cloudwatchlogs_describe_log_groups(logGroupNamePrefix, nextToken, limit)
}
\arguments{
\item{logGroupNamePrefix}{The prefix to match.}

\item{nextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{limit}{The maximum number of items returned. If you don't specify a value, the
default is up to 50 items.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  logGroups = list(
    list(
      logGroupName = "string",
      creationTime = 123,
      retentionInDays = 123,
      metricFilterCount = 123,
      arn = "string",
      storedBytes = 123,
      kmsKeyId = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the specified log groups. You can list all your log groups or
filter the results by prefix. The results are ASCII-sorted by log group
name.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_log_groups(
  logGroupNamePrefix = "string",
  nextToken = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
