% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_lookup_tables}
\alias{cloudwatchlogs_describe_lookup_tables}
\title{Retrieves metadata about lookup tables in your account}
\usage{
cloudwatchlogs_describe_lookup_tables(
  lookupTableNamePrefix = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{lookupTableNamePrefix}{A prefix to filter lookup tables by name. Only tables whose names start
with this prefix are returned. If you don't specify a prefix, all tables
in the account and Region are returned.}

\item{maxResults}{The maximum number of lookup tables to return in the response. The
default value is 50 and the maximum value is 100.}

\item{nextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Retrieves metadata about lookup tables in your account. You can optionally filter the results by table name prefix. Results are sorted by table name in ascending order.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_lookup_tables/} for full documentation.
}
\keyword{internal}
