% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_queries}
\alias{cloudwatchlogs_describe_queries}
\title{Returns a list of CloudWatch Logs Insights queries that are scheduled,
running, or have been run recently in this account}
\usage{
cloudwatchlogs_describe_queries(
  logGroupName = NULL,
  status = NULL,
  maxResults = NULL,
  nextToken = NULL,
  queryLanguage = NULL
)
}
\arguments{
\item{logGroupName}{Limits the returned queries to only those for the specified log group.}

\item{status}{Limits the returned queries to only those that have the specified
status. Valid values are \code{Cancelled}, \code{Complete}, \code{Failed}, \code{Running},
and \code{Scheduled}.}

\item{maxResults}{Limits the number of returned queries to the specified number.}

\item{nextToken}{}

\item{queryLanguage}{Limits the returned queries to only the queries that use the specified
query language.}
}
\description{
Returns a list of CloudWatch Logs Insights queries that are scheduled, running, or have been run recently in this account. You can request all queries or limit it to queries of a specific log group or queries with a certain status.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_queries/} for full documentation.
}
\keyword{internal}
