% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_resource_policies}
\alias{cloudwatchlogs_describe_resource_policies}
\title{Lists the resource policies in this account}
\usage{
cloudwatchlogs_describe_resource_policies(nextToken = NULL, limit = NULL)
}
\arguments{
\item{nextToken}{}

\item{limit}{The maximum number of resource policies to be displayed with one call of
this API.}
}
\description{
Lists the resource policies in this account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_resource_policies/} for full documentation.
}
\keyword{internal}
