% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_disassociate_kms_key}
\alias{cloudwatchlogs_disassociate_kms_key}
\title{Disassociates the associated KMS key from the specified log group}
\usage{
cloudwatchlogs_disassociate_kms_key(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Disassociates the associated KMS key from the specified log group.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_disassociate_kms_key/} for full documentation.
}
\keyword{internal}
