% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_disassociate_source_from_s3_table_integration}
\alias{cloudwatchlogs_disassociate_source_from_s3_table_integration}
\title{Disassociates a data source from an S3 Table Integration, removing query
access and deleting all associated data from the integration}
\usage{
cloudwatchlogs_disassociate_source_from_s3_table_integration(identifier)
}
\arguments{
\item{identifier}{[required] The unique identifier of the association to remove between the data
source and S3 Table Integration.}
}
\description{
Disassociates a data source from an S3 Table Integration, removing query access and deleting all associated data from the integration.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_disassociate_source_from_s3_table_integration/} for full documentation.
}
\keyword{internal}
