% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_data_protection_policy}
\alias{cloudwatchlogs_get_data_protection_policy}
\title{Returns information about a log group data protection policy}
\usage{
cloudwatchlogs_get_data_protection_policy(logGroupIdentifier)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group that contains the data protection
policy that you want to see.}
}
\description{
Returns information about a log group data protection policy.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/get_data_protection_policy.html} for full documentation.
}
\keyword{internal}
