% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_delivery_destination}
\alias{cloudwatchlogs_get_delivery_destination}
\title{Retrieves complete information about one delivery destination}
\usage{
cloudwatchlogs_get_delivery_destination(name)
}
\arguments{
\item{name}{[required] The name of the delivery destination that you want to retrieve.}
}
\description{
Retrieves complete information about one delivery destination.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_delivery_destination/} for full documentation.
}
\keyword{internal}
