% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_delivery_destination_policy}
\alias{cloudwatchlogs_get_delivery_destination_policy}
\title{Retrieves the delivery destination policy assigned to the delivery
destination that you specify}
\usage{
cloudwatchlogs_get_delivery_destination_policy(deliveryDestinationName)
}
\arguments{
\item{deliveryDestinationName}{[required] The name of the delivery destination that you want to retrieve the
policy of.}
}
\description{
Retrieves the delivery destination policy assigned to the delivery destination that you specify. For more information about delivery destinations and their policies, see \code{\link[=cloudwatchlogs_put_delivery_destination_policy]{put_delivery_destination_policy}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_delivery_destination_policy/} for full documentation.
}
\keyword{internal}
