% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_delivery_source}
\alias{cloudwatchlogs_get_delivery_source}
\title{Retrieves complete information about one delivery source}
\usage{
cloudwatchlogs_get_delivery_source(name)
}
\arguments{
\item{name}{[required] The name of the delivery source that you want to retrieve.}
}
\description{
Retrieves complete information about one delivery source.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_delivery_source/} for full documentation.
}
\keyword{internal}
