% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_integration}
\alias{cloudwatchlogs_get_integration}
\title{Returns information about one integration between CloudWatch Logs and
OpenSearch Service}
\usage{
cloudwatchlogs_get_integration(integrationName)
}
\arguments{
\item{integrationName}{[required] The name of the integration that you want to find information about. To
find the name of your integration, use
\code{\link[=cloudwatchlogs_list_integrations]{list_integrations}}}
}
\description{
Returns information about one integration between CloudWatch Logs and OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_integration/} for full documentation.
}
\keyword{internal}
