% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_log_fields}
\alias{cloudwatchlogs_get_log_fields}
\title{Discovers available fields for a specific data source and type}
\usage{
cloudwatchlogs_get_log_fields(dataSourceName, dataSourceType)
}
\arguments{
\item{dataSourceName}{[required] The name of the data source to retrieve log fields for.}

\item{dataSourceType}{[required] The type of the data source to retrieve log fields for.}
}
\description{
Discovers available fields for a specific data source and type. The response includes any field modifications introduced through pipelines, such as new fields or changed field types.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_log_fields/} for full documentation.
}
\keyword{internal}
