% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_log_object}
\alias{cloudwatchlogs_get_log_object}
\title{Retrieves a large logging object (LLO) and streams it back}
\usage{
cloudwatchlogs_get_log_object(unmask = NULL, logObjectPointer)
}
\arguments{
\item{unmask}{A boolean flag that indicates whether to unmask sensitive log data. When
set to true, any masked or redacted data in the log object will be
displayed in its original form. Default is false.}

\item{logObjectPointer}{[required] A pointer to the specific log object to retrieve. This is a required
parameter that uniquely identifies the log object within CloudWatch
Logs. The pointer is typically obtained from a previous query or filter
operation.}
}
\description{
Retrieves a large logging object (LLO) and streams it back. This API is used to fetch the content of large portions of log events that have been ingested through the PutOpenTelemetryLogs API. When log events contain fields that would cause the total event size to exceed 1MB, CloudWatch Logs automatically processes up to 10 fields, starting with the largest fields. Each field is truncated as needed to keep the total event size as close to 1MB as possible. The excess portions are stored as Large Log Objects (LLOs) and these fields are processed separately and LLO reference system fields (in the format \verb{@ptr.$[path.to.field]}) are added. The path in the reference field reflects the original JSON structure where the large field was located. For example, this could be \verb{@ptr.$['input']['message']}, \verb{@ptr.$['AAA']['BBB']['CCC']['DDD']}, \verb{@ptr.$['AAA']}, or any other path matching your log structure.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_log_object/} for full documentation.
}
\keyword{internal}
