% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_lookup_table}
\alias{cloudwatchlogs_get_lookup_table}
\title{Retrieves the full content of a lookup table, including the CSV data}
\usage{
cloudwatchlogs_get_lookup_table(lookupTableArn)
}
\arguments{
\item{lookupTableArn}{[required] The ARN of the lookup table to retrieve.}
}
\description{
Retrieves the full content of a lookup table, including the CSV data.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_lookup_table/} for full documentation.
}
\keyword{internal}
