% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_query_results}
\alias{cloudwatchlogs_get_query_results}
\title{Returns the results from the specified query}
\usage{
cloudwatchlogs_get_query_results(queryId, nextToken = NULL, maxItems = NULL)
}
\arguments{
\item{queryId}{[required] The ID number of the query.}

\item{nextToken}{The token for the next set of items to return. The token expires after 1
hour.}

\item{maxItems}{The maximum number of log events to return in the response. The maximum
is 10,000 log events.}
}
\description{
Returns the results from the specified query.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_query_results/} for full documentation.
}
\keyword{internal}
