% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_scheduled_query}
\alias{cloudwatchlogs_get_scheduled_query}
\title{Retrieves details about a specific scheduled query, including its
configuration, execution status, and metadata}
\usage{
cloudwatchlogs_get_scheduled_query(identifier)
}
\arguments{
\item{identifier}{[required] The ARN or name of the scheduled query to retrieve.}
}
\description{
Retrieves details about a specific scheduled query, including its configuration, execution status, and metadata.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_scheduled_query/} for full documentation.
}
\keyword{internal}
