% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_scheduled_query_history}
\alias{cloudwatchlogs_get_scheduled_query_history}
\title{Retrieves the execution history of a scheduled query within a specified
time range, including query results and destination processing status}
\usage{
cloudwatchlogs_get_scheduled_query_history(
  identifier,
  startTime,
  endTime,
  executionStatuses = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{identifier}{[required] The ARN or name of the scheduled query to retrieve history for.}

\item{startTime}{[required] The start time for the history query in Unix epoch format.}

\item{endTime}{[required] The end time for the history query in Unix epoch format.}

\item{executionStatuses}{An array of execution statuses to filter the history results. Only
executions with the specified statuses are returned.}

\item{maxResults}{The maximum number of history records to return. Valid range is 1 to
1000.}

\item{nextToken}{}
}
\description{
Retrieves the execution history of a scheduled query within a specified time range, including query results and destination processing status.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_scheduled_query_history/} for full documentation.
}
\keyword{internal}
