% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_storage_tier_policy}
\alias{cloudwatchlogs_get_storage_tier_policy}
\title{Returns the storage tier policy for the account}
\usage{
cloudwatchlogs_get_storage_tier_policy()
}
\description{
Returns the storage tier policy for the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_storage_tier_policy/} for full documentation.
}
\keyword{internal}
