% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_transformer}
\alias{cloudwatchlogs_get_transformer}
\title{Returns the information about the log transformer associated with this
log group}
\usage{
cloudwatchlogs_get_transformer(logGroupIdentifier)
}
\arguments{
\item{logGroupIdentifier}{[required] Specify either the name or ARN of the log group to return transformer
information for. If the log group is in a source account and you are
using a monitoring account, you must use the log group ARN.}
}
\description{
Returns the information about the log transformer associated with this log group.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_transformer/} for full documentation.
}
\keyword{internal}
