% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_aggregate_log_group_summaries}
\alias{cloudwatchlogs_list_aggregate_log_group_summaries}
\title{Returns an aggregate summary of all log groups in the Region grouped by
specified data source characteristics}
\usage{
cloudwatchlogs_list_aggregate_log_group_summaries(
  accountIdentifiers = NULL,
  includeLinkedAccounts = NULL,
  logGroupClass = NULL,
  logGroupNamePattern = NULL,
  dataSources = NULL,
  groupBy,
  nextToken = NULL,
  limit = NULL
)
}
\arguments{
\item{accountIdentifiers}{When \code{includeLinkedAccounts} is set to \code{true}, use this parameter to
specify the list of accounts to search. You can specify as many as 20
account IDs in the array.}

\item{includeLinkedAccounts}{If you are using a monitoring account, set this to \code{true} to have the
operation return log groups in the accounts listed in
\code{accountIdentifiers}.

If this parameter is set to \code{true} and \code{accountIdentifiers} contains a
null value, the operation returns all log groups in the monitoring
account and all log groups in all source accounts that are linked to the
monitoring account.

The default for this parameter is \code{false}.}

\item{logGroupClass}{Filters the results by log group class to include only log groups of the
specified class.}

\item{logGroupNamePattern}{Use this parameter to limit the returned log groups to only those with
names that match the pattern that you specify. This parameter is a
regular expression that can match prefixes and substrings, and supports
wildcard matching and matching multiple patterns, as in the following
examples.
\itemize{
\item Use \code{^} to match log group names by prefix.
\item For a substring match, specify the string to match. All matches are
case sensitive
\item To match multiple patterns, separate them with a \code{|} as in the
example \verb{^/aws/lambda|discovery}
}

You can specify as many as five different regular expression patterns in
this field, each of which must be between 3 and 24 characters. You can
include the \code{^} symbol as many as five times, and include the \code{|} symbol
as many as four times.}

\item{dataSources}{Filters the results by data source characteristics to include only log
groups associated with the specified data sources.}

\item{groupBy}{[required] Specifies how to group the log groups in the summary.}

\item{nextToken}{}

\item{limit}{The maximum number of aggregated summaries to return. If you omit this
parameter, the default is up to 50 aggregated summaries.}
}
\description{
Returns an aggregate summary of all log groups in the Region grouped by specified data source characteristics. Supports optional filtering by log group class, name patterns, and data sources. If you perform this action in a monitoring account, you can also return aggregated summaries of log groups from source accounts that are linked to the monitoring account. For more information about using cross-account observability to set up monitoring accounts and source accounts, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Unified-Cross-Account.html}{CloudWatch cross-account observability}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_aggregate_log_group_summaries/} for full documentation.
}
\keyword{internal}
