% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_anomalies}
\alias{cloudwatchlogs_list_anomalies}
\title{Returns a list of anomalies that log anomaly detectors have found}
\usage{
cloudwatchlogs_list_anomalies(
  anomalyDetectorArn = NULL,
  suppressionState = NULL,
  limit = NULL,
  nextToken = NULL
)
}
\arguments{
\item{anomalyDetectorArn}{Use this to optionally limit the results to only the anomalies found by
a certain anomaly detector.}

\item{suppressionState}{You can specify this parameter if you want to the operation to return
only anomalies that are currently either suppressed or unsuppressed.}

\item{limit}{The maximum number of items to return. If you don't specify a value, the
default maximum value of 50 items is used.}

\item{nextToken}{}
}
\description{
Returns a list of anomalies that log anomaly detectors have found. For details about the structure format of each anomaly object that is returned, see the example in this section.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_anomalies/} for full documentation.
}
\keyword{internal}
