% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_integrations}
\alias{cloudwatchlogs_list_integrations}
\title{Returns a list of integrations between CloudWatch Logs and other
services in this account}
\usage{
cloudwatchlogs_list_integrations(
  integrationNamePrefix = NULL,
  integrationType = NULL,
  integrationStatus = NULL
)
}
\arguments{
\item{integrationNamePrefix}{To limit the results to integrations that start with a certain name
prefix, specify that name prefix here.}

\item{integrationType}{To limit the results to integrations of a certain type, specify that
type here.}

\item{integrationStatus}{To limit the results to integrations with a certain status, specify that
status here.}
}
\description{
Returns a list of integrations between CloudWatch Logs and other services in this account. Currently, only one integration can be created in an account, and this integration must be with OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_integrations/} for full documentation.
}
\keyword{internal}
