% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_log_anomaly_detectors}
\alias{cloudwatchlogs_list_log_anomaly_detectors}
\title{Retrieves a list of the log anomaly detectors in the account}
\usage{
cloudwatchlogs_list_log_anomaly_detectors(
  filterLogGroupArn = NULL,
  limit = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filterLogGroupArn}{Use this to optionally filter the results to only include anomaly
detectors that are associated with the specified log group.}

\item{limit}{The maximum number of items to return. If you don't specify a value, the
default maximum value of 50 items is used.}

\item{nextToken}{}
}
\description{
Retrieves a list of the log anomaly detectors in the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_log_anomaly_detectors/} for full documentation.
}
\keyword{internal}
