% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_log_groups}
\alias{cloudwatchlogs_list_log_groups}
\title{Returns a list of log groups in the Region in your account}
\usage{
cloudwatchlogs_list_log_groups(
  logGroupNamePattern = NULL,
  logGroupClass = NULL,
  includeLinkedAccounts = NULL,
  accountIdentifiers = NULL,
  nextToken = NULL,
  limit = NULL,
  dataSources = NULL,
  fieldIndexNames = NULL
)
}
\arguments{
\item{logGroupNamePattern}{Use this parameter to limit the returned log groups to only those with
names that match the pattern that you specify. This parameter is a
regular expression that can match prefixes and substrings, and supports
wildcard matching and matching multiple patterns, as in the following
examples.
\itemize{
\item Use \code{^} to match log group names by prefix.
\item For a substring match, specify the string to match. All matches are
case sensitive
\item To match multiple patterns, separate them with a \code{|} as in the
example \verb{^/aws/lambda|discovery}
}

You can specify as many as five different regular expression patterns in
this field, each of which must be between 3 and 24 characters. You can
include the \code{^} symbol as many as five times, and include the \code{|} symbol
as many as four times.}

\item{logGroupClass}{Use this parameter to limit the results to only those log groups in the
specified log group class. If you omit this parameter, log groups of all
classes can be returned.}

\item{includeLinkedAccounts}{If you are using a monitoring account, set this to \code{true} to have the
operation return log groups in the accounts listed in
\code{accountIdentifiers}.

If this parameter is set to \code{true} and \code{accountIdentifiers} contains a
null value, the operation returns all log groups in the monitoring
account and all log groups in all source accounts that are linked to the
monitoring account.

The default for this parameter is \code{false}.}

\item{accountIdentifiers}{When \code{includeLinkedAccounts} is set to \code{true}, use this parameter to
specify the list of accounts to search. You can specify as many as 20
account IDs in the array.}

\item{nextToken}{}

\item{limit}{The maximum number of log groups to return. If you omit this parameter,
the default is up to 50 log groups.}

\item{dataSources}{An array of data source filters to filter log groups by their associated
data sources. You can filter by data source name, type, or both.
Multiple filters within the same dimension are combined with OR logic,
while filters across different dimensions are combined with AND logic.}

\item{fieldIndexNames}{An array of field index names to filter log groups that have specific
field indexes. Only log groups containing all specified field indexes
are returned. You can specify 1 to 20 field index names, each with 1 to
512 characters.}
}
\description{
Returns a list of log groups in the Region in your account. If you are performing this action in a monitoring account, you can choose to also return log groups from source accounts that are linked to the monitoring account. For more information about using cross-account observability to set up monitoring accounts and source accounts, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Unified-Cross-Account.html}{CloudWatch cross-account observability}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_log_groups/} for full documentation.
}
\keyword{internal}
