% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_log_groups_for_query}
\alias{cloudwatchlogs_list_log_groups_for_query}
\title{Returns a list of the log groups that were analyzed during a single
CloudWatch Logs Insights query}
\usage{
cloudwatchlogs_list_log_groups_for_query(
  queryId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{queryId}{[required] The ID of the query to use. This query ID is from the response to your
\code{\link[=cloudwatchlogs_start_query]{start_query}} operation.}

\item{nextToken}{}

\item{maxResults}{Limits the number of returned log groups to the specified number.}
}
\description{
Returns a list of the log groups that were analyzed during a single CloudWatch Logs Insights query. This can be useful for queries that use log group name prefixes or the \code{filterIndex} command, because the log groups are dynamically selected in these cases.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_log_groups_for_query/} for full documentation.
}
\keyword{internal}
