% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_sources_for_s3_table_integration}
\alias{cloudwatchlogs_list_sources_for_s3_table_integration}
\title{Returns a list of data source associations for a specified S3 Table
Integration, showing which data sources are currently associated for
query access}
\usage{
cloudwatchlogs_list_sources_for_s3_table_integration(
  integrationArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{integrationArn}{[required] The Amazon Resource Name (ARN) of the S3 Table Integration to list
associations for.}

\item{maxResults}{The maximum number of associations to return in a single call. Valid
range is 1 to 100.}

\item{nextToken}{}
}
\description{
Returns a list of data source associations for a specified S3 Table Integration, showing which data sources are currently associated for query access.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_sources_for_s3_table_integration/} for full documentation.
}
\keyword{internal}
