% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_syslog_configurations}
\alias{cloudwatchlogs_list_syslog_configurations}
\title{Returns a list of syslog configurations}
\usage{
cloudwatchlogs_list_syslog_configurations(
  logGroupIdentifier = NULL,
  vpcEndpointId = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{logGroupIdentifier}{The name or ARN of the log group to filter syslog configurations for.}

\item{vpcEndpointId}{The ID of the VPC endpoint to filter syslog configurations for.}

\item{nextToken}{The token for the next set of items to return. You received this token from a previous call.}

\item{maxResults}{The maximum number of syslog configurations to return in the response.}
}
\description{
Returns a list of syslog configurations. You can optionally filter the results by log group or VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_syslog_configurations/} for full documentation.
}
\keyword{internal}
