% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_tags_for_resource}
\alias{cloudwatchlogs_list_tags_for_resource}
\title{Displays the tags associated with a CloudWatch Logs resource}
\usage{
cloudwatchlogs_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that you want to view tags for.

The ARN format of a log group is
\code{arn:aws:logs:Region:account-id:log-group:log-group-name }

The ARN format of a destination is
\code{arn:aws:logs:Region:account-id:destination:destination-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html}{CloudWatch Logs resources and operations}.}
}
\description{
Displays the tags associated with a CloudWatch Logs resource. Currently, log groups and destinations support tagging.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
