% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_bearer_token_authentication}
\alias{cloudwatchlogs_put_bearer_token_authentication}
\title{Enables or disables bearer token authentication for the specified log
group}
\usage{
cloudwatchlogs_put_bearer_token_authentication(
  logGroupIdentifier,
  bearerTokenAuthenticationEnabled
)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group.

Type: String

Length Constraints: Minimum length of 1. Maximum length of 512.

Pattern: \verb{[\\.\\-_/#A-Za-z0-9]+}

Required: Yes}

\item{bearerTokenAuthenticationEnabled}{[required] Whether to enable bearer token authentication.

Type: Boolean

Required: Yes}
}
\description{
Enables or disables bearer token authentication for the specified log group. When enabled on a log group, bearer token authentication is enabled on operations until it is explicitly disabled.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_bearer_token_authentication/} for full documentation.
}
\keyword{internal}
