% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_delivery_destination_policy}
\alias{cloudwatchlogs_put_delivery_destination_policy}
\title{Creates and assigns an IAM policy that grants permissions to CloudWatch
Logs to deliver logs cross-account to a specified destination in this
account}
\usage{
cloudwatchlogs_put_delivery_destination_policy(
  deliveryDestinationName,
  deliveryDestinationPolicy
)
}
\arguments{
\item{deliveryDestinationName}{[required] The name of the delivery destination to assign this policy to.}

\item{deliveryDestinationPolicy}{[required] The contents of the policy.}
}
\description{
Creates and assigns an IAM policy that grants permissions to CloudWatch Logs to deliver logs cross-account to a specified destination in this account. To configure the delivery of logs from an Amazon Web Services service in another account to a logs delivery destination in the current account, you must do the following:

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_delivery_destination_policy/} for full documentation.
}
\keyword{internal}
