% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_integration}
\alias{cloudwatchlogs_put_integration}
\title{Creates an integration between CloudWatch Logs and another service in
this account}
\usage{
cloudwatchlogs_put_integration(
  integrationName,
  resourceConfig,
  integrationType
)
}
\arguments{
\item{integrationName}{[required] A name for the integration.}

\item{resourceConfig}{[required] A structure that contains configuration information for the integration
that you are creating.}

\item{integrationType}{[required] The type of integration. Currently, the only supported type is
\code{OPENSEARCH}.}
}
\description{
Creates an integration between CloudWatch Logs and another service in this account. Currently, only integrations with OpenSearch Service are supported, and currently you can have only one integration in your account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_integration/} for full documentation.
}
\keyword{internal}
