% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_log_group_deletion_protection}
\alias{cloudwatchlogs_put_log_group_deletion_protection}
\title{Enables or disables deletion protection for the specified log group}
\usage{
cloudwatchlogs_put_log_group_deletion_protection(
  logGroupIdentifier,
  deletionProtectionEnabled
)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group.

Type: String

Length Constraints: Minimum length of 1. Maximum length of 512.

Pattern: \verb{[\\.\\-_/#A-Za-z0-9]+}

Required: Yes}

\item{deletionProtectionEnabled}{[required] Whether to enable deletion protection.

Type: Boolean

Required: Yes}
}
\description{
Enables or disables deletion protection for the specified log group. When enabled on a log group, deletion protection blocks all deletion operations until it is explicitly disabled.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_log_group_deletion_protection/} for full documentation.
}
\keyword{internal}
