% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_syslog_configuration}
\alias{cloudwatchlogs_put_syslog_configuration}
\title{Creates or updates a syslog configuration for a log group}
\usage{
cloudwatchlogs_put_syslog_configuration(
  logGroupIdentifier,
  vpcEndpointId = NULL
)
}
\arguments{
\item{logGroupIdentifier}{[required] The name or ARN of the log group to associate with the syslog configuration.}

\item{vpcEndpointId}{The ID of the VPC endpoint to use for syslog ingestion.}
}
\description{
Creates or updates a syslog configuration for a log group. This enables ingestion of syslog data through the specified VPC endpoint into the log group.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_syslog_configuration/} for full documentation.
}
\keyword{internal}
