% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_transformer}
\alias{cloudwatchlogs_put_transformer}
\title{Creates or updates a log transformer for a single log group}
\usage{
cloudwatchlogs_put_transformer(logGroupIdentifier, transformerConfig)
}
\arguments{
\item{logGroupIdentifier}{[required] Specify either the name or ARN of the log group to create the
transformer for.}

\item{transformerConfig}{[required] This structure contains the configuration of this log transformer. A log
transformer is an array of processors, where each processor applies one
type of transformation to the log events that are ingested.}
}
\description{
Creates or updates a \emph{log transformer} for a single log group. You use log transformers to transform log events into a different format, making them easier for you to process and analyze. You can also transform logs from different sources into standardized formats that contains relevant, source-specific information.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_put_transformer/} for full documentation.
}
\keyword{internal}
