% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_stop_query}
\alias{cloudwatchlogs_stop_query}
\title{Stops a CloudWatch Logs Insights query that is in progress}
\usage{
cloudwatchlogs_stop_query(queryId)
}
\arguments{
\item{queryId}{[required] The ID number of the query to stop. To find this ID number, use
\code{\link[=cloudwatchlogs_describe_queries]{describe_queries}}.}
}
\description{
Stops a CloudWatch Logs Insights query that is in progress. If the query has already ended, the operation returns an error indicating that the specified query is not running.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/stop_query.html} for full documentation.
}
\keyword{internal}
