% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_test_transformer}
\alias{cloudwatchlogs_test_transformer}
\title{Use this operation to test a log transformer}
\usage{
cloudwatchlogs_test_transformer(transformerConfig, logEventMessages)
}
\arguments{
\item{transformerConfig}{[required] This structure contains the configuration of this log transformer that
you want to test. A log transformer is an array of processors, where
each processor applies one type of transformation to the log events that
are ingested.}

\item{logEventMessages}{[required] An array of the raw log events that you want to use to test this
transformer.}
}
\description{
Use this operation to test a log transformer. You enter the transformer configuration and a set of log events to test with. The operation responds with an array that includes the original log events and the transformed versions.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_test_transformer/} for full documentation.
}
\keyword{internal}
