% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_delete_link}
\alias{cloudwatchobservabilityaccessmanager_delete_link}
\title{Deletes a link between a monitoring account sink and a source account}
\usage{
cloudwatchobservabilityaccessmanager_delete_link(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the link to delete.}
}
\description{
Deletes a link between a monitoring account sink and a source account. You must run this operation in the source account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_delete_link/} for full documentation.
}
\keyword{internal}
