% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_delete_sink}
\alias{cloudwatchobservabilityaccessmanager_delete_sink}
\title{Deletes a sink}
\usage{
cloudwatchobservabilityaccessmanager_delete_sink(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the sink to delete.}
}
\description{
Deletes a sink. You must delete all links to a sink before you can delete that sink.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_delete_sink/} for full documentation.
}
\keyword{internal}
