% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_get_link}
\alias{cloudwatchobservabilityaccessmanager_get_link}
\title{Returns complete information about one link}
\usage{
cloudwatchobservabilityaccessmanager_get_link(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the link to retrieve information for.}
}
\description{
Returns complete information about one link.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_get_link/} for full documentation.
}
\keyword{internal}
