% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_get_sink}
\alias{cloudwatchobservabilityaccessmanager_get_sink}
\title{Returns complete information about one monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_get_sink(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the sink to retrieve information for.}
}
\description{
Returns complete information about one monitoring account sink.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_get_sink/} for full documentation.
}
\keyword{internal}
