% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_get_sink_policy}
\alias{cloudwatchobservabilityaccessmanager_get_sink_policy}
\title{Returns the current sink policy attached to this sink}
\usage{
cloudwatchobservabilityaccessmanager_get_sink_policy(SinkIdentifier)
}
\arguments{
\item{SinkIdentifier}{[required] The ARN of the sink to retrieve the policy of.}
}
\description{
Returns the current sink policy attached to this sink. The sink policy specifies what accounts can attach to this sink as source accounts, and what types of data they can share.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_get_sink_policy/} for full documentation.
}
\keyword{internal}
