% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_list_attached_links}
\alias{cloudwatchobservabilityaccessmanager_list_attached_links}
\title{Returns a list of source account links that are linked to this
monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_list_attached_links(
  MaxResults = NULL,
  NextToken = NULL,
  SinkIdentifier
)
}
\arguments{
\item{MaxResults}{Limits the number of returned links to the specified number.}

\item{NextToken}{The token for the next set of items to return. You received this token
from a previous call.}

\item{SinkIdentifier}{[required] The ARN of the sink that you want to retrieve links for.}
}
\description{
Returns a list of source account links that are linked to this monitoring account sink.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_list_attached_links/} for full documentation.
}
\keyword{internal}
