% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_list_links}
\alias{cloudwatchobservabilityaccessmanager_list_links}
\title{Use this operation in a source account to return a list of links to
monitoring account sinks that this source account has}
\usage{
cloudwatchobservabilityaccessmanager_list_links(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Limits the number of returned links to the specified number.}

\item{NextToken}{The token for the next set of items to return. You received this token
from a previous call.}
}
\description{
Use this operation in a source account to return a list of links to monitoring account sinks that this source account has.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_list_links/} for full documentation.
}
\keyword{internal}
