% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_list_sinks}
\alias{cloudwatchobservabilityaccessmanager_list_sinks}
\title{Use this operation in a monitoring account to return the list of sinks
created in that account}
\usage{
cloudwatchobservabilityaccessmanager_list_sinks(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Limits the number of returned links to the specified number.}

\item{NextToken}{The token for the next set of items to return. You received this token
from a previous call.}
}
\description{
Use this operation in a monitoring account to return the list of sinks created in that account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_list_sinks/} for full documentation.
}
\keyword{internal}
