% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_list_tags_for_resource}
\alias{cloudwatchobservabilityaccessmanager_list_tags_for_resource}
\title{Displays the tags associated with a resource}
\usage{
cloudwatchobservabilityaccessmanager_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource that you want to view tags for.

The ARN format of a sink is
\code{arn:aws:oam:Region:account-id:sink/sink-id }

The ARN format of a link is
\code{arn:aws:oam:Region:account-id:link/link-id }

For more information about ARN format, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html}{CloudWatch Logs resources and operations}.

Unlike tagging permissions in other Amazon Web Services services, to
retrieve the list of tags for links or sinks you must have the
\code{oam:RequestTag} permission. The \code{aws:ReguestTag} permission does not
allow you to tag and untag links and sinks.}
}
\description{
Displays the tags associated with a resource. Both sinks and links support tagging.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
