% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_put_sink_policy}
\alias{cloudwatchobservabilityaccessmanager_put_sink_policy}
\title{Creates or updates the resource policy that grants permissions to source
accounts to link to the monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_put_sink_policy(SinkIdentifier, Policy)
}
\arguments{
\item{SinkIdentifier}{[required] The ARN of the sink to attach this policy to.}

\item{Policy}{[required] The JSON policy to use. If you are updating an existing policy, the
entire existing policy is replaced by what you specify here.

The policy must be in JSON string format with quotation marks escaped
and no newlines.

For examples of different types of policies, see the \strong{Examples}
section on this page.}
}
\description{
Creates or updates the resource policy that grants permissions to source accounts to link to the monitoring account sink. When you create a sink policy, you can grant permissions to all accounts in an organization or to individual accounts.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_put_sink_policy/} for full documentation.
}
\keyword{internal}
