% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_tag_resource}
\alias{cloudwatchobservabilityaccessmanager_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified resource}
\usage{
cloudwatchobservabilityaccessmanager_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource that you're adding tags to.

The ARN format of a sink is
\code{arn:aws:oam:Region:account-id:sink/sink-id }

The ARN format of a link is
\code{arn:aws:oam:Region:account-id:link/link-id }

For more information about ARN format, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html}{CloudWatch Logs resources and operations}.}

\item{Tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified resource. Both sinks and links can be tagged.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_tag_resource/} for full documentation.
}
\keyword{internal}
