% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_untag_resource}
\alias{cloudwatchobservabilityaccessmanager_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
cloudwatchobservabilityaccessmanager_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource that you're removing tags from.

The ARN format of a sink is
\code{arn:aws:oam:Region:account-id:sink/sink-id }

The ARN format of a link is
\code{arn:aws:oam:Region:account-id:link/link-id }

For more information about ARN format, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html}{CloudWatch Logs resources and operations}.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_untag_resource/} for full documentation.
}
\keyword{internal}
