% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_update_link}
\alias{cloudwatchobservabilityaccessmanager_update_link}
\title{Use this operation to change what types of data are shared from a source
account to its linked monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_update_link(Identifier, ResourceTypes)
}
\arguments{
\item{Identifier}{[required] The ARN of the link that you want to update.}

\item{ResourceTypes}{[required] An array of strings that define which types of data that the source
account will send to the monitoring account.

Your input here replaces the current set of data types that are shared.}
}
\description{
Use this operation to change what types of data are shared from a source account to its linked monitoring account sink. You can't change the sink or change the monitoring account with this operation.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_update_link/} for full documentation.
}
\keyword{internal}
