% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_delete_app_monitor}
\alias{cloudwatchrum_delete_app_monitor}
\title{Deletes an existing app monitor}
\usage{
cloudwatchrum_delete_app_monitor(Name)
}
\arguments{
\item{Name}{[required] The name of the app monitor to delete.}
}
\description{
Deletes an existing app monitor. This immediately stops the collection of data.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_delete_app_monitor/} for full documentation.
}
\keyword{internal}
