% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_delete_resource_policy}
\alias{cloudwatchrum_delete_resource_policy}
\title{Removes the association of a resource-based policy from an app monitor}
\usage{
cloudwatchrum_delete_resource_policy(Name, PolicyRevisionId = NULL)
}
\arguments{
\item{Name}{[required] The app monitor that you want to remove the resource policy from.}

\item{PolicyRevisionId}{Specifies a specific policy revision to delete. Provide a
\code{PolicyRevisionId} to ensure an atomic delete operation. If the revision
ID that you provide doesn't match the latest policy revision ID, the
request will be rejected with an \code{InvalidPolicyRevisionIdException}
error.}
}
\description{
Removes the association of a resource-based policy from an app monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_delete_resource_policy/} for full documentation.
}
\keyword{internal}
