% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_get_resource_policy}
\alias{cloudwatchrum_get_resource_policy}
\title{Use this operation to retrieve information about a resource-based policy
that is attached to an app monitor}
\usage{
cloudwatchrum_get_resource_policy(Name)
}
\arguments{
\item{Name}{[required] The name of the app monitor that is associated with the resource-based
policy that you want to view.}
}
\description{
Use this operation to retrieve information about a resource-based policy that is attached to an app monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_get_resource_policy/} for full documentation.
}
\keyword{internal}
