% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_put_resource_policy}
\alias{cloudwatchrum_put_resource_policy}
\title{Use this operation to assign a resource-based policy to a CloudWatch RUM
app monitor to control access to it}
\usage{
cloudwatchrum_put_resource_policy(
  Name,
  PolicyDocument,
  PolicyRevisionId = NULL
)
}
\arguments{
\item{Name}{[required] The name of the app monitor that you want to apply this resource-based
policy to. To find the names of your app monitors, you can use the
\code{\link[=cloudwatchrum_list_app_monitors]{list_app_monitors}} operation.}

\item{PolicyDocument}{[required] The JSON to use as the resource policy. The document can be up to 4 KB
in size. For more information about the contents and syntax for this
policy, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-RUM-resource-policies.html}{Using resource-based policies with CloudWatch RUM}.}

\item{PolicyRevisionId}{A string value that you can use to conditionally update your policy. You
can provide the revision ID of your existing policy to make mutating
requests against that policy.

When you assign a policy revision ID, then later requests about that
policy will be rejected with an \code{InvalidPolicyRevisionIdException} error
if they don't provide the correct current revision ID.}
}
\description{
Use this operation to assign a resource-based policy to a CloudWatch RUM app monitor to control access to it. Each app monitor can have one resource-based policy. The maximum size of the policy is 4 KB. To learn more about using resource policies with RUM, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-RUM-resource-policies.html}{Using resource-based policies with CloudWatch RUM}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_put_resource_policy/} for full documentation.
}
\keyword{internal}
