% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_tag_resource}
\alias{cloudwatchrum_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified CloudWatch
RUM resource}
\usage{
cloudwatchrum_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the CloudWatch RUM resource that you're adding tags to.}

\item{Tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified CloudWatch RUM resource. Currently, the only resources that can be tagged app monitors.

See \url{https://paws-r.github.io/docs/cloudwatchrum/tag_resource.html} for full documentation.
}
\keyword{internal}
