% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_untag_resource}
\alias{cloudwatchrum_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
cloudwatchrum_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the CloudWatch RUM resource that you're removing tags from.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_untag_resource/} for full documentation.
}
\keyword{internal}
