% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_associate_resource_types}
\alias{configservice_associate_resource_types}
\title{Adds all resource types specified in the ResourceTypes list to the
RecordingGroup of specified configuration recorder and includes those
resource types when recording}
\usage{
configservice_associate_resource_types(ConfigurationRecorderArn, ResourceTypes)
}
\arguments{
\item{ConfigurationRecorderArn}{[required] The Amazon Resource Name (ARN) of the specified configuration recorder.}

\item{ResourceTypes}{[required] The list of resource types you want to add to the recording group of the
specified configuration recorder.}
}
\description{
Adds all resource types specified in the \code{ResourceTypes} list to the \href{https://docs.aws.amazon.com/config/latest/APIReference/API_RecordingGroup.html}{RecordingGroup} of specified configuration recorder and includes those resource types when recording.

See \url{https://www.paws-r-sdk.com/docs/configservice_associate_resource_types/} for full documentation.
}
\keyword{internal}
