% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_batch_get_aggregate_resource_config}
\alias{configservice_batch_get_aggregate_resource_config}
\title{Returns the current configuration items for resources that are present
in your Config aggregator}
\usage{
configservice_batch_get_aggregate_resource_config(
  ConfigurationAggregatorName,
  ResourceIdentifiers
)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{ResourceIdentifiers}{[required] A list of aggregate ResourceIdentifiers objects.}
}
\description{
Returns the current configuration items for resources that are present in your Config aggregator. The operation also returns a list of resources that are not processed in the current request. If there are no unprocessed resources, the operation returns an empty \code{unprocessedResourceIdentifiers} list.

See \url{https://www.paws-r-sdk.com/docs/configservice_batch_get_aggregate_resource_config/} for full documentation.
}
\keyword{internal}
