% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_configuration_aggregator}
\alias{configservice_delete_configuration_aggregator}
\title{Deletes the specified configuration aggregator and the aggregated data
associated with the aggregator}
\usage{
configservice_delete_configuration_aggregator(
  ConfigurationAggregatorName)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}
}
\value{
An empty list.
}
\description{
Deletes the specified configuration aggregator and the aggregated data
associated with the aggregator.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_configuration_aggregator(
  ConfigurationAggregatorName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
