% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_conformance_pack}
\alias{configservice_delete_conformance_pack}
\title{Deletes the specified conformance pack and all the Config rules,
remediation actions, and all evaluation results within that conformance
pack}
\usage{
configservice_delete_conformance_pack(ConformancePackName)
}
\arguments{
\item{ConformancePackName}{[required] Name of the conformance pack you want to delete.}
}
\description{
Deletes the specified conformance pack and all the Config rules, remediation actions, and all evaluation results within that conformance pack.

See \url{https://paws-r.github.io/docs/configservice/delete_conformance_pack.html} for full documentation.
}
\keyword{internal}
