% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_connector}
\alias{configservice_delete_connector}
\title{Deletes the specified connector}
\usage{
configservice_delete_connector(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the connector that you want to delete.}
}
\description{
Deletes the specified connector.

See \url{https://www.paws-r-sdk.com/docs/configservice_delete_connector/} for full documentation.
}
\keyword{internal}
