% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_evaluation_results}
\alias{configservice_delete_evaluation_results}
\title{Deletes the evaluation results for the specified AWS Config rule}
\usage{
configservice_delete_evaluation_results(ConfigRuleName)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the AWS Config rule for which you want to delete the
evaluation results.}
}
\value{
An empty list.
}
\description{
Deletes the evaluation results for the specified AWS Config rule. You
can specify one AWS Config rule per request. After you delete the
evaluation results, you can call the
\code{\link[=configservice_start_config_rules_evaluation]{start_config_rules_evaluation}}
API to start evaluating your AWS resources against the rule.
}
\section{Request syntax}{
\preformatted{svc$delete_evaluation_results(
  ConfigRuleName = "string"
)
}
}

\keyword{internal}
