% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_organization_config_rule}
\alias{configservice_delete_organization_config_rule}
\title{Deletes the specified organization config rule and all of its evaluation
results from all member accounts in that organization}
\usage{
configservice_delete_organization_config_rule(
  OrganizationConfigRuleName)
}
\arguments{
\item{OrganizationConfigRuleName}{[required] The name of organization config rule that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified organization config rule and all of its evaluation
results from all member accounts in that organization.

Only a master account and a delegated administrator account can delete
an organization config rule. When calling this API with a delegated
administrator, you must ensure AWS Organizations
\code{ListDelegatedAdministrator} permissions are added.

AWS Config sets the state of a rule to DELETE_IN_PROGRESS until the
deletion is complete. You cannot update a rule while it is in this
state.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_organization_config_rule(
  OrganizationConfigRuleName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
