% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_pending_aggregation_request}
\alias{configservice_delete_pending_aggregation_request}
\title{Deletes pending authorization requests for a specified aggregator
account in a specified region}
\usage{
configservice_delete_pending_aggregation_request(RequesterAccountId,
  RequesterAwsRegion)
}
\arguments{
\item{RequesterAccountId}{[required] The 12-digit account ID of the account requesting to aggregate data.}

\item{RequesterAwsRegion}{[required] The region requesting to aggregate data.}
}
\description{
Deletes pending authorization requests for a specified aggregator
account in a specified region.
}
\section{Request syntax}{
\preformatted{svc$delete_pending_aggregation_request(
  RequesterAccountId = "string",
  RequesterAwsRegion = "string"
)
}
}

\keyword{internal}
