% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_resource_config}
\alias{configservice_delete_resource_config}
\title{Records the configuration state for a custom resource that has been
deleted}
\usage{
configservice_delete_resource_config(ResourceType, ResourceId)
}
\arguments{
\item{ResourceType}{[required] The type of the resource.}

\item{ResourceId}{[required] Unique identifier of the resource.}
}
\description{
Records the configuration state for a custom resource that has been deleted. This API records a new ConfigurationItem with a ResourceDeleted status. You can retrieve the ConfigurationItems recorded for this resource in your Config History.

See \url{https://paws-r.github.io/docs/configservice/delete_resource_config.html} for full documentation.
}
\keyword{internal}
