% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_stored_query}
\alias{configservice_delete_stored_query}
\title{Deletes the stored query for a single AWS account and a single AWS
Region}
\usage{
configservice_delete_stored_query(QueryName)
}
\arguments{
\item{QueryName}{[required] The name of the query that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes the stored query for a single AWS account and a single AWS
Region.
}
\section{Request syntax}{
\preformatted{svc$delete_stored_query(
  QueryName = "string"
)
}
}

\keyword{internal}
