% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_deliver_config_snapshot}
\alias{configservice_deliver_config_snapshot}
\title{Schedules delivery of a configuration snapshot to the Amazon S3 bucket
in the specified delivery channel}
\usage{
configservice_deliver_config_snapshot(deliveryChannelName)
}
\arguments{
\item{deliveryChannelName}{[required] The name of the delivery channel through which the snapshot is
delivered.}
}
\description{
Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel. After the delivery has started, Config sends the following notifications using an Amazon SNS topic that you have specified.

See \url{https://www.paws-r-sdk.com/docs/configservice_deliver_config_snapshot/} for full documentation.
}
\keyword{internal}
