% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_organization_conformance_pack_statuses}
\alias{configservice_describe_organization_conformance_pack_statuses}
\title{Provides organization conformance pack deployment status for an
organization}
\usage{
configservice_describe_organization_conformance_pack_statuses(
  OrganizationConformancePackNames = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OrganizationConformancePackNames}{The names of organization conformance packs for which you want status
details. If you do not specify any names, Config returns details for all
your organization conformance packs.}

\item{Limit}{The maximum number of OrganizationConformancePackStatuses returned on
each page. If you do no specify a number, Config uses the default. The
default is 100.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}
}
\description{
Provides organization conformance pack deployment status for an organization.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_organization_conformance_pack_statuses/} for full documentation.
}
\keyword{internal}
