% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_aggregate_compliance_by_config_rules}
\alias{configservice_describe_aggregate_compliance_by_config_rules}
\title{Returns a list of compliant and noncompliant rules with the number of
resources for compliant and noncompliant rules}
\usage{
configservice_describe_aggregate_compliance_by_config_rules(
  ConfigurationAggregatorName, Filters, Limit, NextToken)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{Filters}{Filters the results by ConfigRuleComplianceFilters object.}

\item{Limit}{The maximum number of evaluation results returned on each page. The
default is maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns a list of compliant and noncompliant rules with the number of
resources for compliant and noncompliant rules.

The results can return an empty result page, but if you have a
\code{nextToken}, the results are displayed on the next page.
}
\section{Request syntax}{
\preformatted{svc$describe_aggregate_compliance_by_config_rules(
  ConfigurationAggregatorName = "string",
  Filters = list(
    ConfigRuleName = "string",
    ComplianceType = "COMPLIANT"|"NON_COMPLIANT"|"NOT_APPLICABLE"|"INSUFFICIENT_DATA",
    AccountId = "string",
    AwsRegion = "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
