% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_aggregate_compliance_by_conformance_packs}
\alias{configservice_describe_aggregate_compliance_by_conformance_packs}
\title{Returns a list of the conformance packs and their associated compliance
status with the count of compliant and noncompliant AWS Config rules
within each conformance pack}
\usage{
configservice_describe_aggregate_compliance_by_conformance_packs(
  ConfigurationAggregatorName, Filters, Limit, NextToken)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{Filters}{Filters the result by \code{AggregateConformancePackComplianceFilters}
object.}

\item{Limit}{The maximum number of conformance packs compliance details returned on
each page. The default is maximum. If you specify 0, AWS Config uses the
default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\value{
A list with the following syntax:\preformatted{list(
  AggregateComplianceByConformancePacks = list(
    list(
      ConformancePackName = "string",
      Compliance = list(
        ComplianceType = "COMPLIANT"|"NON_COMPLIANT"|"INSUFFICIENT_DATA",
        CompliantRuleCount = 123,
        NonCompliantRuleCount = 123,
        TotalRuleCount = 123
      ),
      AccountId = "string",
      AwsRegion = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of the conformance packs and their associated compliance
status with the count of compliant and noncompliant AWS Config rules
within each conformance pack. Also returns the total rule count which
includes compliant rules, noncompliant rules, and rules that cannot be
evaluated due to insufficient data.

The results can return an empty result page, but if you have a
\code{nextToken}, the results are displayed on the next page.
}
\section{Request syntax}{
\preformatted{svc$describe_aggregate_compliance_by_conformance_packs(
  ConfigurationAggregatorName = "string",
  Filters = list(
    ConformancePackName = "string",
    ComplianceType = "COMPLIANT"|"NON_COMPLIANT"|"INSUFFICIENT_DATA",
    AccountId = "string",
    AwsRegion = "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
