% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_compliance_by_config_rule}
\alias{configservice_describe_compliance_by_config_rule}
\title{Indicates whether the specified Config rules are compliant}
\usage{
configservice_describe_compliance_by_config_rule(
  ConfigRuleNames = NULL,
  ComplianceTypes = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ConfigRuleNames}{Specify one or more Config rule names to filter the results by rule.}

\item{ComplianceTypes}{Filters the results by compliance.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Indicates whether the specified Config rules are compliant. If a rule is noncompliant, this action returns the number of Amazon Web Services resources that do not comply with the rule.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_compliance_by_config_rule/} for full documentation.
}
\keyword{internal}
