% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_config_rules}
\alias{configservice_describe_config_rules}
\title{Returns details about your Config rules}
\usage{
configservice_describe_config_rules(
  ConfigRuleNames = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ConfigRuleNames}{The names of the Config rules for which you want details. If you do not specify any names, Config returns details for all your rules.}

\item{Filters}{Returns a list of Detective or Proactive Config rules. By default, this API returns an unfiltered list. For more information on Detective or Proactive Config rules, see \href{https://docs.aws.amazon.com/config/latest/developerguide/evaluate-config_components.html}{\strong{Evaluation Mode}} in the \emph{Config Developer Guide}.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns details about your Config rules.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_config_rules/} for full documentation.
}
\keyword{internal}
