% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_aggregators}
\alias{configservice_describe_configuration_aggregators}
\title{Returns the details of one or more configuration aggregators}
\usage{
configservice_describe_configuration_aggregators(
  ConfigurationAggregatorNames, NextToken, Limit)
}
\arguments{
\item{ConfigurationAggregatorNames}{The name of the configuration aggregators.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}

\item{Limit}{The maximum number of configuration aggregators returned on each page.
The default is maximum. If you specify 0, AWS Config uses the default.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ConfigurationAggregators = list(
    list(
      ConfigurationAggregatorName = "string",
      ConfigurationAggregatorArn = "string",
      AccountAggregationSources = list(
        list(
          AccountIds = list(
            "string"
          ),
          AllAwsRegions = TRUE|FALSE,
          AwsRegions = list(
            "string"
          )
        )
      ),
      OrganizationAggregationSource = list(
        RoleArn = "string",
        AwsRegions = list(
          "string"
        ),
        AllAwsRegions = TRUE|FALSE
      ),
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedTime = as.POSIXct(
        "2015-01-01"
      ),
      CreatedBy = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the details of one or more configuration aggregators. If the
configuration aggregator is not specified, this action returns the
details for all the configuration aggregators associated with the
account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_configuration_aggregators(
  ConfigurationAggregatorNames = list(
    "string"
  ),
  NextToken = "string",
  Limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
