% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_recorders}
\alias{configservice_describe_configuration_recorders}
\title{Returns details for the configuration recorder you specify}
\usage{
configservice_describe_configuration_recorders(
  ConfigurationRecorderNames = NULL,
  ServicePrincipal = NULL,
  Arn = NULL
)
}
\arguments{
\item{ConfigurationRecorderNames}{A list of names of the configuration recorders that you want to specify.

When making a request to this operation, you can only specify one configuration recorder.}

\item{ServicePrincipal}{For service-linked configuration recorders, you can use the service principal of the linked Amazon Web Services service to specify the configuration recorder. This field is only supported for Amazon Web Services service principals. For third-party service-linked configuration recorders, use \code{Arn} instead.}

\item{Arn}{The Amazon Resource Name (ARN) of the configuration recorder that you want to specify.}
}
\description{
Returns details for the configuration recorder you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_configuration_recorders/} for full documentation.
}
\keyword{internal}
