% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_conformance_pack_status}
\alias{configservice_describe_conformance_pack_status}
\title{Provides one or more conformance packs deployment status}
\usage{
configservice_describe_conformance_pack_status(ConformancePackNames,
  Limit, NextToken)
}
\arguments{
\item{ConformancePackNames}{Comma-separated list of conformance pack names.}

\item{Limit}{The maximum number of conformance packs status returned on each page.}

\item{NextToken}{The \code{nextToken} string returned in a previous request that you use to
request the next page of results in a paginated response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ConformancePackStatusDetails = list(
    list(
      ConformancePackName = "string",
      ConformancePackId = "string",
      ConformancePackArn = "string",
      ConformancePackState = "CREATE_IN_PROGRESS"|"CREATE_COMPLETE"|"CREATE_FAILED"|"DELETE_IN_PROGRESS"|"DELETE_FAILED",
      StackArn = "string",
      ConformancePackStatusReason = "string",
      LastUpdateRequestedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdateCompletedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Provides one or more conformance packs deployment status.

If there are no conformance packs then you will see an empty result.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_conformance_pack_status(
  ConformancePackNames = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
