% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_organization_conformance_pack_statuses}
\alias{configservice_describe_organization_conformance_pack_statuses}
\title{Provides organization conformance pack deployment status for an
organization}
\usage{
configservice_describe_organization_conformance_pack_statuses(
  OrganizationConformancePackNames, Limit, NextToken)
}
\arguments{
\item{OrganizationConformancePackNames}{The names of organization conformance packs for which you want status
details. If you do not specify any names, AWS Config returns details for
all your organization conformance packs.}

\item{Limit}{The maximum number of OrganizationConformancePackStatuses returned on
each page. If you do no specify a number, AWS Config uses the default.
The default is 100.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrganizationConformancePackStatuses = list(
    list(
      OrganizationConformancePackName = "string",
      Status = "CREATE_SUCCESSFUL"|"CREATE_IN_PROGRESS"|"CREATE_FAILED"|"DELETE_SUCCESSFUL"|"DELETE_FAILED"|"DELETE_IN_PROGRESS"|"UPDATE_SUCCESSFUL"|"UPDATE_IN_PROGRESS"|"UPDATE_FAILED",
      ErrorCode = "string",
      ErrorMessage = "string",
      LastUpdateTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Provides organization conformance pack deployment status for an
organization.

The status is not considered successful until organization conformance
pack is successfully deployed in all the member accounts with an
exception of excluded accounts.

When you specify the limit and the next token, you receive a paginated
response. Limit and next token are not applicable if you specify
organization conformance pack names. They are only applicable, when you
request all the organization conformance packs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_organization_conformance_pack_statuses(
  OrganizationConformancePackNames = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
