% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_organization_conformance_packs}
\alias{configservice_describe_organization_conformance_packs}
\title{Returns a list of organization conformance packs}
\usage{
configservice_describe_organization_conformance_packs(
  OrganizationConformancePackNames = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OrganizationConformancePackNames}{The name that you assign to an organization conformance pack.}

\item{Limit}{The maximum number of organization config packs returned on each page.
If you do no specify a number, Config uses the default. The default is
100.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}
}
\description{
Returns a list of organization conformance packs.

See \url{https://paws-r.github.io/docs/configservice/describe_organization_conformance_packs.html} for full documentation.
}
\keyword{internal}
