% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_pending_aggregation_requests}
\alias{configservice_describe_pending_aggregation_requests}
\title{Returns a list of all pending aggregation requests}
\usage{
configservice_describe_pending_aggregation_requests(Limit, NextToken)
}
\arguments{
\item{Limit}{The maximum number of evaluation results returned on each page. The
default is maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PendingAggregationRequests = list(
    list(
      RequesterAccountId = "string",
      RequesterAwsRegion = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of all pending aggregation requests.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_pending_aggregation_requests(
  Limit = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
