% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_remediation_configurations}
\alias{configservice_describe_remediation_configurations}
\title{Returns the details of one or more remediation configurations}
\usage{
configservice_describe_remediation_configurations(ConfigRuleNames)
}
\arguments{
\item{ConfigRuleNames}{[required] A list of Config rule names of remediation configurations for which you
want details.}
}
\description{
Returns the details of one or more remediation configurations.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_remediation_configurations/} for full documentation.
}
\keyword{internal}
