% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_remediation_exceptions}
\alias{configservice_describe_remediation_exceptions}
\title{Returns the details of one or more remediation exceptions}
\usage{
configservice_describe_remediation_exceptions(
  ConfigRuleName,
  ResourceKeys = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule.}

\item{ResourceKeys}{An exception list of resource exception keys to be processed with the
current request. Config adds exception for each resource key. For
example, Config adds 3 exceptions for 3 resource keys.}

\item{Limit}{The maximum number of RemediationExceptionResourceKey returned on each
page. The default is 25. If you specify 0, Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned in a previous request that you use to
request the next page of results in a paginated response.}
}
\description{
Returns the details of one or more remediation exceptions. A detailed view of a remediation exception for a set of resources that includes an explanation of an exception and the time when the exception will be deleted. When you specify the limit and the next token, you receive a paginated response.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_remediation_exceptions/} for full documentation.
}
\keyword{internal}
