% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_disassociate_resource_types}
\alias{configservice_disassociate_resource_types}
\title{Removes all resource types specified in the ResourceTypes list from the
RecordingGroup of configuration recorder and excludes these resource
types when recording}
\usage{
configservice_disassociate_resource_types(
  ConfigurationRecorderArn,
  ResourceTypes
)
}
\arguments{
\item{ConfigurationRecorderArn}{[required] The Amazon Resource Name (ARN) of the specified configuration recorder.}

\item{ResourceTypes}{[required] The list of resource types you want to remove from the recording group
of the specified configuration recorder.}
}
\description{
Removes all resource types specified in the \code{ResourceTypes} list from the \href{https://docs.aws.amazon.com/config/latest/APIReference/API_RecordingGroup.html}{RecordingGroup} of configuration recorder and excludes these resource types when recording.

See \url{https://www.paws-r-sdk.com/docs/configservice_disassociate_resource_types/} for full documentation.
}
\keyword{internal}
