% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_conformance_pack_compliance_summary}
\alias{configservice_get_conformance_pack_compliance_summary}
\title{Returns compliance details for the conformance pack based on the
cumulative compliance results of all the rules in that conformance pack}
\usage{
configservice_get_conformance_pack_compliance_summary(
  ConformancePackNames,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ConformancePackNames}{[required] Names of conformance packs.}

\item{Limit}{The maximum number of conformance packs returned on each page.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}
}
\description{
Returns compliance details for the conformance pack based on the cumulative compliance results of all the rules in that conformance pack.

See \url{https://www.paws-r-sdk.com/docs/configservice_get_conformance_pack_compliance_summary/} for full documentation.
}
\keyword{internal}
