% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_connector}
\alias{configservice_get_connector}
\title{Returns the details of the specified connector}
\usage{
configservice_get_connector(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the connector.}
}
\description{
Returns the details of the specified connector.

See \url{https://www.paws-r-sdk.com/docs/configservice_get_connector/} for full documentation.
}
\keyword{internal}
