% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_custom_rule_policy}
\alias{configservice_get_custom_rule_policy}
\title{Returns the policy definition containing the logic for your Config
Custom Policy rule}
\usage{
configservice_get_custom_rule_policy(ConfigRuleName = NULL)
}
\arguments{
\item{ConfigRuleName}{The name of your Config Custom Policy rule.}
}
\description{
Returns the policy definition containing the logic for your Config Custom Policy rule.

See \url{https://www.paws-r-sdk.com/docs/configservice_get_custom_rule_policy/} for full documentation.
}
\keyword{internal}
