% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_organization_config_rule_detailed_status}
\alias{configservice_get_organization_config_rule_detailed_status}
\title{Returns detailed status for each member account within an organization
for a given organization Config rule}
\usage{
configservice_get_organization_config_rule_detailed_status(
  OrganizationConfigRuleName,
  Filters = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OrganizationConfigRuleName}{[required] The name of your organization Config rule for which you want status
details for member accounts.}

\item{Filters}{A \code{StatusDetailFilters} object.}

\item{Limit}{The maximum number of \code{OrganizationConfigRuleDetailedStatus} returned on
each page. If you do not specify a number, Config uses the default. The
default is 100.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns detailed status for each member account within an organization for a given organization Config rule.

See \url{https://www.paws-r-sdk.com/docs/configservice_get_organization_config_rule_detailed_status/} for full documentation.
}
\keyword{internal}
