% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_resource_evaluation_summary}
\alias{configservice_get_resource_evaluation_summary}
\title{Returns a summary of resource evaluation for the specified resource
evaluation ID from the proactive rules that were run}
\usage{
configservice_get_resource_evaluation_summary(ResourceEvaluationId)
}
\arguments{
\item{ResourceEvaluationId}{[required] The unique \code{ResourceEvaluationId} of Amazon Web Services resource
execution for which you want to retrieve the evaluation summary.}
}
\description{
Returns a summary of resource evaluation for the specified resource evaluation ID from the proactive rules that were run. The results indicate which evaluation context was used to evaluate the rules, which resource details were evaluated, the evaluation mode that was run, and whether the resource details comply with the configuration of the proactive rules.

See \url{https://www.paws-r-sdk.com/docs/configservice_get_resource_evaluation_summary/} for full documentation.
}
\keyword{internal}
