% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_list_configuration_recorders}
\alias{configservice_list_configuration_recorders}
\title{Returns a list of configuration recorders depending on the filters you
specify}
\usage{
configservice_list_configuration_recorders(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Filters the results based on a list of \code{ConfigurationRecorderFilter}
objects that you specify.}

\item{MaxResults}{The maximum number of results to include in the response.}

\item{NextToken}{The \code{NextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns a list of configuration recorders depending on the filters you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_list_configuration_recorders/} for full documentation.
}
\keyword{internal}
