% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_list_connectors}
\alias{configservice_list_connectors}
\title{Returns a list of connectors depending on the filters you specify}
\usage{
configservice_list_connectors(
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response.}

\item{NextToken}{The \code{NextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}

\item{Filters}{Filters the results based on a list of \code{ConnectorFilter} objects that you specify.}
}
\description{
Returns a list of connectors depending on the filters you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_list_connectors/} for full documentation.
}
\keyword{internal}
