% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_list_resource_evaluations}
\alias{configservice_list_resource_evaluations}
\title{Returns a list of proactive resource evaluations}
\usage{
configservice_list_resource_evaluations(
  Filters = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Returns a \code{ResourceEvaluationFilters} object.}

\item{Limit}{The maximum number of evaluations returned on each page. The default is
10. You cannot specify a number greater than 100. If you specify 0,
Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns a list of proactive resource evaluations.

See \url{https://paws-r.github.io/docs/configservice/list_resource_evaluations.html} for full documentation.
}
\keyword{internal}
