% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_list_stored_queries}
\alias{configservice_list_stored_queries}
\title{Lists the stored queries for a single Amazon Web Services account and a
single Amazon Web Services Region}
\usage{
configservice_list_stored_queries(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The nextToken string returned in a previous request that you use to
request the next page of results in a paginated response.}

\item{MaxResults}{The maximum number of results to be returned with a single call.}
}
\description{
Lists the stored queries for a single Amazon Web Services account and a single Amazon Web Services Region. The default is 100.

See \url{https://www.paws-r-sdk.com/docs/configservice_list_stored_queries/} for full documentation.
}
\keyword{internal}
