% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_list_tags_for_resource}
\alias{configservice_list_tags_for_resource}
\title{List the tags for Config resource}
\usage{
configservice_list_tags_for_resource(
  ResourceArn,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. The following resources are supported:
\itemize{
\item \code{ConfigurationRecorder}
\item \code{ConfigRule}
\item \code{OrganizationConfigRule}
\item \code{ConformancePack}
\item \code{OrganizationConformancePack}
\item \code{ConfigurationAggregator}
\item \code{AggregationAuthorization}
\item \code{StoredQuery}
\item \code{Connector}
}}

\item{Limit}{The maximum number of tags returned on each page. The limit maximum is 50. You cannot specify a number greater than 50. If you specify 0, Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
List the tags for Config resource.

See \url{https://www.paws-r-sdk.com/docs/configservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
