% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_config_rule}
\alias{configservice_put_config_rule}
\title{Adds or updates an Config rule to evaluate if your Amazon Web Services
resources comply with your desired configurations}
\usage{
configservice_put_config_rule(ConfigRule, Tags = NULL)
}
\arguments{
\item{ConfigRule}{[required] The rule that you want to add to your account.}

\item{Tags}{An array of tag object.}
}
\description{
Adds or updates an Config rule to evaluate if your Amazon Web Services resources comply with your desired configurations. For information on how many Config rules you can have per account, see \href{https://docs.aws.amazon.com/config/latest/developerguide/configlimits.html}{\strong{Service Limits}} in the \emph{Config Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_config_rule/} for full documentation.
}
\keyword{internal}
