% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_configuration_recorder}
\alias{configservice_put_configuration_recorder}
\title{Creates or updates the customer managed configuration recorder}
\usage{
configservice_put_configuration_recorder(ConfigurationRecorder, Tags = NULL)
}
\arguments{
\item{ConfigurationRecorder}{[required] An object for the configuration recorder. A configuration recorder
records configuration changes for the resource types in scope.}

\item{Tags}{The tags for the customer managed configuration recorder. Each tag
consists of a key and an optional value, both of which you define.}
}
\description{
Creates or updates the customer managed configuration recorder.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_configuration_recorder/} for full documentation.
}
\keyword{internal}
