% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_connector}
\alias{configservice_put_connector}
\title{Creates a connector that specifies the connection between a third-party
cloud service provider and Config}
\usage{
configservice_put_connector(ConnectorConfiguration, Tags = NULL)
}
\arguments{
\item{ConnectorConfiguration}{[required] The provider-specific configuration for connecting to the third-party cloud service provider.}

\item{Tags}{The tags for the connector. Each tag consists of a key and an optional value, both of which you define.}
}
\description{
Creates a connector that specifies the connection between a third-party cloud service provider and Config.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_connector/} for full documentation.
}
\keyword{internal}
