% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_external_evaluation}
\alias{configservice_put_external_evaluation}
\title{Add or updates the evaluations for process checks}
\usage{
configservice_put_external_evaluation(ConfigRuleName, ExternalEvaluation)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule.}

\item{ExternalEvaluation}{[required] An \code{ExternalEvaluation} object that provides details about compliance.}
}
\description{
Add or updates the evaluations for process checks. This API checks if the rule is a process check when the name of the Config rule is provided.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_external_evaluation/} for full documentation.
}
\keyword{internal}
