% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_remediation_exceptions}
\alias{configservice_put_remediation_exceptions}
\title{A remediation exception is when a specific resource is no longer
considered for auto-remediation}
\usage{
configservice_put_remediation_exceptions(
  ConfigRuleName,
  ResourceKeys,
  Message = NULL,
  ExpirationTime = NULL
)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule for which you want to create remediation
exception.}

\item{ResourceKeys}{[required] An exception list of resource exception keys to be processed with the
current request. Config adds exception for each resource key. For
example, Config adds 3 exceptions for 3 resource keys.}

\item{Message}{The message contains an explanation of the exception.}

\item{ExpirationTime}{The exception is automatically deleted after the expiration date.}
}
\description{
A remediation exception is when a specific resource is no longer considered for auto-remediation. This API adds a new exception or updates an existing exception for a specific resource with a specific Config rule.

See \url{https://paws-r.github.io/docs/configservice/put_remediation_exceptions.html} for full documentation.
}
\keyword{internal}
