% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_service_linked_configuration_recorder}
\alias{configservice_put_service_linked_configuration_recorder}
\title{Creates a service-linked configuration recorder that is linked to a
specific Amazon Web Services service based on the ServicePrincipal you
specify}
\usage{
configservice_put_service_linked_configuration_recorder(
  ServicePrincipal,
  Tags = NULL
)
}
\arguments{
\item{ServicePrincipal}{[required] The service principal of the Amazon Web Services service for the
service-linked configuration recorder that you want to create.}

\item{Tags}{The tags for a service-linked configuration recorder. Each tag consists
of a key and an optional value, both of which you define.}
}
\description{
Creates a service-linked configuration recorder that is linked to a specific Amazon Web Services service based on the \code{ServicePrincipal} you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_service_linked_configuration_recorder/} for full documentation.
}
\keyword{internal}
