% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_third_party_service_linked_configuration_recorder}
\alias{configservice_put_third_party_service_linked_configuration_recorder}
\title{Creates or updates a service-linked configuration recorder that is
linked to a third-party cloud service provider based on the ConnectorArn
you specify}
\usage{
configservice_put_third_party_service_linked_configuration_recorder(
  ServicePrincipal,
  ConnectorArn,
  ScopeConfiguration,
  Tags = NULL
)
}
\arguments{
\item{ServicePrincipal}{[required] The service principal of the Amazon Web Services service for the service-linked configuration recorder that you want to create.}

\item{ConnectorArn}{[required] The Amazon Resource Name (ARN) of the connector that specifies the connection between the third-party cloud service provider and Config. The specified connector must exist.}

\item{ScopeConfiguration}{[required] Specifies the scope of resources to record from the third-party cloud service provider.}

\item{Tags}{The tags for a service-linked configuration recorder. Each tag consists of a key and an optional value, both of which you define.}
}
\description{
Creates or updates a service-linked configuration recorder that is linked to a third-party cloud service provider based on the \code{ConnectorArn} you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_third_party_service_linked_configuration_recorder/} for full documentation.
}
\keyword{internal}
