% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_start_config_rules_evaluation}
\alias{configservice_start_config_rules_evaluation}
\title{Runs an on-demand evaluation for the specified Config rules against the
last known configuration state of the resources}
\usage{
configservice_start_config_rules_evaluation(ConfigRuleNames = NULL)
}
\arguments{
\item{ConfigRuleNames}{The list of names of Config rules that you want to run evaluations for.}
}
\description{
Runs an on-demand evaluation for the specified Config rules against the last known configuration state of the resources. Use \code{\link[=configservice_start_config_rules_evaluation]{start_config_rules_evaluation}} when you want to test that a rule you updated is working as expected. \code{\link[=configservice_start_config_rules_evaluation]{start_config_rules_evaluation}} does not re-record the latest configuration state for your resources. It re-runs an evaluation against the last known state of your resources.

See \url{https://www.paws-r-sdk.com/docs/configservice_start_config_rules_evaluation/} for full documentation.
}
\keyword{internal}
