% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_start_configuration_recorder}
\alias{configservice_start_configuration_recorder}
\title{Starts the customer managed configuration recorder}
\usage{
configservice_start_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{[required] The name of the customer managed configuration recorder that you want to
start.}
}
\description{
Starts the customer managed configuration recorder. The customer managed configuration recorder will begin recording configuration changes for the resource types you specify.

See \url{https://www.paws-r-sdk.com/docs/configservice_start_configuration_recorder/} for full documentation.
}
\keyword{internal}
