% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_start_remediation_execution}
\alias{configservice_start_remediation_execution}
\title{Runs an on-demand remediation for the specified Config rules against the
last known remediation configuration}
\usage{
configservice_start_remediation_execution(ConfigRuleName, ResourceKeys)
}
\arguments{
\item{ConfigRuleName}{[required] The list of names of Config rules that you want to run remediation
execution for.}

\item{ResourceKeys}{[required] A list of resource keys to be processed with the current request. Each
element in the list consists of the resource type and resource ID.}
}
\description{
Runs an on-demand remediation for the specified Config rules against the last known remediation configuration. It runs an execution against the current state of your resources. Remediation execution is asynchronous.

See \url{https://www.paws-r-sdk.com/docs/configservice_start_remediation_execution/} for full documentation.
}
\keyword{internal}
