% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_stop_configuration_recorder}
\alias{configservice_stop_configuration_recorder}
\title{Stops the customer managed configuration recorder}
\usage{
configservice_stop_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{[required] The name of the customer managed configuration recorder that you want to
stop.}
}
\description{
Stops the customer managed configuration recorder. The customer managed configuration recorder will stop recording configuration changes for the resource types you have specified.

See \url{https://www.paws-r-sdk.com/docs/configservice_stop_configuration_recorder/} for full documentation.
}
\keyword{internal}
