% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_delete_landing_zone}
\alias{controltower_delete_landing_zone}
\title{Decommissions a landing zone}
\usage{
controltower_delete_landing_zone(landingZoneIdentifier)
}
\arguments{
\item{landingZoneIdentifier}{[required] The unique identifier of the landing zone.}
}
\description{
Decommissions a landing zone. This API call starts an asynchronous operation that deletes Amazon Web Services Control Tower resources deployed in accounts managed by Amazon Web Services Control Tower.

See \url{https://www.paws-r-sdk.com/docs/controltower_delete_landing_zone/} for full documentation.
}
\keyword{internal}
