% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_baseline}
\alias{controltower_get_baseline}
\title{Retrieve details about an existing Baseline resource by specifying its
identifier}
\usage{
controltower_get_baseline(baselineIdentifier)
}
\arguments{
\item{baselineIdentifier}{[required] The ARN of the \code{Baseline} resource to be retrieved.}
}
\description{
Retrieve details about an existing \code{Baseline} resource by specifying its identifier. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/baseline-api-examples.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_get_baseline/} for full documentation.
}
\keyword{internal}
