% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_control_operation}
\alias{controltower_get_control_operation}
\title{Returns the status of a particular EnableControl or DisableControl
operation}
\usage{
controltower_get_control_operation(operationIdentifier)
}
\arguments{
\item{operationIdentifier}{[required] The ID of the asynchronous operation, which is used to track status. The
operation is available for 90 days.}
}
\description{
Returns the status of a particular \code{\link[=controltower_enable_control]{enable_control}} or \code{\link[=controltower_disable_control]{disable_control}} operation. Displays a message in case of error. Details for an operation are available for 90 days. For usage examples, see the \href{https://docs.aws.amazon.com/controltower/latest/controlreference/control-api-examples-short.html}{\emph{Controls Reference Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_get_control_operation/} for full documentation.
}
\keyword{internal}
