% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_enabled_baseline}
\alias{controltower_get_enabled_baseline}
\title{Retrieve details of an EnabledBaseline resource by specifying its
identifier}
\usage{
controltower_get_enabled_baseline(enabledBaselineIdentifier)
}
\arguments{
\item{enabledBaselineIdentifier}{[required] Identifier of the \code{EnabledBaseline} resource to be retrieved, in ARN
format.}
}
\description{
Retrieve details of an \code{EnabledBaseline} resource by specifying its identifier.

See \url{https://www.paws-r-sdk.com/docs/controltower_get_enabled_baseline/} for full documentation.
}
\keyword{internal}
