% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_enabled_control}
\alias{controltower_get_enabled_control}
\title{Retrieves details about an enabled control}
\usage{
controltower_get_enabled_control(enabledControlIdentifier)
}
\arguments{
\item{enabledControlIdentifier}{[required] The \code{controlIdentifier} of the enabled control.}
}
\description{
Retrieves details about an enabled control. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-api-examples-short.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_get_enabled_control/} for full documentation.
}
\keyword{internal}
