% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_landing_zone}
\alias{controltower_get_landing_zone}
\title{Returns details about the landing zone}
\usage{
controltower_get_landing_zone(landingZoneIdentifier)
}
\arguments{
\item{landingZoneIdentifier}{[required] The unique identifier of the landing zone.}
}
\description{
Returns details about the landing zone. Displays a message in case of error.

See \url{https://www.paws-r-sdk.com/docs/controltower_get_landing_zone/} for full documentation.
}
\keyword{internal}
