% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_get_landing_zone_operation}
\alias{controltower_get_landing_zone_operation}
\title{Returns the status of the specified landing zone operation}
\usage{
controltower_get_landing_zone_operation(operationIdentifier)
}
\arguments{
\item{operationIdentifier}{[required] A unique identifier assigned to a landing zone operation.}
}
\description{
Returns the status of the specified landing zone operation. Details for an operation are available for 60 days.

See \url{https://www.paws-r-sdk.com/docs/controltower_get_landing_zone_operation/} for full documentation.
}
\keyword{internal}
