% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_baselines}
\alias{controltower_list_baselines}
\title{Returns a summary list of all available baselines}
\usage{
controltower_list_baselines(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to be shown.}

\item{nextToken}{A pagination token.}
}
\description{
Returns a summary list of all available baselines.

See \url{https://www.paws-r-sdk.com/docs/controltower_list_baselines/} for full documentation.
}
\keyword{internal}
