% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_control_operations}
\alias{controltower_list_control_operations}
\title{Provides a list of operations in progress or queued}
\usage{
controltower_list_control_operations(
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filter}{An input filter for the
\code{\link[=controltower_list_control_operations]{list_control_operations}} API
that lets you select the types of control operations to view.}

\item{nextToken}{A pagination token.}

\item{maxResults}{The maximum number of results to be shown.}
}
\description{
Provides a list of operations in progress or queued. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/controlreference/control-api-examples-short.html#list-control-operations-api-examples}{ListControlOperation examples}.

See \url{https://www.paws-r-sdk.com/docs/controltower_list_control_operations/} for full documentation.
}
\keyword{internal}
