% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_tags_for_resource}
\alias{controltower_list_tags_for_resource}
\title{Returns a list of tags associated with the resource}
\usage{
controltower_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}
}
\description{
Returns a list of tags associated with the resource. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/controlreference/control-api-examples-short.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
