% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_reset_enabled_baseline}
\alias{controltower_reset_enabled_baseline}
\title{Re-enables an EnabledBaseline resource}
\usage{
controltower_reset_enabled_baseline(enabledBaselineIdentifier)
}
\arguments{
\item{enabledBaselineIdentifier}{[required] Specifies the ID of the \code{EnabledBaseline} resource to be re-enabled, in
ARN format.}
}
\description{
Re-enables an \code{EnabledBaseline} resource. For example, this API can re-apply the existing \code{Baseline} after a new member account is moved to the target OU. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/baseline-api-examples.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_reset_enabled_baseline/} for full documentation.
}
\keyword{internal}
