% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_reset_enabled_control}
\alias{controltower_reset_enabled_control}
\title{Resets an enabled control}
\usage{
controltower_reset_enabled_control(enabledControlIdentifier)
}
\arguments{
\item{enabledControlIdentifier}{[required] The ARN of the enabled control to be reset.}
}
\description{
Resets an enabled control. Does not work for controls implemented with SCPs.

See \url{https://www.paws-r-sdk.com/docs/controltower_reset_enabled_control/} for full documentation.
}
\keyword{internal}
