% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_reset_landing_zone}
\alias{controltower_reset_landing_zone}
\title{This API call resets a landing zone}
\usage{
controltower_reset_landing_zone(landingZoneIdentifier)
}
\arguments{
\item{landingZoneIdentifier}{[required] The unique identifier of the landing zone.}
}
\description{
This API call resets a landing zone. It starts an asynchronous operation that resets the landing zone to the parameters specified in its original configuration.

See \url{https://www.paws-r-sdk.com/docs/controltower_reset_landing_zone/} for full documentation.
}
\keyword{internal}
