% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_tag_resource}
\alias{controltower_tag_resource}
\title{Applies tags to a resource}
\usage{
controltower_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be tagged.}

\item{tags}{[required] Tags to be applied to the resource.}
}
\description{
Applies tags to a resource. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-api-examples-short.html}{\emph{the AWS Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_tag_resource/} for full documentation.
}
\keyword{internal}
