% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_untag_resource}
\alias{controltower_untag_resource}
\title{Removes tags from a resource}
\usage{
controltower_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] Tag keys to be removed from the resource.}
}
\description{
Removes tags from a resource. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/control-api-examples-short.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_untag_resource/} for full documentation.
}
\keyword{internal}
