% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_update_enabled_baseline}
\alias{controltower_update_enabled_baseline}
\title{Updates an EnabledBaseline resource's applied parameters or version}
\usage{
controltower_update_enabled_baseline(
  baselineVersion,
  parameters = NULL,
  enabledBaselineIdentifier
)
}
\arguments{
\item{baselineVersion}{[required] Specifies the new \code{Baseline} version, to which the \code{EnabledBaseline}
should be updated.}

\item{parameters}{Parameters to apply when making an update.}

\item{enabledBaselineIdentifier}{[required] Specifies the \code{EnabledBaseline} resource to be updated.}
}
\description{
Updates an \code{EnabledBaseline} resource's applied parameters or version. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/baseline-api-examples.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_update_enabled_baseline/} for full documentation.
}
\keyword{internal}
