% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_update_enabled_control}
\alias{controltower_update_enabled_control}
\title{Updates the configuration of an already enabled control}
\usage{
controltower_update_enabled_control(parameters, enabledControlIdentifier)
}
\arguments{
\item{parameters}{[required] A key/value pair, where \code{Key} is of type \code{String} and \code{Value} is of type
\code{Document}.}

\item{enabledControlIdentifier}{[required] The ARN of the enabled control that will be updated.}
}
\description{
Updates the configuration of an already enabled control.

See \url{https://www.paws-r-sdk.com/docs/controltower_update_enabled_control/} for full documentation.
}
\keyword{internal}
