% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_update_landing_zone}
\alias{controltower_update_landing_zone}
\title{This API call updates the landing zone}
\usage{
controltower_update_landing_zone(landingZoneIdentifier, manifest, version)
}
\arguments{
\item{landingZoneIdentifier}{[required] The unique identifier of the landing zone.}

\item{manifest}{[required] The \code{manifest.yaml} file is a text file that describes your Amazon Web
Services resources. For examples, review \href{https://docs.aws.amazon.com/controltower/latest/userguide/the-manifest-file.html}{The manifest file}.}

\item{version}{[required] The landing zone version, for example, 3.2.}
}
\description{
This API call updates the landing zone. It starts an asynchronous operation that updates the landing zone based on the new landing zone version, or on the changed parameters specified in the updated manifest file.

See \url{https://www.paws-r-sdk.com/docs/controltower_update_landing_zone/} for full documentation.
}
\keyword{internal}
