% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_create_kx_database}
\alias{finspace_create_kx_database}
\title{Creates a new kdb database in the environment}
\usage{
finspace_create_kx_database(
  environmentId,
  databaseName,
  description = NULL,
  tags = NULL,
  clientToken
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{databaseName}{[required] The name of the kdb database.}

\item{description}{A description of the database.}

\item{tags}{A list of key-value pairs to label the kdb database. You can add up to
50 tags to your kdb database}

\item{clientToken}{[required] A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Creates a new kdb database in the environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_create_kx_database/} for full documentation.
}
\keyword{internal}
