% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_delete_kx_cluster}
\alias{finspace_delete_kx_cluster}
\title{Deletes a kdb cluster}
\usage{
finspace_delete_kx_cluster(environmentId, clusterName, clientToken = NULL)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{clusterName}{[required] The name of the cluster that you want to delete.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Deletes a kdb cluster.

See \url{https://www.paws-r-sdk.com/docs/finspace_delete_kx_cluster/} for full documentation.
}
\keyword{internal}
