% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_delete_kx_dataview}
\alias{finspace_delete_kx_dataview}
\title{Deletes the specified dataview}
\usage{
finspace_delete_kx_dataview(
  environmentId,
  databaseName,
  dataviewName,
  clientToken
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment, from where you want to
delete the dataview.}

\item{databaseName}{[required] The name of the database whose dataview you want to delete.}

\item{dataviewName}{[required] The name of the dataview that you want to delete.}

\item{clientToken}{[required] A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Deletes the specified dataview. Before deleting a dataview, make sure that it is not in use by any cluster.

See \url{https://www.paws-r-sdk.com/docs/finspace_delete_kx_dataview/} for full documentation.
}
\keyword{internal}
