% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_environment}
\alias{finspace_get_environment}
\title{Returns the FinSpace environment object}
\usage{
finspace_get_environment(environmentId)
}
\arguments{
\item{environmentId}{[required] The identifier of the FinSpace environment.}
}
\description{
Returns the FinSpace environment object.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_environment/} for full documentation.
}
\keyword{internal}
