% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_changeset}
\alias{finspace_get_kx_changeset}
\title{Returns information about a kdb changeset}
\usage{
finspace_get_kx_changeset(environmentId, databaseName, changesetId)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{databaseName}{[required] The name of the kdb database.}

\item{changesetId}{[required] A unique identifier of the changeset for which you want to retrieve
data.}
}
\description{
Returns information about a kdb changeset.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_changeset/} for full documentation.
}
\keyword{internal}
