% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_cluster}
\alias{finspace_get_kx_cluster}
\title{Retrieves information about a kdb cluster}
\usage{
finspace_get_kx_cluster(environmentId, clusterName)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{clusterName}{[required] The name of the cluster that you want to retrieve.}
}
\description{
Retrieves information about a kdb cluster.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_cluster/} for full documentation.
}
\keyword{internal}
