% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_database}
\alias{finspace_get_kx_database}
\title{Returns database information for the specified environment ID}
\usage{
finspace_get_kx_database(environmentId, databaseName)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{databaseName}{[required] The name of the kdb database.}
}
\description{
Returns database information for the specified environment ID.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_database/} for full documentation.
}
\keyword{internal}
