% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_dataview}
\alias{finspace_get_kx_dataview}
\title{Retrieves details of the dataview}
\usage{
finspace_get_kx_dataview(environmentId, databaseName, dataviewName)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment, from where you want to
retrieve the dataview details.}

\item{databaseName}{[required] The name of the database where you created the dataview.}

\item{dataviewName}{[required] A unique identifier for the dataview.}
}
\description{
Retrieves details of the dataview.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_dataview/} for full documentation.
}
\keyword{internal}
