% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_environment}
\alias{finspace_get_kx_environment}
\title{Retrieves all the information for the specified kdb environment}
\usage{
finspace_get_kx_environment(environmentId)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}
}
\description{
Retrieves all the information for the specified kdb environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_environment/} for full documentation.
}
\keyword{internal}
