% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_scaling_group}
\alias{finspace_get_kx_scaling_group}
\title{Retrieves details of a scaling group}
\usage{
finspace_get_kx_scaling_group(environmentId, scalingGroupName)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{scalingGroupName}{[required] A unique identifier for the kdb scaling group.}
}
\description{
Retrieves details of a scaling group.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_scaling_group/} for full documentation.
}
\keyword{internal}
