% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_user}
\alias{finspace_get_kx_user}
\title{Retrieves information about the specified kdb user}
\usage{
finspace_get_kx_user(userName, environmentId)
}
\arguments{
\item{userName}{[required] A unique identifier for the user.}

\item{environmentId}{[required] A unique identifier for the kdb environment.}
}
\description{
Retrieves information about the specified kdb user.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_user/} for full documentation.
}
\keyword{internal}
