% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_get_kx_volume}
\alias{finspace_get_kx_volume}
\title{Retrieves the information about the volume}
\usage{
finspace_get_kx_volume(environmentId, volumeName)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment, whose clusters can attach
to the volume.}

\item{volumeName}{[required] A unique identifier for the volume.}
}
\description{
Retrieves the information about the volume.

See \url{https://www.paws-r-sdk.com/docs/finspace_get_kx_volume/} for full documentation.
}
\keyword{internal}
