% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_list_kx_changesets}
\alias{finspace_list_kx_changesets}
\title{Returns a list of all the changesets for a database}
\usage{
finspace_list_kx_changesets(
  environmentId,
  databaseName,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{databaseName}{[required] The name of the kdb database.}

\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{The maximum number of results to return in this request.}
}
\description{
Returns a list of all the changesets for a database.

See \url{https://www.paws-r-sdk.com/docs/finspace_list_kx_changesets/} for full documentation.
}
\keyword{internal}
