% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_list_kx_environments}
\alias{finspace_list_kx_environments}
\title{Returns a list of kdb environments created in an account}
\usage{
finspace_list_kx_environments(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{A token that indicates where a results page should begin.}

\item{maxResults}{The maximum number of results to return in this request.}
}
\description{
Returns a list of kdb environments created in an account.

See \url{https://www.paws-r-sdk.com/docs/finspace_list_kx_environments/} for full documentation.
}
\keyword{internal}
