% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_list_kx_scaling_groups}
\alias{finspace_list_kx_scaling_groups}
\title{Returns a list of scaling groups in a kdb environment}
\usage{
finspace_list_kx_scaling_groups(
  environmentId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment, for which you want to
retrieve a list of scaling groups.}

\item{maxResults}{The maximum number of results to return in this request.}

\item{nextToken}{A token that indicates where a results page should begin.}
}
\description{
Returns a list of scaling groups in a kdb environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_list_kx_scaling_groups/} for full documentation.
}
\keyword{internal}
