% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_list_kx_volumes}
\alias{finspace_list_kx_volumes}
\title{Lists all the volumes in a kdb environment}
\usage{
finspace_list_kx_volumes(
  environmentId,
  maxResults = NULL,
  nextToken = NULL,
  volumeType = NULL
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment, whose clusters can attach
to the volume.}

\item{maxResults}{The maximum number of results to return in this request.}

\item{nextToken}{A token that indicates where a results page should begin.}

\item{volumeType}{The type of file system volume. Currently, FinSpace only supports
\code{NAS_1} volume type.}
}
\description{
Lists all the volumes in a kdb environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_list_kx_volumes/} for full documentation.
}
\keyword{internal}
