% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_untag_resource}
\alias{finspace_untag_resource}
\title{Removes metadata tags from a FinSpace resource}
\usage{
finspace_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] A FinSpace resource from which you want to remove a tag or tags. The
value for this parameter is an Amazon Resource Name (ARN).}

\item{tagKeys}{[required] The tag keys (names) of one or more tags to be removed.}
}
\description{
Removes metadata tags from a FinSpace resource.

See \url{https://www.paws-r-sdk.com/docs/finspace_untag_resource/} for full documentation.
}
\keyword{internal}
