% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_update_kx_database}
\alias{finspace_update_kx_database}
\title{Updates information for the given kdb database}
\usage{
finspace_update_kx_database(
  environmentId,
  databaseName,
  description = NULL,
  clientToken
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{databaseName}{[required] The name of the kdb database.}

\item{description}{A description of the database.}

\item{clientToken}{[required] A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Updates information for the given kdb database.

See \url{https://www.paws-r-sdk.com/docs/finspace_update_kx_database/} for full documentation.
}
\keyword{internal}
