% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_update_kx_environment}
\alias{finspace_update_kx_environment}
\title{Updates information for the given kdb environment}
\usage{
finspace_update_kx_environment(
  environmentId,
  name = NULL,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{name}{The name of the kdb environment.}

\item{description}{A description of the kdb environment.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Updates information for the given kdb environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_update_kx_environment/} for full documentation.
}
\keyword{internal}
