% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_update_kx_user}
\alias{finspace_update_kx_user}
\title{Updates the user details}
\usage{
finspace_update_kx_user(environmentId, userName, iamRole, clientToken = NULL)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment.}

\item{userName}{[required] A unique identifier for the user.}

\item{iamRole}{[required] The IAM role ARN that is associated with the user.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}
}
\description{
Updates the user details. You can only update the IAM role associated with a user.

See \url{https://www.paws-r-sdk.com/docs/finspace_update_kx_user/} for full documentation.
}
\keyword{internal}
