% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_update_kx_volume}
\alias{finspace_update_kx_volume}
\title{Updates the throughput or capacity of a volume}
\usage{
finspace_update_kx_volume(
  environmentId,
  volumeName,
  description = NULL,
  clientToken = NULL,
  nas1Configuration = NULL
)
}
\arguments{
\item{environmentId}{[required] A unique identifier for the kdb environment where you created the
storage volume.}

\item{volumeName}{[required] A unique identifier for the volume.}

\item{description}{A description of the volume.}

\item{clientToken}{A token that ensures idempotency. This token expires in 10 minutes.}

\item{nas1Configuration}{Specifies the configuration for the Network attached storage (NAS_1)
file system volume.}
}
\description{
Updates the throughput or capacity of a volume. During the update process, the filesystem might be unavailable for a few minutes. You can retry any operations after the update is complete.

See \url{https://www.paws-r-sdk.com/docs/finspace_update_kx_volume/} for full documentation.
}
\keyword{internal}
