% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_entity_aggregates_for_organization}
\alias{health_describe_entity_aggregates_for_organization}
\title{Returns a list of entity aggregates for your Organizations that are
affected by each of the specified events}
\usage{
health_describe_entity_aggregates_for_organization(
  eventArns,
  awsAccountIds = NULL
)
}
\arguments{
\item{eventArns}{[required] A list of event ARNs (unique identifiers). For example:
\verb{"arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"}}

\item{awsAccountIds}{A list of 12-digit Amazon Web Services account numbers that contains the
affected entities.}
}
\description{
Returns a list of entity aggregates for your Organizations that are affected by each of the specified events.

See \url{https://www.paws-r-sdk.com/docs/health_describe_entity_aggregates_for_organization/} for full documentation.
}
\keyword{internal}
