% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_event_types}
\alias{health_describe_event_types}
\title{Returns the event types that meet the specified filter criteria}
\usage{
health_describe_event_types(
  filter = NULL,
  locale = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filter}{Values to narrow the results returned.}

\item{locale}{The locale (language) to return information in. English (en) is the
default and the only supported value at this time.}

\item{nextToken}{If the results of a search are large, only a portion of the results are
returned, and a \code{nextToken} pagination token is returned in the
response. To retrieve the next batch of results, reissue the search
request and include the returned token. When all results have been
returned, the response does not contain a pagination token value.}

\item{maxResults}{The maximum number of items to return in one batch, between 10 and 100,
inclusive.

If you don't specify the \code{maxResults} parameter, this operation returns
a maximum of 30 items by default.}
}
\description{
Returns the event types that meet the specified filter criteria. You can use this API operation to find information about the Health event, such as the category, Amazon Web Service, and event code. The metadata for each event appears in the \href{https://docs.aws.amazon.com/health/latest/APIReference/API_EventType.html}{EventType} object.

See \url{https://paws-r.github.io/docs/health/describe_event_types.html} for full documentation.
}
\keyword{internal}
