% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_health_service_status_for_organization}
\alias{health_describe_health_service_status_for_organization}
\title{This operation provides status information on enabling or disabling
Health to work with your organization}
\usage{
health_describe_health_service_status_for_organization()
}
\description{
This operation provides status information on enabling or disabling Health to work with your organization. To call this operation, you must use the organization's management account.

See \url{https://paws-r.github.io/docs/health/describe_health_service_status_for_organization.html} for full documentation.
}
\keyword{internal}
