% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_service_lifecycle}
\alias{health_describe_service_lifecycle}
\title{Returns lifecycle information for Amazon Web Services services,
including end-of-life dates, version recommendations, and lifecycle
events}
\usage{
health_describe_service_lifecycle(
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filter}{Values to narrow the results returned.}

\item{nextToken}{If the results of a search are large, only a portion of the results are returned, and a \code{nextToken} pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.}

\item{maxResults}{The maximum number of items to return in one batch, between 1 and 20, inclusive.}
}
\description{
Returns lifecycle information for Amazon Web Services services, including end-of-life dates, version recommendations, and lifecycle events.

See \url{https://www.paws-r-sdk.com/docs/health_describe_service_lifecycle/} for full documentation.
}
\keyword{internal}
