% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_accept_grant}
\alias{licensemanager_accept_grant}
\title{Accepts the specified grant}
\usage{
licensemanager_accept_grant(GrantArn)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}
}
\value{
A list with the following syntax:\preformatted{list(
  GrantArn = "string",
  Status = "PENDING_WORKFLOW"|"PENDING_ACCEPT"|"REJECTED"|"ACTIVE"|"FAILED_WORKFLOW"|"DELETED"|"PENDING_DELETE"|"DISABLED",
  Version = "string"
)
}
}
\description{
Accepts the specified grant.
}
\section{Request syntax}{
\preformatted{svc$accept_grant(
  GrantArn = "string"
)
}
}

\keyword{internal}
