% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_check_in_license}
\alias{licensemanager_check_in_license}
\title{Checks in the specified license}
\usage{
licensemanager_check_in_license(LicenseConsumptionToken, Beneficiary)
}
\arguments{
\item{LicenseConsumptionToken}{[required] License consumption token.}

\item{Beneficiary}{License beneficiary.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Checks in the specified license. Check in a license when it is no longer
in use.
}
\section{Request syntax}{
\preformatted{svc$check_in_license(
  LicenseConsumptionToken = "string",
  Beneficiary = "string"
)
}
}

\keyword{internal}
