% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_create_license_asset_group}
\alias{licensemanager_create_license_asset_group}
\title{Creates a license asset group}
\usage{
licensemanager_create_license_asset_group(
  Name,
  Description = NULL,
  LicenseAssetGroupConfigurations,
  AssociatedLicenseAssetRulesetARNs,
  Properties = NULL,
  Tags = NULL,
  ClientToken
)
}
\arguments{
\item{Name}{[required] License asset group name.}

\item{Description}{License asset group description.}

\item{LicenseAssetGroupConfigurations}{[required] License asset group configurations.}

\item{AssociatedLicenseAssetRulesetARNs}{[required] ARNs of associated license asset rulesets.}

\item{Properties}{License asset group properties.}

\item{Tags}{Tags to add to the license asset group.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Creates a license asset group.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_create_license_asset_group/} for full documentation.
}
\keyword{internal}
