% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_create_license_asset_ruleset}
\alias{licensemanager_create_license_asset_ruleset}
\title{Creates a license asset ruleset}
\usage{
licensemanager_create_license_asset_ruleset(
  Name,
  Description = NULL,
  Rules,
  Tags = NULL,
  ClientToken
)
}
\arguments{
\item{Name}{[required] License asset ruleset name.}

\item{Description}{License asset ruleset description.}

\item{Rules}{[required] License asset rules.}

\item{Tags}{Tags to add to the license asset ruleset.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Creates a license asset ruleset.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_create_license_asset_ruleset/} for full documentation.
}
\keyword{internal}
