% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_create_license_manager_report_generator}
\alias{licensemanager_create_license_manager_report_generator}
\title{Creates a report generator}
\usage{
licensemanager_create_license_manager_report_generator(
  ReportGeneratorName,
  Type,
  ReportContext,
  ReportFrequency,
  ClientToken,
  Description = NULL,
  Tags = NULL
)
}
\arguments{
\item{ReportGeneratorName}{[required] Name of the report generator.}

\item{Type}{[required] Type of reports to generate. The following report types an be generated:
\itemize{
\item License configuration report - Reports the number and details of
consumed licenses for a license configuration.
\item Resource report - Reports the tracked licenses and resource
consumption for a license configuration.
}}

\item{ReportContext}{[required] Defines the type of license configuration the report generator tracks.}

\item{ReportFrequency}{[required] Frequency by which reports are generated. Reports can be generated
daily, monthly, or weekly.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{Description}{Description of the report generator.}

\item{Tags}{Tags to add to the report generator.}
}
\description{
Creates a report generator.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_create_license_manager_report_generator/} for full documentation.
}
\keyword{internal}
