% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_license}
\alias{licensemanager_delete_license}
\title{Deletes the specified license}
\usage{
licensemanager_delete_license(LicenseArn, SourceVersion)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}

\item{SourceVersion}{[required] Current version of the license.}
}
\description{
Deletes the specified license.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_delete_license/} for full documentation.
}
\keyword{internal}
