% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_license_asset_group}
\alias{licensemanager_delete_license_asset_group}
\title{Deletes a license asset group}
\usage{
licensemanager_delete_license_asset_group(LicenseAssetGroupArn)
}
\arguments{
\item{LicenseAssetGroupArn}{[required] Amazon Resource Name (ARN) of the license asset group.}
}
\description{
Deletes a license asset group.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_delete_license_asset_group/} for full documentation.
}
\keyword{internal}
