% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_license_manager_report_generator}
\alias{licensemanager_delete_license_manager_report_generator}
\title{Deletes the specified report generator}
\usage{
licensemanager_delete_license_manager_report_generator(
  LicenseManagerReportGeneratorArn
)
}
\arguments{
\item{LicenseManagerReportGeneratorArn}{[required] Amazon Resource Name (ARN) of the report generator to be deleted.}
}
\description{
Deletes the specified report generator.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_delete_license_manager_report_generator/} for full documentation.
}
\keyword{internal}
