% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_access_token}
\alias{licensemanager_get_access_token}
\title{Gets a temporary access token to use with AssumeRoleWithWebIdentity}
\usage{
licensemanager_get_access_token(Token, TokenProperties = NULL)
}
\arguments{
\item{Token}{[required] Refresh token, encoded as a JWT token.}

\item{TokenProperties}{Token properties to validate against those present in the JWT token.}
}
\description{
Gets a temporary access token to use with AssumeRoleWithWebIdentity. Access tokens are valid for one hour.

See \url{https://paws-r.github.io/docs/licensemanager/get_access_token.html} for full documentation.
}
\keyword{internal}
