% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_asset_ruleset}
\alias{licensemanager_get_license_asset_ruleset}
\title{Gets a license asset ruleset}
\usage{
licensemanager_get_license_asset_ruleset(LicenseAssetRulesetArn)
}
\arguments{
\item{LicenseAssetRulesetArn}{[required] Amazon Resource Name (ARN) of the license asset ruleset.}
}
\description{
Gets a license asset ruleset.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_get_license_asset_ruleset/} for full documentation.
}
\keyword{internal}
