% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_service_settings}
\alias{licensemanager_get_service_settings}
\title{Gets the License Manager settings for the current Region}
\usage{
licensemanager_get_service_settings()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  S3BucketArn = "string",
  SnsTopicArn = "string",
  OrganizationConfiguration = list(
    EnableIntegration = TRUE|FALSE
  ),
  EnableCrossAccountsDiscovery = TRUE|FALSE,
  LicenseManagerResourceShareArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the License Manager settings for the current Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_service_settings()
}\if{html}{\out{</div>}}
}

\keyword{internal}
