% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_assets_for_license_asset_group}
\alias{licensemanager_list_assets_for_license_asset_group}
\title{Lists assets for a license asset group}
\usage{
licensemanager_list_assets_for_license_asset_group(
  LicenseAssetGroupArn,
  AssetType,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{LicenseAssetGroupArn}{[required] Amazon Resource Name (ARN) of the license asset group.}

\item{AssetType}{[required] Asset type. The possible values are \code{Instance} | \code{License} |
\code{LicenseConfiguration}.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists assets for a license asset group.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_assets_for_license_asset_group/} for full documentation.
}
\keyword{internal}
