% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_distributed_grants}
\alias{licensemanager_list_distributed_grants}
\title{Lists the grants distributed for the specified license}
\usage{
licensemanager_list_distributed_grants(GrantArns, Filters, NextToken,
  MaxResults)
}
\arguments{
\item{GrantArns}{Amazon Resource Names (ARNs) of the grants.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{LicenseARN}
\item \code{Status}
\item \code{PrincipalARN}
\item \code{ParentARN}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists the grants distributed for the specified license.
}
\section{Request syntax}{
\preformatted{svc$list_distributed_grants(
  GrantArns = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
