% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_license_asset_groups}
\alias{licensemanager_list_license_asset_groups}
\title{Lists license asset groups}
\usage{
licensemanager_list_license_asset_groups(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Filters to scope the results. Following filters are supported
\itemize{
\item \code{LicenseAssetRulesetArn}
}}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists license asset groups.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_license_asset_groups/} for full documentation.
}
\keyword{internal}
