% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_license_asset_rulesets}
\alias{licensemanager_list_license_asset_rulesets}
\title{Lists license asset rulesets}
\usage{
licensemanager_list_license_asset_rulesets(
  Filters = NULL,
  ShowAWSManagedLicenseAssetRulesets = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Filters to scope the results. Following filters are supported
\itemize{
\item \code{Name}
}}

\item{ShowAWSManagedLicenseAssetRulesets}{Specifies whether to show License Manager managed license asset
rulesets.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists license asset rulesets.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_license_asset_rulesets/} for full documentation.
}
\keyword{internal}
