% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_license_configurations_for_organization}
\alias{licensemanager_list_license_configurations_for_organization}
\title{Lists license configurations for an organization}
\usage{
licensemanager_list_license_configurations_for_organization(
  LicenseConfigurationArns = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{LicenseConfigurationArns}{License configuration ARNs.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}

\item{Filters}{Filters to scope the results.}
}
\description{
Lists license configurations for an organization.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_license_configurations_for_organization/} for full documentation.
}
\keyword{internal}
