% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_licenses}
\alias{licensemanager_list_licenses}
\title{Lists the licenses for your account}
\usage{
licensemanager_list_licenses(LicenseArns, Filters, NextToken,
  MaxResults)
}
\arguments{
\item{LicenseArns}{Amazon Resource Names (ARNs) of the licenses.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{Beneficiary}
\item \code{ProductSKU}
\item \code{KeyFingerprint}
\item \code{Status}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Licenses = list(
    list(
      LicenseArn = "string",
      LicenseName = "string",
      ProductName = "string",
      ProductSKU = "string",
      Issuer = list(
        Name = "string",
        SignKey = "string",
        KeyFingerprint = "string"
      ),
      HomeRegion = "string",
      Status = "AVAILABLE"|"PENDING_AVAILABLE"|"DEACTIVATED"|"SUSPENDED"|"EXPIRED"|"PENDING_DELETE"|"DELETED",
      Validity = list(
        Begin = "string",
        End = "string"
      ),
      Beneficiary = "string",
      Entitlements = list(
        list(
          Name = "string",
          Value = "string",
          MaxCount = 123,
          Overage = TRUE|FALSE,
          Unit = "Count"|"None"|"Seconds"|"Microseconds"|"Milliseconds"|"Bytes"|"Kilobytes"|"Megabytes"|"Gigabytes"|"Terabytes"|"Bits"|"Kilobits"|"Megabits"|"Gigabits"|"Terabits"|"Percent"|"Bytes/Second"|"Kilobytes/Second"|"Megabytes/Second"|"Gigabytes/Second"|"Terabytes/Second"|"Bits/Second"|"Kilobits/Second"|"Megabits/Second"|"Gigabits/Second"|"Terabits/Second"|"Count/Second",
          AllowCheckIn = TRUE|FALSE
        )
      ),
      ConsumptionConfiguration = list(
        RenewType = "None"|"Weekly"|"Monthly",
        ProvisionalConfiguration = list(
          MaxTimeToLiveInMinutes = 123
        ),
        BorrowConfiguration = list(
          AllowEarlyCheckIn = TRUE|FALSE,
          MaxTimeToLiveInMinutes = 123
        )
      ),
      LicenseMetadata = list(
        list(
          Name = "string",
          Value = "string"
        )
      ),
      CreateTime = "string",
      Version = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the licenses for your account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_licenses(
  LicenseArns = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
