% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_received_grants_for_organization}
\alias{licensemanager_list_received_grants_for_organization}
\title{Lists the grants received for all accounts in the organization}
\usage{
licensemanager_list_received_grants_for_organization(
  LicenseArn,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LicenseArn}{[required] The Amazon Resource Name (ARN) of the received license.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{ParentArn}
\item \code{GranteePrincipalArn}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists the grants received for all accounts in the organization.

See \url{https://paws-r.github.io/docs/licensemanager/list_received_grants_for_organization.html} for full documentation.
}
\keyword{internal}
