% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_reject_grant}
\alias{licensemanager_reject_grant}
\title{Rejects the specified grant}
\usage{
licensemanager_reject_grant(GrantArn)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}
}
\description{
Rejects the specified grant.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_reject_grant/} for full documentation.
}
\keyword{internal}
