% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_tag_resource}
\alias{licensemanager_tag_resource}
\title{Adds the specified tags to the specified license configuration}
\usage{
licensemanager_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the license configuration.}

\item{Tags}{[required] One or more tags.}
}
\description{
Adds the specified tags to the specified license configuration.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_tag_resource/} for full documentation.
}
\keyword{internal}
