% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_untag_resource}
\alias{licensemanager_untag_resource}
\title{Removes the specified tags from the specified license configuration}
\usage{
licensemanager_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the license configuration.}

\item{TagKeys}{[required] Keys identifying the tags to remove.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Removes the specified tags from the specified license configuration.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
