% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_update_license_asset_group}
\alias{licensemanager_update_license_asset_group}
\title{Updates a license asset group}
\usage{
licensemanager_update_license_asset_group(
  Name = NULL,
  Description = NULL,
  LicenseAssetGroupConfigurations = NULL,
  AssociatedLicenseAssetRulesetARNs,
  Properties = NULL,
  LicenseAssetGroupArn,
  Status = NULL,
  ClientToken
)
}
\arguments{
\item{Name}{License asset group name.}

\item{Description}{License asset group description.}

\item{LicenseAssetGroupConfigurations}{License asset group configurations.}

\item{AssociatedLicenseAssetRulesetARNs}{[required] ARNs of associated license asset rulesets.}

\item{Properties}{License asset group properties.}

\item{LicenseAssetGroupArn}{[required] Amazon Resource Name (ARN) of the license asset group.}

\item{Status}{License asset group status. The possible values are \code{ACTIVE} |
\code{DISABLED}.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates a license asset group.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_update_license_asset_group/} for full documentation.
}
\keyword{internal}
