% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_update_license_asset_ruleset}
\alias{licensemanager_update_license_asset_ruleset}
\title{Updates a license asset ruleset}
\usage{
licensemanager_update_license_asset_ruleset(
  Name = NULL,
  Description = NULL,
  Rules,
  LicenseAssetRulesetArn,
  ClientToken
)
}
\arguments{
\item{Name}{License asset ruleset name.}

\item{Description}{License asset ruleset description.}

\item{Rules}{[required] License asset rules.}

\item{LicenseAssetRulesetArn}{[required] Amazon Resource Name (ARN) of the license asset ruleset.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates a license asset ruleset.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_update_license_asset_ruleset/} for full documentation.
}
\keyword{internal}
