% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_update_license_manager_report_generator}
\alias{licensemanager_update_license_manager_report_generator}
\title{Updates a report generator}
\usage{
licensemanager_update_license_manager_report_generator(
  LicenseManagerReportGeneratorArn,
  ReportGeneratorName,
  Type,
  ReportContext,
  ReportFrequency,
  ClientToken,
  Description = NULL
)
}
\arguments{
\item{LicenseManagerReportGeneratorArn}{[required] Amazon Resource Name (ARN) of the report generator to update.}

\item{ReportGeneratorName}{[required] Name of the report generator.}

\item{Type}{[required] Type of reports to generate. The following report types are supported:
\itemize{
\item License configuration report - Reports the number and details of
consumed licenses for a license configuration.
\item Resource report - Reports the tracked licenses and resource
consumption for a license configuration.
}}

\item{ReportContext}{[required] The report context.}

\item{ReportFrequency}{[required] Frequency by which reports are generated.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{Description}{Description of the report generator.}
}
\description{
Updates a report generator.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_update_license_manager_report_generator/} for full documentation.
}
\keyword{internal}
