% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_deregister_subscription_provider}
\alias{licensemanagerlinuxsubscriptions_deregister_subscription_provider}
\title{Remove a third-party subscription provider from the Bring Your Own
License (BYOL) subscriptions registered to your account}
\usage{
licensemanagerlinuxsubscriptions_deregister_subscription_provider(
  SubscriptionProviderArn
)
}
\arguments{
\item{SubscriptionProviderArn}{[required] The Amazon Resource Name (ARN) of the subscription provider resource to
deregister.}
}
\description{
Remove a third-party subscription provider from the Bring Your Own License (BYOL) subscriptions registered to your account.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_deregister_subscription_provider/} for full documentation.
}
\keyword{internal}
