% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_get_registered_subscription_provider}
\alias{licensemanagerlinuxsubscriptions_get_registered_subscription_provider}
\title{Get details for a Bring Your Own License (BYOL) subscription that's
registered to your account}
\usage{
licensemanagerlinuxsubscriptions_get_registered_subscription_provider(
  SubscriptionProviderArn
)
}
\arguments{
\item{SubscriptionProviderArn}{[required] The Amazon Resource Name (ARN) of the BYOL registration resource to get
details for.}
}
\description{
Get details for a Bring Your Own License (BYOL) subscription that's registered to your account.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_get_registered_subscription_provider/} for full documentation.
}
\keyword{internal}
