% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_get_service_settings}
\alias{licensemanagerlinuxsubscriptions_get_service_settings}
\title{Lists the Linux subscriptions service settings}
\usage{
licensemanagerlinuxsubscriptions_get_service_settings()
}
\description{
Lists the Linux subscriptions service settings.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_get_service_settings/} for full documentation.
}
\keyword{internal}
