% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_list_linux_subscription_instances}
\alias{licensemanagerlinuxsubscriptions_list_linux_subscription_instances}
\title{Lists the running Amazon EC2 instances that were discovered with
commercial Linux subscriptions}
\usage{
licensemanagerlinuxsubscriptions_list_linux_subscription_instances(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results by your
specified criteria. For example, you can specify \code{Region} in the \code{Name},
with the \code{contains} operator to list all subscriptions that match a
partial string in the \code{Value}, such as \code{us-west}.

For each filter, you can specify one of the following values for the
\code{Name} key to streamline results:
\itemize{
\item \code{AccountID}
\item \code{AmiID}
\item \code{DualSubscription}
\item \code{InstanceID}
\item \code{InstanceType}
\item \code{ProductCode}
\item \code{Region}
\item \code{Status}
\item \code{UsageOperation}
}

For each filter, you can use one of the following \code{Operator} values to
define the behavior of the filter:
\itemize{
\item \code{contains}
\item \code{equals}
\item \code{Notequal}
}}

\item{MaxResults}{The maximum items to return in a request.}

\item{NextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Lists the running Amazon EC2 instances that were discovered with commercial Linux subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_list_linux_subscription_instances/} for full documentation.
}
\keyword{internal}
