% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_list_registered_subscription_providers}
\alias{licensemanagerlinuxsubscriptions_list_registered_subscription_providers}
\title{List Bring Your Own License (BYOL) subscription registration resources
for your account}
\usage{
licensemanagerlinuxsubscriptions_list_registered_subscription_providers(
  MaxResults = NULL,
  NextToken = NULL,
  SubscriptionProviderSources = NULL
)
}
\arguments{
\item{MaxResults}{The maximum items to return in a request.}

\item{NextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}

\item{SubscriptionProviderSources}{To filter your results, specify which subscription providers to return
in the list.}
}
\description{
List Bring Your Own License (BYOL) subscription registration resources for your account.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_list_registered_subscription_providers/} for full documentation.
}
\keyword{internal}
