% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_list_linux_subscriptions}
\alias{licensemanagerlinuxsubscriptions_list_linux_subscriptions}
\title{Lists the Linux subscriptions that have been discovered}
\usage{
licensemanagerlinuxsubscriptions_list_linux_subscriptions(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results to those
that match one or more sets of key-value pairs that you specify. For
example, you can filter by the name of \code{Subscription} with an optional
operator to see subscriptions that match, partially match, or don't
match a certain subscription's name.

The valid names for this filter are:
\itemize{
\item \code{Subscription}
}

The valid Operators for this filter are:
\itemize{
\item \code{contains}
\item \code{equals}
\item \code{Notequal}
}}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists the Linux subscriptions that have been discovered. If you have linked your organization, the returned results will include data aggregated across your accounts in Organizations.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_list_linux_subscriptions/} for full documentation.
}
\keyword{internal}
