% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_list_tags_for_resource}
\alias{licensemanagerlinuxsubscriptions_list_tags_for_resource}
\title{List the metadata tags that are assigned to the specified Amazon Web
Services resource}
\usage{
licensemanagerlinuxsubscriptions_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to list
metadata tags.}
}
\description{
List the metadata tags that are assigned to the specified Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
