% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_register_subscription_provider}
\alias{licensemanagerlinuxsubscriptions_register_subscription_provider}
\title{Register the supported third-party subscription provider for your Bring
Your Own License (BYOL) subscription}
\usage{
licensemanagerlinuxsubscriptions_register_subscription_provider(
  SecretArn,
  SubscriptionProviderSource,
  Tags = NULL
)
}
\arguments{
\item{SecretArn}{[required] The Amazon Resource Name (ARN) of the secret where you've stored your
subscription provider's access token. For RHEL subscriptions managed
through the Red Hat Subscription Manager (RHSM), the secret contains
your Red Hat Offline token.}

\item{SubscriptionProviderSource}{[required] The supported Linux subscription provider to register.}

\item{Tags}{The metadata tags to assign to your registered Linux subscription
provider resource.}
}
\description{
Register the supported third-party subscription provider for your Bring Your Own License (BYOL) subscription.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_register_subscription_provider/} for full documentation.
}
\keyword{internal}
