% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_tag_resource}
\alias{licensemanagerlinuxsubscriptions_tag_resource}
\title{Add metadata tags to the specified Amazon Web Services resource}
\usage{
licensemanagerlinuxsubscriptions_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to
which to add the specified metadata tags.}

\item{tags}{[required] The metadata tags to assign to the Amazon Web Services resource. Tags
are formatted as key value pairs.}
}
\description{
Add metadata tags to the specified Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_tag_resource/} for full documentation.
}
\keyword{internal}
