% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_untag_resource}
\alias{licensemanagerlinuxsubscriptions_untag_resource}
\title{Remove one or more metadata tag from the specified Amazon Web Services
resource}
\usage{
licensemanagerlinuxsubscriptions_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services resource to
remove the metadata tags from.}

\item{tagKeys}{[required] A list of metadata tag keys to remove from the requested resource.}
}
\description{
Remove one or more metadata tag from the specified Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_untag_resource/} for full documentation.
}
\keyword{internal}
