% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_update_service_settings}
\alias{licensemanagerlinuxsubscriptions_update_service_settings}
\title{Updates the service settings for Linux subscriptions}
\usage{
licensemanagerlinuxsubscriptions_update_service_settings(
  AllowUpdate = NULL,
  LinuxSubscriptionsDiscovery,
  LinuxSubscriptionsDiscoverySettings
)
}
\arguments{
\item{AllowUpdate}{Describes if updates are allowed to the service settings for Linux
subscriptions. If you allow updates, you can aggregate Linux
subscription data in more than one home Region.}

\item{LinuxSubscriptionsDiscovery}{[required] Describes if the discovery of Linux subscriptions is enabled.}

\item{LinuxSubscriptionsDiscoverySettings}{[required] The settings defined for Linux subscriptions discovery. The settings
include if Organizations integration has been enabled, and which Regions
data will be aggregated from.}
}
\description{
Updates the service settings for Linux subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_update_service_settings/} for full documentation.
}
\keyword{internal}
