% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_associate_user}
\alias{licensemanagerusersubscriptions_associate_user}
\title{Associates the user to an EC2 instance to utilize user-based
subscriptions}
\usage{
licensemanagerusersubscriptions_associate_user(
  Domain = NULL,
  IdentityProvider,
  InstanceId,
  Username
)
}
\arguments{
\item{Domain}{The domain name of the user.}

\item{IdentityProvider}{[required] The identity provider of the user.}

\item{InstanceId}{[required] The ID of the EC2 instance, which provides user-based subscriptions.}

\item{Username}{[required] The user name from the identity provider for the user.}
}
\description{
Associates the user to an EC2 instance to utilize user-based subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_associate_user/} for full documentation.
}
\keyword{internal}
