% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_delete_license_server_endpoint}
\alias{licensemanagerusersubscriptions_delete_license_server_endpoint}
\title{Deletes a LicenseServerEndpoint resource}
\usage{
licensemanagerusersubscriptions_delete_license_server_endpoint(
  LicenseServerEndpointArn,
  ServerType
)
}
\arguments{
\item{LicenseServerEndpointArn}{[required] The Amazon Resource Name (ARN) that identifies the
\code{LicenseServerEndpoint} resource to delete.}

\item{ServerType}{[required] The type of License Server that the delete request refers to.}
}
\description{
Deletes a \code{LicenseServerEndpoint} resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_delete_license_server_endpoint/} for full documentation.
}
\keyword{internal}
