% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_license_server_endpoints}
\alias{licensemanagerusersubscriptions_list_license_server_endpoints}
\title{List the Remote Desktop Services (RDS) License Server endpoints}
\usage{
licensemanagerusersubscriptions_list_license_server_endpoints(
  MaxResults = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return from a single request.}

\item{Filters}{You can use the following filters to streamline results:
\itemize{
\item IdentityProviderArn
}}

\item{NextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
List the Remote Desktop Services (RDS) License Server endpoints

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_license_server_endpoints/} for full documentation.
}
\keyword{internal}
