% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_tags_for_resource}
\alias{licensemanagerusersubscriptions_list_tags_for_resource}
\title{Returns the list of tags for the specified resource}
\usage{
licensemanagerusersubscriptions_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags you want to
retrieve.}
}
\description{
Returns the list of tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
