% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_user_associations}
\alias{licensemanagerusersubscriptions_list_user_associations}
\title{Lists user associations for an identity provider}
\usage{
licensemanagerusersubscriptions_list_user_associations(
  Filters = NULL,
  IdentityProvider,
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results to those
that match one or more sets of key-value pairs that you specify.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{InstanceId}{[required] The ID of the EC2 instance, which provides user-based subscriptions.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists user associations for an identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_user_associations/} for full documentation.
}
\keyword{internal}
