% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_tag_resource}
\alias{licensemanagerusersubscriptions_tag_resource}
\title{Adds tags to a resource}
\usage{
licensemanagerusersubscriptions_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{Tags}{[required] The tags to apply to the specified resource.}
}
\description{
Adds tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_tag_resource/} for full documentation.
}
\keyword{internal}
