% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_untag_resource}
\alias{licensemanagerusersubscriptions_untag_resource}
\title{Removes tags from a resource}
\usage{
licensemanagerusersubscriptions_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
tags from.}

\item{TagKeys}{[required] The tag keys to remove from the resource.}
}
\description{
Removes tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_untag_resource/} for full documentation.
}
\keyword{internal}
