% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_delete_workspace}
\alias{managedgrafana_delete_workspace}
\title{Deletes an Amazon Managed Grafana workspace}
\usage{
managedgrafana_delete_workspace(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to delete.}
}
\description{
Deletes an Amazon Managed Grafana workspace.

See \url{https://paws-r.github.io/docs/managedgrafana/delete_workspace.html} for full documentation.
}
\keyword{internal}
