% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_delete_workspace_api_key}
\alias{managedgrafana_delete_workspace_api_key}
\title{Deletes a Grafana API key for the workspace}
\usage{
managedgrafana_delete_workspace_api_key(keyName, workspaceId)
}
\arguments{
\item{keyName}{[required] The name of the API key to delete.}

\item{workspaceId}{[required] The ID of the workspace to delete.}
}
\description{
Deletes a Grafana API key for the workspace.

See \url{https://paws-r.github.io/docs/managedgrafana/delete_workspace_api_key.html} for full documentation.
}
\keyword{internal}
