% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_delete_workspace_service_account}
\alias{managedgrafana_delete_workspace_service_account}
\title{Deletes a workspace service account from the workspace}
\usage{
managedgrafana_delete_workspace_service_account(serviceAccountId, workspaceId)
}
\arguments{
\item{serviceAccountId}{[required] The ID of the service account to delete.}

\item{workspaceId}{[required] The ID of the workspace where the service account resides.}
}
\description{
Deletes a workspace service account from the workspace.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_delete_workspace_service_account/} for full documentation.
}
\keyword{internal}
