% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_delete_workspace_service_account_token}
\alias{managedgrafana_delete_workspace_service_account_token}
\title{Deletes a token for the workspace service account}
\usage{
managedgrafana_delete_workspace_service_account_token(
  serviceAccountId,
  tokenId,
  workspaceId
)
}
\arguments{
\item{serviceAccountId}{[required] The ID of the service account from which to delete the token.}

\item{tokenId}{[required] The ID of the token to delete.}

\item{workspaceId}{[required] The ID of the workspace from which to delete the token.}
}
\description{
Deletes a token for the workspace service account.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_delete_workspace_service_account_token/} for full documentation.
}
\keyword{internal}
