% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_describe_workspace_authentication}
\alias{managedgrafana_describe_workspace_authentication}
\title{Displays information about the authentication methods used in one Amazon
Managed Grafana workspace}
\usage{
managedgrafana_describe_workspace_authentication(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to return authentication information about.}
}
\description{
Displays information about the authentication methods used in one Amazon Managed Grafana workspace.

See \url{https://paws-r.github.io/docs/managedgrafana/describe_workspace_authentication.html} for full documentation.
}
\keyword{internal}
