% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_disassociate_license}
\alias{managedgrafana_disassociate_license}
\title{Removes the Grafana Enterprise license from a workspace}
\usage{
managedgrafana_disassociate_license(licenseType, workspaceId)
}
\arguments{
\item{licenseType}{[required] The type of license to remove from the workspace.}

\item{workspaceId}{[required] The ID of the workspace to remove the Grafana Enterprise license from.}
}
\description{
Removes the Grafana Enterprise license from a workspace.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_disassociate_license/} for full documentation.
}
\keyword{internal}
