% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_list_versions}
\alias{managedgrafana_list_versions}
\title{Lists available versions of Grafana}
\usage{
managedgrafana_list_versions(
  maxResults = NULL,
  nextToken = NULL,
  workspaceId = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token to use when requesting the next set of results. You receive
this token from a previous
\code{\link[=managedgrafana_list_versions]{list_versions}} operation.}

\item{workspaceId}{The ID of the workspace to list the available upgrade versions. If not
included, lists all versions of Grafana that are supported for
\code{\link[=managedgrafana_create_workspace]{create_workspace}}.}
}
\description{
Lists available versions of Grafana. These are available when calling \code{\link[=managedgrafana_create_workspace]{create_workspace}}. Optionally, include a workspace to list the versions to which it can be upgraded.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_list_versions/} for full documentation.
}
\keyword{internal}
