% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_list_workspace_service_account_tokens}
\alias{managedgrafana_list_workspace_service_account_tokens}
\title{Returns a list of tokens for a workspace service account}
\usage{
managedgrafana_list_workspace_service_account_tokens(
  maxResults = NULL,
  nextToken = NULL,
  serviceAccountId,
  workspaceId
)
}
\arguments{
\item{maxResults}{The maximum number of tokens to include in the results.}

\item{nextToken}{The token for the next set of service accounts to return. (You receive
this token from a previous
\code{\link[=managedgrafana_list_workspace_service_account_tokens]{list_workspace_service_account_tokens}}
operation.)}

\item{serviceAccountId}{[required] The ID of the service account for which to return tokens.}

\item{workspaceId}{[required] The ID of the workspace for which to return tokens.}
}
\description{
Returns a list of tokens for a workspace service account.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_list_workspace_service_account_tokens/} for full documentation.
}
\keyword{internal}
