% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_list_workspace_service_accounts}
\alias{managedgrafana_list_workspace_service_accounts}
\title{Returns a list of service accounts for a workspace}
\usage{
managedgrafana_list_workspace_service_accounts(
  maxResults = NULL,
  nextToken = NULL,
  workspaceId
)
}
\arguments{
\item{maxResults}{The maximum number of service accounts to include in the results.}

\item{nextToken}{The token for the next set of service accounts to return. (You receive
this token from a previous
\code{\link[=managedgrafana_list_workspace_service_accounts]{list_workspace_service_accounts}}
operation.)}

\item{workspaceId}{[required] The workspace for which to list service accounts.}
}
\description{
Returns a list of service accounts for a workspace.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_list_workspace_service_accounts/} for full documentation.
}
\keyword{internal}
