% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_list_workspaces}
\alias{managedgrafana_list_workspaces}
\title{Returns a list of Amazon Managed Grafana workspaces in the account, with
some information about each workspace}
\usage{
managedgrafana_list_workspaces(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of workspaces to include in the results.}

\item{nextToken}{The token for the next set of workspaces to return. (You receive this
token from a previous
\code{\link[=managedgrafana_list_workspaces]{list_workspaces}} operation.)}
}
\description{
Returns a list of Amazon Managed Grafana workspaces in the account, with some information about each workspace. For more complete information about one workspace, use \code{\link[=managedgrafana_describe_workspace]{describe_workspace}}.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_list_workspaces/} for full documentation.
}
\keyword{internal}
