% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_assign_instance}
\alias{opsworks_assign_instance}
\title{Assign a registered instance to a layer}
\usage{
opsworks_assign_instance(InstanceId, LayerIds)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}

\item{LayerIds}{[required] The layer ID, which must correspond to a custom layer. You cannot assign
a registered instance to a built-in layer.}
}
\description{
Assign a registered instance to a layer.

See \url{https://www.paws-r-sdk.com/docs/opsworks_assign_instance/} for full documentation.
}
\keyword{internal}
