% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_associate_elastic_ip}
\alias{opsworks_associate_elastic_ip}
\title{Associates one of the stack's registered Elastic IP addresses with a
specified instance}
\usage{
opsworks_associate_elastic_ip(ElasticIp, InstanceId = NULL)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}

\item{InstanceId}{The instance ID.}
}
\description{
Associates one of the stack's registered Elastic IP addresses with a specified instance. The address must first be registered with the stack by calling \code{\link[=opsworks_register_elastic_ip]{register_elastic_ip}}. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://paws-r.github.io/docs/opsworks/associate_elastic_ip.html} for full documentation.
}
\keyword{internal}
