% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_attach_elastic_load_balancer}
\alias{opsworks_attach_elastic_load_balancer}
\title{Attaches an Elastic Load Balancing load balancer to a specified layer}
\usage{
opsworks_attach_elastic_load_balancer(ElasticLoadBalancerName, LayerId)
}
\arguments{
\item{ElasticLoadBalancerName}{[required] The Elastic Load Balancing instance's name.}

\item{LayerId}{[required] The ID of the layer to which the Elastic Load Balancing instance is to
be attached.}
}
\description{
Attaches an Elastic Load Balancing load balancer to a specified layer. AWS OpsWorks Stacks does not support Application Load Balancer. You can only use Classic Load Balancer with AWS OpsWorks Stacks. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/layers-elb.html}{Elastic Load Balancing}.

See \url{https://paws-r.github.io/docs/opsworks/attach_elastic_load_balancer.html} for full documentation.
}
\keyword{internal}
