% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_app}
\alias{opsworks_delete_app}
\title{Deletes a specified app}
\usage{
opsworks_delete_app(AppId)
}
\arguments{
\item{AppId}{[required] The app ID.}
}
\description{
Deletes a specified app.

See \url{https://www.paws-r-sdk.com/docs/opsworks_delete_app/} for full documentation.
}
\keyword{internal}
