% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_instance}
\alias{opsworks_delete_instance}
\title{Deletes a specified instance, which terminates the associated Amazon EC2
instance}
\usage{
opsworks_delete_instance(
  InstanceId,
  DeleteElasticIp = NULL,
  DeleteVolumes = NULL
)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}

\item{DeleteElasticIp}{Whether to delete the instance Elastic IP address.}

\item{DeleteVolumes}{Whether to delete the instance's Amazon EBS volumes.}
}
\description{
Deletes a specified instance, which terminates the associated Amazon EC2 instance. You must stop an instance before you can delete it.

See \url{https://paws-r.github.io/docs/opsworks/delete_instance.html} for full documentation.
}
\keyword{internal}
