% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_stack}
\alias{opsworks_delete_stack}
\title{Deletes a specified stack}
\usage{
opsworks_delete_stack(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\description{
Deletes a specified stack. You must first delete all instances, layers, and apps or deregister registered instances. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workingstacks-shutting.html}{Shut Down a Stack}.

See \url{https://paws-r.github.io/docs/opsworks/delete_stack.html} for full documentation.
}
\keyword{internal}
