% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_elastic_ip}
\alias{opsworks_deregister_elastic_ip}
\title{Deregisters a specified Elastic IP address}
\usage{
opsworks_deregister_elastic_ip(ElasticIp)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}
}
\description{
Deregisters a specified Elastic IP address. The address can be registered by another stack after it is deregistered. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_deregister_elastic_ip/} for full documentation.
}
\keyword{internal}
