% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_instance}
\alias{opsworks_deregister_instance}
\title{Deregister an instance from OpsWorks Stacks}
\usage{
opsworks_deregister_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}
}
\description{
Deregister an instance from OpsWorks Stacks. The instance can be a registered instance (Amazon EC2 or on-premises) or an instance created with OpsWorks. This action removes the instance from the stack and returns it to your control.

See \url{https://www.paws-r-sdk.com/docs/opsworks_deregister_instance/} for full documentation.
}
\keyword{internal}
