% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_rds_db_instance}
\alias{opsworks_deregister_rds_db_instance}
\title{Deregisters an Amazon RDS instance}
\usage{
opsworks_deregister_rds_db_instance(RdsDbInstanceArn)
}
\arguments{
\item{RdsDbInstanceArn}{[required] The Amazon RDS instance's ARN.}
}
\description{
Deregisters an Amazon RDS instance.

See \url{https://www.paws-r-sdk.com/docs/opsworks_deregister_rds_db_instance/} for full documentation.
}
\keyword{internal}
