% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_commands}
\alias{opsworks_describe_commands}
\title{Describes the results of specified commands}
\usage{
opsworks_describe_commands(
  DeploymentId = NULL,
  InstanceId = NULL,
  CommandIds = NULL
)
}
\arguments{
\item{DeploymentId}{The deployment ID. If you include this parameter,
\code{\link[=opsworks_describe_commands]{describe_commands}} returns a description
of the commands associated with the specified deployment.}

\item{InstanceId}{The instance ID. If you include this parameter,
\code{\link[=opsworks_describe_commands]{describe_commands}} returns a description
of the commands associated with the specified instance.}

\item{CommandIds}{An array of command IDs. If you include this parameter,
\code{\link[=opsworks_describe_commands]{describe_commands}} returns a description
of the specified commands. Otherwise, it returns a description of every
command.}
}
\description{
Describes the results of specified commands.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_commands/} for full documentation.
}
\keyword{internal}
