% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_deployments}
\alias{opsworks_describe_deployments}
\title{Requests a description of a specified set of deployments}
\usage{
opsworks_describe_deployments(
  StackId = NULL,
  AppId = NULL,
  DeploymentIds = NULL
)
}
\arguments{
\item{StackId}{The stack ID. If you include this parameter, the command returns a
description of the commands associated with the specified stack.}

\item{AppId}{The app ID. If you include this parameter, the command returns a
description of the commands associated with the specified app.}

\item{DeploymentIds}{An array of deployment IDs to be described. If you include this
parameter, the command returns a description of the specified
deployments. Otherwise, it returns a description of every deployment.}
}
\description{
Requests a description of a specified set of deployments.

See \url{https://paws-r.github.io/docs/opsworks/describe_deployments.html} for full documentation.
}
\keyword{internal}
