% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_load_based_auto_scaling}
\alias{opsworks_describe_load_based_auto_scaling}
\title{Describes load-based auto scaling configurations for specified layers}
\usage{
opsworks_describe_load_based_auto_scaling(LayerIds)
}
\arguments{
\item{LayerIds}{[required] An array of layer IDs.}
}
\description{
Describes load-based auto scaling configurations for specified layers.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_load_based_auto_scaling/} for full documentation.
}
\keyword{internal}
