% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_my_user_profile}
\alias{opsworks_describe_my_user_profile}
\title{Describes a user's SSH information}
\usage{
opsworks_describe_my_user_profile()
}
\description{
Describes a user's SSH information.

\strong{Required Permissions}: To use this action, an IAM user must have
self-management enabled or an attached policy that explicitly grants
permissions. For more information about user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\keyword{internal}
