% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_permissions}
\alias{opsworks_describe_permissions}
\title{Describes the permissions for a specified stack}
\usage{
opsworks_describe_permissions(IamUserArn = NULL, StackId = NULL)
}
\arguments{
\item{IamUserArn}{The user's IAM ARN. This can also be a federated user's ARN. For more
information about IAM ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html}{Using Identifiers}.}

\item{StackId}{The stack ID.}
}
\description{
Describes the permissions for a specified stack.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_permissions/} for full documentation.
}
\keyword{internal}
