% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_raid_arrays}
\alias{opsworks_describe_raid_arrays}
\title{Describe an instance's RAID arrays}
\usage{
opsworks_describe_raid_arrays(
  InstanceId = NULL,
  StackId = NULL,
  RaidArrayIds = NULL
)
}
\arguments{
\item{InstanceId}{The instance ID. If you use this parameter,
\code{\link[=opsworks_describe_raid_arrays]{describe_raid_arrays}} returns
descriptions of the RAID arrays associated with the specified instance.}

\item{StackId}{The stack ID.}

\item{RaidArrayIds}{An array of RAID array IDs. If you use this parameter,
\code{\link[=opsworks_describe_raid_arrays]{describe_raid_arrays}} returns
descriptions of the specified arrays. Otherwise, it returns a
description of every array.}
}
\description{
Describe an instance's RAID arrays.

See \url{https://paws-r.github.io/docs/opsworks/describe_raid_arrays.html} for full documentation.
}
\keyword{internal}
