% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_rds_db_instances}
\alias{opsworks_describe_rds_db_instances}
\title{Describes Amazon RDS instances}
\usage{
opsworks_describe_rds_db_instances(StackId, RdsDbInstanceArns = NULL)
}
\arguments{
\item{StackId}{[required] The ID of the stack with which the instances are registered. The
operation returns descriptions of all registered Amazon RDS instances.}

\item{RdsDbInstanceArns}{An array containing the ARNs of the instances to be described.}
}
\description{
Describes Amazon RDS instances.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_rds_db_instances/} for full documentation.
}
\keyword{internal}
