% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_service_errors}
\alias{opsworks_describe_service_errors}
\title{Describes OpsWorks Stacks service errors}
\usage{
opsworks_describe_service_errors(
  StackId = NULL,
  InstanceId = NULL,
  ServiceErrorIds = NULL
)
}
\arguments{
\item{StackId}{The stack ID. If you use this parameter,
\code{\link[=opsworks_describe_service_errors]{describe_service_errors}} returns
descriptions of the errors associated with the specified stack.}

\item{InstanceId}{The instance ID. If you use this parameter,
\code{\link[=opsworks_describe_service_errors]{describe_service_errors}} returns
descriptions of the errors associated with the specified instance.}

\item{ServiceErrorIds}{An array of service error IDs. If you use this parameter,
\code{\link[=opsworks_describe_service_errors]{describe_service_errors}} returns
descriptions of the specified errors. Otherwise, it returns a
description of every error.}
}
\description{
Describes OpsWorks Stacks service errors.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_service_errors/} for full documentation.
}
\keyword{internal}
