% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stack_provisioning_parameters}
\alias{opsworks_describe_stack_provisioning_parameters}
\title{Requests a description of a stack's provisioning parameters}
\usage{
opsworks_describe_stack_provisioning_parameters(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\description{
Requests a description of a stack's provisioning parameters.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_stack_provisioning_parameters/} for full documentation.
}
\keyword{internal}
