% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stack_summary}
\alias{opsworks_describe_stack_summary}
\title{Describes the number of layers and apps in a specified stack, and the
number of instances in each state, such as running_setup or online}
\usage{
opsworks_describe_stack_summary(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StackSummary = list(
    StackId = "string",
    Name = "string",
    Arn = "string",
    LayersCount = 123,
    AppsCount = 123,
    InstancesCount = list(
      Assigning = 123,
      Booting = 123,
      ConnectionLost = 123,
      Deregistering = 123,
      Online = 123,
      Pending = 123,
      Rebooting = 123,
      Registered = 123,
      Registering = 123,
      Requested = 123,
      RunningSetup = 123,
      SetupFailed = 123,
      ShuttingDown = 123,
      StartFailed = 123,
      StopFailed = 123,
      Stopped = 123,
      Stopping = 123,
      Terminated = 123,
      Terminating = 123,
      Unassigning = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the number of layers and apps in a specified stack, and the
number of instances in each state, such as \code{running_setup} or \code{online}.

\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack, or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_stack_summary(
  StackId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
