% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stacks}
\alias{opsworks_describe_stacks}
\title{Requests a description of one or more stacks}
\usage{
opsworks_describe_stacks(StackIds = NULL)
}
\arguments{
\item{StackIds}{An array of stack IDs that specify the stacks to be described. If you
omit this parameter, and have permissions to get information about all
stacks, \code{\link[=opsworks_describe_stacks]{describe_stacks}} returns a
description of every stack. If the IAM policy that is attached to an IAM
user limits the \code{\link[=opsworks_describe_stacks]{describe_stacks}} action to
specific stack ARNs, this parameter is required, and the user must
specify a stack ARN that is allowed by the policy. Otherwise,
\code{\link[=opsworks_describe_stacks]{describe_stacks}} returns an \code{AccessDenied}
error.}
}
\description{
Requests a description of one or more stacks.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_stacks/} for full documentation.
}
\keyword{internal}
