% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_volumes}
\alias{opsworks_describe_volumes}
\title{Describes an instance's Amazon EBS volumes}
\usage{
opsworks_describe_volumes(InstanceId, StackId, RaidArrayId, VolumeIds)
}
\arguments{
\item{InstanceId}{The instance ID. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the volumes associated with the specified instance.}

\item{StackId}{A stack ID. The action describes the stack's registered Amazon EBS
volumes.}

\item{RaidArrayId}{The RAID array ID. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the volumes associated with the specified RAID array.}

\item{VolumeIds}{Am array of volume IDs. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the specified volumes. Otherwise, it returns a description of every
volume.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Volumes = list(
    list(
      VolumeId = "string",
      Ec2VolumeId = "string",
      Name = "string",
      RaidArrayId = "string",
      InstanceId = "string",
      Status = "string",
      Size = 123,
      Device = "string",
      MountPoint = "string",
      Region = "string",
      AvailabilityZone = "string",
      VolumeType = "string",
      Iops = 123,
      Encrypted = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes an instance's Amazon EBS volumes.

This call accepts only one resource-identifying parameter.

\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack, or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_volumes(
  InstanceId = "string",
  StackId = "string",
  RaidArrayId = "string",
  VolumeIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
