% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_disassociate_elastic_ip}
\alias{opsworks_disassociate_elastic_ip}
\title{Disassociates an Elastic IP address from its instance}
\usage{
opsworks_disassociate_elastic_ip(ElasticIp)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}
}
\description{
Disassociates an Elastic IP address from its instance. The address remains registered with the stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_disassociate_elastic_ip/} for full documentation.
}
\keyword{internal}
