% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_register_ecs_cluster}
\alias{opsworks_register_ecs_cluster}
\title{Registers a specified Amazon ECS cluster with a stack}
\usage{
opsworks_register_ecs_cluster(EcsClusterArn, StackId)
}
\arguments{
\item{EcsClusterArn}{[required] The cluster's ARN.}

\item{StackId}{[required] The stack ID.}
}
\description{
Registers a specified Amazon ECS cluster with a stack. You can register only one cluster with a stack. A cluster can be registered with only one stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workinglayers-ecscluster.html}{Resource Management}.

See \url{https://paws-r.github.io/docs/opsworks/register_ecs_cluster.html} for full documentation.
}
\keyword{internal}
