% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_register_elastic_ip}
\alias{opsworks_register_elastic_ip}
\title{Registers an Elastic IP address with a specified stack}
\usage{
opsworks_register_elastic_ip(ElasticIp, StackId)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}

\item{StackId}{[required] The stack ID.}
}
\description{
Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling \code{\link[=opsworks_deregister_elastic_ip]{deregister_elastic_ip}}. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_register_elastic_ip/} for full documentation.
}
\keyword{internal}
