% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_set_time_based_auto_scaling}
\alias{opsworks_set_time_based_auto_scaling}
\title{Specify the time-based auto scaling configuration for a specified
instance}
\usage{
opsworks_set_time_based_auto_scaling(InstanceId, AutoScalingSchedule = NULL)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}

\item{AutoScalingSchedule}{An \code{AutoScalingSchedule} with the instance schedule.}
}
\description{
Specify the time-based auto scaling configuration for a specified instance. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-autoscaling.html}{Managing Load with Time-based and Load-based Instances}.

See \url{https://paws-r.github.io/docs/opsworks/set_time_based_auto_scaling.html} for full documentation.
}
\keyword{internal}
