% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_unassign_instance}
\alias{opsworks_unassign_instance}
\title{Unassigns a registered instance from all layers that are using the
instance}
\usage{
opsworks_unassign_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}
}
\description{
Unassigns a registered instance from all layers that are using the instance. The instance remains in the stack as an unassigned instance, and can be assigned to another layer as needed. You cannot use this action with instances that were created with AWS OpsWorks Stacks.

See \url{https://www.paws-r-sdk.com/docs/opsworks_unassign_instance/} for full documentation.
}
\keyword{internal}
