% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_unassign_volume}
\alias{opsworks_unassign_volume}
\title{Unassigns an assigned Amazon EBS volume}
\usage{
opsworks_unassign_volume(VolumeId)
}
\arguments{
\item{VolumeId}{[required] The volume ID.}
}
\value{
An empty list.
}
\description{
Unassigns an assigned Amazon EBS volume. The volume remains registered
with the stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack, or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$unassign_volume(
  VolumeId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
