% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_update_elastic_ip}
\alias{opsworks_update_elastic_ip}
\title{Updates a registered Elastic IP address's name}
\usage{
opsworks_update_elastic_ip(ElasticIp, Name)
}
\arguments{
\item{ElasticIp}{[required] The IP address for which you want to update the name.}

\item{Name}{The new name.}
}
\value{
An empty list.
}
\description{
Updates a registered Elastic IP address's name. For more information,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack, or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_elastic_ip(
  ElasticIp = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
