% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_update_my_user_profile}
\alias{opsworks_update_my_user_profile}
\title{Updates a user's SSH public key}
\usage{
opsworks_update_my_user_profile(SshPublicKey = NULL)
}
\arguments{
\item{SshPublicKey}{The user's SSH public key.}
}
\description{
Updates a user's SSH public key.

See \url{https://www.paws-r-sdk.com/docs/opsworks_update_my_user_profile/} for full documentation.
}
\keyword{internal}
