% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_update_rds_db_instance}
\alias{opsworks_update_rds_db_instance}
\title{Updates an Amazon RDS instance}
\usage{
opsworks_update_rds_db_instance(
  RdsDbInstanceArn,
  DbUser = NULL,
  DbPassword = NULL
)
}
\arguments{
\item{RdsDbInstanceArn}{[required] The Amazon RDS instance's ARN.}

\item{DbUser}{The master user name.}

\item{DbPassword}{The database password.}
}
\description{
Updates an Amazon RDS instance.

See \url{https://www.paws-r-sdk.com/docs/opsworks_update_rds_db_instance/} for full documentation.
}
\keyword{internal}
