% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_delete_server}
\alias{opsworkscm_delete_server}
\title{Deletes the server and the underlying AWS CloudFormation stacks
(including the server's EC2 instance)}
\usage{
opsworkscm_delete_server(ServerName)
}
\arguments{
\item{ServerName}{[required] The ID of the server to delete.}
}
\description{
Deletes the server and the underlying AWS CloudFormation stacks
(including the server's EC2 instance). When you run this command, the
server state is updated to \code{DELETING}. After the server is deleted, it
is no longer returned by \code{DescribeServer} requests. If the AWS
CloudFormation stack cannot be deleted, the server cannot be deleted.
}
\details{
This operation is asynchronous.

An \code{InvalidStateException} is thrown when a server deletion is already
in progress. A \code{ResourceNotFoundException} is thrown when the server
does not exist. A \code{ValidationException} is raised when parameters of the
request are not valid.
}
\section{Request syntax}{
\preformatted{svc$delete_server(
  ServerName = "string"
)
}
}

\keyword{internal}
