% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_describe_backups}
\alias{opsworkscm_describe_backups}
\title{Describes backups}
\usage{
opsworkscm_describe_backups(BackupId, ServerName, NextToken, MaxResults)
}
\arguments{
\item{BackupId}{Describes a single backup.}

\item{ServerName}{Returns backups for the server with the specified ServerName.}

\item{NextToken}{This is not currently implemented for
\code{\link[=opsworkscm_describe_backups]{describe_backups}} requests.}

\item{MaxResults}{This is not currently implemented for
\code{\link[=opsworkscm_describe_backups]{describe_backups}} requests.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Backups = list(
    list(
      BackupArn = "string",
      BackupId = "string",
      BackupType = "AUTOMATED"|"MANUAL",
      CreatedAt = as.POSIXct(
        "2015-01-01"
      ),
      Description = "string",
      Engine = "string",
      EngineModel = "string",
      EngineVersion = "string",
      InstanceProfileArn = "string",
      InstanceType = "string",
      KeyPair = "string",
      PreferredBackupWindow = "string",
      PreferredMaintenanceWindow = "string",
      S3DataSize = 123,
      S3DataUrl = "string",
      S3LogUrl = "string",
      SecurityGroupIds = list(
        "string"
      ),
      ServerName = "string",
      ServiceRoleArn = "string",
      Status = "IN_PROGRESS"|"OK"|"FAILED"|"DELETING",
      StatusDescription = "string",
      SubnetIds = list(
        "string"
      ),
      ToolsVersion = "string",
      UserArn = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes backups. The results are ordered by time, with newest backups
first. If you do not specify a BackupId or ServerName, the command
returns all backups.

This operation is synchronous.

A \code{ResourceNotFoundException} is thrown when the backup does not exist.
A \code{ValidationException} is raised when parameters of the request are not
valid.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_backups(
  BackupId = "string",
  ServerName = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
