% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_untag_resource}
\alias{opsworkscm_untag_resource}
\title{Removes specified tags from an AWS OpsWorks-CM server or backup}
\usage{
opsworkscm_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Number (ARN) of a resource from which you want to
remove tags. For example,
\verb{arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1a2bEXAMPLE}.}

\item{TagKeys}{[required] The keys of tags that you want to remove.}
}
\value{
An empty list.
}
\description{
Removes specified tags from an AWS OpsWorks-CM server or backup.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
