% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_update_server}
\alias{opsworkscm_update_server}
\title{Updates settings for a server}
\usage{
opsworkscm_update_server(DisableAutomatedBackup, BackupRetentionCount,
  ServerName, PreferredMaintenanceWindow, PreferredBackupWindow)
}
\arguments{
\item{DisableAutomatedBackup}{Setting DisableAutomatedBackup to \code{true} disables automated or scheduled
backups. Automated backups are enabled by default.}

\item{BackupRetentionCount}{Sets the number of automated backups that you want to keep.}

\item{ServerName}{[required] The name of the server to update.}

\item{PreferredMaintenanceWindow}{}

\item{PreferredBackupWindow}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Server = list(
    AssociatePublicIpAddress = TRUE|FALSE,
    BackupRetentionCount = 123,
    ServerName = "string",
    CreatedAt = as.POSIXct(
      "2015-01-01"
    ),
    CloudFormationStackArn = "string",
    CustomDomain = "string",
    DisableAutomatedBackup = TRUE|FALSE,
    Endpoint = "string",
    Engine = "string",
    EngineModel = "string",
    EngineAttributes = list(
      list(
        Name = "string",
        Value = "string"
      )
    ),
    EngineVersion = "string",
    InstanceProfileArn = "string",
    InstanceType = "string",
    KeyPair = "string",
    MaintenanceStatus = "SUCCESS"|"FAILED",
    PreferredMaintenanceWindow = "string",
    PreferredBackupWindow = "string",
    SecurityGroupIds = list(
      "string"
    ),
    ServiceRoleArn = "string",
    Status = "BACKING_UP"|"CONNECTION_LOST"|"CREATING"|"DELETING"|"MODIFYING"|"FAILED"|"HEALTHY"|"RUNNING"|"RESTORING"|"SETUP"|"UNDER_MAINTENANCE"|"UNHEALTHY"|"TERMINATED",
    StatusReason = "string",
    SubnetIds = list(
      "string"
    ),
    ServerArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates settings for a server.

This operation is synchronous.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_server(
  DisableAutomatedBackup = TRUE|FALSE,
  BackupRetentionCount = 123,
  ServerName = "string",
  PreferredMaintenanceWindow = "string",
  PreferredBackupWindow = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
