% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_cancel_handshake}
\alias{organizations_cancel_handshake}
\title{Cancels a Handshake}
\usage{
organizations_cancel_handshake(HandshakeId)
}
\arguments{
\item{HandshakeId}{[required] ID for the handshake that you want to cancel. You can get the ID from
the
\code{\link[=organizations_list_handshakes_for_organization]{list_handshakes_for_organization}}
operation.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for handshake
ID string requires "h-" followed by from 8 to 32 lowercase letters or
digits.}
}
\description{
Cancels a Handshake.

See \url{https://www.paws-r-sdk.com/docs/organizations_cancel_handshake/} for full documentation.
}
\keyword{internal}
