% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_delete_organization}
\alias{organizations_delete_organization}
\title{Deletes the organization}
\usage{
organizations_delete_organization()
}
\description{
Deletes the organization. You can delete an organization only by using credentials from the management account. The organization must be empty of member accounts.

See \url{https://paws-r.github.io/docs/organizations/delete_organization.html} for full documentation.
}
\keyword{internal}
