% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_delete_organizational_unit}
\alias{organizations_delete_organizational_unit}
\title{Deletes an organizational unit (OU) from a root or another OU}
\usage{
organizations_delete_organizational_unit(OrganizationalUnitId)
}
\arguments{
\item{OrganizationalUnitId}{[required] The unique identifier (ID) of the organizational unit that you want to
delete. You can get the ID from the ListOrganizationalUnitsForParent
operation.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for an
organizational unit ID string requires "ou-" followed by from 4 to 32
lower-case letters or digits (the ID of the root that contains the OU)
followed by a second "-" dash and from 8 to 32 additional lower-case
letters or digits.}
}
\description{
Deletes an organizational unit (OU) from a root or another OU. You must
first remove all accounts and child OUs from the OU that you want to
delete.
}
\details{
This operation can be called only from the organization's master
account.
}
\section{Request syntax}{
\preformatted{svc$delete_organizational_unit(
  OrganizationalUnitId = "string"
)
}
}

\examples{
# The following example shows how to delete an OU. The example assumes
# that you previously removed all accounts and other OUs from the OU:
# 
# 
\donttest{svc$delete_organizational_unit(
  OrganizationalUnitId = "ou-examplerootid111-exampleouid111"
)}

}
\keyword{internal}
