% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_deregister_delegated_administrator}
\alias{organizations_deregister_delegated_administrator}
\title{Removes the specified member Amazon Web Services account as a delegated
administrator for the specified Amazon Web Services service}
\usage{
organizations_deregister_delegated_administrator(AccountId, ServicePrincipal)
}
\arguments{
\item{AccountId}{[required] The account ID number of the member account in the organization that you
want to deregister as a delegated administrator.}

\item{ServicePrincipal}{[required] The service principal name of an Amazon Web Services service for which
the account is a delegated administrator.

Delegated administrator privileges are revoked for only the specified
Amazon Web Services service from the member account. If the specified
service is the only service for which the member account is a delegated
administrator, the operation also revokes Organizations read action
permissions.}
}
\description{
Removes the specified member Amazon Web Services account as a delegated administrator for the specified Amazon Web Services service.

See \url{https://www.paws-r-sdk.com/docs/organizations_deregister_delegated_administrator/} for full documentation.
}
\keyword{internal}
