% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_account}
\alias{organizations_describe_account}
\title{Retrieves Organizations-related information about the specified account}
\usage{
organizations_describe_account(AccountId)
}
\arguments{
\item{AccountId}{[required] The unique identifier (ID) of the Amazon Web Services account that you
want information about. You can get the ID from the
\code{\link[=organizations_list_accounts]{list_accounts}} or
\code{\link[=organizations_list_accounts_for_parent]{list_accounts_for_parent}}
operations.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for an account
ID string requires exactly 12 digits.}
}
\description{
Retrieves Organizations-related information about the specified account.

See \url{https://www.paws-r-sdk.com/docs/organizations_describe_account/} for full documentation.
}
\keyword{internal}
