% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_policy}
\alias{organizations_describe_policy}
\title{Retrieves information about a policy}
\usage{
organizations_describe_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The unique identifier (ID) of the policy that you want details about.
You can get the ID from the
\code{\link[=organizations_list_policies]{list_policies}} or
\code{\link[=organizations_list_policies_for_target]{list_policies_for_target}}
operations.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for a policy ID
string requires "p-" followed by from 8 to 128 lowercase or uppercase
letters, digits, or the underscore character (_).}
}
\value{
A list with the following syntax:\preformatted{list(
  Policy = list(
    PolicySummary = list(
      Id = "string",
      Arn = "string",
      Name = "string",
      Description = "string",
      Type = "SERVICE_CONTROL_POLICY"|"TAG_POLICY"|"BACKUP_POLICY"|"AISERVICES_OPT_OUT_POLICY",
      AwsManaged = TRUE|FALSE
    ),
    Content = "string"
  )
)
}
}
\description{
Retrieves information about a policy.

This operation can be called only from the organization's management
account or by a member account that is a delegated administrator for an
AWS service.
}
\section{Request syntax}{
\preformatted{svc$describe_policy(
  PolicyId = "string"
)
}
}

\examples{
\dontrun{
# The following example shows how to request information about a
# policy:/n/n
svc$describe_policy(
  PolicyId = "p-examplepolicyid111"
)
}

}
\keyword{internal}
