% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_resource_policy}
\alias{organizations_describe_resource_policy}
\title{Retrieves information about a resource policy}
\usage{
organizations_describe_resource_policy()
}
\description{
Retrieves information about a resource policy.

See \url{https://www.paws-r-sdk.com/docs/organizations_describe_resource_policy/} for full documentation.
}
\keyword{internal}
