% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_responsibility_transfer}
\alias{organizations_describe_responsibility_transfer}
\title{Returns details for a transfer}
\usage{
organizations_describe_responsibility_transfer(Id)
}
\arguments{
\item{Id}{[required] ID for the transfer.}
}
\description{
Returns details for a transfer. A \emph{transfer} is an arrangement between two management accounts where one account designates the other with specified responsibilities for their organization.

See \url{https://www.paws-r-sdk.com/docs/organizations_describe_responsibility_transfer/} for full documentation.
}
\keyword{internal}
