% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_invite_account_to_organization}
\alias{organizations_invite_account_to_organization}
\title{Sends an invitation to another account to join your organization as a
member account}
\usage{
organizations_invite_account_to_organization(Target, Notes, Tags)
}
\arguments{
\item{Target}{[required] The identifier (ID) of the AWS account that you want to invite to join
your organization. This is a JSON object that contains the following
elements:

\verb{\{ "Type": "ACCOUNT", "Id": "< account id number >" \}}

If you use the AWS CLI, you can submit this as a single string, similar
to the following example:

\verb{--target Id=123456789012,Type=ACCOUNT}

If you specify \code{"Type": "ACCOUNT"}, you must provide the AWS account ID
number as the \code{Id}. If you specify \code{"Type": "EMAIL"}, you must specify
the email address that is associated with the account.

\verb{--target Id=diego@example.com,Type=EMAIL}}

\item{Notes}{Additional information that you want to include in the generated email
to the recipient account owner.}

\item{Tags}{A list of tags that you want to attach to the account when it becomes a
member of the organization. For each tag in the list, you must specify
both a tag key and a value. You can set the value to an empty string,
but you can't set it to \code{null}. For more information about tagging, see
\href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_tagging.html}{Tagging AWS Organizations resources}
in the AWS Organizations User Guide.

Any tags in the request are checked for compliance with any applicable
tag policies when the request is made. The request is rejected if the
tags in the request don't match the requirements of the policy at that
time. Tag policy compliance is \emph{\strong{not}} checked again when the
invitation is accepted and the tags are actually attached to the
account. That means that if the tag policy changes between the
invitation and the acceptance, then that tags could potentially be
non-compliant.

If any one of the tags is invalid or if you exceed the allowed number of
tags for an account, then the entire request fails and invitations are
not sent.}
}
\value{
A list with the following syntax:\preformatted{list(
  Handshake = list(
    Id = "string",
    Arn = "string",
    Parties = list(
      list(
        Id = "string",
        Type = "ACCOUNT"|"ORGANIZATION"|"EMAIL"
      )
    ),
    State = "REQUESTED"|"OPEN"|"CANCELED"|"ACCEPTED"|"DECLINED"|"EXPIRED",
    RequestedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    ExpirationTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Action = "INVITE"|"ENABLE_ALL_FEATURES"|"APPROVE_ALL_FEATURES"|"ADD_ORGANIZATIONS_SERVICE_LINKED_ROLE",
    Resources = list(
      list(
        Value = "string",
        Type = "ACCOUNT"|"ORGANIZATION"|"ORGANIZATION_FEATURE_SET"|"EMAIL"|"MASTER_EMAIL"|"MASTER_NAME"|"NOTES"|"PARENT_HANDSHAKE",
        Resources = list()
      )
    )
  )
)
}
}
\description{
Sends an invitation to another account to join your organization as a
member account. AWS Organizations sends email on your behalf to the
email address that is associated with the other account's owner. The
invitation is implemented as a Handshake whose details are in the
response.
\itemize{
\item You can invite AWS accounts only from the same seller as the
management account. For example, if your organization's management
account was created by Amazon Internet Services Pvt. Ltd (AISPL), an
AWS seller in India, you can invite only other AISPL accounts to
your organization. You can't combine accounts from AISPL and AWS or
from any other AWS seller. For more information, see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/}{Consolidated Billing in India}.
\item If you receive an exception that indicates that you exceeded your
account limits for the organization or that the operation failed
because your organization is still initializing, wait one hour and
then try again. If the error persists after an hour, contact AWS
Support.
}

If the request includes tags, then the requester must have the
\code{organizations:TagResource} permission.

This operation can be called only from the organization's management
account.
}
\section{Request syntax}{
\preformatted{svc$invite_account_to_organization(
  Target = list(
    Id = "string",
    Type = "ACCOUNT"|"ORGANIZATION"|"EMAIL"
  ),
  Notes = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
\dontrun{
# The following example shows the admin of the master account owned by
# bill@example.com inviting the account owned by juan@example.com to join
# an organization.
svc$invite_account_to_organization(
  Notes = "This is a request for Juan's account to join Bill's organization",
  Target = list(
    Id = "juan@example.com",
    Type = "EMAIL"
  )
)
}

}
\keyword{internal}
