% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_invite_organization_to_transfer_responsibility}
\alias{organizations_invite_organization_to_transfer_responsibility}
\title{Sends an invitation to another organization's management account to
designate your account with the specified responsibilities for their
organization}
\usage{
organizations_invite_organization_to_transfer_responsibility(
  Type,
  Target,
  Notes = NULL,
  StartTimestamp,
  SourceName,
  Tags = NULL
)
}
\arguments{
\item{Type}{[required] The type of responsibility you want to designate to your organization.
Currently, only \code{BILLING} is supported.}

\item{Target}{[required] A \code{HandshakeParty} object. Contains details for the account you want to
invite. Currently, only \code{ACCOUNT} and \code{EMAIL} are supported.}

\item{Notes}{Additional information that you want to include in the invitation.}

\item{StartTimestamp}{[required] Timestamp when the recipient will begin managing the specified
responsibilities.}

\item{SourceName}{[required] Name you want to assign to the transfer.}

\item{Tags}{A list of tags that you want to attach to the transfer. For each tag in
the list, you must specify both a tag key and a value. You can set the
value to an empty string, but you can't set it to \code{null}. For more
information about tagging, see \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_tagging.html}{Tagging Organizations resources}
in the Organizations User Guide.

Any tags in the request are checked for compliance with any applicable
tag policies when the request is made. The request is rejected if the
tags in the request don't match the requirements of the policy at that
time. Tag policy compliance is \emph{not} checked again when the invitation
is accepted and the tags are actually attached to the transfer. That
means that if the tag policy changes between the invitation and the
acceptance, then that tags could potentially be non-compliant.

If any one of the tags is not valid or if you exceed the allowed number
of tags for a transfer, then the entire request fails and invitations
are not sent.}
}
\description{
Sends an invitation to another organization's management account to designate your account with the specified responsibilities for their organization. The invitation is implemented as a Handshake whose details are in the response.

See \url{https://www.paws-r-sdk.com/docs/organizations_invite_organization_to_transfer_responsibility/} for full documentation.
}
\keyword{internal}
