% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_leave_organization}
\alias{organizations_leave_organization}
\title{Removes a member account from its parent organization}
\usage{
organizations_leave_organization()
}
\description{
Removes a member account from its parent organization. This version of the operation is performed by the account that wants to leave. To remove a member account as a user in the management account, use \code{\link[=organizations_remove_account_from_organization]{remove_account_from_organization}} instead.

See \url{https://paws-r.github.io/docs/organizations/leave_organization.html} for full documentation.
}
\keyword{internal}
