% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_accounts}
\alias{organizations_list_accounts}
\title{Lists all the accounts in the organization}
\usage{
organizations_list_accounts(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The total number of results that you want included on each page of the
response. If you do not include this parameter, it defaults to a value
that is specific to the operation. If additional items exist beyond the
maximum you specify, the \code{NextToken} response element is present and has
a value (is not null). Include that value as the \code{NextToken} request
parameter in the next call to the operation to get the next part of the
results. Note that Organizations might return fewer results than the
maximum even when there are more results available. You should check
\code{NextToken} after every operation to ensure that you receive all of the
results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Accounts = list(
    list(
      Id = "string",
      Arn = "string",
      Email = "string",
      Name = "string",
      Status = "ACTIVE"|"SUSPENDED",
      JoinedMethod = "INVITED"|"CREATED",
      JoinedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all the accounts in the organization. To request only the accounts
in a specified root or organizational unit (OU), use the
\code{\link[=organizations_list_accounts_for_parent]{list_accounts_for_parent}}
operation instead.

Always check the \code{NextToken} response parameter for a \code{null} value when
calling a \verb{List*} operation. These operations can occasionally return an
empty set of results even when there are more results available. The
\code{NextToken} response parameter value is \code{null} \emph{only} when there are no
more results to display.

This operation can be called only from the organization's management
account or by a member account that is a delegated administrator for an
AWS service.
}
\section{Request syntax}{
\preformatted{svc$list_accounts(
  NextToken = "string",
  MaxResults = 123
)
}
}

\examples{
\dontrun{
# The following example shows you how to request a list of the accounts in
# an organization:
svc$list_accounts()
}

}
\keyword{internal}
