% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_delegated_administrators}
\alias{organizations_list_delegated_administrators}
\title{Lists the Amazon Web Services accounts that are designated as delegated
administrators in this organization}
\usage{
organizations_list_delegated_administrators(
  ServicePrincipal = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ServicePrincipal}{Specifies a service principal name. If specified, then the operation
lists the delegated administrators only for the specified service.

If you don't specify a service principal, the operation lists all
delegated administrators for all services in your organization.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists the Amazon Web Services accounts that are designated as delegated administrators in this organization.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_delegated_administrators/} for full documentation.
}
\keyword{internal}
