% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_inbound_responsibility_transfers}
\alias{organizations_list_inbound_responsibility_transfers}
\title{Lists transfers that allow you to manage the specified responsibilities
for another organization}
\usage{
organizations_list_inbound_responsibility_transfers(
  Type,
  Id = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Type}{[required] The type of responsibility. Currently, only \code{BILLING} is supported.}

\item{Id}{ID for the transfer.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists transfers that allow you to manage the specified responsibilities for another organization. This operation returns both transfer invitations and transfers.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_inbound_responsibility_transfers/} for full documentation.
}
\keyword{internal}
