% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_tags_for_resource}
\alias{organizations_list_tags_for_resource}
\title{Lists tags for the specified resource}
\usage{
organizations_list_tags_for_resource(ResourceId, NextToken)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource that you want to retrieve tags for.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}
}
\description{
Lists tags for the specified resource.
}
\details{
Currently, you can list tags on an account in AWS Organizations.

This operation can be called only from the organization's master account
or by a member account that is a delegated administrator for an AWS
service.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
