% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_terminate_responsibility_transfer}
\alias{organizations_terminate_responsibility_transfer}
\title{Ends a transfer}
\usage{
organizations_terminate_responsibility_transfer(Id, EndTimestamp = NULL)
}
\arguments{
\item{Id}{[required] ID for the transfer.}

\item{EndTimestamp}{Timestamp when the responsibility transfer is to end.}
}
\description{
Ends a transfer. A \emph{transfer} is an arrangement between two management accounts where one account designates the other with specified responsibilities for their organization.

See \url{https://www.paws-r-sdk.com/docs/organizations_terminate_responsibility_transfer/} for full documentation.
}
\keyword{internal}
