% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_update_responsibility_transfer}
\alias{organizations_update_responsibility_transfer}
\title{Updates a transfer}
\usage{
organizations_update_responsibility_transfer(Id, Name)
}
\arguments{
\item{Id}{[required] ID for the transfer.}

\item{Name}{[required] New name you want to assign to the transfer.}
}
\description{
Updates a transfer. A \emph{transfer} is the arrangement between two management accounts where one account designates the other with specified responsibilities for their organization.

See \url{https://www.paws-r-sdk.com/docs/organizations_update_responsibility_transfer/} for full documentation.
}
\keyword{internal}
