% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_operations.R
\name{pi_get_resource_metadata}
\alias{pi_get_resource_metadata}
\title{Retrieve the metadata for different features}
\usage{
pi_get_resource_metadata(ServiceType, Identifier)
}
\arguments{
\item{ServiceType}{[required] The Amazon Web Services service for which Performance Insights returns
metrics.}

\item{Identifier}{[required] An immutable identifier for a data source that is unique for an Amazon
Web Services Region. Performance Insights gathers metrics from this data
source. To use a DB instance as a data source, specify its
\code{DbiResourceId} value. For example, specify
\code{db-ABCDEFGHIJKLMNOPQRSTU1VW2X}.}
}
\description{
Retrieve the metadata for different features. For example, the metadata might indicate that a feature is turned on or off on a specific DB instance.

See \url{https://www.paws-r-sdk.com/docs/pi_get_resource_metadata/} for full documentation.
}
\keyword{internal}
