% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_operations.R
\name{pi_list_tags_for_resource}
\alias{pi_list_tags_for_resource}
\title{Retrieves all the metadata tags associated with Amazon RDS Performance
Insights resource}
\usage{
pi_list_tags_for_resource(ServiceType, ResourceARN)
}
\arguments{
\item{ServiceType}{[required] List the tags for the Amazon Web Services service for which Performance
Insights returns metrics. Valid value is \code{RDS}.}

\item{ResourceARN}{[required] Lists all the tags for the Amazon RDS Performance Insights resource.
This value is an Amazon Resource Name (ARN). For information about
creating an ARN, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Tagging.ARN.html#USER_Tagging.ARN.Constructing}{Constructing an RDS Amazon Resource Name (ARN)}.}
}
\description{
Retrieves all the metadata tags associated with Amazon RDS Performance Insights resource.

See \url{https://www.paws-r-sdk.com/docs/pi_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
