% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_create_rule_groups_namespace}
\alias{prometheusservice_create_rule_groups_namespace}
\title{Create a rule group namespace}
\usage{
prometheusservice_create_rule_groups_namespace(
  workspaceId,
  name,
  data,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace in which to create the rule group namespace.}

\item{name}{[required] The rule groups namespace name.}

\item{data}{[required] The namespace data that define the rule groups.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{tags}{Optional, user-provided tags for this rule groups namespace.}
}
\description{
Create a rule group namespace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_create_rule_groups_namespace/} for full documentation.
}
\keyword{internal}
