% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_create_scraper}
\alias{prometheusservice_create_scraper}
\title{Creates a scraper to collect metrics from Prometheus-compatible sources}
\usage{
prometheusservice_create_scraper(
  alias = NULL,
  scrapeConfiguration,
  source,
  destination,
  roleConfiguration = NULL,
  clientToken = NULL,
  tags = NULL,
  exporters = NULL
)
}
\arguments{
\item{alias}{(optional) An alias to associate with the scraper. This is for your use, and does not need to be unique.}

\item{scrapeConfiguration}{[required] The configuration file to use in the new scraper. For more information, see \href{https://docs.aws.amazon.com/prometheus/latest/userguide/AMP-collector-how-to.html#AMP-collector-configuration}{Scraper configuration} in the \emph{Amazon Managed Service for Prometheus User Guide}.}

\item{source}{[required] The Amazon EKS or Amazon Web Services cluster from which the scraper will collect metrics.}

\item{destination}{[required] The destination where the scraper sends the collected metrics. Valid destinations are Amazon Managed Service for Prometheus workspaces and CloudWatch datasets.}

\item{roleConfiguration}{Use this structure to enable cross-account access, so that you can use a target account to access Prometheus metrics from source accounts.}

\item{clientToken}{(Optional) A unique, case-sensitive identifier that you can provide to ensure the idempotency of the request.}

\item{tags}{(Optional) The list of tag keys and values to associate with the scraper.}

\item{exporters}{The exporter configurations for the scraper. You can configure at most one Amazon OpenSearch Service domain. If you don't specify a value, the scraper is created without an exporter configuration.}
}
\description{
Creates a scraper to collect metrics from Prometheus-compatible sources. The scraper sends the collected metrics to Amazon Managed Service for Prometheus workspaces or CloudWatch datasets. You can configure scrapers to collect metrics from Amazon EKS clusters, Amazon MSK clusters, or from VPC-based sources that support DNS-based service discovery. Scrapers are flexible. You can configure a scraper to control which metrics to collect, the frequency of collection, which transformations to apply to the metrics, and more.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_create_scraper/} for full documentation.
}
\keyword{internal}
