% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_anomaly_detector}
\alias{prometheusservice_delete_anomaly_detector}
\title{Removes an anomaly detector from a workspace}
\usage{
prometheusservice_delete_anomaly_detector(
  workspaceId,
  anomalyDetectorId,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The identifier of the workspace containing the anomaly detector to
delete.}

\item{anomalyDetectorId}{[required] The identifier of the anomaly detector to delete.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Removes an anomaly detector from a workspace. This operation is idempotent.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_anomaly_detector/} for full documentation.
}
\keyword{internal}
