% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_logging_configuration}
\alias{prometheusservice_delete_logging_configuration}
\title{Delete logging configuration}
\usage{
prometheusservice_delete_logging_configuration(workspaceId, clientToken = NULL)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to vend logs to.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}
}
\description{
Delete logging configuration.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_logging_configuration/} for full documentation.
}
\keyword{internal}
