% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_query_logging_configuration}
\alias{prometheusservice_delete_query_logging_configuration}
\title{Deletes the query logging configuration for the specified workspace}
\usage{
prometheusservice_delete_query_logging_configuration(
  workspaceId,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace from which to delete the query logging
configuration.}

\item{clientToken}{(Optional) A unique, case-sensitive identifier that you can provide to
ensure the idempotency of the request.}
}
\description{
Deletes the query logging configuration for the specified workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_query_logging_configuration/} for full documentation.
}
\keyword{internal}
