% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_resource_policy}
\alias{prometheusservice_delete_resource_policy}
\title{Deletes the resource-based policy attached to an Amazon Managed Service
for Prometheus workspace}
\usage{
prometheusservice_delete_resource_policy(
  workspaceId,
  clientToken = NULL,
  revisionId = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace from which to delete the resource-based policy.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
request is safe to retry (idempotent).}

\item{revisionId}{The revision ID of the policy to delete. Use this parameter to ensure
that you are deleting the correct version of the policy.}
}
\description{
Deletes the resource-based policy attached to an Amazon Managed Service for Prometheus workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_resource_policy/} for full documentation.
}
\keyword{internal}
