% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_rule_groups_namespace}
\alias{prometheusservice_delete_rule_groups_namespace}
\title{Delete a rule groups namespace}
\usage{
prometheusservice_delete_rule_groups_namespace(
  workspaceId,
  name,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to delete rule group definition.}

\item{name}{[required] The rule groups namespace name.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}
}
\description{
Delete a rule groups namespace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_rule_groups_namespace/} for full documentation.
}
\keyword{internal}
