% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_scraper_logging_configuration}
\alias{prometheusservice_delete_scraper_logging_configuration}
\title{Deletes the logging configuration for a Amazon Managed Service for
Prometheus scraper}
\usage{
prometheusservice_delete_scraper_logging_configuration(
  scraperId,
  clientToken = NULL
)
}
\arguments{
\item{scraperId}{[required] The ID of the scraper whose logging configuration will be deleted.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
request is processed exactly once.}
}
\description{
Deletes the logging configuration for a Amazon Managed Service for Prometheus scraper.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_scraper_logging_configuration/} for full documentation.
}
\keyword{internal}
