% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_alert_manager_definition}
\alias{prometheusservice_describe_alert_manager_definition}
\title{Retrieves the full information about the alert manager definition for a
workspace}
\usage{
prometheusservice_describe_alert_manager_definition(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to retrieve the alert manager definition from.}
}
\description{
Retrieves the full information about the alert manager definition for a workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_alert_manager_definition/} for full documentation.
}
\keyword{internal}
