% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_anomaly_detector}
\alias{prometheusservice_describe_anomaly_detector}
\title{Retrieves detailed information about a specific anomaly detector,
including its status and configuration}
\usage{
prometheusservice_describe_anomaly_detector(workspaceId, anomalyDetectorId)
}
\arguments{
\item{workspaceId}{[required] The identifier of the workspace containing the anomaly detector.}

\item{anomalyDetectorId}{[required] The identifier of the anomaly detector to describe.}
}
\description{
Retrieves detailed information about a specific anomaly detector, including its status and configuration.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_anomaly_detector/} for full documentation.
}
\keyword{internal}
