% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_query_logging_configuration}
\alias{prometheusservice_describe_query_logging_configuration}
\title{Retrieves the details of the query logging configuration for the
specified workspace}
\usage{
prometheusservice_describe_query_logging_configuration(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace for which to retrieve the query logging
configuration.}
}
\description{
Retrieves the details of the query logging configuration for the specified workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_query_logging_configuration/} for full documentation.
}
\keyword{internal}
