% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_resource_policy}
\alias{prometheusservice_describe_resource_policy}
\title{Returns information about the resource-based policy attached to an
Amazon Managed Service for Prometheus workspace}
\usage{
prometheusservice_describe_resource_policy(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to describe the resource-based policy for.}
}
\description{
Returns information about the resource-based policy attached to an Amazon Managed Service for Prometheus workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_resource_policy/} for full documentation.
}
\keyword{internal}
