% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_rule_groups_namespace}
\alias{prometheusservice_describe_rule_groups_namespace}
\title{Describe a rule groups namespace}
\usage{
prometheusservice_describe_rule_groups_namespace(name, workspaceId)
}
\arguments{
\item{name}{[required] The rule groups namespace.}

\item{workspaceId}{[required] The ID of the workspace to describe.}
}
\description{
Describe a rule groups namespace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_rule_groups_namespace/} for full documentation.
}
\keyword{internal}
