% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_scraper_logging_configuration}
\alias{prometheusservice_describe_scraper_logging_configuration}
\title{Describes the logging configuration for a Amazon Managed Service for
Prometheus scraper}
\usage{
prometheusservice_describe_scraper_logging_configuration(scraperId)
}
\arguments{
\item{scraperId}{[required] The ID of the scraper whose logging configuration will be described.}
}
\description{
Describes the logging configuration for a Amazon Managed Service for Prometheus scraper.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_scraper_logging_configuration/} for full documentation.
}
\keyword{internal}
