% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_workspace}
\alias{prometheusservice_describe_workspace}
\title{Describes an existing AMP workspace}
\usage{
prometheusservice_describe_workspace(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to describe.}
}
\description{
Describes an existing AMP workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_workspace/} for full documentation.
}
\keyword{internal}
