% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_workspace_configuration}
\alias{prometheusservice_describe_workspace_configuration}
\title{Use this operation to return information about the configuration of a
workspace}
\usage{
prometheusservice_describe_workspace_configuration(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace that you want to retrieve information for. To
find the IDs of your workspaces, use the
\code{\link[=prometheusservice_list_workspaces]{list_workspaces}} operation.}
}
\description{
Use this operation to return information about the configuration of a workspace. The configuration details returned include workspace configuration status, label set limits, and retention period.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_workspace_configuration/} for full documentation.
}
\keyword{internal}
