% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_list_anomaly_detectors}
\alias{prometheusservice_list_anomaly_detectors}
\title{Returns a paginated list of anomaly detectors for a workspace with
optional filtering by alias}
\usage{
prometheusservice_list_anomaly_detectors(
  workspaceId,
  alias = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The identifier of the workspace containing the anomaly detectors to
list.}

\item{alias}{Filters the results to anomaly detectors with the specified alias.}

\item{maxResults}{The maximum number of results to return in a single call. Valid range is
1 to 1000.}

\item{nextToken}{The pagination token to continue retrieving results.}
}
\description{
Returns a paginated list of anomaly detectors for a workspace with optional filtering by alias.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_list_anomaly_detectors/} for full documentation.
}
\keyword{internal}
