% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_list_rule_groups_namespaces}
\alias{prometheusservice_list_rule_groups_namespaces}
\title{Lists rule groups namespaces}
\usage{
prometheusservice_list_rule_groups_namespaces(
  workspaceId,
  name = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace.}

\item{name}{Optional filter for rule groups namespace name. Only the rule groups
namespace that begin with this value will be returned.}

\item{nextToken}{Pagination token to request the next page in a paginated list. This
token is obtained from the output of the previous
ListRuleGroupsNamespaces request.}

\item{maxResults}{Maximum results to return in response (default=100, maximum=1000).}
}
\description{
Lists rule groups namespaces.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_list_rule_groups_namespaces/} for full documentation.
}
\keyword{internal}
