% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_list_scrapers}
\alias{prometheusservice_list_scrapers}
\title{Lists all scrapers in a customer account, including scrapers being
created or deleted}
\usage{
prometheusservice_list_scrapers(
  filters = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filters}{A list of scraper filters.}

\item{nextToken}{Pagination token to request the next page in a paginated list. This
token is obtained from the output of the previous ListScrapers request.}

\item{maxResults}{Maximum results to return in response (default=100, maximum=1000).}
}
\description{
Lists all scrapers in a customer account, including scrapers being created or deleted. You may provide filters to return a more specific list of results.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_list_scrapers/} for full documentation.
}
\keyword{internal}
