% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_put_anomaly_detector}
\alias{prometheusservice_put_anomaly_detector}
\title{When you call PutAnomalyDetector, the operation creates a new anomaly
detector if one doesn't exist, or updates an existing one}
\usage{
prometheusservice_put_anomaly_detector(
  workspaceId,
  anomalyDetectorId,
  evaluationIntervalInSeconds = NULL,
  missingDataAction = NULL,
  configuration,
  labels = NULL,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The identifier of the workspace containing the anomaly detector to
update.}

\item{anomalyDetectorId}{[required] The identifier of the anomaly detector to update.}

\item{evaluationIntervalInSeconds}{The frequency, in seconds, at which the anomaly detector evaluates
metrics.}

\item{missingDataAction}{Specifies the action to take when data is missing during evaluation.}

\item{configuration}{[required] The algorithm configuration for the anomaly detector.}

\item{labels}{The Amazon Managed Service for Prometheus metric labels to associate
with the anomaly detector.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
When you call \code{\link[=prometheusservice_put_anomaly_detector]{put_anomaly_detector}}, the operation creates a new anomaly detector if one doesn't exist, or updates an existing one. Each call to this operation triggers a complete retraining of the detector, which includes querying the minimum required samples and backfilling the detector with historical data. This process occurs regardless of whether you're making a minor change like updating the evaluation interval or making more substantial modifications. The operation serves as the single method for creating, updating, and retraining anomaly detectors.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_put_anomaly_detector/} for full documentation.
}
\keyword{internal}
