% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_put_rule_groups_namespace}
\alias{prometheusservice_put_rule_groups_namespace}
\title{Update a rule groups namespace}
\usage{
prometheusservice_put_rule_groups_namespace(
  clientToken = NULL,
  data,
  name,
  workspaceId
)
}
\arguments{
\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{data}{[required] The namespace data that define the rule groups.}

\item{name}{[required] The rule groups namespace name.}

\item{workspaceId}{[required] The ID of the workspace in which to update the rule group namespace.}
}
\description{
Update a rule groups namespace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_put_rule_groups_namespace/} for full documentation.
}
\keyword{internal}
