% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_logging_configuration}
\alias{prometheusservice_update_logging_configuration}
\title{Update logging configuration}
\usage{
prometheusservice_update_logging_configuration(
  workspaceId,
  logGroupArn,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to vend logs to.}

\item{logGroupArn}{[required] The ARN of the CW log group to which the vended log data will be
published.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}
}
\description{
Update logging configuration.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_logging_configuration/} for full documentation.
}
\keyword{internal}
