% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_query_logging_configuration}
\alias{prometheusservice_update_query_logging_configuration}
\title{Updates the query logging configuration for the specified workspace}
\usage{
prometheusservice_update_query_logging_configuration(
  workspaceId,
  destinations,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace for which to update the query logging
configuration.}

\item{destinations}{[required] The destinations where query logs will be sent. Only CloudWatch Logs
destination is supported. The list must contain exactly one element.}

\item{clientToken}{(Optional) A unique, case-sensitive identifier that you can provide to
ensure the idempotency of the request.}
}
\description{
Updates the query logging configuration for the specified workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_query_logging_configuration/} for full documentation.
}
\keyword{internal}
