% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_scraper}
\alias{prometheusservice_update_scraper}
\title{Updates an existing scraper}
\usage{
prometheusservice_update_scraper(
  scraperId,
  alias = NULL,
  scrapeConfiguration = NULL,
  destination = NULL,
  roleConfiguration = NULL,
  clientToken = NULL,
  exporters = NULL
)
}
\arguments{
\item{scraperId}{[required] The ID of the scraper to update.}

\item{alias}{The new alias of the scraper.}

\item{scrapeConfiguration}{Contains the base-64 encoded YAML configuration for the scraper.

For more information about configuring a scraper, see \href{https://docs.aws.amazon.com/prometheus/latest/userguide/AMP-collector-how-to.html}{Using an Amazon Web Services managed collector} in the \emph{Amazon Managed Service for Prometheus User Guide}.}

\item{destination}{The new destination where the scraper sends metrics. Valid destinations are Amazon Managed Service for Prometheus workspaces and CloudWatch datasets.}

\item{roleConfiguration}{Use this structure to enable cross-account access, so that you can use a target account to access Prometheus metrics from source accounts.}

\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of the request. Case-sensitive.}

\item{exporters}{The exporter configurations for the scraper. You can configure at most one Amazon OpenSearch Service domain. If you don't specify a value, the existing exporter configuration remains unchanged.}
}
\description{
Updates an existing scraper.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_scraper/} for full documentation.
}
\keyword{internal}
