% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_scraper_logging_configuration}
\alias{prometheusservice_update_scraper_logging_configuration}
\title{Updates the logging configuration for a Amazon Managed Service for
Prometheus scraper}
\usage{
prometheusservice_update_scraper_logging_configuration(
  scraperId,
  loggingDestination,
  scraperComponents = NULL
)
}
\arguments{
\item{scraperId}{[required] The ID of the scraper whose logging configuration will be updated.}

\item{loggingDestination}{[required] The destination where scraper logs will be sent.}

\item{scraperComponents}{The list of scraper components to configure for logging.}
}
\description{
Updates the logging configuration for a Amazon Managed Service for Prometheus scraper.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_scraper_logging_configuration/} for full documentation.
}
\keyword{internal}
