% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_workspace_configuration}
\alias{prometheusservice_update_workspace_configuration}
\title{Use this operation to create or update the label sets, label set limits,
and retention period of a workspace}
\usage{
prometheusservice_update_workspace_configuration(
  workspaceId,
  clientToken = NULL,
  limitsPerLabelSet = NULL,
  retentionPeriodInDays = NULL,
  outOfOrderTimeWindowInSeconds = NULL,
  ruleQueryOffsetInSeconds = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace that you want to update. To find the IDs of your workspaces, use the \code{\link[=prometheusservice_list_workspaces]{list_workspaces}} operation.}

\item{clientToken}{You can include a token in your operation to make it an idempotent opeartion.}

\item{limitsPerLabelSet}{This is an array of structures, where each structure defines a label set for the workspace, and defines the active time series limit for each of those label sets. Each label name in a label set must be unique.}

\item{retentionPeriodInDays}{Specifies how many days that metrics will be retained in the workspace.}

\item{outOfOrderTimeWindowInSeconds}{Specifies the time window in seconds for accepting out of order samples. Out of order samples older than this window are rejected.}

\item{ruleQueryOffsetInSeconds}{Specifies the duration in seconds to offset rule evaluation queries into the past. This allows ingested samples to be available before rule evaluation.}
}
\description{
Use this operation to create or update the label sets, label set limits, and retention period of a workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_workspace_configuration/} for full documentation.
}
\keyword{internal}
