% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_app_version_app_component}
\alias{resiliencehub_describe_app_version_app_component}
\title{Describes an Application Component in the Resilience Hub application}
\usage{
resiliencehub_describe_app_version_app_component(appArn, appVersion, id)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format for this ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon Resource Names (ARNs)} in the \emph{Amazon Web Services General Reference} guide.}

\item{appVersion}{[required] Resilience Hub application version.}

\item{id}{[required] Identifier of the Application Component.}
}
\description{
Describes an Application Component in the Resilience Hub application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_describe_app_version_app_component/} for full documentation.
}
\keyword{internal}
