% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_metrics_export}
\alias{resiliencehub_describe_metrics_export}
\title{Describes the metrics of the application configuration being exported}
\usage{
resiliencehub_describe_metrics_export(metricsExportId)
}
\arguments{
\item{metricsExportId}{[required] Identifier of the metrics export task.}
}
\description{
Describes the metrics of the application configuration being exported.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_describe_metrics_export/} for full documentation.
}
\keyword{internal}
