% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_resource_grouping_recommendation_task}
\alias{resiliencehub_describe_resource_grouping_recommendation_task}
\title{Describes the resource grouping recommendation tasks run by Resilience
Hub for your application}
\usage{
resiliencehub_describe_resource_grouping_recommendation_task(
  appArn,
  groupingId = NULL
)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format for this ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon Resource Names (ARNs)} in the \emph{Amazon Web Services General Reference} guide.}

\item{groupingId}{Identifier of the grouping recommendation task.}
}
\description{
Describes the resource grouping recommendation tasks run by Resilience Hub for your application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_describe_resource_grouping_recommendation_task/} for full documentation.
}
\keyword{internal}
