% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_list_metrics}
\alias{resiliencehub_list_metrics}
\title{Lists the metrics that can be exported}
\usage{
resiliencehub_list_metrics(
  conditions = NULL,
  dataSource = NULL,
  fields = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sorts = NULL
)
}
\arguments{
\item{conditions}{Indicates the list of all the conditions that were applied on the
metrics.}

\item{dataSource}{Indicates the data source of the metrics.}

\item{fields}{Indicates the list of fields in the data source.}

\item{maxResults}{Maximum number of results to include in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that the remaining results can be retrieved.}

\item{nextToken}{Null, or the token from a previous call to get the next set of results.}

\item{sorts}{(Optional) Indicates the order in which you want to sort the fields in
the metrics. By default, the fields are sorted in the ascending order.}
}
\description{
Lists the metrics that can be exported.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_list_metrics/} for full documentation.
}
\keyword{internal}
