% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_list_suggested_resiliency_policies}
\alias{resiliencehub_list_suggested_resiliency_policies}
\title{Lists the suggested resiliency policies for the Resilience Hub
applications}
\usage{
resiliencehub_list_suggested_resiliency_policies(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{Maximum number of results to include in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that the remaining results can be retrieved.}

\item{nextToken}{Null, or the token from a previous call to get the next set of results.}
}
\description{
Lists the suggested resiliency policies for the Resilience Hub applications.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_list_suggested_resiliency_policies/} for full documentation.
}
\keyword{internal}
