% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_publish_app_version}
\alias{resiliencehub_publish_app_version}
\title{Publishes a new version of a specific Resilience Hub application}
\usage{
resiliencehub_publish_app_version(appArn, versionName = NULL)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format for this ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon Resource Names (ARNs)} in the \emph{Amazon Web Services General Reference} guide.}

\item{versionName}{Name of the application version.}
}
\description{
Publishes a new version of a specific Resilience Hub application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_publish_app_version/} for full documentation.
}
\keyword{internal}
