% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_start_metrics_export}
\alias{resiliencehub_start_metrics_export}
\title{Initiates the export task of metrics}
\usage{
resiliencehub_start_metrics_export(bucketName = NULL, clientToken = NULL)
}
\arguments{
\item{bucketName}{(Optional) Specifies the name of the Amazon Simple Storage Service
bucket where the exported metrics will be stored.}

\item{clientToken}{Used for an idempotency token. A client token is a unique,
case-sensitive string of up to 64 ASCII characters. You should not reuse
the same client token for other API requests.}
}
\description{
Initiates the export task of metrics.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_start_metrics_export/} for full documentation.
}
\keyword{internal}
