% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_tag_resource}
\alias{resiliencehub_tag_resource}
\title{Applies one or more tags to a resource}
\usage{
resiliencehub_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags to assign to the resource. Each tag consists of a key/value
pair.}
}
\description{
Applies one or more tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_tag_resource/} for full documentation.
}
\keyword{internal}
