% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_untag_resource}
\alias{resiliencehub_untag_resource}
\title{Removes one or more tags from a resource}
\usage{
resiliencehub_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The keys of the tags you want to remove.}
}
\description{
Removes one or more tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_untag_resource/} for full documentation.
}
\keyword{internal}
