% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_cancel_tag_sync_task}
\alias{resourcegroups_cancel_tag_sync_task}
\title{Cancels the specified tag-sync task}
\usage{
resourcegroups_cancel_tag_sync_task(TaskArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon resource name (ARN) of the tag-sync task.}
}
\description{
Cancels the specified tag-sync task.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_cancel_tag_sync_task/} for full documentation.
}
\keyword{internal}
