% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_delete_group}
\alias{resourcegroups_delete_group}
\title{Deletes a specified resource group}
\usage{
resourcegroups_delete_group(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the resource group to delete.}
}
\description{
Deletes a specified resource group. Deleting a resource group does not
delete resources that are members of the group; it only deletes the
group structure.
}
\section{Request syntax}{
\preformatted{svc$delete_group(
  GroupName = "string"
)
}
}

\keyword{internal}
