% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_account_settings}
\alias{resourcegroups_get_account_settings}
\title{Retrieves the current status of optional features in Resource Groups}
\usage{
resourcegroups_get_account_settings()
}
\description{
Retrieves the current status of optional features in Resource Groups.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_get_account_settings/} for full documentation.
}
\keyword{internal}
