% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group}
\alias{resourcegroups_get_group}
\title{Returns information about a specified resource group}
\usage{
resourcegroups_get_group(GroupName = NULL, Group = NULL)
}
\arguments{
\item{GroupName}{Deprecated - don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the Amazon resource name (ARN) of the resource group to
retrieve.}
}
\description{
Returns information about a specified resource group.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_get_group/} for full documentation.
}
\keyword{internal}
