% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group_configuration}
\alias{resourcegroups_get_group_configuration}
\title{Retrieves the service configuration associated with the specified
resource group}
\usage{
resourcegroups_get_group_configuration(Group = NULL)
}
\arguments{
\item{Group}{The name or the Amazon resource name (ARN) of the resource group for
which you want to retrive the service configuration.}
}
\description{
Retrieves the service configuration associated with the specified resource group. For details about the service configuration syntax, see \href{https://docs.aws.amazon.com/ARG/latest/userguide/about-slg.html}{Service configurations for Resource Groups}.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_get_group_configuration/} for full documentation.
}
\keyword{internal}
