% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_tags}
\alias{resourcegroups_get_tags}
\title{Returns a list of tags that are associated with a resource group,
specified by an ARN}
\usage{
resourcegroups_get_tags(Arn)
}
\arguments{
\item{Arn}{[required] The ARN of the resource group whose tags you want to retrieve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of tags that are associated with a resource group,
specified by an ARN.

\strong{Minimum permissions}

To run this command, you must have the following permissions:
\itemize{
\item \code{resource-groups:GetTags}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_tags(
  Arn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
