% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_list_grouping_statuses}
\alias{resourcegroups_list_grouping_statuses}
\title{Returns the status of the last grouping or ungrouping action for each
resource in the specified application group}
\usage{
resourcegroups_list_grouping_statuses(
  Group,
  MaxResults = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Group}{[required] The application group identifier, expressed as an Amazon resource name
(ARN) or the application group name.}

\item{MaxResults}{The maximum number of resources and their statuses returned in the
response.}

\item{Filters}{The filter name and value pair that is used to return more specific
results from a list of resources.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
provided by a previous call's \code{NextToken} response to indicate where the
output should continue from.}
}
\description{
Returns the status of the last grouping or ungrouping action for each resource in the specified application group.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_list_grouping_statuses/} for full documentation.
}
\keyword{internal}
