% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_list_tag_sync_tasks}
\alias{resourcegroups_list_tag_sync_tasks}
\title{Returns a list of tag-sync tasks}
\usage{
resourcegroups_list_tag_sync_tasks(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{The Amazon resource name (ARN) or name of the application group for
which you want to return a list of tag-sync tasks.}

\item{MaxResults}{The maximum number of results to be included in the response.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
provided by a previous call's \code{NextToken} response to indicate where the
output should continue from.}
}
\description{
Returns a list of tag-sync tasks.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_list_tag_sync_tasks/} for full documentation.
}
\keyword{internal}
