% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_start_tag_sync_task}
\alias{resourcegroups_start_tag_sync_task}
\title{Creates a new tag-sync task to onboard and sync resources tagged with a
specific tag key-value pair to an application}
\usage{
resourcegroups_start_tag_sync_task(
  Group,
  TagKey = NULL,
  TagValue = NULL,
  ResourceQuery = NULL,
  RoleArn
)
}
\arguments{
\item{Group}{[required] The Amazon resource name (ARN) or name of the application group for
which you want to create a tag-sync task.}

\item{TagKey}{The tag key. Resources tagged with this tag key-value pair will be added
to the application. If a resource with this tag is later untagged, the
tag-sync task removes the resource from the application.

When using the \code{TagKey} parameter, you must also specify the \code{TagValue}
parameter. If you specify a tag key-value pair, you can't use the
\code{ResourceQuery} parameter.}

\item{TagValue}{The tag value. Resources tagged with this tag key-value pair will be
added to the application. If a resource with this tag is later untagged,
the tag-sync task removes the resource from the application.

When using the \code{TagValue} parameter, you must also specify the \code{TagKey}
parameter. If you specify a tag key-value pair, you can't use the
\code{ResourceQuery} parameter.}

\item{ResourceQuery}{The query you can use to create the tag-sync task. With this method, all
resources matching the query are added to the specified application
group. A \code{ResourceQuery} specifies both a query \code{Type} and a \code{Query}
string as JSON string objects. For more information on defining a
resource query for a tag-sync task, see the tag-based query type in
\href{https://docs.aws.amazon.com/ARG/latest/userguide/gettingstarted-query.html#getting_started-query_types}{Types of resource group queries}
in \emph{Resource Groups User Guide}.

When using the \code{ResourceQuery} parameter, you cannot use the \code{TagKey}
and \code{TagValue} parameters.

When you combine all of the elements together into a single string, any
double quotes that are embedded inside another double quote pair must be
escaped by preceding the embedded double quote with a backslash
character (\\). For example, a complete \code{ResourceQuery} parameter must
be formatted like the following CLI parameter example:

\verb{--resource-query '\{"Type":"TAG_FILTERS_1_0","Query":"\{\\"ResourceTypeFilters\\":[\\"AWS::AllSupported\\"],\\"TagFilters\\":[\{\\"Key\\":\\"Stage\\",\\"Values\\":[\\"Test\\"]\}]\}"\}'}

In the preceding example, all of the double quote characters in the
value part of the \code{Query} element must be escaped because the value
itself is surrounded by double quotes. For more information, see
\href{https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-parameters-quoting-strings.html}{Quoting strings}
in the \emph{Command Line Interface User Guide}.

For the complete list of resource types that you can use in the array
value for \code{ResourceTypeFilters}, see \href{https://docs.aws.amazon.com/ARG/latest/userguide/supported-resources.html}{Resources you can use with Resource Groups and Tag Editor}
in the \emph{Resource Groups User Guide}. For example:

\verb{"ResourceTypeFilters":["AWS::S3::Bucket", "AWS::EC2::Instance"]}}

\item{RoleArn}{[required] The Amazon resource name (ARN) of the role assumed by the service to tag
and untag resources on your behalf.}
}
\description{
Creates a new tag-sync task to onboard and sync resources tagged with a specific tag key-value pair to an application. To start a tag-sync task, you need a \href{https://docs.aws.amazon.com/servicecatalog/latest/arguide/app-tag-sync.html#tag-sync-role}{resource tagging role}. The resource tagging role grants permissions to tag and untag applications resources and must include a trust policy that allows Resource Groups to assume the role and perform resource tagging tasks on your behalf.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_start_tag_sync_task/} for full documentation.
}
\keyword{internal}
