% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_update_group}
\alias{resourcegroups_update_group}
\title{Updates the description for an existing group}
\usage{
resourcegroups_update_group(
  GroupName = NULL,
  Group = NULL,
  Description = NULL,
  Criticality = NULL,
  Owner = NULL,
  DisplayName = NULL
)
}
\arguments{
\item{GroupName}{Don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to update.}

\item{Description}{The new description that you want to update the resource group with.
Descriptions can contain letters, numbers, hyphens, underscores,
periods, and spaces.}

\item{Criticality}{The critical rank of the application group on a scale of 1 to 10, with a
rank of 1 being the most critical, and a rank of 10 being least
critical.}

\item{Owner}{A name, email address or other identifier for the person or group who is
considered as the owner of this application group within your
organization.}

\item{DisplayName}{The name of the application group, which you can change at any time.}
}
\description{
Updates the description for an existing group. You cannot update the name of a resource group.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_update_group/} for full documentation.
}
\keyword{internal}
