% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_list_required_tags}
\alias{resourcegroupstaggingapi_list_required_tags}
\title{Lists the required tags for supported resource types in an Amazon Web
Services account}
\usage{
resourcegroupstaggingapi_list_required_tags(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{A token for requesting another page of required tags if the \code{NextToken}
response element indicates that more required tags are available. Use
the value of the returned \code{NextToken} element in your request until the
token comes back as null. Pass null if this is the first call.}

\item{MaxResults}{The maximum number of required tags.}
}
\description{
Lists the required tags for supported resource types in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/resourcegroupstaggingapi_list_required_tags/} for full documentation.
}
\keyword{internal}
