% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_tag_resources}
\alias{resourcegroupstaggingapi_tag_resources}
\title{Applies one or more tags to the specified resources}
\usage{
resourcegroupstaggingapi_tag_resources(ResourceARNList, Tags)
}
\arguments{
\item{ResourceARNList}{[required] Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces} in the \emph{Amazon Web Services General Reference}.}

\item{Tags}{[required] Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.}
}
\description{
Applies one or more tags to the specified resources. Note the following:

See \url{https://www.paws-r-sdk.com/docs/resourcegroupstaggingapi_tag_resources/} for full documentation.
}
\keyword{internal}
