% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_service_action_with_provisioning_artifact}
\alias{servicecatalog_associate_service_action_with_provisioning_artifact}
\title{Associates a self-service action with a provisioning artifact}
\usage{
servicecatalog_associate_service_action_with_provisioning_artifact(
  ProductId,
  ProvisioningArtifactId,
  ServiceActionId,
  AcceptLanguage = NULL,
  IdempotencyToken = NULL
)
}
\arguments{
\item{ProductId}{[required] The product identifier. For example, \code{prod-abcdzk7xy33qa}.}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact. For example,
\verb{pa-4abcdjnxjj6ne}.}

\item{ServiceActionId}{[required] The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{IdempotencyToken}{A unique identifier that you provide to ensure idempotency. If multiple
requests from the same Amazon Web Services account use the same
idempotency token, the same response is returned for each repeated
request.}
}
\description{
Associates a self-service action with a provisioning artifact.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_associate_service_action_with_provisioning_artifact/} for full documentation.
}
\keyword{internal}
