% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_budget_with_resource}
\alias{servicecatalog_associate_budget_with_resource}
\title{Associates the specified budget with the specified resource}
\usage{
servicecatalog_associate_budget_with_resource(BudgetName, ResourceId)
}
\arguments{
\item{BudgetName}{[required] The name of the budget you want to associate.}

\item{ResourceId}{[required] The resource identifier. Either a portfolio-id or a product-id.}
}
\description{
Associates the specified budget with the specified resource.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_associate_budget_with_resource/} for full documentation.
}
\keyword{internal}
