% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_principal_with_portfolio}
\alias{servicecatalog_associate_principal_with_portfolio}
\title{Associates the specified principal ARN with the specified portfolio}
\usage{
servicecatalog_associate_principal_with_portfolio(
  AcceptLanguage = NULL,
  PortfolioId,
  PrincipalARN,
  PrincipalType
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{PrincipalARN}{[required] The ARN of the principal (user, role, or group). This field allows an
ARN with no \code{accountID} if \code{PrincipalType} is \code{IAM_PATTERN}.

You can associate multiple \code{IAM} patterns even if the account has no
principal with that name. This is useful in Principal Name Sharing if
you want to share a principal without creating it in the account that
owns the portfolio.}

\item{PrincipalType}{[required] The principal type. The supported value is \code{IAM} if you use a fully
defined ARN, or \code{IAM_PATTERN} if you use an ARN with no \code{accountID}.}
}
\description{
Associates the specified principal ARN with the specified portfolio.

See \url{https://paws-r.github.io/docs/servicecatalog/associate_principal_with_portfolio.html} for full documentation.
}
\keyword{internal}
