% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_product_with_portfolio}
\alias{servicecatalog_associate_product_with_portfolio}
\title{Associates the specified product with the specified portfolio}
\usage{
servicecatalog_associate_product_with_portfolio(AcceptLanguage,
  ProductId, PortfolioId, SourcePortfolioId)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{SourcePortfolioId}{The identifier of the source portfolio.}
}
\value{
An empty list.
}
\description{
Associates the specified product with the specified portfolio.

A delegated admin is authorized to invoke this command.
}
\section{Request syntax}{
\preformatted{svc$associate_product_with_portfolio(
  AcceptLanguage = "string",
  ProductId = "string",
  PortfolioId = "string",
  SourcePortfolioId = "string"
)
}
}

\keyword{internal}
