% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_tag_option_with_resource}
\alias{servicecatalog_associate_tag_option_with_resource}
\title{Associate the specified TagOption with the specified portfolio or
product}
\usage{
servicecatalog_associate_tag_option_with_resource(ResourceId,
  TagOptionId)
}
\arguments{
\item{ResourceId}{[required] The resource identifier.}

\item{TagOptionId}{[required] The TagOption identifier.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Associate the specified TagOption with the specified portfolio or
product.
}
\section{Request syntax}{
\preformatted{svc$associate_tag_option_with_resource(
  ResourceId = "string",
  TagOptionId = "string"
)
}
}

\keyword{internal}
