% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_batch_disassociate_service_action_from_provisioning_artifact}
\alias{servicecatalog_batch_disassociate_service_action_from_provisioning_artifact}
\title{Disassociates a batch of self-service actions from the specified
provisioning artifact}
\usage{
servicecatalog_batch_disassociate_service_action_from_provisioning_artifact(
  ServiceActionAssociations, AcceptLanguage)
}
\arguments{
\item{ServiceActionAssociations}{[required] One or more associations, each consisting of the Action ID, the Product
ID, and the Provisioning Artifact ID.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\value{
A list with the following syntax:\preformatted{list(
  FailedServiceActionAssociations = list(
    list(
      ServiceActionId = "string",
      ProductId = "string",
      ProvisioningArtifactId = "string",
      ErrorCode = "DUPLICATE_RESOURCE"|"INTERNAL_FAILURE"|"LIMIT_EXCEEDED"|"RESOURCE_NOT_FOUND"|"THROTTLING",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Disassociates a batch of self-service actions from the specified
provisioning artifact.
}
\section{Request syntax}{
\preformatted{svc$batch_disassociate_service_action_from_provisioning_artifact(
  ServiceActionAssociations = list(
    list(
      ServiceActionId = "string",
      ProductId = "string",
      ProvisioningArtifactId = "string"
    )
  ),
  AcceptLanguage = "string"
)
}
}

\keyword{internal}
