% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_portfolio}
\alias{servicecatalog_delete_portfolio}
\title{Deletes the specified portfolio}
\usage{
servicecatalog_delete_portfolio(AcceptLanguage = NULL, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}
}
\description{
Deletes the specified portfolio.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_delete_portfolio/} for full documentation.
}
\keyword{internal}
