% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_portfolio_share}
\alias{servicecatalog_delete_portfolio_share}
\title{Stops sharing the specified portfolio with the specified account or
organization node}
\usage{
servicecatalog_delete_portfolio_share(
  AcceptLanguage = NULL,
  PortfolioId,
  AccountId = NULL,
  OrganizationNode = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{AccountId}{The Amazon Web Services account ID.}

\item{OrganizationNode}{The organization node to whom you are going to stop sharing.}
}
\description{
Stops sharing the specified portfolio with the specified account or organization node. Shares to an organization node can only be deleted by the management account of an organization or by a delegated administrator.

See \url{https://paws-r.github.io/docs/servicecatalog/delete_portfolio_share.html} for full documentation.
}
\keyword{internal}
