% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_service_action}
\alias{servicecatalog_delete_service_action}
\title{Deletes a self-service action}
\usage{
servicecatalog_delete_service_action(Id, AcceptLanguage = NULL)
}
\arguments{
\item{Id}{[required] The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Deletes a self-service action.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_delete_service_action/} for full documentation.
}
\keyword{internal}
