% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_tag_option}
\alias{servicecatalog_delete_tag_option}
\title{Deletes the specified TagOption}
\usage{
servicecatalog_delete_tag_option(Id)
}
\arguments{
\item{Id}{[required] The TagOption identifier.}
}
\value{
An empty list.
}
\description{
Deletes the specified TagOption.

You cannot delete a TagOption if it is associated with a product or
portfolio.
}
\section{Request syntax}{
\preformatted{svc$delete_tag_option(
  Id = "string"
)
}
}

\keyword{internal}
