% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_constraint}
\alias{servicecatalog_describe_constraint}
\title{Gets information about the specified constraint}
\usage{
servicecatalog_describe_constraint(AcceptLanguage, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The identifier of the constraint.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ConstraintDetail = list(
    ConstraintId = "string",
    Type = "string",
    Description = "string",
    Owner = "string",
    ProductId = "string",
    PortfolioId = "string"
  ),
  ConstraintParameters = "string",
  Status = "AVAILABLE"|"CREATING"|"FAILED"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the specified constraint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_constraint(
  AcceptLanguage = "string",
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
