% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_copy_product_status}
\alias{servicecatalog_describe_copy_product_status}
\title{Gets the status of the specified copy product operation}
\usage{
servicecatalog_describe_copy_product_status(
  AcceptLanguage = NULL,
  CopyProductToken
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{CopyProductToken}{[required] The token for the copy product operation. This token is returned by
\code{\link[=servicecatalog_copy_product]{copy_product}}.}
}
\description{
Gets the status of the specified copy product operation.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_copy_product_status.html} for full documentation.
}
\keyword{internal}
