% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_provisioned_product}
\alias{servicecatalog_describe_provisioned_product}
\title{Gets information about the specified provisioned product}
\usage{
servicecatalog_describe_provisioned_product(AcceptLanguage, Id, Name)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{The provisioned product identifier. You must provide the name or ID, but
not both.

If you do not provide a name or ID, or you provide both name and ID, an
\code{InvalidParametersException} will occur.}

\item{Name}{The name of the provisioned product. You must provide the name or ID,
but not both.

If you do not provide a name or ID, or you provide both name and ID, an
\code{InvalidParametersException} will occur.}
}
\value{
A list with the following syntax:\preformatted{list(
  ProvisionedProductDetail = list(
    Name = "string",
    Arn = "string",
    Type = "string",
    Id = "string",
    Status = "AVAILABLE"|"UNDER_CHANGE"|"TAINTED"|"ERROR"|"PLAN_IN_PROGRESS",
    StatusMessage = "string",
    CreatedTime = as.POSIXct(
      "2015-01-01"
    ),
    IdempotencyToken = "string",
    LastRecordId = "string",
    LastProvisioningRecordId = "string",
    LastSuccessfulProvisioningRecordId = "string",
    ProductId = "string",
    ProvisioningArtifactId = "string",
    LaunchRoleArn = "string"
  ),
  CloudWatchDashboards = list(
    list(
      Name = "string"
    )
  )
)
}
}
\description{
Gets information about the specified provisioned product.
}
\section{Request syntax}{
\preformatted{svc$describe_provisioned_product(
  AcceptLanguage = "string",
  Id = "string",
  Name = "string"
)
}
}

\keyword{internal}
