% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_service_action}
\alias{servicecatalog_describe_service_action}
\title{Describes a self-service action}
\usage{
servicecatalog_describe_service_action(Id, AcceptLanguage = NULL)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Describes a self-service action.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_describe_service_action/} for full documentation.
}
\keyword{internal}
