% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_service_action_execution_parameters}
\alias{servicecatalog_describe_service_action_execution_parameters}
\title{Finds the default parameters for a specific self-service action on a
specific provisioned product and returns a map of the results to the
user}
\usage{
servicecatalog_describe_service_action_execution_parameters(
  ProvisionedProductId, ServiceActionId, AcceptLanguage)
}
\arguments{
\item{ProvisionedProductId}{[required] The identifier of the provisioned product.}

\item{ServiceActionId}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ServiceActionParameters = list(
    list(
      Name = "string",
      Type = "string",
      DefaultValues = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Finds the default parameters for a specific self-service action on a
specific provisioned product and returns a map of the results to the
user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_service_action_execution_parameters(
  ProvisionedProductId = "string",
  ServiceActionId = "string",
  AcceptLanguage = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
