% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_tag_option}
\alias{servicecatalog_describe_tag_option}
\title{Gets information about the specified TagOption}
\usage{
servicecatalog_describe_tag_option(Id)
}
\arguments{
\item{Id}{[required] The TagOption identifier.}
}
\value{
A list with the following syntax:\preformatted{list(
  TagOptionDetail = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE,
    Id = "string",
    Owner = "string"
  )
)
}
}
\description{
Gets information about the specified TagOption.
}
\section{Request syntax}{
\preformatted{svc$describe_tag_option(
  Id = "string"
)
}
}

\keyword{internal}
