% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_enable_aws_organizations_access}
\alias{servicecatalog_enable_aws_organizations_access}
\title{Enable portfolio sharing feature through AWS Organizations}
\usage{
servicecatalog_enable_aws_organizations_access()
}
\value{
An empty list.
}
\description{
Enable portfolio sharing feature through AWS Organizations. This API
will allow Service Catalog to receive updates on your organization in
order to sync your shares with the current structure. This API can only
be called by the management account in the organization.

By calling this API Service Catalog will make a call to
organizations:EnableAWSServiceAccess on your behalf so that your shares
can be in sync with any changes in your AWS Organizations structure.

Note that a delegated administrator is not authorized to invoke
\code{\link[=servicecatalog_enable_aws_organizations_access]{enable_aws_organizations_access}}.
}
\section{Request syntax}{
\preformatted{svc$enable_aws_organizations_access()
}
}

\keyword{internal}
