% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_get_aws_organizations_access_status}
\alias{servicecatalog_get_aws_organizations_access_status}
\title{Get the Access Status for Organizations portfolio share feature}
\usage{
servicecatalog_get_aws_organizations_access_status()
}
\description{
Get the Access Status for Organizations portfolio share feature. This API can only be called by the management account in the organization or by a delegated admin.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_get_aws_organizations_access_status/} for full documentation.
}
\keyword{internal}
