% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_launch_paths}
\alias{servicecatalog_list_launch_paths}
\title{Lists the paths to the specified product}
\usage{
servicecatalog_list_launch_paths(
  AcceptLanguage = NULL,
  ProductId,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists the paths to the specified product. A path describes how the user gets access to a specified product and is necessary when provisioning a product. A path also determines the constraints that are put on a product. A path is dependent on a specific product, porfolio, and principal.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_list_launch_paths/} for full documentation.
}
\keyword{internal}
