% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_organization_portfolio_access}
\alias{servicecatalog_list_organization_portfolio_access}
\title{Lists the organization nodes that have access to the specified portfolio}
\usage{
servicecatalog_list_organization_portfolio_access(
  AcceptLanguage = NULL,
  PortfolioId,
  OrganizationNodeType,
  PageToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier. For example, \verb{port-2abcdext3y5fk}.}

\item{OrganizationNodeType}{[required] The organization node type that will be returned in the output.
\itemize{
\item \code{ORGANIZATION} - Organization that has access to the portfolio.
\item \code{ORGANIZATIONAL_UNIT} - Organizational unit that has access to the
portfolio within your organization.
\item \code{ACCOUNT} - Account that has access to the portfolio within your
organization.
}}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Lists the organization nodes that have access to the specified portfolio. This API can only be called by the management account in the organization or by a delegated admin.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_list_organization_portfolio_access/} for full documentation.
}
\keyword{internal}
