% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_portfolios}
\alias{servicecatalog_list_portfolios}
\title{Lists all portfolios in the catalog}
\usage{
servicecatalog_list_portfolios(
  AcceptLanguage = NULL,
  PageToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Lists all portfolios in the catalog.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_list_portfolios/} for full documentation.
}
\keyword{internal}
