% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_provisioned_product_plans}
\alias{servicecatalog_list_provisioned_product_plans}
\title{Lists the plans for the specified provisioned product or all plans to
which the user has access}
\usage{
servicecatalog_list_provisioned_product_plans(AcceptLanguage,
  ProvisionProductId, PageSize, PageToken, AccessLevelFilter)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisionProductId}{The product identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{AccessLevelFilter}{The access level to use to obtain results. The default is \code{User}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProvisionedProductPlans = list(
    list(
      PlanName = "string",
      PlanId = "string",
      ProvisionProductId = "string",
      ProvisionProductName = "string",
      PlanType = "CLOUDFORMATION",
      ProvisioningArtifactId = "string"
    )
  ),
  NextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the plans for the specified provisioned product or all plans to
which the user has access.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_provisioned_product_plans(
  AcceptLanguage = "string",
  ProvisionProductId = "string",
  PageSize = 123,
  PageToken = "string",
  AccessLevelFilter = list(
    Key = "Account"|"Role"|"User",
    Value = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
