% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_service_actions}
\alias{servicecatalog_list_service_actions}
\title{Lists all self-service actions}
\usage{
servicecatalog_list_service_actions(
  AcceptLanguage = NULL,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists all self-service actions.

See \url{https://paws-r.github.io/docs/servicecatalog/list_service_actions.html} for full documentation.
}
\keyword{internal}
