% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_tag_options}
\alias{servicecatalog_list_tag_options}
\title{Lists the specified TagOptions or all TagOptions}
\usage{
servicecatalog_list_tag_options(Filters, PageSize, PageToken)
}
\arguments{
\item{Filters}{The search filters. If no search filters are specified, the output
includes all TagOptions.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagOptionDetails = list(
    list(
      Key = "string",
      Value = "string",
      Active = TRUE|FALSE,
      Id = "string",
      Owner = "string"
    )
  ),
  PageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the specified TagOptions or all TagOptions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tag_options(
  Filters = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE
  ),
  PageSize = 123,
  PageToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
