% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_notify_terminate_provisioned_product_engine_workflow_result}
\alias{servicecatalog_notify_terminate_provisioned_product_engine_workflow_result}
\title{Notifies the result of the terminate engine execution}
\usage{
servicecatalog_notify_terminate_provisioned_product_engine_workflow_result(
  WorkflowToken,
  RecordId,
  Status,
  FailureReason = NULL,
  IdempotencyToken
)
}
\arguments{
\item{WorkflowToken}{[required] The encrypted contents of the terminate engine execution payload that
Service Catalog sends after the Terraform product terminate workflow
starts.}

\item{RecordId}{[required] The identifier of the record.}

\item{Status}{[required] The status of the terminate engine execution.}

\item{FailureReason}{The reason why the terminate engine execution failed.}

\item{IdempotencyToken}{[required] The idempotency token that identifies the terminate engine execution.}
}
\description{
Notifies the result of the terminate engine execution.

See \url{https://paws-r.github.io/docs/servicecatalog/notify_terminate_provisioned_product_engine_workflow_result.html} for full documentation.
}
\keyword{internal}
