% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_notify_update_provisioned_product_engine_workflow_result}
\alias{servicecatalog_notify_update_provisioned_product_engine_workflow_result}
\title{Notifies the result of the update engine execution}
\usage{
servicecatalog_notify_update_provisioned_product_engine_workflow_result(
  WorkflowToken,
  RecordId,
  Status,
  FailureReason = NULL,
  Outputs = NULL,
  IdempotencyToken
)
}
\arguments{
\item{WorkflowToken}{[required] The encrypted contents of the update engine execution payload that
Service Catalog sends after the Terraform product update workflow
starts.}

\item{RecordId}{[required] The identifier of the record.}

\item{Status}{[required] The status of the update engine execution.}

\item{FailureReason}{The reason why the update engine execution failed.}

\item{Outputs}{The output of the update engine execution.}

\item{IdempotencyToken}{[required] The idempotency token that identifies the update engine execution.}
}
\description{
Notifies the result of the update engine execution.

See \url{https://paws-r.github.io/docs/servicecatalog/notify_update_provisioned_product_engine_workflow_result.html} for full documentation.
}
\keyword{internal}
