% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_update_portfolio}
\alias{servicecatalog_update_portfolio}
\title{Updates the specified portfolio}
\usage{
servicecatalog_update_portfolio(
  AcceptLanguage = NULL,
  Id,
  DisplayName = NULL,
  Description = NULL,
  ProviderName = NULL,
  AddTags = NULL,
  RemoveTags = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}

\item{DisplayName}{The name to use for display purposes.}

\item{Description}{The updated description of the portfolio.}

\item{ProviderName}{The updated name of the portfolio provider.}

\item{AddTags}{The tags to add.}

\item{RemoveTags}{The tags to remove.}
}
\description{
Updates the specified portfolio.

See \url{https://paws-r.github.io/docs/servicecatalog/update_portfolio.html} for full documentation.
}
\keyword{internal}
