% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_update_tag_option}
\alias{servicecatalog_update_tag_option}
\title{Updates the specified TagOption}
\usage{
servicecatalog_update_tag_option(Id, Value, Active)
}
\arguments{
\item{Id}{[required] The TagOption identifier.}

\item{Value}{The updated value.}

\item{Active}{The updated active state.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagOptionDetail = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE,
    Id = "string",
    Owner = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified TagOption.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_tag_option(
  Id = "string",
  Value = "string",
  Active = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
