% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_associate_service_quota_template}
\alias{servicequotas_associate_service_quota_template}
\title{Associates the Service Quotas template with your organization so that
when new accounts are created in your organization, the template submits
increase requests for the specified service quotas}
\usage{
servicequotas_associate_service_quota_template()
}
\description{
Associates the Service Quotas template with your organization so that
when new accounts are created in your organization, the template submits
increase requests for the specified service quotas. Use the Service
Quotas template to request an increase for any adjustable quota value.
After you define the Service Quotas template, use this operation to
associate, or enable, the template.
}
\section{Request syntax}{
\preformatted{svc$associate_service_quota_template()
}
}

\keyword{internal}
