% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_create_support_case}
\alias{servicequotas_create_support_case}
\title{Creates a Support case for an existing quota increase request}
\usage{
servicequotas_create_support_case(RequestId)
}
\arguments{
\item{RequestId}{[required] The ID of the pending quota increase request for which you want to open
a Support case.}
}
\description{
Creates a Support case for an existing quota increase request. This call only creates a Support case if the request has a \code{Pending} status.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_create_support_case/} for full documentation.
}
\keyword{internal}
