% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_delete_service_quota_increase_request_from_template}
\alias{servicequotas_delete_service_quota_increase_request_from_template}
\title{Deletes the quota increase request for the specified quota from your
quota request template}
\usage{
servicequotas_delete_service_quota_increase_request_from_template(
  ServiceCode,
  QuotaCode,
  AwsRegion
)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service identifier. To find the service code value for an
Amazon Web Services service, use the
\code{\link[=servicequotas_list_services]{list_services}} operation.}

\item{QuotaCode}{[required] Specifies the quota identifier. To find the quota code for a specific
quota, use the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} operation,
and look for the \code{QuotaCode} response in the output for the quota you
want.}

\item{AwsRegion}{[required] Specifies the Amazon Web Services Region for which the request was made.}
}
\description{
Deletes the quota increase request for the specified quota from your quota request template.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_delete_service_quota_increase_request_from_template/} for full documentation.
}
\keyword{internal}
