% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_auto_management_configuration}
\alias{servicequotas_get_auto_management_configuration}
\title{Retrieves information about your Service Quotas Automatic Management
configuration}
\usage{
servicequotas_get_auto_management_configuration()
}
\description{
Retrieves information about your \href{https://docs.aws.amazon.com/servicequotas/latest/userguide/automatic-management.html}{Service Quotas Automatic Management} configuration. Automatic Management monitors your Service Quotas utilization and notifies you before you run out of your allocated quotas.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_get_auto_management_configuration/} for full documentation.
}
\keyword{internal}
