% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_aws_default_service_quota}
\alias{servicequotas_get_aws_default_service_quota}
\title{Retrieves the default value for the specified quota}
\usage{
servicequotas_get_aws_default_service_quota(ServiceCode, QuotaCode)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service identifier. To find the service code value for an
Amazon Web Services service, use the
\code{\link[=servicequotas_list_services]{list_services}} operation.}

\item{QuotaCode}{[required] Specifies the quota identifier. To find the quota code for a specific
quota, use the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} operation,
and look for the \code{QuotaCode} response in the output for the quota you
want.}
}
\description{
Retrieves the default value for the specified quota. The default value does not reflect any quota increases.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_get_aws_default_service_quota/} for full documentation.
}
\keyword{internal}
