% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_requested_service_quota_change}
\alias{servicequotas_get_requested_service_quota_change}
\title{Retrieves information about the specified quota increase request}
\usage{
servicequotas_get_requested_service_quota_change(RequestId)
}
\arguments{
\item{RequestId}{[required] The ID of the quota increase request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RequestedQuota = list(
    Id = "string",
    CaseId = "string",
    ServiceCode = "string",
    ServiceName = "string",
    QuotaCode = "string",
    QuotaName = "string",
    DesiredValue = 123.0,
    Status = "PENDING"|"CASE_OPENED"|"APPROVED"|"DENIED"|"CASE_CLOSED",
    Created = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdated = as.POSIXct(
      "2015-01-01"
    ),
    Requester = "string",
    QuotaArn = "string",
    GlobalQuota = TRUE|FALSE,
    Unit = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about the specified quota increase request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_requested_service_quota_change(
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
