% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_service_quota}
\alias{servicequotas_get_service_quota}
\title{Returns the details for the specified service quota}
\usage{
servicequotas_get_service_quota(ServiceCode, QuotaCode)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service that you want to use.}

\item{QuotaCode}{[required] Identifies the service quota you want to select.}
}
\description{
Returns the details for the specified service quota. This operation
provides a different Value than the \code{GetAWSDefaultServiceQuota}
operation. This operation returns the applied value for each quota.
\code{GetAWSDefaultServiceQuota} returns the default AWS value for each
quota.
}
\section{Request syntax}{
\preformatted{svc$get_service_quota(
  ServiceCode = "string",
  QuotaCode = "string"
)
}
}

\keyword{internal}
