% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_service_quota_increase_request_from_template}
\alias{servicequotas_get_service_quota_increase_request_from_template}
\title{Returns the details of the service quota increase request in your
template}
\usage{
servicequotas_get_service_quota_increase_request_from_template(
  ServiceCode, QuotaCode, AwsRegion)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service that you want to use.}

\item{QuotaCode}{[required] Specifies the quota you want.}

\item{AwsRegion}{[required] Specifies the AWS Region for the quota that you want to use.}
}
\description{
Returns the details of the service quota increase request in your
template.
}
\section{Request syntax}{
\preformatted{svc$get_service_quota_increase_request_from_template(
  ServiceCode = "string",
  QuotaCode = "string",
  AwsRegion = "string"
)
}
}

\keyword{internal}
