% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_requested_service_quota_change_history_by_quota}
\alias{servicequotas_list_requested_service_quota_change_history_by_quota}
\title{Retrieves the quota increase requests for the specified quota}
\usage{
servicequotas_list_requested_service_quota_change_history_by_quota(
  ServiceCode,
  QuotaCode,
  Status = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{QuotaCode}{[required] The quota identifier.}

\item{Status}{The status value of the quota increase request.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, if any, make another call with the token returned
from this call.}
}
\description{
Retrieves the quota increase requests for the specified quota.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_list_requested_service_quota_change_history_by_quota/} for full documentation.
}
\keyword{internal}
