% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_service_quota_increase_requests_in_template}
\alias{servicequotas_list_service_quota_increase_requests_in_template}
\title{Lists the quota increase requests in the specified quota request
template}
\usage{
servicequotas_list_service_quota_increase_requests_in_template(
  ServiceCode, AwsRegion, NextToken, MaxResults)
}
\arguments{
\item{ServiceCode}{The service identifier.}

\item{AwsRegion}{The AWS Region.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, if any, make another call with the token returned
from this call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ServiceQuotaIncreaseRequestInTemplateList = list(
    list(
      ServiceCode = "string",
      ServiceName = "string",
      QuotaCode = "string",
      QuotaName = "string",
      DesiredValue = 123.0,
      AwsRegion = "string",
      Unit = "string",
      GlobalQuota = TRUE|FALSE
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the quota increase requests in the specified quota request
template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_service_quota_increase_requests_in_template(
  ServiceCode = "string",
  AwsRegion = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
