% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_tags_for_resource}
\alias{servicequotas_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified applied quota}
\usage{
servicequotas_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) for the applied quota for which you want
to list tags. You can get this information by using the Service Quotas
console, or by listing the quotas using the
\href{https://docs.aws.amazon.com/cli/latest/reference/service-quotas/list-service-quotas.html}{list-service-quotas}
AWS CLI command or the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} AWS API
operation.}
}
\description{
Returns a list of the tags assigned to the specified applied quota.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
