% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_put_service_quota_increase_request_into_template}
\alias{servicequotas_put_service_quota_increase_request_into_template}
\title{Adds a quota increase request to your quota request template}
\usage{
servicequotas_put_service_quota_increase_request_into_template(
  QuotaCode,
  ServiceCode,
  AwsRegion,
  DesiredValue
)
}
\arguments{
\item{QuotaCode}{[required] Specifies the quota identifier. To find the quota code for a specific
quota, use the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} operation,
and look for the \code{QuotaCode} response in the output for the quota you
want.}

\item{ServiceCode}{[required] Specifies the service identifier. To find the service code value for an
Amazon Web Services service, use the
\code{\link[=servicequotas_list_services]{list_services}} operation.}

\item{AwsRegion}{[required] Specifies the Amazon Web Services Region to which the template applies.}

\item{DesiredValue}{[required] Specifies the new, increased value for the quota.}
}
\description{
Adds a quota increase request to your quota request template.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_put_service_quota_increase_request_into_template/} for full documentation.
}
\keyword{internal}
