% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_start_auto_management}
\alias{servicequotas_start_auto_management}
\title{Starts Service Quotas Automatic Management for an Amazon Web Services
account, including notification preferences and excluded quotas
configurations}
\usage{
servicequotas_start_auto_management(
  OptInLevel,
  OptInType,
  NotificationArn = NULL,
  ExclusionList = NULL
)
}
\arguments{
\item{OptInLevel}{[required] Sets the opt-in level for Automatic Management. Only Amazon Web Services
account level is supported.}

\item{OptInType}{[required] Sets the opt-in type for Automatic Management. There are two modes:
Notify only and Notify and Auto-Adjust. Currently, only NotifyOnly is
available.}

\item{NotificationArn}{The \href{https://docs.aws.amazon.com/notifications/latest/userguide/resource-level-permissions.html#rlp-table}{User Notifications}
Amazon Resource Name (ARN) for Automatic Management notifications.}

\item{ExclusionList}{List of Amazon Web Services services excluded from Automatic Management.
You won't be notified of Service Quotas utilization for Amazon Web
Services services added to the Automatic Management exclusion list.}
}
\description{
Starts \href{https://docs.aws.amazon.com/servicequotas/latest/userguide/automatic-management.html}{Service Quotas Automatic Management} for an Amazon Web Services account, including notification preferences and excluded quotas configurations. Automatic Management monitors your Service Quotas utilization and notifies you before you run out of your allocated quotas.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_start_auto_management/} for full documentation.
}
\keyword{internal}
