% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_tag_resource}
\alias{servicequotas_tag_resource}
\title{Adds tags to the specified applied quota}
\usage{
servicequotas_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) for the applied quota. You can get this
information by using the Service Quotas console, or by listing the
quotas using the
\href{https://docs.aws.amazon.com/cli/latest/reference/service-quotas/list-service-quotas.html}{list-service-quotas}
AWS CLI command or the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} AWS API
operation.}

\item{Tags}{[required] The tags that you want to add to the resource.}
}
\description{
Adds tags to the specified applied quota. You can include one or more tags to add to the quota.

See \url{https://paws-r.github.io/docs/servicequotas/tag_resource.html} for full documentation.
}
\keyword{internal}
