% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_untag_resource}
\alias{servicequotas_untag_resource}
\title{Removes tags from the specified applied quota}
\usage{
servicequotas_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) for the applied quota that you want to
untag. You can get this information by using the Service Quotas console,
or by listing the quotas using the
\href{https://docs.aws.amazon.com/cli/latest/reference/service-quotas/list-service-quotas.html}{list-service-quotas}
AWS CLI command or the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} AWS API
operation.}

\item{TagKeys}{[required] The keys of the tags that you want to remove from the resource.}
}
\value{
An empty list.
}
\description{
Removes tags from the specified applied quota. You can specify one or
more tags to remove.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
