% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_create_cloud_connector}
\alias{ssm_create_cloud_connector}
\title{Creates a cloud connector that establishes a connection between Systems
Manager and a third-party cloud environment}
\usage{
ssm_create_cloud_connector(
  DisplayName,
  RoleArn,
  Description = NULL,
  Configuration,
  ConfigConnectorArn,
  Tags = NULL
)
}
\arguments{
\item{DisplayName}{[required] A friendly name for the cloud connector.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that the cloud connector uses to communicate with the third-party cloud environment.}

\item{Description}{A description for the cloud connector.}

\item{Configuration}{[required] The configuration details for connecting to the third-party cloud environment.}

\item{ConfigConnectorArn}{[required] The ARN of the Amazon Web Services Config connector associated with this cloud connector.}

\item{Tags}{Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment.}
}
\description{
Creates a cloud connector that establishes a connection between Systems Manager and a third-party cloud environment.

See \url{https://www.paws-r-sdk.com/docs/ssm_create_cloud_connector/} for full documentation.
}
\keyword{internal}
