% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_cloud_connector}
\alias{ssm_delete_cloud_connector}
\title{Deletes a cloud connector}
\usage{
ssm_delete_cloud_connector(CloudConnectorId)
}
\arguments{
\item{CloudConnectorId}{[required] The ID of the cloud connector to delete.}
}
\description{
Deletes a cloud connector.

See \url{https://www.paws-r-sdk.com/docs/ssm_delete_cloud_connector/} for full documentation.
}
\keyword{internal}
