% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_ops_item}
\alias{ssm_delete_ops_item}
\title{Delete an OpsItem}
\usage{
ssm_delete_ops_item(OpsItemId)
}
\arguments{
\item{OpsItemId}{[required] The ID of the OpsItem that you want to delete.}
}
\description{
Delete an OpsItem. You must have permission in Identity and Access Management (IAM) to delete an OpsItem.

See \url{https://www.paws-r-sdk.com/docs/ssm_delete_ops_item/} for full documentation.
}
\keyword{internal}
