% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_ops_metadata}
\alias{ssm_delete_ops_metadata}
\title{Delete OpsMetadata related to an application}
\usage{
ssm_delete_ops_metadata(OpsMetadataArn)
}
\arguments{
\item{OpsMetadataArn}{[required] The Amazon Resource Name (ARN) of an OpsMetadata Object to delete.}
}
\value{
An empty list.
}
\description{
Delete OpsMetadata related to an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_ops_metadata(
  OpsMetadataArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
