% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_resource_data_sync}
\alias{ssm_delete_resource_data_sync}
\title{Deletes a Resource Data Sync configuration}
\usage{
ssm_delete_resource_data_sync(SyncName, SyncType)
}
\arguments{
\item{SyncName}{[required] The name of the configuration to delete.}

\item{SyncType}{Specify the type of resource data sync to delete.}
}
\value{
An empty list.
}
\description{
Deletes a Resource Data Sync configuration. After the configuration is
deleted, changes to data on managed instances are no longer synced to or
from the target. Deleting a sync configuration does not delete data.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_resource_data_sync(
  SyncName = "string",
  SyncType = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
