% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_deregister_managed_instance}
\alias{ssm_deregister_managed_instance}
\title{Removes the server or virtual machine from the list of registered
servers}
\usage{
ssm_deregister_managed_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The ID assigned to the managed node when you registered it using the
activation process.}
}
\description{
Removes the server or virtual machine from the list of registered servers.

See \url{https://www.paws-r-sdk.com/docs/ssm_deregister_managed_instance/} for full documentation.
}
\keyword{internal}
