% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_deregister_patch_baseline_for_patch_group}
\alias{ssm_deregister_patch_baseline_for_patch_group}
\title{Removes a patch group from a patch baseline}
\usage{
ssm_deregister_patch_baseline_for_patch_group(BaselineId, PatchGroup)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to deregister the patch group from.}

\item{PatchGroup}{[required] The name of the patch group that should be deregistered from the patch
baseline.}
}
\description{
Removes a patch group from a patch baseline.

See \url{https://www.paws-r-sdk.com/docs/ssm_deregister_patch_baseline_for_patch_group/} for full documentation.
}
\keyword{internal}
