% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_deregister_target_from_maintenance_window}
\alias{ssm_deregister_target_from_maintenance_window}
\title{Removes a target from a maintenance window}
\usage{
ssm_deregister_target_from_maintenance_window(WindowId, WindowTargetId,
  Safe)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window the target should be removed from.}

\item{WindowTargetId}{[required] The ID of the target definition to remove.}

\item{Safe}{The system checks if the target is being referenced by a task. If the
target is being referenced, the system returns an error and does not
deregister the target from the maintenance window.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  WindowId = "string",
  WindowTargetId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Removes a target from a maintenance window.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$deregister_target_from_maintenance_window(
  WindowId = "string",
  WindowTargetId = "string",
  Safe = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
