% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_association}
\alias{ssm_describe_association}
\title{Describes the association for the specified target or managed node}
\usage{
ssm_describe_association(
  Name = NULL,
  InstanceId = NULL,
  AssociationId = NULL,
  AssociationVersion = NULL
)
}
\arguments{
\item{Name}{The name of the SSM document.}

\item{InstanceId}{The managed node ID.}

\item{AssociationId}{The association ID for which you want information.}

\item{AssociationVersion}{Specify the association version to retrieve. To view the latest version,
either specify \verb{$LATEST} for this parameter, or omit this parameter. To
view a list of all associations for a managed node, use
\code{\link[=ssm_list_associations]{list_associations}}. To get a list of versions
for a specific association, use
\code{\link[=ssm_list_association_versions]{list_association_versions}}.}
}
\description{
Describes the association for the specified target or managed node. If you created the association by using the \code{Targets} parameter, then you must retrieve the association by using the association ID.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_association/} for full documentation.
}
\keyword{internal}
