% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_automation_step_executions}
\alias{ssm_describe_automation_step_executions}
\title{Information about all active and terminated step executions in an
Automation workflow}
\usage{
ssm_describe_automation_step_executions(AutomationExecutionId, Filters,
  NextToken, MaxResults, ReverseOrder)
}
\arguments{
\item{AutomationExecutionId}{[required] The Automation execution ID for which you want step execution
descriptions.}

\item{Filters}{One or more filters to limit the number of step executions returned by
the request.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{ReverseOrder}{Indicates whether to list step executions in reverse order by start
time. The default value is 'false'.}
}
\value{
A list with the following syntax:\preformatted{list(
  StepExecutions = list(
    list(
      StepName = "string",
      Action = "string",
      TimeoutSeconds = 123,
      OnFailure = "string",
      MaxAttempts = 123,
      ExecutionStartTime = as.POSIXct(
        "2015-01-01"
      ),
      ExecutionEndTime = as.POSIXct(
        "2015-01-01"
      ),
      StepStatus = "Pending"|"InProgress"|"Waiting"|"Success"|"TimedOut"|"Cancelling"|"Cancelled"|"Failed"|"PendingApproval"|"Approved"|"Rejected"|"Scheduled"|"RunbookInProgress"|"PendingChangeCalendarOverride"|"ChangeCalendarOverrideApproved"|"ChangeCalendarOverrideRejected"|"CompletedWithSuccess"|"CompletedWithFailure",
      ResponseCode = "string",
      Inputs = list(
        "string"
      ),
      Outputs = list(
        list(
          "string"
        )
      ),
      Response = "string",
      FailureMessage = "string",
      FailureDetails = list(
        FailureStage = "string",
        FailureType = "string",
        Details = list(
          list(
            "string"
          )
        )
      ),
      StepExecutionId = "string",
      OverriddenParameters = list(
        list(
          "string"
        )
      ),
      IsEnd = TRUE|FALSE,
      NextStep = "string",
      IsCritical = TRUE|FALSE,
      ValidNextSteps = list(
        "string"
      ),
      Targets = list(
        list(
          Key = "string",
          Values = list(
            "string"
          )
        )
      ),
      TargetLocation = list(
        Accounts = list(
          "string"
        ),
        Regions = list(
          "string"
        ),
        TargetLocationMaxConcurrency = "string",
        TargetLocationMaxErrors = "string",
        ExecutionRoleName = "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Information about all active and terminated step executions in an
Automation workflow.
}
\section{Request syntax}{
\preformatted{svc$describe_automation_step_executions(
  AutomationExecutionId = "string",
  Filters = list(
    list(
      Key = "StartTimeBefore"|"StartTimeAfter"|"StepExecutionStatus"|"StepExecutionId"|"StepName"|"Action",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string",
  MaxResults = 123,
  ReverseOrder = TRUE|FALSE
)
}
}

\keyword{internal}
