% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_effective_instance_associations}
\alias{ssm_describe_effective_instance_associations}
\title{All associations for the managed nodes}
\usage{
ssm_describe_effective_instance_associations(
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The managed node ID for which you want to view all associations.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
All associations for the managed nodes.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_effective_instance_associations/} for full documentation.
}
\keyword{internal}
