% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_instance_properties}
\alias{ssm_describe_instance_properties}
\title{An API operation used by the Systems Manager console to display
information about Systems Manager managed nodes}
\usage{
ssm_describe_instance_properties(
  InstancePropertyFilterList = NULL,
  FiltersWithOperator = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstancePropertyFilterList}{An array of instance property filters.}

\item{FiltersWithOperator}{The request filters to use with the operator.}

\item{MaxResults}{The maximum number of items to return for the call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token provided by a previous request to use to return the next set
of properties.}
}
\description{
An API operation used by the Systems Manager console to display information about Systems Manager managed nodes.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_instance_properties/} for full documentation.
}
\keyword{internal}
