% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_inventory_deletions}
\alias{ssm_describe_inventory_deletions}
\title{Describes a specific delete inventory operation}
\usage{
ssm_describe_inventory_deletions(
  DeletionId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DeletionId}{Specify the delete inventory ID for which you want information. This ID
was returned by the \code{\link[=ssm_delete_inventory]{delete_inventory}}
operation.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}
}
\description{
Describes a specific delete inventory operation.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_inventory_deletions/} for full documentation.
}
\keyword{internal}
