% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_maintenance_windows}
\alias{ssm_describe_maintenance_windows}
\title{Retrieves the maintenance windows in an Amazon Web Services account}
\usage{
ssm_describe_maintenance_windows(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Optional filters used to narrow down the scope of the returned
maintenance windows. Supported filter keys are \code{Name} and \code{Enabled}. For
example, \code{Name=MyMaintenanceWindow} and \code{Enabled=True}.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Retrieves the maintenance windows in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_maintenance_windows/} for full documentation.
}
\keyword{internal}
