% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_baselines}
\alias{ssm_describe_patch_baselines}
\title{Lists the patch baselines in your Amazon Web Services account}
\usage{
ssm_describe_patch_baselines(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{Each element in the array is a structure containing a key-value pair.

Supported keys for
\code{\link[=ssm_describe_patch_baselines]{describe_patch_baselines}} include the
following:
\itemize{
\item \strong{\code{NAME_PREFIX}}

Sample values: \verb{AWS-} | \verb{My-}
\item \strong{\code{OWNER}}

Sample values: \code{AWS} | \code{Self}
\item \strong{\code{OPERATING_SYSTEM}}

Sample values: \code{AMAZON_LINUX} | \code{SUSE} | \code{WINDOWS}
}}

\item{MaxResults}{The maximum number of patch baselines to return (per page).}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Lists the patch baselines in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_patch_baselines/} for full documentation.
}
\keyword{internal}
