% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_group_state}
\alias{ssm_describe_patch_group_state}
\title{Returns high-level aggregated patch compliance state for a patch group}
\usage{
ssm_describe_patch_group_state(PatchGroup)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch snapshot should be retrieved.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Instances = 123,
  InstancesWithInstalledPatches = 123,
  InstancesWithInstalledOtherPatches = 123,
  InstancesWithInstalledPendingRebootPatches = 123,
  InstancesWithInstalledRejectedPatches = 123,
  InstancesWithMissingPatches = 123,
  InstancesWithFailedPatches = 123,
  InstancesWithNotApplicablePatches = 123,
  InstancesWithUnreportedNotApplicablePatches = 123
)
}\if{html}{\out{</div>}}
}
\description{
Returns high-level aggregated patch compliance state for a patch group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_patch_group_state(
  PatchGroup = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
