% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_disassociate_ops_item_related_item}
\alias{ssm_disassociate_ops_item_related_item}
\title{Deletes the association between an OpsItem and a related resource}
\usage{
ssm_disassociate_ops_item_related_item(OpsItemId, AssociationId)
}
\arguments{
\item{OpsItemId}{[required] The ID of the OpsItem for which you want to delete an association
between the OpsItem and a related resource.}

\item{AssociationId}{[required] The ID of the association for which you want to delete an association
between the OpsItem and a related resource.}
}
\value{
An empty list.
}
\description{
Deletes the association between an OpsItem and a related resource. For
example, this API action can delete an Incident Manager incident from an
OpsItem. Incident Manager is a capability of AWS Systems Manager.
}
\section{Request syntax}{
\preformatted{svc$disassociate_ops_item_related_item(
  OpsItemId = "string",
  AssociationId = "string"
)
}
}

\keyword{internal}
