% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_access_token}
\alias{ssm_get_access_token}
\title{Returns a credentials set to be used with just-in-time node access}
\usage{
ssm_get_access_token(AccessRequestId)
}
\arguments{
\item{AccessRequestId}{[required] The ID of a just-in-time node access request.}
}
\description{
Returns a credentials set to be used with just-in-time node access.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_access_token/} for full documentation.
}
\keyword{internal}
