% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_command_invocation}
\alias{ssm_get_command_invocation}
\title{Returns detailed information about command execution for an invocation
or plugin}
\usage{
ssm_get_command_invocation(CommandId, InstanceId, PluginName)
}
\arguments{
\item{CommandId}{[required] (Required) The parent command ID of the invocation plugin.}

\item{InstanceId}{[required] (Required) The ID of the managed instance targeted by the command. A
managed instance can be an Amazon Elastic Compute Cloud (Amazon EC2)
instance or an instance in your hybrid environment that is configured
for AWS Systems Manager.}

\item{PluginName}{The name of the plugin for which you want detailed results. If the
document contains only one plugin, you can omit the name and details for
that plugin. If the document contains more than one plugin, you must
specify the name of the plugin for which you want to view details.

Plugin names are also referred to as \emph{step names} in Systems Manager
documents. For example, \code{aws:RunShellScript} is a plugin.

To find the \code{PluginName}, check the document content and find the name
of the plugin. Alternatively, use
\code{\link[=ssm_list_command_invocations]{list_command_invocations}} with the
\code{CommandId} and \code{Details} parameters. The \code{PluginName} is the \code{Name}
attribute of the \code{CommandPlugin} object in the \code{CommandPlugins} list.}
}
\value{
A list with the following syntax:\preformatted{list(
  CommandId = "string",
  InstanceId = "string",
  Comment = "string",
  DocumentName = "string",
  DocumentVersion = "string",
  PluginName = "string",
  ResponseCode = 123,
  ExecutionStartDateTime = "string",
  ExecutionElapsedTime = "string",
  ExecutionEndDateTime = "string",
  Status = "Pending"|"InProgress"|"Delayed"|"Success"|"Cancelled"|"TimedOut"|"Failed"|"Cancelling",
  StatusDetails = "string",
  StandardOutputContent = "string",
  StandardOutputUrl = "string",
  StandardErrorContent = "string",
  StandardErrorUrl = "string",
  CloudWatchOutputConfig = list(
    CloudWatchLogGroupName = "string",
    CloudWatchOutputEnabled = TRUE|FALSE
  )
)
}
}
\description{
Returns detailed information about command execution for an invocation
or plugin.

\code{\link[=ssm_get_command_invocation]{get_command_invocation}} only gives the
execution status of a plugin in a document. To get the command execution
status on a specific instance, use
\code{\link[=ssm_list_command_invocations]{list_command_invocations}}. To get the
command execution status across instances, use
\code{\link[=ssm_list_commands]{list_commands}}.
}
\section{Request syntax}{
\preformatted{svc$get_command_invocation(
  CommandId = "string",
  InstanceId = "string",
  PluginName = "string"
)
}
}

\keyword{internal}
