% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_default_patch_baseline}
\alias{ssm_get_default_patch_baseline}
\title{Retrieves the default patch baseline}
\usage{
ssm_get_default_patch_baseline(OperatingSystem)
}
\arguments{
\item{OperatingSystem}{Returns the default patch baseline for the specified operating system.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BaselineId = "string",
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"|"MACOS"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the default patch baseline. Note that Systems Manager supports
creating multiple default patch baselines. For example, you can create a
default patch baseline for each operating system.

If you do not specify an operating system value, the default patch
baseline for Windows is returned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_default_patch_baseline(
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"|"MACOS"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
