% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_execution_preview}
\alias{ssm_get_execution_preview}
\title{Initiates the process of retrieving an existing preview that shows the
effects that running a specified Automation runbook would have on the
targeted resources}
\usage{
ssm_get_execution_preview(ExecutionPreviewId)
}
\arguments{
\item{ExecutionPreviewId}{[required] The ID of the existing execution preview.}
}
\description{
Initiates the process of retrieving an existing preview that shows the effects that running a specified Automation runbook would have on the targeted resources.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_execution_preview/} for full documentation.
}
\keyword{internal}
