% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_execution}
\alias{ssm_get_maintenance_window_execution}
\title{Retrieves details about a specific a maintenance window execution}
\usage{
ssm_get_maintenance_window_execution(WindowExecutionId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the maintenance window execution that includes the task.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  WindowExecutionId = "string",
  TaskIds = list(
    "string"
  ),
  Status = "PENDING"|"IN_PROGRESS"|"SUCCESS"|"FAILED"|"TIMED_OUT"|"CANCELLING"|"CANCELLED"|"SKIPPED_OVERLAPPING",
  StatusDetails = "string",
  StartTime = as.POSIXct(
    "2015-01-01"
  ),
  EndTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves details about a specific a maintenance window execution.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_maintenance_window_execution(
  WindowExecutionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
