% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_parameters}
\alias{ssm_get_parameters}
\title{Get details of a parameter}
\usage{
ssm_get_parameters(Names, WithDecryption)
}
\arguments{
\item{Names}{[required] Names of the parameters for which you want to query information.}

\item{WithDecryption}{Return decrypted secure string value. Return decrypted values for secure
string parameters. This flag is ignored for String and StringList
parameter types.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Parameters = list(
    list(
      Name = "string",
      Type = "String"|"StringList"|"SecureString",
      Value = "string",
      Version = 123,
      Selector = "string",
      SourceResult = "string",
      LastModifiedDate = as.POSIXct(
        "2015-01-01"
      ),
      ARN = "string",
      DataType = "string"
    )
  ),
  InvalidParameters = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Get details of a parameter. Don't confuse this API action with the
\code{\link[=ssm_get_parameter]{get_parameter}} API action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_parameters(
  Names = list(
    "string"
  ),
  WithDecryption = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
