% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_patch_baseline}
\alias{ssm_get_patch_baseline}
\title{Retrieves information about a patch baseline}
\usage{
ssm_get_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to retrieve.

To retrieve information about an Amazon Web Services managed patch
baseline, specify the full Amazon Resource Name (ARN) of the baseline.
For example, for the baseline \code{AWS-AmazonLinuxDefaultPatchBaseline},
specify
\verb{arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0e392de35e7c563b7}
instead of \verb{pb-0e392de35e7c563b7}.}
}
\description{
Retrieves information about a patch baseline.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_patch_baseline/} for full documentation.
}
\keyword{internal}
