% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_patch_baseline_for_patch_group}
\alias{ssm_get_patch_baseline_for_patch_group}
\title{Retrieves the patch baseline that should be used for the specified patch
group}
\usage{
ssm_get_patch_baseline_for_patch_group(PatchGroup, OperatingSystem = NULL)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch baseline should be retrieved.}

\item{OperatingSystem}{Returns the operating system rule specified for patch groups using the patch baseline. The default value is \code{WINDOWS}.}
}
\description{
Retrieves the patch baseline that should be used for the specified patch group.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_patch_baseline_for_patch_group/} for full documentation.
}
\keyword{internal}
