% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_associations}
\alias{ssm_list_associations}
\title{Returns all State Manager associations in the current Amazon Web
Services account and Amazon Web Services Region}
\usage{
ssm_list_associations(
  AssociationFilterList = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssociationFilterList}{One or more filters. Use a filter to return a more specific list of
results.

Filtering associations using the \code{InstanceID} attribute only returns
legacy associations created using the \code{InstanceID} attribute.
Associations targeting the managed node that are part of the Target
Attributes \code{ResourceGroup} or \code{Tags} aren't returned.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Returns all State Manager associations in the current Amazon Web Services account and Amazon Web Services Region. You can limit the results to a specific State Manager association document or managed node by specifying a filter. State Manager is a capability of Amazon Web Services Systems Manager.

See \url{https://www.paws-r-sdk.com/docs/ssm_list_associations/} for full documentation.
}
\keyword{internal}
