% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_cloud_connectors}
\alias{ssm_list_cloud_connectors}
\title{Returns a list of cloud connectors in the current Amazon Web Services
account and Amazon Web Services Region}
\usage{
ssm_list_cloud_connectors(MaxResults = NULL, NextToken = NULL, Filters = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of items to return for this call.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{Filters}{One or more filters to limit the cloud connectors returned in the response.}
}
\description{
Returns a list of cloud connectors in the current Amazon Web Services account and Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/ssm_list_cloud_connectors/} for full documentation.
}
\keyword{internal}
