% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_document_versions}
\alias{ssm_list_document_versions}
\title{List all versions for a document}
\usage{
ssm_list_document_versions(Name, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Name}{[required] The name of the document. You can specify an Amazon Resource Name (ARN).}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
List all versions for a document.

See \url{https://paws-r.github.io/docs/ssm/list_document_versions.html} for full documentation.
}
\keyword{internal}
