% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_nodes}
\alias{ssm_list_nodes}
\title{Takes in filters and returns a list of managed nodes matching the filter
criteria}
\usage{
ssm_list_nodes(
  SyncName = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SyncName}{The name of the resource data sync to retrieve information about.
Required for cross-account/cross-Region configurations. Optional for
single account/single-Region configurations.}

\item{Filters}{One or more filters. Use a filter to return a more specific list of
managed nodes.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}
}
\description{
Takes in filters and returns a list of managed nodes matching the filter criteria.

See \url{https://www.paws-r-sdk.com/docs/ssm_list_nodes/} for full documentation.
}
\keyword{internal}
