% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_ops_metadata}
\alias{ssm_list_ops_metadata}
\title{Amazon Web Services Systems Manager calls this API operation when
displaying all Application Manager OpsMetadata objects or blobs}
\usage{
ssm_list_ops_metadata(Filters = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Filters}{One or more filters to limit the number of OpsMetadata objects returned
by the call.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Amazon Web Services Systems Manager calls this API operation when displaying all Application Manager OpsMetadata objects or blobs.

See \url{https://www.paws-r-sdk.com/docs/ssm_list_ops_metadata/} for full documentation.
}
\keyword{internal}
