% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_tags_for_resource}
\alias{ssm_list_tags_for_resource}
\title{Returns a list of the tags assigned to the specified resource}
\usage{
ssm_list_tags_for_resource(ResourceType, ResourceId)
}
\arguments{
\item{ResourceType}{[required] Returns a list of tags for a specific resource type.}

\item{ResourceId}{[required] The resource ID for which you want to see a list of tags.}
}
\description{
Returns a list of the tags assigned to the specified resource.

See \url{https://paws-r.github.io/docs/ssm/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
