% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_put_inventory}
\alias{ssm_put_inventory}
\title{Bulk update custom inventory items on one or more managed nodes}
\usage{
ssm_put_inventory(InstanceId, Items)
}
\arguments{
\item{InstanceId}{[required] An managed node ID where you want to add or update inventory items.}

\item{Items}{[required] The inventory items that you want to add or update on managed nodes.}
}
\description{
Bulk update custom inventory items on one or more managed nodes. The request adds an inventory item, if it doesn't already exist, or updates an inventory item, if it does exist.

See \url{https://paws-r.github.io/docs/ssm/put_inventory.html} for full documentation.
}
\keyword{internal}
