% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_put_resource_policy}
\alias{ssm_put_resource_policy}
\title{Creates or updates a Systems Manager resource policy}
\usage{
ssm_put_resource_policy(
  ResourceArn,
  Policy,
  PolicyId = NULL,
  PolicyHash = NULL
)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the resource to which you want to attach a
policy.}

\item{Policy}{[required] A policy you want to associate with a resource.}

\item{PolicyId}{The policy ID.}

\item{PolicyHash}{ID of the current policy version. The hash helps to prevent a situation
where multiple users attempt to overwrite a policy. You must provide
this hash when updating or deleting a policy.}
}
\description{
Creates or updates a Systems Manager resource policy. A resource policy helps you to define the IAM entity (for example, an Amazon Web Services account) that can manage your Systems Manager resources. Currently, \code{OpsItemGroup} is the only resource that supports Systems Manager resource policies. The resource policy for \code{OpsItemGroup} enables Amazon Web Services accounts to view and interact with OpsCenter operational work items (OpsItems).

See \url{https://paws-r.github.io/docs/ssm/put_resource_policy.html} for full documentation.
}
\keyword{internal}
