% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_register_target_with_maintenance_window}
\alias{ssm_register_target_with_maintenance_window}
\title{Registers a target with a maintenance window}
\usage{
ssm_register_target_with_maintenance_window(WindowId, ResourceType,
  Targets, OwnerInformation, Name, Description, ClientToken)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window the target should be registered with.}

\item{ResourceType}{[required] The type of target being registered with the maintenance window.}

\item{Targets}{[required] The targets to register with the maintenance window. In other words, the
instances to run commands on when the maintenance window runs.

You can specify targets using instance IDs, resource group names, or
tags that have been applied to instances.

\strong{Example 1}: Specify instance IDs

\verb{Key=InstanceIds,Values=instance-id-1,instance-id-2,instance-id-3 }

\strong{Example 2}: Use tag key-pairs applied to instances

\verb{Key=tag:my-tag-key,Values=my-tag-value-1,my-tag-value-2 }

\strong{Example 3}: Use tag-keys applied to instances

\verb{Key=tag-key,Values=my-tag-key-1,my-tag-key-2 }

\strong{Example 4}: Use resource group names

\verb{Key=resource-groups:Name,Values=resource-group-name }

\strong{Example 5}: Use filters for resource group types

\verb{Key=resource-groups:ResourceTypeFilters,Values=resource-type-1,resource-type-2 }

For \code{Key=resource-groups:ResourceTypeFilters}, specify resource types in
the following format

\verb{Key=resource-groups:ResourceTypeFilters,Values=AWS::EC2::INSTANCE,AWS::EC2::VPC }

For more information about these examples formats, including the best
use case for each one, see \href{https://docs.aws.amazon.com/systems-manager/latest/userguide/mw-cli-tutorial-targets-examples.html}{Examples: Register targets with a maintenance window}
in the \emph{AWS Systems Manager User Guide}.}

\item{OwnerInformation}{User-provided value that will be included in any CloudWatch events
raised while running tasks for these targets in this maintenance window.}

\item{Name}{An optional name for the target.}

\item{Description}{An optional description for the target.}

\item{ClientToken}{User-provided idempotency token.}
}
\value{
A list with the following syntax:\preformatted{list(
  WindowTargetId = "string"
)
}
}
\description{
Registers a target with a maintenance window.
}
\section{Request syntax}{
\preformatted{svc$register_target_with_maintenance_window(
  WindowId = "string",
  ResourceType = "INSTANCE"|"RESOURCE_GROUP",
  Targets = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  OwnerInformation = "string",
  Name = "string",
  Description = "string",
  ClientToken = "string"
)
}
}

\keyword{internal}
