% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_start_associations_once}
\alias{ssm_start_associations_once}
\title{Runs an association immediately and only one time}
\usage{
ssm_start_associations_once(AssociationIds)
}
\arguments{
\item{AssociationIds}{[required] The association IDs that you want to run immediately and only one time.}
}
\description{
Runs an association immediately and only one time. This operation can be helpful when troubleshooting associations.

See \url{https://www.paws-r-sdk.com/docs/ssm_start_associations_once/} for full documentation.
}
\keyword{internal}
