% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_start_execution_preview}
\alias{ssm_start_execution_preview}
\title{Initiates the process of creating a preview showing the effects that
running a specified Automation runbook would have on the targeted
resources}
\usage{
ssm_start_execution_preview(
  DocumentName,
  DocumentVersion = NULL,
  ExecutionInputs = NULL
)
}
\arguments{
\item{DocumentName}{[required] The name of the Automation runbook to run. The result of the execution
preview indicates what the impact would be of running this runbook.}

\item{DocumentVersion}{The version of the Automation runbook to run. The default value is
\verb{$DEFAULT}.}

\item{ExecutionInputs}{Information about the inputs that can be specified for the preview
operation.}
}
\description{
Initiates the process of creating a preview showing the effects that running a specified Automation runbook would have on the targeted resources.

See \url{https://www.paws-r-sdk.com/docs/ssm_start_execution_preview/} for full documentation.
}
\keyword{internal}
