% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_stop_automation_execution}
\alias{ssm_stop_automation_execution}
\title{Stop an Automation that is currently running}
\usage{
ssm_stop_automation_execution(AutomationExecutionId, Type)
}
\arguments{
\item{AutomationExecutionId}{[required] The execution ID of the Automation to stop.}

\item{Type}{The stop request type. Valid types include the following: Cancel and
Complete. The default type is Cancel.}
}
\value{
An empty list.
}
\description{
Stop an Automation that is currently running.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_automation_execution(
  AutomationExecutionId = "string",
  Type = "Complete"|"Cancel"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
