% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_terminate_session}
\alias{ssm_terminate_session}
\title{Permanently ends a session and closes the data connection between the
Session Manager client and SSM Agent on the managed node}
\usage{
ssm_terminate_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The ID of the session to terminate.}
}
\description{
Permanently ends a session and closes the data connection between the Session Manager client and SSM Agent on the managed node. A terminated session can't be resumed.

See \url{https://www.paws-r-sdk.com/docs/ssm_terminate_session/} for full documentation.
}
\keyword{internal}
