% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_unlabel_parameter_version}
\alias{ssm_unlabel_parameter_version}
\title{Remove a label or labels from a parameter}
\usage{
ssm_unlabel_parameter_version(Name, ParameterVersion, Labels)
}
\arguments{
\item{Name}{[required] The name of the parameter from which you want to delete one or more
labels.

You can't enter the Amazon Resource Name (ARN) for a parameter, only the
parameter name itself.}

\item{ParameterVersion}{[required] The specific version of the parameter which you want to delete one or
more labels from. If it isn't present, the call will fail.}

\item{Labels}{[required] One or more labels to delete from the specified parameter version.}
}
\description{
Remove a label or labels from a parameter.

See \url{https://www.paws-r-sdk.com/docs/ssm_unlabel_parameter_version/} for full documentation.
}
\keyword{internal}
