% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_association_status}
\alias{ssm_update_association_status}
\title{Updates the status of the Amazon Web Services Systems Manager document
(SSM document) associated with the specified managed node}
\usage{
ssm_update_association_status(Name, InstanceId, AssociationStatus)
}
\arguments{
\item{Name}{[required] The name of the SSM document.}

\item{InstanceId}{[required] The managed node ID.}

\item{AssociationStatus}{[required] The association status.}
}
\description{
Updates the status of the Amazon Web Services Systems Manager document (SSM document) associated with the specified managed node.

See \url{https://www.paws-r-sdk.com/docs/ssm_update_association_status/} for full documentation.
}
\keyword{internal}
