% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_cloud_connector}
\alias{ssm_update_cloud_connector}
\title{Updates an existing cloud connector with new configuration details}
\usage{
ssm_update_cloud_connector(
  CloudConnectorId,
  DisplayName = NULL,
  Configuration = NULL,
  Description = NULL
)
}
\arguments{
\item{CloudConnectorId}{[required] The ID of the cloud connector to update.}

\item{DisplayName}{A new friendly name for the cloud connector.}

\item{Configuration}{The updated configuration details for connecting to the third-party cloud environment.}

\item{Description}{A new description for the cloud connector.}
}
\description{
Updates an existing cloud connector with new configuration details.

See \url{https://www.paws-r-sdk.com/docs/ssm_update_cloud_connector/} for full documentation.
}
\keyword{internal}
