% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_document_default_version}
\alias{ssm_update_document_default_version}
\title{Set the default version of a document}
\usage{
ssm_update_document_default_version(Name, DocumentVersion)
}
\arguments{
\item{Name}{[required] The name of a custom document that you want to set as the default
version.}

\item{DocumentVersion}{[required] The version of a custom document that you want to set as the default
version.}
}
\value{
A list with the following syntax:\preformatted{list(
  Description = list(
    Name = "string",
    DefaultVersion = "string",
    DefaultVersionName = "string"
  )
)
}
}
\description{
Set the default version of a document.
}
\section{Request syntax}{
\preformatted{svc$update_document_default_version(
  Name = "string",
  DocumentVersion = "string"
)
}
}

\keyword{internal}
