% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_ops_metadata}
\alias{ssm_update_ops_metadata}
\title{Amazon Web Services Systems Manager calls this API operation when you
edit OpsMetadata in Application Manager}
\usage{
ssm_update_ops_metadata(
  OpsMetadataArn,
  MetadataToUpdate = NULL,
  KeysToDelete = NULL
)
}
\arguments{
\item{OpsMetadataArn}{[required] The Amazon Resource Name (ARN) of the OpsMetadata Object to update.}

\item{MetadataToUpdate}{Metadata to add to an OpsMetadata object.}

\item{KeysToDelete}{The metadata keys to delete from the OpsMetadata object.}
}
\description{
Amazon Web Services Systems Manager calls this API operation when you edit OpsMetadata in Application Manager.

See \url{https://www.paws-r-sdk.com/docs/ssm_update_ops_metadata/} for full documentation.
}
\keyword{internal}
