% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_validate_cloud_connector}
\alias{ssm_validate_cloud_connector}
\title{Validates the configuration and connectivity of a cloud connector}
\usage{
ssm_validate_cloud_connector(
  CloudConnectorId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CloudConnectorId}{[required] The ID of the cloud connector to validate.}

\item{MaxResults}{The maximum number of validation findings to return.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Validates the configuration and connectivity of a cloud connector.

See \url{https://www.paws-r-sdk.com/docs/ssm_validate_cloud_connector/} for full documentation.
}
\keyword{internal}
