% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_activate_contact_channel}
\alias{ssmcontacts_activate_contact_channel}
\title{Activates a contact's contact channel}
\usage{
ssmcontacts_activate_contact_channel(ContactChannelId, ActivationCode)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel.}

\item{ActivationCode}{[required] The code sent to the contact channel when it was created in the contact.}
}
\description{
Activates a contact's contact channel. Incident Manager can't engage a contact until the contact channel has been activated.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_activate_contact_channel/} for full documentation.
}
\keyword{internal}
