% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_deactivate_contact_channel}
\alias{ssmcontacts_deactivate_contact_channel}
\title{To no longer receive Incident Manager engagements to a contact channel,
you can deactivate the channel}
\usage{
ssmcontacts_deactivate_contact_channel(ContactChannelId)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel you're
deactivating.}
}
\description{
To no longer receive Incident Manager engagements to a contact channel, you can deactivate the channel.

See \url{https://paws-r.github.io/docs/ssmcontacts/deactivate_contact_channel.html} for full documentation.
}
\keyword{internal}
