% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_contact}
\alias{ssmcontacts_delete_contact}
\title{To remove a contact from Incident Manager, you can delete the contact}
\usage{
ssmcontacts_delete_contact(ContactId)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact that you're deleting.}
}
\description{
To remove a contact from Incident Manager, you can delete the contact. Deleting a contact removes them from all escalation plans and related response plans. Deleting an escalation plan removes it from all related response plans. You will have to recreate the contact and its contact channels before you can use it again.

See \url{https://paws-r.github.io/docs/ssmcontacts/delete_contact.html} for full documentation.
}
\keyword{internal}
