% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_contact_channel}
\alias{ssmcontacts_delete_contact_channel}
\title{To stop receiving engagements on a contact channel, you can delete the
channel from a contact}
\usage{
ssmcontacts_delete_contact_channel(ContactChannelId)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel.}
}
\description{
To stop receiving engagements on a contact channel, you can delete the channel from a contact. Deleting the contact channel does not remove it from the contact's engagement plan, but the stage that includes the channel will be ignored. If you delete the only contact channel for a contact, you'll no longer be able to engage that contact during an incident.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_delete_contact_channel/} for full documentation.
}
\keyword{internal}
