% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_rotation}
\alias{ssmcontacts_delete_rotation}
\title{Deletes a rotation from the system}
\usage{
ssmcontacts_delete_rotation(RotationId)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the on-call rotation to delete.}
}
\description{
Deletes a rotation from the system. If a rotation belongs to more than one on-call schedule, this operation deletes it from all of them.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_delete_rotation/} for full documentation.
}
\keyword{internal}
