% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_rotation_override}
\alias{ssmcontacts_delete_rotation_override}
\title{Deletes an existing override for an on-call rotation}
\usage{
ssmcontacts_delete_rotation_override(RotationId, RotationOverrideId)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the rotation that was overridden.}

\item{RotationOverrideId}{[required] The Amazon Resource Name (ARN) of the on-call rotation override to
delete.}
}
\description{
Deletes an existing override for an on-call rotation.

See \url{https://paws-r.github.io/docs/ssmcontacts/delete_rotation_override.html} for full documentation.
}
\keyword{internal}
