% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_describe_engagement}
\alias{ssmcontacts_describe_engagement}
\title{Incident Manager uses engagements to engage contacts and escalation
plans during an incident}
\usage{
ssmcontacts_describe_engagement(EngagementId)
}
\arguments{
\item{EngagementId}{[required] The Amazon Resource Name (ARN) of the engagement you want the details
of.}
}
\description{
Incident Manager uses engagements to engage contacts and escalation plans during an incident. Use this command to describe the engagement that occurred during an incident.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_describe_engagement/} for full documentation.
}
\keyword{internal}
