% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_get_contact}
\alias{ssmcontacts_get_contact}
\title{Retrieves information about the specified contact or escalation plan}
\usage{
ssmcontacts_get_contact(ContactId)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}
}
\description{
Retrieves information about the specified contact or escalation plan.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_get_contact/} for full documentation.
}
\keyword{internal}
