% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_get_contact_channel}
\alias{ssmcontacts_get_contact_channel}
\title{List details about a specific contact channel}
\usage{
ssmcontacts_get_contact_channel(ContactChannelId)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel you want
information about.}
}
\description{
List details about a specific contact channel.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_get_contact_channel/} for full documentation.
}
\keyword{internal}
