% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_get_contact_policy}
\alias{ssmcontacts_get_contact_policy}
\title{Retrieves the resource policies attached to the specified contact or
escalation plan}
\usage{
ssmcontacts_get_contact_policy(ContactArn)
}
\arguments{
\item{ContactArn}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}
}
\description{
Retrieves the resource policies attached to the specified contact or escalation plan.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_get_contact_policy/} for full documentation.
}
\keyword{internal}
