% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_get_rotation_override}
\alias{ssmcontacts_get_rotation_override}
\title{Retrieves information about an override to an on-call rotation}
\usage{
ssmcontacts_get_rotation_override(RotationId, RotationOverrideId)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the overridden rotation to retrieve
information about.}

\item{RotationOverrideId}{[required] The Amazon Resource Name (ARN) of the on-call rotation override to
retrieve information about.}
}
\description{
Retrieves information about an override to an on-call rotation.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_get_rotation_override/} for full documentation.
}
\keyword{internal}
