% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_contact_channels}
\alias{ssmcontacts_list_contact_channels}
\title{Lists all contact channels for the specified contact}
\usage{
ssmcontacts_list_contact_channels(
  ContactId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact.}

\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of contact channels per page.}
}
\description{
Lists all contact channels for the specified contact.

See \url{https://paws-r.github.io/docs/ssmcontacts/list_contact_channels.html} for full documentation.
}
\keyword{internal}
