% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_contacts}
\alias{ssmcontacts_list_contacts}
\title{Lists all contacts and escalation plans in Incident Manager}
\usage{
ssmcontacts_list_contacts(
  NextToken = NULL,
  MaxResults = NULL,
  AliasPrefix = NULL,
  Type = NULL
)
}
\arguments{
\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of contacts and escalation plans per page of results.}

\item{AliasPrefix}{Used to list only contacts who's aliases start with the specified
prefix.}

\item{Type}{The type of contact.}
}
\description{
Lists all contacts and escalation plans in Incident Manager.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_contacts/} for full documentation.
}
\keyword{internal}
