% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_engagements}
\alias{ssmcontacts_list_engagements}
\title{Lists all engagements that have happened in an incident}
\usage{
ssmcontacts_list_engagements(
  NextToken = NULL,
  MaxResults = NULL,
  IncidentId = NULL,
  TimeRangeValue = NULL
)
}
\arguments{
\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of engagements per page of results.}

\item{IncidentId}{The Amazon Resource Name (ARN) of the incident you're listing
engagements for.}

\item{TimeRangeValue}{The time range to lists engagements for an incident.}
}
\description{
Lists all engagements that have happened in an incident.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_engagements/} for full documentation.
}
\keyword{internal}
