% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_page_receipts}
\alias{ssmcontacts_list_page_receipts}
\title{Lists all of the engagements to contact channels that have been
acknowledged}
\usage{
ssmcontacts_list_page_receipts(PageId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{PageId}{[required] The Amazon Resource Name (ARN) of the engagement to a specific contact
channel.}

\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of acknowledgements per page of results.}
}
\description{
Lists all of the engagements to contact channels that have been acknowledged.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_page_receipts/} for full documentation.
}
\keyword{internal}
