% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_page_resolutions}
\alias{ssmcontacts_list_page_resolutions}
\title{Returns the resolution path of an engagement}
\usage{
ssmcontacts_list_page_resolutions(NextToken = NULL, PageId)
}
\arguments{
\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}

\item{PageId}{[required] The Amazon Resource Name (ARN) of the contact engaged for the incident.}
}
\description{
Returns the resolution path of an engagement. For example, the escalation plan engaged in an incident might target an on-call schedule that includes several contacts in a rotation, but just one contact on-call when the incident starts. The resolution path indicates the hierarchy of \emph{escalation plan \> on-call schedule \> contact}.

See \url{https://paws-r.github.io/docs/ssmcontacts/list_page_resolutions.html} for full documentation.
}
\keyword{internal}
