% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_pages_by_contact}
\alias{ssmcontacts_list_pages_by_contact}
\title{Lists the engagements to a contact's contact channels}
\usage{
ssmcontacts_list_pages_by_contact(
  ContactId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact you are retrieving
engagements for.}

\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of engagements to contact channels to list per page
of results.}
}
\description{
Lists the engagements to a contact's contact channels.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_pages_by_contact/} for full documentation.
}
\keyword{internal}
