% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_pages_by_engagement}
\alias{ssmcontacts_list_pages_by_engagement}
\title{Lists the engagements to contact channels that occurred by engaging a
contact}
\usage{
ssmcontacts_list_pages_by_engagement(
  EngagementId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{EngagementId}{[required] The Amazon Resource Name (ARN) of the engagement.}

\item{NextToken}{The pagination token to continue to the next page of results.}

\item{MaxResults}{The maximum number of engagements to contact channels to list per page
of results.}
}
\description{
Lists the engagements to contact channels that occurred by engaging a contact.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_pages_by_engagement/} for full documentation.
}
\keyword{internal}
