% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_rotation_overrides}
\alias{ssmcontacts_list_rotation_overrides}
\title{Retrieves a list of overrides currently specified for an on-call
rotation}
\usage{
ssmcontacts_list_rotation_overrides(
  RotationId,
  StartTime,
  EndTime,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the rotation to retrieve information
about.}

\item{StartTime}{[required] The date and time for the beginning of a time range for listing
overrides.}

\item{EndTime}{[required] The date and time for the end of a time range for listing overrides.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}
}
\description{
Retrieves a list of overrides currently specified for an on-call rotation.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_rotation_overrides/} for full documentation.
}
\keyword{internal}
