% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_rotation_shifts}
\alias{ssmcontacts_list_rotation_shifts}
\title{Returns a list of shifts generated by an existing rotation in the system}
\usage{
ssmcontacts_list_rotation_shifts(
  RotationId,
  StartTime = NULL,
  EndTime,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the rotation to retrieve shift
information about.}

\item{StartTime}{The date and time for the beginning of the time range to list shifts
for.}

\item{EndTime}{[required] The date and time for the end of the time range to list shifts for.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}
}
\description{
Returns a list of shifts generated by an existing rotation in the system.

See \url{https://paws-r.github.io/docs/ssmcontacts/list_rotation_shifts.html} for full documentation.
}
\keyword{internal}
