% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_list_tags_for_resource}
\alias{ssmcontacts_list_tags_for_resource}
\title{Lists the tags of an escalation plan or contact}
\usage{
ssmcontacts_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}
}
\description{
Lists the tags of an escalation plan or contact.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
