% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_put_contact_policy}
\alias{ssmcontacts_put_contact_policy}
\title{Adds a resource policy to the specified contact or escalation plan}
\usage{
ssmcontacts_put_contact_policy(ContactArn, Policy)
}
\arguments{
\item{ContactArn}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}

\item{Policy}{[required] Details of the resource policy.}
}
\description{
Adds a resource policy to the specified contact or escalation plan. The resource policy is used to share the contact or escalation plan using Resource Access Manager (RAM). For more information about cross-account sharing, see \href{https://docs.aws.amazon.com/incident-manager/latest/userguide/}{Setting up cross-account functionality}.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_put_contact_policy/} for full documentation.
}
\keyword{internal}
