% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_send_activation_code}
\alias{ssmcontacts_send_activation_code}
\title{Sends an activation code to a contact channel}
\usage{
ssmcontacts_send_activation_code(ContactChannelId)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel.}
}
\description{
Sends an activation code to a contact channel. The contact can use this code to activate the contact channel in the console or with the \code{ActivateChannel} operation. Incident Manager can't engage a contact channel until it has been activated.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_send_activation_code/} for full documentation.
}
\keyword{internal}
