% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_tag_resource}
\alias{ssmcontacts_tag_resource}
\title{Tags a contact or escalation plan}
\usage{
ssmcontacts_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}

\item{Tags}{[required] A list of tags that you are adding to the contact or escalation plan.}
}
\description{
Tags a contact or escalation plan. You can tag only contacts and escalation plans in the first region of your replication set.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_tag_resource/} for full documentation.
}
\keyword{internal}
