% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_update_contact}
\alias{ssmcontacts_update_contact}
\title{Updates the contact or escalation plan specified}
\usage{
ssmcontacts_update_contact(ContactId, DisplayName = NULL, Plan = NULL)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan you're
updating.}

\item{DisplayName}{The full name of the contact or escalation plan.}

\item{Plan}{A list of stages. A contact has an engagement plan with stages for
specified contact channels. An escalation plan uses these stages to
contact specified contacts.}
}
\description{
Updates the contact or escalation plan specified.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_update_contact/} for full documentation.
}
\keyword{internal}
