% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_batch_get_incident_findings}
\alias{ssmincidents_batch_get_incident_findings}
\title{Retrieves details about all specified findings for an incident,
including descriptive details about each finding}
\usage{
ssmincidents_batch_get_incident_findings(findingIds, incidentRecordArn)
}
\arguments{
\item{findingIds}{[required] A list of IDs of findings for which you want to view details.}

\item{incidentRecordArn}{[required] The Amazon Resource Name (ARN) of the incident for which you want to
view finding details.}
}
\description{
Retrieves details about all specified findings for an incident, including descriptive details about each finding. A finding represents a recent application environment change made by an CodeDeploy deployment or an CloudFormation stack creation or update that can be investigated as a potential cause of the incident.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_batch_get_incident_findings/} for full documentation.
}
\keyword{internal}
