% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_delete_incident_record}
\alias{ssmincidents_delete_incident_record}
\title{Delete an incident record from Incident Manager}
\usage{
ssmincidents_delete_incident_record(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the incident record you are deleting.}
}
\description{
Delete an incident record from Incident Manager.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_delete_incident_record/} for full documentation.
}
\keyword{internal}
