% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_delete_replication_set}
\alias{ssmincidents_delete_replication_set}
\title{Deletes all Regions in your replication set}
\usage{
ssmincidents_delete_replication_set(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the replication set you're deleting.}
}
\description{
Deletes all Regions in your replication set. Deleting the replication set deletes all Incident Manager data.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_delete_replication_set/} for full documentation.
}
\keyword{internal}
