% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_replication_set}
\alias{ssmincidents_get_replication_set}
\title{Retrieve your Incident Manager replication set}
\usage{
ssmincidents_get_replication_set(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the replication set you want to
retrieve.}
}
\description{
Retrieve your Incident Manager replication set.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_get_replication_set/} for full documentation.
}
\keyword{internal}
