% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_resource_policies}
\alias{ssmincidents_get_resource_policies}
\title{Retrieves the resource policies attached to the specified response plan}
\usage{
ssmincidents_get_resource_policies(
  maxResults = NULL,
  nextToken = NULL,
  resourceArn
)
}
\arguments{
\item{maxResults}{The maximum number of resource policies to display for each page of
results.}

\item{nextToken}{The pagination token to continue to the next page of results.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan with the attached
resource policy.}
}
\description{
Retrieves the resource policies attached to the specified response plan.

See \url{https://paws-r.github.io/docs/ssmincidents/get_resource_policies.html} for full documentation.
}
\keyword{internal}
