% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_response_plan}
\alias{ssmincidents_get_response_plan}
\title{Retrieves the details of the specified response plan}
\usage{
ssmincidents_get_response_plan(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the response plan.}
}
\description{
Retrieves the details of the specified response plan.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_get_response_plan/} for full documentation.
}
\keyword{internal}
