% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_timeline_event}
\alias{ssmincidents_get_timeline_event}
\title{Retrieves a timeline event based on its ID and incident record}
\usage{
ssmincidents_get_timeline_event(eventId, incidentRecordArn)
}
\arguments{
\item{eventId}{[required] The ID of the event. You can get an event's ID when you create it, or by
using \code{\link[=ssmincidents_list_timeline_events]{list_timeline_events}}.}

\item{incidentRecordArn}{[required] The Amazon Resource Name (ARN) of the incident that includes the
timeline event.}
}
\description{
Retrieves a timeline event based on its ID and incident record.

See \url{https://paws-r.github.io/docs/ssmincidents/get_timeline_event.html} for full documentation.
}
\keyword{internal}
