% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_list_replication_sets}
\alias{ssmincidents_list_replication_sets}
\title{Lists details about the replication set configured in your account}
\usage{
ssmincidents_list_replication_sets(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results per page.}

\item{nextToken}{The pagination token for the next set of items to return. (You received
this token from a previous call.)}
}
\description{
Lists details about the replication set configured in your account.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_list_replication_sets/} for full documentation.
}
\keyword{internal}
