% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_list_response_plans}
\alias{ssmincidents_list_response_plans}
\title{Lists all response plans in your account}
\usage{
ssmincidents_list_response_plans(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of response plans per page.}

\item{nextToken}{The pagination token to continue to the next page of results.}
}
\description{
Lists all response plans in your account.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_list_response_plans/} for full documentation.
}
\keyword{internal}
