% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_list_tags_for_resource}
\alias{ssmincidents_list_tags_for_resource}
\title{Lists the tags that are attached to the specified response plan}
\usage{
ssmincidents_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan.}
}
\description{
Lists the tags that are attached to the specified response plan.

See \url{https://paws-r.github.io/docs/ssmincidents/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
