% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_put_resource_policy}
\alias{ssmincidents_put_resource_policy}
\title{Adds a resource policy to the specified response plan}
\usage{
ssmincidents_put_resource_policy(policy, resourceArn)
}
\arguments{
\item{policy}{[required] Details of the resource policy.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan you're adding the
resource policy to.}
}
\description{
Adds a resource policy to the specified response plan. The resource policy is used to share the response plan using Resource Access Manager (RAM). For more information about cross-account sharing, see \href{https://docs.aws.amazon.com/incident-manager/latest/userguide/}{Setting up cross-account functionality}.

See \url{https://paws-r.github.io/docs/ssmincidents/put_resource_policy.html} for full documentation.
}
\keyword{internal}
