% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_tag_resource}
\alias{ssmincidents_tag_resource}
\title{Adds a tag to a response plan}
\usage{
ssmincidents_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan you're adding the
tags to.}

\item{tags}{[required] A list of tags to add to the response plan.}
}
\description{
Adds a tag to a response plan.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_tag_resource/} for full documentation.
}
\keyword{internal}
