% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_untag_resource}
\alias{ssmincidents_untag_resource}
\title{Removes a tag from a resource}
\usage{
ssmincidents_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan you're removing a
tag from.}

\item{tagKeys}{[required] The name of the tag to remove from the response plan.}
}
\description{
Removes a tag from a resource.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_untag_resource/} for full documentation.
}
\keyword{internal}
