% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_update_replication_set}
\alias{ssmincidents_update_replication_set}
\title{Add or delete Regions from your replication set}
\usage{
ssmincidents_update_replication_set(actions, arn, clientToken = NULL)
}
\arguments{
\item{actions}{[required] An action to add or delete a Region.}

\item{arn}{[required] The Amazon Resource Name (ARN) of the replication set you're updating.}

\item{clientToken}{A token that ensures that the operation is called only once with the
specified details.}
}
\description{
Add or delete Regions from your replication set.

See \url{https://paws-r.github.io/docs/ssmincidents/update_replication_set.html} for full documentation.
}
\keyword{internal}
