% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_delete_resource_permission}
\alias{ssmsap_delete_resource_permission}
\title{Removes permissions associated with the target database}
\usage{
ssmsap_delete_resource_permission(
  ActionType = NULL,
  SourceResourceArn = NULL,
  ResourceArn
)
}
\arguments{
\item{ActionType}{Delete or restore the permissions on the target database.}

\item{SourceResourceArn}{The Amazon Resource Name (ARN) of the source resource.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Removes permissions associated with the target database.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_delete_resource_permission/} for full documentation.
}
\keyword{internal}
