% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_deregister_application}
\alias{ssmsap_deregister_application}
\title{Deregister an SAP application with AWS Systems Manager for SAP}
\usage{
ssmsap_deregister_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}
}
\description{
Deregister an SAP application with AWS Systems Manager for SAP. This action does not aﬀect the existing setup of your SAP workloads on Amazon EC2.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_deregister_application/} for full documentation.
}
\keyword{internal}
