% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_application}
\alias{ssmsap_get_application}
\title{Gets an application registered with AWS Systems Manager for SAP}
\usage{
ssmsap_get_application(
  ApplicationId = NULL,
  ApplicationArn = NULL,
  AppRegistryArn = NULL
)
}
\arguments{
\item{ApplicationId}{The ID of the application.}

\item{ApplicationArn}{The Amazon Resource Name (ARN) of the application.}

\item{AppRegistryArn}{The Amazon Resource Name (ARN) of the application registry.}
}
\description{
Gets an application registered with AWS Systems Manager for SAP. It also returns the components of the application.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_application/} for full documentation.
}
\keyword{internal}
