% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_component}
\alias{ssmsap_get_component}
\title{Gets the component of an application registered with AWS Systems Manager
for SAP}
\usage{
ssmsap_get_component(ApplicationId, ComponentId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{ComponentId}{[required] The ID of the component.}
}
\description{
Gets the component of an application registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_component/} for full documentation.
}
\keyword{internal}
