% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_configuration_check_operation}
\alias{ssmsap_get_configuration_check_operation}
\title{Gets the details of a configuration check operation by specifying the
operation ID}
\usage{
ssmsap_get_configuration_check_operation(OperationId)
}
\arguments{
\item{OperationId}{[required] The ID of the configuration check operation.}
}
\description{
Gets the details of a configuration check operation by specifying the operation ID.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_configuration_check_operation/} for full documentation.
}
\keyword{internal}
