% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_database}
\alias{ssmsap_get_database}
\title{Gets the SAP HANA database of an application registered with AWS Systems
Manager for SAP}
\usage{
ssmsap_get_database(
  ApplicationId = NULL,
  ComponentId = NULL,
  DatabaseId = NULL,
  DatabaseArn = NULL
)
}
\arguments{
\item{ApplicationId}{The ID of the application.}

\item{ComponentId}{The ID of the component.}

\item{DatabaseId}{The ID of the database.}

\item{DatabaseArn}{The Amazon Resource Name (ARN) of the database.}
}
\description{
Gets the SAP HANA database of an application registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_database/} for full documentation.
}
\keyword{internal}
