% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_operation}
\alias{ssmsap_get_operation}
\title{Gets the details of an operation by specifying the operation ID}
\usage{
ssmsap_get_operation(OperationId)
}
\arguments{
\item{OperationId}{[required] The ID of the operation.}
}
\description{
Gets the details of an operation by specifying the operation ID.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_operation/} for full documentation.
}
\keyword{internal}
