% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_get_resource_permission}
\alias{ssmsap_get_resource_permission}
\title{Gets permissions associated with the target database}
\usage{
ssmsap_get_resource_permission(ActionType = NULL, ResourceArn)
}
\arguments{
\item{ActionType}{}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Gets permissions associated with the target database.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_get_resource_permission/} for full documentation.
}
\keyword{internal}
