% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_applications}
\alias{ssmsap_list_applications}
\title{Lists all the applications registered with AWS Systems Manager for SAP}
\usage{
ssmsap_list_applications(NextToken = NULL, MaxResults = NULL, Filters = NULL)
}
\arguments{
\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{Filters}{The filter of name, value, and operator.}
}
\description{
Lists all the applications registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_applications/} for full documentation.
}
\keyword{internal}
