% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_configuration_check_definitions}
\alias{ssmsap_list_configuration_check_definitions}
\title{Lists all configuration check types supported by AWS Systems Manager for
SAP}
\usage{
ssmsap_list_configuration_check_definitions(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists all configuration check types supported by AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_configuration_check_definitions/} for full documentation.
}
\keyword{internal}
