% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_configuration_check_operations}
\alias{ssmsap_list_configuration_check_operations}
\title{Lists the configuration check operations performed by AWS Systems
Manager for SAP}
\usage{
ssmsap_list_configuration_check_operations(
  ApplicationId,
  ListMode = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{ListMode}{The mode for listing configuration check operations. Defaults to
"LATEST_PER_CHECK".
\itemize{
\item LATEST_PER_CHECK - Will list the latest configuration check
operation per check type.
\item ALL_OPERATIONS - Will list all configuration check operations
performed on the application.
}}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{NextToken}{The token for the next page of results.}

\item{Filters}{The filters of an operation.}
}
\description{
Lists the configuration check operations performed by AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_configuration_check_operations/} for full documentation.
}
\keyword{internal}
