% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_databases}
\alias{ssmsap_list_databases}
\title{Lists the SAP HANA databases of an application registered with AWS
Systems Manager for SAP}
\usage{
ssmsap_list_databases(
  ApplicationId = NULL,
  ComponentId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ApplicationId}{The ID of the application.}

\item{ComponentId}{The ID of the component.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value. If you do not specify a value for MaxResults, the request returns
50 items per page by default.}
}
\description{
Lists the SAP HANA databases of an application registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_databases/} for full documentation.
}
\keyword{internal}
