% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_operations}
\alias{ssmsap_list_operations}
\title{Lists the operations performed by AWS Systems Manager for SAP}
\usage{
ssmsap_list_operations(
  ApplicationId,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value. If you do not specify a value for MaxResults, the request returns
50 items per page by default.}

\item{NextToken}{The token for the next page of results.}

\item{Filters}{The filters of an operation.}
}
\description{
Lists the operations performed by AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_operations/} for full documentation.
}
\keyword{internal}
