% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_sub_check_results}
\alias{ssmsap_list_sub_check_results}
\title{Lists the sub-check results of a specified configuration check operation}
\usage{
ssmsap_list_sub_check_results(OperationId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{OperationId}{[required] The ID of the configuration check operation.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists the sub-check results of a specified configuration check operation.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_sub_check_results/} for full documentation.
}
\keyword{internal}
