% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_list_tags_for_resource}
\alias{ssmsap_list_tags_for_resource}
\title{Lists all tags on an SAP HANA application and/or database registered
with AWS Systems Manager for SAP}
\usage{
ssmsap_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Lists all tags on an SAP HANA application and/or database registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
