% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_put_resource_permission}
\alias{ssmsap_put_resource_permission}
\title{Adds permissions to the target database}
\usage{
ssmsap_put_resource_permission(ActionType, SourceResourceArn, ResourceArn)
}
\arguments{
\item{ActionType}{[required]}

\item{SourceResourceArn}{[required]}

\item{ResourceArn}{[required]}
}
\description{
Adds permissions to the target database.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_put_resource_permission/} for full documentation.
}
\keyword{internal}
