% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_start_application}
\alias{ssmsap_start_application}
\title{Request is an operation which starts an application}
\usage{
ssmsap_start_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}
}
\description{
Request is an operation which starts an application.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_start_application/} for full documentation.
}
\keyword{internal}
