% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_start_application_refresh}
\alias{ssmsap_start_application_refresh}
\title{Refreshes a registered application}
\usage{
ssmsap_start_application_refresh(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}
}
\description{
Refreshes a registered application.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_start_application_refresh/} for full documentation.
}
\keyword{internal}
