% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_start_configuration_checks}
\alias{ssmsap_start_configuration_checks}
\title{Initiates configuration check operations against a specified application}
\usage{
ssmsap_start_configuration_checks(ApplicationId, ConfigurationCheckIds = NULL)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{ConfigurationCheckIds}{The list of configuration checks to perform.}
}
\description{
Initiates configuration check operations against a specified application.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_start_configuration_checks/} for full documentation.
}
\keyword{internal}
