% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_stop_application}
\alias{ssmsap_stop_application}
\title{Request is an operation to stop an application}
\usage{
ssmsap_stop_application(
  ApplicationId,
  StopConnectedEntity = NULL,
  IncludeEc2InstanceShutdown = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{StopConnectedEntity}{Specify the \code{ConnectedEntityType}. Accepted type is \code{DBMS}.

If this parameter is included, the connected DBMS (Database Management
System) will be stopped.}

\item{IncludeEc2InstanceShutdown}{Boolean. If included and if set to \code{True}, the StopApplication operation
will shut down the associated Amazon EC2 instance in addition to the
application.}
}
\description{
Request is an operation to stop an application.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_stop_application/} for full documentation.
}
\keyword{internal}
