% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_tag_resource}
\alias{ssmsap_tag_resource}
\title{Creates tag for a resource by specifying the ARN}
\usage{
ssmsap_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags on a resource.}
}
\description{
Creates tag for a resource by specifying the ARN.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_tag_resource/} for full documentation.
}
\keyword{internal}
