% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_untag_resource}
\alias{ssmsap_untag_resource}
\title{Delete the tags for a resource}
\usage{
ssmsap_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] Adds/updates or removes credentials for applications registered with AWS
Systems Manager for SAP.}
}
\description{
Delete the tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_untag_resource/} for full documentation.
}
\keyword{internal}
