% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmsap_operations.R
\name{ssmsap_update_application_settings}
\alias{ssmsap_update_application_settings}
\title{Updates the settings of an application registered with AWS Systems
Manager for SAP}
\usage{
ssmsap_update_application_settings(
  ApplicationId,
  CredentialsToAddOrUpdate = NULL,
  CredentialsToRemove = NULL,
  Backint = NULL,
  DatabaseArn = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The ID of the application.}

\item{CredentialsToAddOrUpdate}{The credentials to be added or updated.}

\item{CredentialsToRemove}{The credentials to be removed.}

\item{Backint}{Installation of AWS Backint Agent for SAP HANA.}

\item{DatabaseArn}{The Amazon Resource Name of the SAP HANA database that replaces the
current SAP HANA connection with the SAP_ABAP application.}
}
\description{
Updates the settings of an application registered with AWS Systems Manager for SAP.

See \url{https://www.paws-r-sdk.com/docs/ssmsap_update_application_settings/} for full documentation.
}
\keyword{internal}
