% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_services}
\alias{support_describe_services}
\title{Returns the current list of Amazon Web Services services and a list of
service categories for each service}
\usage{
support_describe_services(
  serviceCodeList = NULL,
  language = NULL,
  dryRun = NULL
)
}
\arguments{
\item{serviceCodeList}{A JSON-formatted list of service codes available for Amazon Web Services services.}

\item{language}{The language in which Amazon Web Services Support handles the case. Amazon Web Services Support currently supports Chinese (“zh”), English ("en"), Japanese ("ja") , Chinese ("zh"), Spanish ("es"), Portuguese ("pt"), French ("fr"), Korean (“ko”), and Turkish ("tr"). You must specify the ISO 639-1 code for the \code{language} parameter if you want support in that language.}

\item{dryRun}{Specifies whether to validate the request without actually returning the list of services. When set to \code{true}, the request is validated but no services are returned, and the operation returns a \code{DryRunOperationException}. When omitted or set to \code{false}, the request runs normally.}
}
\description{
Returns the current list of Amazon Web Services services and a list of service categories for each service. You then use service names and categories in your \code{\link[=support_create_case]{create_case}} requests. Each Amazon Web Services service has its own set of categories.

See \url{https://www.paws-r-sdk.com/docs/support_describe_services/} for full documentation.
}
\keyword{internal}
