% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_check_summaries}
\alias{support_describe_trusted_advisor_check_summaries}
\title{Returns the results for the Trusted Advisor check summaries for the
check IDs that you specified}
\usage{
support_describe_trusted_advisor_check_summaries(checkIds)
}
\arguments{
\item{checkIds}{[required] The IDs of the Trusted Advisor checks.}
}
\description{
Returns the results for the Trusted Advisor check summaries for the check IDs that you specified. You can get the check IDs by calling the \code{\link[=support_describe_trusted_advisor_checks]{describe_trusted_advisor_checks}} operation.

See \url{https://www.paws-r-sdk.com/docs/support_describe_trusted_advisor_check_summaries/} for full documentation.
}
\keyword{internal}
