% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_checks}
\alias{support_describe_trusted_advisor_checks}
\title{Returns information about all available Trusted Advisor checks,
including the name, ID, category, description, and metadata}
\usage{
support_describe_trusted_advisor_checks(language)
}
\arguments{
\item{language}{[required] The ISO 639-1 code for the language that you want your checks to appear
in.

The Amazon Web Services Support API currently supports the following
languages for Trusted Advisor:
\itemize{
\item Chinese, Simplified - \code{zh}
\item Chinese, Traditional - \code{zh_TW}
\item English - \code{en}
\item French - \code{fr}
\item German - \code{de}
\item Indonesian - \code{id}
\item Italian - \code{it}
\item Japanese - \code{ja}
\item Korean - \code{ko}
\item Portuguese, Brazilian - \code{pt_BR}
\item Spanish - \code{es}
}}
}
\description{
Returns information about all available Trusted Advisor checks, including the name, ID, category, description, and metadata. You must specify a language code.

See \url{https://paws-r.github.io/docs/support/describe_trusted_advisor_checks.html} for full documentation.
}
\keyword{internal}
