% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_delete_account_alias}
\alias{supportapp_delete_account_alias}
\title{Deletes an alias for an Amazon Web Services account ID}
\usage{
supportapp_delete_account_alias()
}
\description{
Deletes an alias for an Amazon Web Services account ID. The alias appears in the Amazon Web Services Support App page of the Amazon Web Services Support Center. The alias also appears in Slack messages from the Amazon Web Services Support App.

See \url{https://www.paws-r-sdk.com/docs/supportapp_delete_account_alias/} for full documentation.
}
\keyword{internal}
