% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_delete_slack_channel_configuration}
\alias{supportapp_delete_slack_channel_configuration}
\title{Deletes a Slack channel configuration from your Amazon Web Services
account}
\usage{
supportapp_delete_slack_channel_configuration(channelId, teamId)
}
\arguments{
\item{channelId}{[required] The channel ID in Slack. This ID identifies a channel within a Slack
workspace.}

\item{teamId}{[required] The team ID in Slack. This ID uniquely identifies a Slack workspace,
such as \code{T012ABCDEFG}.}
}
\description{
Deletes a Slack channel configuration from your Amazon Web Services account. This operation doesn't delete your Slack channel.

See \url{https://www.paws-r-sdk.com/docs/supportapp_delete_slack_channel_configuration/} for full documentation.
}
\keyword{internal}
