% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_delete_slack_workspace_configuration}
\alias{supportapp_delete_slack_workspace_configuration}
\title{Deletes a Slack workspace configuration from your Amazon Web Services
account}
\usage{
supportapp_delete_slack_workspace_configuration(teamId)
}
\arguments{
\item{teamId}{[required] The team ID in Slack. This ID uniquely identifies a Slack workspace,
such as \code{T012ABCDEFG}.}
}
\description{
Deletes a Slack workspace configuration from your Amazon Web Services account. This operation doesn't delete your Slack workspace.

See \url{https://www.paws-r-sdk.com/docs/supportapp_delete_slack_workspace_configuration/} for full documentation.
}
\keyword{internal}
