% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_get_account_alias}
\alias{supportapp_get_account_alias}
\title{Retrieves the alias from an Amazon Web Services account ID}
\usage{
supportapp_get_account_alias()
}
\description{
Retrieves the alias from an Amazon Web Services account ID. The alias appears in the Amazon Web Services Support App page of the Amazon Web Services Support Center. The alias also appears in Slack messages from the Amazon Web Services Support App.

See \url{https://www.paws-r-sdk.com/docs/supportapp_get_account_alias/} for full documentation.
}
\keyword{internal}
