% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_list_slack_workspace_configurations}
\alias{supportapp_list_slack_workspace_configurations}
\title{Lists the Slack workspace configurations for an Amazon Web Services
account}
\usage{
supportapp_list_slack_workspace_configurations(nextToken = NULL)
}
\arguments{
\item{nextToken}{If the results of a search are large, the API only returns a portion of
the results and includes a \code{nextToken} pagination token in the response.
To retrieve the next batch of results, reissue the search request and
include the returned token. When the API returns the last set of
results, the response doesn't include a pagination token value.}
}
\description{
Lists the Slack workspace configurations for an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/supportapp_list_slack_workspace_configurations/} for full documentation.
}
\keyword{internal}
