% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_put_account_alias}
\alias{supportapp_put_account_alias}
\title{Creates or updates an individual alias for each Amazon Web Services
account ID}
\usage{
supportapp_put_account_alias(accountAlias)
}
\arguments{
\item{accountAlias}{[required] An alias or short name for an Amazon Web Services account.}
}
\description{
Creates or updates an individual alias for each Amazon Web Services account ID. The alias appears in the Amazon Web Services Support App page of the Amazon Web Services Support Center. The alias also appears in Slack messages from the Amazon Web Services Support App.

See \url{https://www.paws-r-sdk.com/docs/supportapp_put_account_alias/} for full documentation.
}
\keyword{internal}
