% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportapp_operations.R
\name{supportapp_register_slack_workspace_for_organization}
\alias{supportapp_register_slack_workspace_for_organization}
\title{Registers a Slack workspace for your Amazon Web Services account}
\usage{
supportapp_register_slack_workspace_for_organization(teamId)
}
\arguments{
\item{teamId}{[required] The team ID in Slack. This ID uniquely identifies a Slack workspace,
such as \code{T012ABCDEFG}. Specify the Slack workspace that you want to use
for your organization.}
}
\description{
Registers a Slack workspace for your Amazon Web Services account. To call this API, your account must be part of an organization in Organizations.

See \url{https://www.paws-r-sdk.com/docs/supportapp_register_slack_workspace_for_organization/} for full documentation.
}
\keyword{internal}
