% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_disassociate_resource}
\alias{synthetics_disassociate_resource}
\title{Removes a canary from a group}
\usage{
synthetics_disassociate_resource(GroupIdentifier, ResourceArn)
}
\arguments{
\item{GroupIdentifier}{[required] Specifies the group. You can specify the group name, the ARN, or the
group ID as the \code{GroupIdentifier}.}

\item{ResourceArn}{[required] The ARN of the canary that you want to remove from the specified group.}
}
\description{
Removes a canary from a group. You must run this operation in the Region where the canary exists.

See \url{https://paws-r.github.io/docs/synthetics/disassociate_resource.html} for full documentation.
}
\keyword{internal}
