% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_get_group}
\alias{synthetics_get_group}
\title{Returns information about one group}
\usage{
synthetics_get_group(GroupIdentifier)
}
\arguments{
\item{GroupIdentifier}{[required] Specifies the group to return information for. You can specify the group
name, the ARN, or the group ID as the \code{GroupIdentifier}.}
}
\description{
Returns information about one group. Groups are a global resource, so you can use this operation from any Region.

See \url{https://www.paws-r-sdk.com/docs/synthetics_get_group/} for full documentation.
}
\keyword{internal}
