% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_list_associated_groups}
\alias{synthetics_list_associated_groups}
\title{Returns a list of the groups that the specified canary is associated
with}
\usage{
synthetics_list_associated_groups(
  NextToken = NULL,
  MaxResults = NULL,
  ResourceArn
)
}
\arguments{
\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent operation to retrieve the next set of
results.}

\item{MaxResults}{Specify this parameter to limit how many groups are returned each time
you use the
\code{\link[=synthetics_list_associated_groups]{list_associated_groups}} operation.
If you omit this parameter, the default of 20 is used.}

\item{ResourceArn}{[required] The ARN of the canary that you want to view groups for.}
}
\description{
Returns a list of the groups that the specified canary is associated with. The canary that you specify must be in the current Region.

See \url{https://www.paws-r-sdk.com/docs/synthetics_list_associated_groups/} for full documentation.
}
\keyword{internal}
