% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_list_groups}
\alias{synthetics_list_groups}
\title{Returns a list of all groups in the account, displaying their names,
unique IDs, and ARNs}
\usage{
synthetics_list_groups(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent operation to retrieve the next set of
results.}

\item{MaxResults}{Specify this parameter to limit how many groups are returned each time
you use the \code{\link[=synthetics_list_groups]{list_groups}} operation. If you
omit this parameter, the default of 20 is used.}
}
\description{
Returns a list of all groups in the account, displaying their names, unique IDs, and ARNs. The groups from all Regions are returned.

See \url{https://www.paws-r-sdk.com/docs/synthetics_list_groups/} for full documentation.
}
\keyword{internal}
