% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_start_canary_dry_run}
\alias{synthetics_start_canary_dry_run}
\title{Use this operation to start a dry run for a canary that has already been
created}
\usage{
synthetics_start_canary_dry_run(
  Name,
  Code = NULL,
  RuntimeVersion = NULL,
  RunConfig = NULL,
  VpcConfig = NULL,
  ExecutionRoleArn = NULL,
  SuccessRetentionPeriodInDays = NULL,
  FailureRetentionPeriodInDays = NULL,
  VisualReference = NULL,
  ArtifactS3Location = NULL,
  ArtifactConfig = NULL,
  ProvisionedResourceCleanup = NULL,
  BrowserConfigs = NULL,
  VisualReferences = NULL
)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to dry run. To find canary names, use \code{\link[=synthetics_describe_canaries]{describe_canaries}}.}

\item{Code}{Use this structure to input your script code for the canary. This structure contains the Lambda handler with the location where the canary should start running the script. If the script is stored in an Amazon S3 bucket, the bucket name, key, and version are also included. If the script was passed into the canary directly, the script code is contained in the value of \code{Zipfile}.

If you are uploading your canary scripts with an Amazon S3 bucket, your zip file should include your script in a certain folder structure.
\itemize{
\item For Node.js canaries, the folder structure must be \code{nodejs/node_modules/myCanaryFilename.js } For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/#CloudWatch_Synthetics_Canaries_package}{Packaging your Node.js canary files}
\item For Python canaries, the folder structure must be \code{python/myCanaryFilename.py } or \code{python/myFolder/myCanaryFilename.py } For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_WritingCanary_Python.html#CloudWatch_Synthetics_Canaries_WritingCanary_Python_package}{Packaging your Python canary files}
}}

\item{RuntimeVersion}{Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.}

\item{RunConfig}{A structure that contains input information for a canary run.}

\item{VpcConfig}{If this canary is to test an endpoint in a VPC, this structure contains information about the subnets and security groups of the VPC endpoint. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_VPC.html}{Running a Canary in a VPC}.}

\item{ExecutionRoleArn}{The ARN of the IAM role to be used to run the canary. This role must already exist, and must include \code{lambda.amazonaws.com} as a principal in the trust policy. The role must also have the following permissions:}

\item{SuccessRetentionPeriodInDays}{The number of days to retain data about successful runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

This setting affects the range of information returned by \code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of information displayed in the Synthetics console.}

\item{FailureRetentionPeriodInDays}{The number of days to retain data about failed runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

This setting affects the range of information returned by \code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of information displayed in the Synthetics console.}

\item{VisualReference}{An object that specifies what screenshots to use as a baseline for visual monitoring by this canary. It can optionally also specify parts of the screenshots to ignore during the visual monitoring comparison.

Visual monitoring is supported only on canaries running the \strong{syn-puppeteer-node-3.2} runtime or later. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/}{Visual monitoring} and \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/}{Visual monitoring blueprint}}

\item{ArtifactS3Location}{The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the Amazon S3 bucket can't include a period (.).}

\item{ArtifactConfig}{A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.}

\item{ProvisionedResourceCleanup}{Specifies whether to also delete the Lambda functions and layers used by this canary when the canary is deleted. If you omit this parameter, the default of \code{AUTOMATIC} is used, which means that the Lambda functions and layers will be deleted when the canary is deleted.

If the value of this parameter is \code{OFF}, then the value of the \code{DeleteLambda} parameter of the \code{\link[=synthetics_delete_canary]{delete_canary}} operation determines whether the Lambda functions and layers will be deleted.}

\item{BrowserConfigs}{A structure that specifies the browser type to use for a canary run. CloudWatch Synthetics supports running canaries on both \code{CHROME} and \code{FIREFOX} browsers.

If not specified, \code{browserConfigs} defaults to Chrome.}

\item{VisualReferences}{A list of visual reference configurations for the canary, one for each browser type that the canary is configured to run on. Visual references are used for visual monitoring comparisons.

\code{syn-nodejs-puppeteer-11.0} and above, and \code{syn-nodejs-playwright-3.0} and above, only supports \code{visualReferences}. \code{visualReference} field is not supported.

Versions older than \code{syn-nodejs-puppeteer-11.0} supports both \code{visualReference} and \code{visualReferences} for backward compatibility. It is recommended to use \code{visualReferences} for consistency and future compatibility.}
}
\description{
Use this operation to start a dry run for a canary that has already been created

See \url{https://www.paws-r-sdk.com/docs/synthetics_start_canary_dry_run/} for full documentation.
}
\keyword{internal}
