% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_stop_canary}
\alias{synthetics_stop_canary}
\title{Stops the canary to prevent all future runs}
\usage{
synthetics_stop_canary(Name)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to stop. To find the names of your
canaries, use
\href{https://docs.aws.amazon.com/AmazonSynthetics/latest/APIReference/API_DescribeCanaries.html}{ListCanaries}.}
}
\description{
Stops the canary to prevent all future runs. If the canary is currently running,the run that is in progress completes on its own, publishes metrics, and uploads artifacts, but it is not recorded in Synthetics as a completed run.

See \url{https://paws-r.github.io/docs/synthetics/stop_canary.html} for full documentation.
}
\keyword{internal}
