% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_untag_resource}
\alias{synthetics_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
synthetics_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary or group that you're removing tags from.

The ARN format of a canary is
\code{arn:aws:synthetics:Region:account-id:canary:canary-name }.

The ARN format of a group is
\code{arn:aws:synthetics:Region:account-id:group:group-name }}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/synthetics_untag_resource/} for full documentation.
}
\keyword{internal}
