% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_update_canary}
\alias{synthetics_update_canary}
\title{Updates the configuration of a canary that has already been created}
\usage{
synthetics_update_canary(
  Name,
  Code = NULL,
  ExecutionRoleArn = NULL,
  RuntimeVersion = NULL,
  Schedule = NULL,
  RunConfig = NULL,
  SuccessRetentionPeriodInDays = NULL,
  FailureRetentionPeriodInDays = NULL,
  VpcConfig = NULL,
  VisualReference = NULL,
  ArtifactS3Location = NULL,
  ArtifactConfig = NULL,
  ProvisionedResourceCleanup = NULL,
  DryRunId = NULL,
  VisualReferences = NULL,
  BrowserConfigs = NULL,
  AddReplicaLocations = NULL,
  RemoveReplicaLocations = NULL,
  KmsKeyArn = NULL
)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to update. To find the names of your canaries, use \code{\link[=synthetics_describe_canaries]{describe_canaries}}.

You cannot change the name of a canary that has already been created.}

\item{Code}{A structure that includes the entry point from which the canary should start running your script. If the script is stored in an Amazon S3 bucket, the bucket name, key, and version are also included.}

\item{ExecutionRoleArn}{The ARN of the IAM role to be used to run the canary. This role must already exist, and must include \code{lambda.amazonaws.com} as a principal in the trust policy. The role must also have the following permissions:
\itemize{
\item \code{s3:PutObject}
\item \code{s3:GetBucketLocation}
\item \code{s3:ListAllMyBuckets}
\item \code{cloudwatch:PutMetricData}
\item \code{logs:CreateLogGroup}
\item \code{logs:CreateLogStream}
\item \code{logs:CreateLogStream}
}}

\item{RuntimeVersion}{Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.}

\item{Schedule}{A structure that contains information about how often the canary is to run, and when these runs are to stop.}

\item{RunConfig}{A structure that contains the timeout value that is used for each individual run of the canary.

Environment variable keys and values are encrypted at rest using Amazon Web Services owned KMS keys. However, the environment variables are not encrypted on the client side. Do not store sensitive information in them.}

\item{SuccessRetentionPeriodInDays}{The number of days to retain data about successful runs of this canary.

This setting affects the range of information returned by \code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of information displayed in the Synthetics console.}

\item{FailureRetentionPeriodInDays}{The number of days to retain data about failed runs of this canary.

This setting affects the range of information returned by \code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of information displayed in the Synthetics console.}

\item{VpcConfig}{If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_VPC.html}{Running a Canary in a VPC}.}

\item{VisualReference}{Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the \strong{syn-puppeteer-node-3.2} runtime or later. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/}{Visual monitoring} and \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/}{Visual monitoring blueprint}}

\item{ArtifactS3Location}{The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the Amazon S3 bucket can't include a period (.).}

\item{ArtifactConfig}{A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.}

\item{ProvisionedResourceCleanup}{Specifies whether to also delete the Lambda functions and layers used by this canary when the canary is deleted.

If the value of this parameter is \code{OFF}, then the value of the \code{DeleteLambda} parameter of the \code{\link[=synthetics_delete_canary]{delete_canary}} operation determines whether the Lambda functions and layers will be deleted.}

\item{DryRunId}{Update the existing canary using the updated configurations from the DryRun associated with the DryRunId.

When you use the \code{dryRunId} field when updating a canary, the only other field you can provide is the \code{Schedule}. Adding any other field will thrown an exception.}

\item{VisualReferences}{A list of visual reference configurations for the canary, one for each browser type that the canary is configured to run on. Visual references are used for visual monitoring comparisons.

\code{syn-nodejs-puppeteer-11.0} and above, and \code{syn-nodejs-playwright-3.0} and above, only supports \code{visualReferences}. \code{visualReference} field is not supported.

Versions older than \code{syn-nodejs-puppeteer-11.0} supports both \code{visualReference} and \code{visualReferences} for backward compatibility. It is recommended to use \code{visualReferences} for consistency and future compatibility.

For multibrowser visual monitoring, you can update the baseline for all configured browsers in a single update call by specifying a list of VisualReference objects, one per browser. Each VisualReference object maps to a specific browser configuration, allowing you to manage visual baselines for multiple browsers simultaneously.

For single configuration canaries using Chrome browser (default browser), use visualReferences for \code{syn-nodejs-puppeteer-11.0} and above, and \code{syn-nodejs-playwright-3.0} and above canaries. The browserType in the visualReference object is not mandatory.}

\item{BrowserConfigs}{A structure that specifies the browser type to use for a canary run. CloudWatch Synthetics supports running canaries on both \code{CHROME} and \code{FIREFOX} browsers.

If not specified, \code{browserConfigs} defaults to Chrome.}

\item{AddReplicaLocations}{A list of locations (Amazon Web Services Regions) to add as replicas for the canary. Each location specifies a Region and optional VPC configuration for the replica. You can add up to 50 replica locations.}

\item{RemoveReplicaLocations}{A list of locations (Amazon Web Services Regions) to remove as replicas for the canary. You must specify at least one location to remove. All replicas can be removed in a single API call and you cannot remove the primary location.}

\item{KmsKeyArn}{The Amazon Resource Name (ARN) of the customer-managed AWS Key Management Service (AWS KMS) key used to encrypt the canary's AWS Lambda function environment variables at rest. If you don't specify a value, the service uses an AWS-managed key. If you omit this parameter, the service retains the existing value. To revert to the AWS-managed key, set this parameter to an empty string.}
}
\description{
Updates the configuration of a canary that has already been created.

See \url{https://www.paws-r-sdk.com/docs/synthetics_update_canary/} for full documentation.
}
\keyword{internal}
