% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_create_component}
\alias{applicationinsights_create_component}
\title{Creates a custom component by grouping similar standalone instances to
monitor}
\usage{
applicationinsights_create_component(ResourceGroupName, ComponentName,
  ResourceList)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{ResourceList}{[required] The list of resource ARNs that belong to the component.}
}
\value{
An empty list.
}
\description{
Creates a custom component by grouping similar standalone instances to
monitor.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_component(
  ResourceGroupName = "string",
  ComponentName = "string",
  ResourceList = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
