% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_delete_application}
\alias{applicationinsights_delete_application}
\title{Removes the specified application from monitoring}
\usage{
applicationinsights_delete_application(ResourceGroupName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}
}
\value{
An empty list.
}
\description{
Removes the specified application from monitoring. Does not delete the
application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_application(
  ResourceGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
