% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_delete_component}
\alias{applicationinsights_delete_component}
\title{Ungroups a custom component}
\usage{
applicationinsights_delete_component(ResourceGroupName, ComponentName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}
}
\description{
Ungroups a custom component. When you ungroup custom components, all
applicable monitors that are set up for the component are removed and
the instances revert to their standalone status.
}
\section{Request syntax}{
\preformatted{svc$delete_component(
  ResourceGroupName = "string",
  ComponentName = "string"
)
}
}

\keyword{internal}
