% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_component_configuration_recommendation}
\alias{applicationinsights_describe_component_configuration_recommendation}
\title{Describes the recommended monitoring configuration of the component}
\usage{
applicationinsights_describe_component_configuration_recommendation(
  ResourceGroupName,
  ComponentName,
  Tier,
  WorkloadName = NULL,
  RecommendationType = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{Tier}{[required] The tier of the application component.}

\item{WorkloadName}{The name of the workload. The name of the workload is required when the
tier of the application component is \code{SAP_ASE_SINGLE_NODE} or
\code{SAP_ASE_HIGH_AVAILABILITY}.}

\item{RecommendationType}{The recommended configuration type.}
}
\description{
Describes the recommended monitoring configuration of the component.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_component_configuration_recommendation/} for full documentation.
}
\keyword{internal}
