% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_application}
\alias{applicationinsights_describe_application}
\title{Describes the application}
\usage{
applicationinsights_describe_application(ResourceGroupName, AccountId = NULL)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{AccountId}{The AWS account ID for the resource group owner.}
}
\description{
Describes the application.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_application/} for full documentation.
}
\keyword{internal}
