% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_application}
\alias{applicationinsights_describe_application}
\title{Describes the application}
\usage{
applicationinsights_describe_application(ResourceGroupName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApplicationInfo = list(
    ResourceGroupName = "string",
    LifeCycle = "string",
    OpsItemSNSTopicArn = "string",
    OpsCenterEnabled = TRUE|FALSE,
    CWEMonitorEnabled = TRUE|FALSE,
    Remarks = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_application(
  ResourceGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
