% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_component}
\alias{applicationinsights_describe_component}
\title{Describes a component and lists the resources that are grouped together
in a component}
\usage{
applicationinsights_describe_component(ResourceGroupName, ComponentName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}
}
\description{
Describes a component and lists the resources that are grouped together in a component.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_component/} for full documentation.
}
\keyword{internal}
