% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_component_configuration}
\alias{applicationinsights_describe_component_configuration}
\title{Describes the monitoring configuration of the component}
\usage{
applicationinsights_describe_component_configuration(ResourceGroupName,
  ComponentName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Monitor = TRUE|FALSE,
  Tier = "CUSTOM"|"DEFAULT"|"DOT_NET_CORE"|"DOT_NET_WORKER"|"DOT_NET_WEB_TIER"|"DOT_NET_WEB"|"SQL_SERVER"|"SQL_SERVER_ALWAYSON_AVAILABILITY_GROUP"|"MYSQL"|"POSTGRESQL"|"JAVA_JMX"|"ORACLE",
  ComponentConfiguration = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the monitoring configuration of the component.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_component_configuration(
  ResourceGroupName = "string",
  ComponentName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
