% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_log_pattern}
\alias{applicationinsights_describe_log_pattern}
\title{Describe a specific log pattern from a LogPatternSet}
\usage{
applicationinsights_describe_log_pattern(ResourceGroupName,
  PatternSetName, PatternName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{PatternSetName}{[required] The name of the log pattern set.}

\item{PatternName}{[required] The name of the log pattern.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceGroupName = "string",
  LogPattern = list(
    PatternSetName = "string",
    PatternName = "string",
    Pattern = "string",
    Rank = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describe a specific log pattern from a \code{LogPatternSet}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_log_pattern(
  ResourceGroupName = "string",
  PatternSetName = "string",
  PatternName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
