% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_problem}
\alias{applicationinsights_describe_problem}
\title{Describes an application problem}
\usage{
applicationinsights_describe_problem(ProblemId, AccountId = NULL)
}
\arguments{
\item{ProblemId}{[required] The ID of the problem.}

\item{AccountId}{The Amazon Web Services account ID for the owner of the resource group
affected by the problem.}
}
\description{
Describes an application problem.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_problem/} for full documentation.
}
\keyword{internal}
