% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_problem_observations}
\alias{applicationinsights_describe_problem_observations}
\title{Describes the anomalies or errors associated with the problem}
\usage{
applicationinsights_describe_problem_observations(ProblemId)
}
\arguments{
\item{ProblemId}{[required] The ID of the problem.}
}
\description{
Describes the anomalies or errors associated with the problem.
}
\section{Request syntax}{
\preformatted{svc$describe_problem_observations(
  ProblemId = "string"
)
}
}

\keyword{internal}
