% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_problem_observations}
\alias{applicationinsights_describe_problem_observations}
\title{Describes the anomalies or errors associated with the problem}
\usage{
applicationinsights_describe_problem_observations(ProblemId, AccountId = NULL)
}
\arguments{
\item{ProblemId}{[required] The ID of the problem.}

\item{AccountId}{The Amazon Web Services account ID for the resource group owner.}
}
\description{
Describes the anomalies or errors associated with the problem.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_describe_problem_observations/} for full documentation.
}
\keyword{internal}
