% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_components}
\alias{applicationinsights_list_components}
\title{Lists the auto-grouped, standalone, and custom components of the
application}
\usage{
applicationinsights_list_components(
  ResourceGroupName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}
}
\description{
Lists the auto-grouped, standalone, and custom components of the application.

See \url{https://paws-r.github.io/docs/applicationinsights/list_components.html} for full documentation.
}
\keyword{internal}
