% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_log_patterns}
\alias{applicationinsights_list_log_patterns}
\title{Lists the log patterns in the specific log LogPatternSet}
\usage{
applicationinsights_list_log_patterns(ResourceGroupName, PatternSetName,
  MaxResults, NextToken)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{PatternSetName}{The name of the log pattern set.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceGroupName = "string",
  LogPatterns = list(
    list(
      PatternSetName = "string",
      PatternName = "string",
      Pattern = "string",
      Rank = 123
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the log patterns in the specific log \code{LogPatternSet}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_log_patterns(
  ResourceGroupName = "string",
  PatternSetName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
