% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_update_component}
\alias{applicationinsights_update_component}
\title{Updates the custom component name and/or the list of resources that make
up the component}
\usage{
applicationinsights_update_component(ResourceGroupName, ComponentName,
  NewComponentName, ResourceList)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{NewComponentName}{The new name of the component.}

\item{ResourceList}{The list of resource ARNs that belong to the component.}
}
\value{
An empty list.
}
\description{
Updates the custom component name and/or the list of resources that make
up the component.
}
\section{Request syntax}{
\preformatted{svc$update_component(
  ResourceGroupName = "string",
  ComponentName = "string",
  NewComponentName = "string",
  ResourceList = list(
    "string"
  )
)
}
}

\keyword{internal}
