% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_update_component}
\alias{applicationinsights_update_component}
\title{Updates the custom component name and/or the list of resources that make
up the component}
\usage{
applicationinsights_update_component(
  ResourceGroupName,
  ComponentName,
  NewComponentName = NULL,
  ResourceList = NULL
)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}

\item{ComponentName}{[required] The name of the component.}

\item{NewComponentName}{The new name of the component.}

\item{ResourceList}{The list of resource ARNs that belong to the component.}
}
\description{
Updates the custom component name and/or the list of resources that make up the component.

See \url{https://www.paws-r-sdk.com/docs/applicationinsights_update_component/} for full documentation.
}
\keyword{internal}
