% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_get_associated_resource}
\alias{appregistry_get_associated_resource}
\title{Gets the resource associated with the application}
\usage{
appregistry_get_associated_resource(application, resourceType, resource)
}
\arguments{
\item{application}{[required] The name, ID, or ARN of the application.}

\item{resourceType}{[required] The type of resource associated with the application.}

\item{resource}{[required] The name or ID of the resource associated with the application.}
}
\description{
Gets the resource associated with the application.

See \url{https://paws-r.github.io/docs/appregistry/get_associated_resource.html} for full documentation.
}
\keyword{internal}
