% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_get_attribute_group}
\alias{appregistry_get_attribute_group}
\title{Retrieves an attribute group by its ARN, ID, or name}
\usage{
appregistry_get_attribute_group(attributeGroup)
}
\arguments{
\item{attributeGroup}{[required] The name, ID, or ARN of the attribute group that holds the attributes to
describe the application.}
}
\description{
Retrieves an attribute group by its ARN, ID, or name. The attribute group can be specified by its ARN, ID, or name.

See \url{https://paws-r.github.io/docs/appregistry/get_attribute_group.html} for full documentation.
}
\keyword{internal}
