% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_get_configuration}
\alias{appregistry_get_configuration}
\title{Retrieves a TagKey configuration from an account}
\usage{
appregistry_get_configuration()
}
\description{
Retrieves a \code{TagKey} configuration from an account.

See \url{https://www.paws-r-sdk.com/docs/appregistry_get_configuration/} for full documentation.
}
\keyword{internal}
