% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_put_configuration}
\alias{appregistry_put_configuration}
\title{Associates a TagKey configuration to an account}
\usage{
appregistry_put_configuration(configuration)
}
\arguments{
\item{configuration}{[required] Associates a \code{TagKey} configuration to an account.}
}
\description{
Associates a \code{TagKey} configuration to an account.

See \url{https://www.paws-r-sdk.com/docs/appregistry_put_configuration/} for full documentation.
}
\keyword{internal}
