% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_import_evidence_to_assessment_control}
\alias{auditmanager_batch_import_evidence_to_assessment_control}
\title{Adds one or more pieces of evidence to a control in an Audit Manager
assessment}
\usage{
auditmanager_batch_import_evidence_to_assessment_control(
  assessmentId,
  controlSetId,
  controlId,
  manualEvidence
)
}
\arguments{
\item{assessmentId}{[required] The identifier for the assessment.}

\item{controlSetId}{[required] The identifier for the control set.}

\item{controlId}{[required] The identifier for the control.}

\item{manualEvidence}{[required] The list of manual evidence objects.}
}
\description{
Adds one or more pieces of evidence to a control in an Audit Manager assessment.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_batch_import_evidence_to_assessment_control/} for full documentation.
}
\keyword{internal}
