% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment}
\alias{auditmanager_create_assessment}
\title{Creates an assessment in Audit Manager}
\usage{
auditmanager_create_assessment(
  name,
  description = NULL,
  assessmentReportsDestination,
  scope,
  roles,
  frameworkId,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the assessment to be created.}

\item{description}{The optional description of the assessment to be created.}

\item{assessmentReportsDestination}{[required] The assessment report storage destination for the assessment that's
being created.}

\item{scope}{[required]}

\item{roles}{[required] The list of roles for the assessment.}

\item{frameworkId}{[required] The identifier for the framework that the assessment will be created
from.}

\item{tags}{The tags that are associated with the assessment.}
}
\description{
Creates an assessment in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_create_assessment/} for full documentation.
}
\keyword{internal}
