% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_evidence_folder}
\alias{auditmanager_get_evidence_folder}
\title{Returns an evidence folder from the specified assessment in Audit
Manager}
\usage{
auditmanager_get_evidence_folder(assessmentId, controlSetId, evidenceFolderId)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{controlSetId}{[required] The unique identifier for the control set.}

\item{evidenceFolderId}{[required] The unique identifier for the folder that the evidence is stored in.}
}
\description{
Returns an evidence folder from the specified assessment in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/get_evidence_folder.html} for full documentation.
}
\keyword{internal}
