% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_control_insights_by_control_domain}
\alias{auditmanager_list_control_insights_by_control_domain}
\title{Lists the latest analytics data for controls within a specific control
domain across all active assessments}
\usage{
auditmanager_list_control_insights_by_control_domain(
  controlDomainId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{controlDomainId}{[required] The unique identifier for the control domain.

Audit Manager supports the control domains that are provided by Amazon
Web Services Control Catalog. For information about how to find a list
of available control domains, see
\href{https://docs.aws.amazon.com/controlcatalog/latest/APIReference/API_ListDomains.html}{\code{ListDomains}}
in the Amazon Web Services Control Catalog API Reference.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Lists the latest analytics data for controls within a specific control domain across all active assessments.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_list_control_insights_by_control_domain/} for full documentation.
}
\keyword{internal}
