% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_tags_for_resource}
\alias{auditmanager_list_tags_for_resource}
\title{Returns a list of tags for the specified resource in Audit Manager}
\usage{
auditmanager_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Returns a list of tags for the specified resource in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
