% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_tag_resource}
\alias{auditmanager_tag_resource}
\title{Tags the specified resource in Audit Manager}
\usage{
auditmanager_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags that are associated with the resource.}
}
\description{
Tags the specified resource in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_tag_resource/} for full documentation.
}
\keyword{internal}
