% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_assessment_framework}
\alias{auditmanager_update_assessment_framework}
\title{Updates a custom framework in Audit Manager}
\usage{
auditmanager_update_assessment_framework(
  frameworkId,
  name,
  description = NULL,
  complianceType = NULL,
  controlSets
)
}
\arguments{
\item{frameworkId}{[required] The unique identifier for the framework.}

\item{name}{[required] The name of the framework to be updated.}

\item{description}{The description of the updated framework.}

\item{complianceType}{The compliance type that the new custom framework supports, such as CIS
or HIPAA.}

\item{controlSets}{[required] The control sets that are associated with the framework.

The \code{Controls} object returns a partial response when called through
Framework APIs. For a complete \code{Controls} object, use
\code{\link[=auditmanager_get_control]{get_control}}.}
}
\description{
Updates a custom framework in Audit Manager.

See \url{https://www.paws-r-sdk.com/docs/auditmanager_update_assessment_framework/} for full documentation.
}
\keyword{internal}
