% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_attach_instances}
\alias{autoscaling_attach_instances}
\title{Attaches one or more EC2 instances to the specified Auto Scaling group}
\usage{
autoscaling_attach_instances(InstanceIds = NULL, AutoScalingGroupName)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}
}
\description{
Attaches one or more EC2 instances to the specified Auto Scaling group.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_attach_instances/} for full documentation.
}
\keyword{internal}
