% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_cancel_instance_refresh}
\alias{autoscaling_cancel_instance_refresh}
\title{Cancels an instance refresh or rollback that is in progress}
\usage{
autoscaling_cancel_instance_refresh(
  AutoScalingGroupName,
  WaitForTransitioningInstances = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{WaitForTransitioningInstances}{When cancelling an instance refresh, this indicates whether to wait for
in-flight launches and terminations to complete. The default is true.

When set to false, Amazon EC2 Auto Scaling cancels the instance refresh
without waiting for any pending launches or terminations to complete.}
}
\description{
Cancels an instance refresh or rollback that is in progress. If an instance refresh or rollback is not in progress, an \code{ActiveInstanceRefreshNotFound} error occurs.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_cancel_instance_refresh/} for full documentation.
}
\keyword{internal}
