% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_scheduled_action}
\alias{autoscaling_delete_scheduled_action}
\title{Deletes the specified scheduled action}
\usage{
autoscaling_delete_scheduled_action(AutoScalingGroupName,
  ScheduledActionName)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScheduledActionName}{[required] The name of the action to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified scheduled action.
}
\section{Request syntax}{
\preformatted{svc$delete_scheduled_action(
  AutoScalingGroupName = "string",
  ScheduledActionName = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the specified scheduled action from the specified
# Auto Scaling group.
svc$delete_scheduled_action(
  AutoScalingGroupName = "my-auto-scaling-group",
  ScheduledActionName = "my-scheduled-action"
)
}

}
\keyword{internal}
