% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_delete_warm_pool}
\alias{autoscaling_delete_warm_pool}
\title{Deletes the warm pool for the specified Auto Scaling group}
\usage{
autoscaling_delete_warm_pool(AutoScalingGroupName, ForceDelete = NULL)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ForceDelete}{Specifies that the warm pool is to be deleted along with all of its
associated instances, without waiting for all instances to be
terminated. This parameter also deletes any outstanding lifecycle
actions associated with the warm pool instances.}
}
\description{
Deletes the warm pool for the specified Auto Scaling group.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_delete_warm_pool/} for full documentation.
}
\keyword{internal}
