% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_adjustment_types}
\alias{autoscaling_describe_adjustment_types}
\title{Describes the available adjustment types for Amazon EC2 Auto Scaling
scaling policies}
\usage{
autoscaling_describe_adjustment_types()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AdjustmentTypes = list(
    list(
      AdjustmentType = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the available adjustment types for Amazon EC2 Auto Scaling
scaling policies. These settings apply to step scaling policies and
simple scaling policies; they do not apply to target tracking scaling
policies.

The following adjustment types are supported:
\itemize{
\item ChangeInCapacity
\item ExactCapacity
\item PercentChangeInCapacity
}
}
\examples{
\dontrun{
# This example describes the available adjustment types.
svc$describe_adjustment_types()
}

}
\keyword{internal}
