% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_adjustment_types}
\alias{autoscaling_describe_adjustment_types}
\title{Describes the available adjustment types for step scaling and simple
scaling policies}
\usage{
autoscaling_describe_adjustment_types()
}
\description{
Describes the available adjustment types for step scaling and simple scaling policies.

See \url{https://paws-r.github.io/docs/autoscaling/describe_adjustment_types.html} for full documentation.
}
\keyword{internal}
