% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_auto_scaling_groups}
\alias{autoscaling_describe_auto_scaling_groups}
\title{Gets information about the Auto Scaling groups in the account and Region}
\usage{
autoscaling_describe_auto_scaling_groups(
  AutoScalingGroupNames = NULL,
  IncludeInstances = NULL,
  NextToken = NULL,
  MaxRecords = NULL,
  Filters = NULL
)
}
\arguments{
\item{AutoScalingGroupNames}{The names of the Auto Scaling groups. By default, you can only specify
up to 50 names. You can optionally increase this limit using the
\code{MaxRecords} property.

If you omit this property, all Auto Scaling groups are described.}

\item{IncludeInstances}{Specifies whether to include information about Amazon EC2 instances in
the response. When set to \code{true} (default), the response includes
instance details.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{100}.}

\item{Filters}{One or more filters to limit the results based on specific tags.}
}
\description{
Gets information about the Auto Scaling groups in the account and Region.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_auto_scaling_groups/} for full documentation.
}
\keyword{internal}
