% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_auto_scaling_notification_types}
\alias{autoscaling_describe_auto_scaling_notification_types}
\title{Describes the notification types that are supported by Amazon EC2 Auto
Scaling}
\usage{
autoscaling_describe_auto_scaling_notification_types()
}
\description{
Describes the notification types that are supported by Amazon EC2 Auto Scaling.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_auto_scaling_notification_types/} for full documentation.
}
\keyword{internal}
