% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_instance_refreshes}
\alias{autoscaling_describe_instance_refreshes}
\title{Gets information about the instance refreshes for the specified Auto
Scaling group from the previous six weeks}
\usage{
autoscaling_describe_instance_refreshes(
  AutoScalingGroupName,
  InstanceRefreshIds = NULL,
  NextToken = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{InstanceRefreshIds}{One or more instance refresh IDs.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{100}.}
}
\description{
Gets information about the instance refreshes for the specified Auto Scaling group from the previous six weeks.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_instance_refreshes/} for full documentation.
}
\keyword{internal}
