% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_lifecycle_hook_types}
\alias{autoscaling_describe_lifecycle_hook_types}
\title{Describes the available types of lifecycle hooks}
\usage{
autoscaling_describe_lifecycle_hook_types()
}
\description{
Describes the available types of lifecycle hooks.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_lifecycle_hook_types/} for full documentation.
}
\keyword{internal}
