% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_lifecycle_hook_types}
\alias{autoscaling_describe_lifecycle_hook_types}
\title{Describes the available types of lifecycle hooks}
\usage{
autoscaling_describe_lifecycle_hook_types()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LifecycleHookTypes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the available types of lifecycle hooks.

The following hook types are supported:
\itemize{
\item autoscaling:EC2_INSTANCE_LAUNCHING
\item autoscaling:EC2_INSTANCE_TERMINATING
}
}
\examples{
\dontrun{
# This example describes the available lifecycle hook types.
svc$describe_lifecycle_hook_types()
}

}
\keyword{internal}
