% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_lifecycle_hooks}
\alias{autoscaling_describe_lifecycle_hooks}
\title{Gets information about the lifecycle hooks for the specified Auto
Scaling group}
\usage{
autoscaling_describe_lifecycle_hooks(
  AutoScalingGroupName,
  LifecycleHookNames = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LifecycleHookNames}{The names of one or more lifecycle hooks. If you omit this property, all
lifecycle hooks are described.}
}
\description{
Gets information about the lifecycle hooks for the specified Auto Scaling group.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_describe_lifecycle_hooks/} for full documentation.
}
\keyword{internal}
