% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_metric_collection_types}
\alias{autoscaling_describe_metric_collection_types}
\title{Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling}
\usage{
autoscaling_describe_metric_collection_types()
}
\description{
Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling.

The \code{GroupStandbyInstances} metric is not returned by default. You must
explicitly request this metric when calling the
\code{\link[=autoscaling_enable_metrics_collection]{enable_metrics_collection}}
API.
}
\examples{
\dontrun{
# This example describes the available metric collection types.
svc$describe_metric_collection_types()
}

}
\keyword{internal}
