% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_termination_policy_types}
\alias{autoscaling_describe_termination_policy_types}
\title{Describes the termination policies supported by Amazon EC2 Auto Scaling}
\usage{
autoscaling_describe_termination_policy_types()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TerminationPolicyTypes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the termination policies supported by Amazon EC2 Auto Scaling.

For more information, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-instance-termination.html}{Controlling which Auto Scaling instances terminate during scale in}
in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
\dontrun{
# This example describes the available termination policy types.
svc$describe_termination_policy_types()
}

}
\keyword{internal}
