% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_detach_load_balancer_target_groups}
\alias{autoscaling_detach_load_balancer_target_groups}
\title{This API operation is superseded by DetachTrafficSources, which can
detach multiple traffic sources types}
\usage{
autoscaling_detach_load_balancer_target_groups(
  AutoScalingGroupName,
  TargetGroupARNs
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{TargetGroupARNs}{[required] The Amazon Resource Names (ARN) of the target groups. You can specify up
to 10 target groups.}
}
\description{
This API operation is superseded by \code{\link[=autoscaling_detach_traffic_sources]{detach_traffic_sources}}, which can detach multiple traffic sources types. We recommend using \code{\link[=autoscaling_detach_traffic_sources]{detach_traffic_sources}} to simplify how you manage traffic sources. However, we continue to support \code{\link[=autoscaling_detach_load_balancer_target_groups]{detach_load_balancer_target_groups}}. You can use both the original \code{\link[=autoscaling_detach_load_balancer_target_groups]{detach_load_balancer_target_groups}} API operation and \code{\link[=autoscaling_detach_traffic_sources]{detach_traffic_sources}} on the same Auto Scaling group.

See \url{https://paws-r.github.io/docs/autoscaling/detach_load_balancer_target_groups.html} for full documentation.
}
\keyword{internal}
