% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_enable_metrics_collection}
\alias{autoscaling_enable_metrics_collection}
\title{Enables group metrics for the specified Auto Scaling group}
\usage{
autoscaling_enable_metrics_collection(AutoScalingGroupName, Metrics,
  Granularity)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{Metrics}{Specifies which group-level metrics to start collecting. You can specify
one or more of the following metrics:
\itemize{
\item \code{GroupMinSize}
\item \code{GroupMaxSize}
\item \code{GroupDesiredCapacity}
\item \code{GroupInServiceInstances}
\item \code{GroupPendingInstances}
\item \code{GroupStandbyInstances}
\item \code{GroupTerminatingInstances}
\item \code{GroupTotalInstances}
}

The instance weighting feature supports the following additional
metrics:
\itemize{
\item \code{GroupInServiceCapacity}
\item \code{GroupPendingCapacity}
\item \code{GroupStandbyCapacity}
\item \code{GroupTerminatingCapacity}
\item \code{GroupTotalCapacity}
}

If you omit this parameter, all metrics are enabled.}

\item{Granularity}{[required] The granularity to associate with the metrics to collect. The only valid
value is \verb{1Minute}.}
}
\value{
An empty list.
}
\description{
Enables group metrics for the specified Auto Scaling group. For more
information, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-instance-monitoring.html}{Monitoring CloudWatch metrics for your Auto Scaling groups and instances}
in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\section{Request syntax}{
\preformatted{svc$enable_metrics_collection(
  AutoScalingGroupName = "string",
  Metrics = list(
    "string"
  ),
  Granularity = "string"
)
}
}

\examples{
\dontrun{
# This example enables data collection for the specified Auto Scaling
# group.
svc$enable_metrics_collection(
  AutoScalingGroupName = "my-auto-scaling-group",
  Granularity = "1Minute"
)
}

}
\keyword{internal}
