% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_exit_standby}
\alias{autoscaling_exit_standby}
\title{Moves the specified instances out of the standby state}
\usage{
autoscaling_exit_standby(InstanceIds = NULL, AutoScalingGroupName)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}
}
\description{
Moves the specified instances out of the standby state.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_exit_standby/} for full documentation.
}
\keyword{internal}
