% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_set_desired_capacity}
\alias{autoscaling_set_desired_capacity}
\title{Sets the size of the specified Auto Scaling group}
\usage{
autoscaling_set_desired_capacity(
  AutoScalingGroupName,
  DesiredCapacity,
  HonorCooldown = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{DesiredCapacity}{[required] The desired capacity is the initial capacity of the Auto Scaling group
after this operation completes and the capacity it attempts to maintain.}

\item{HonorCooldown}{Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period
to complete before initiating a scaling activity to set your Auto
Scaling group to its new capacity. By default, Amazon EC2 Auto Scaling
does not honor the cooldown period during manual scaling activities.}
}
\description{
Sets the size of the specified Auto Scaling group.

See \url{https://www.paws-r-sdk.com/docs/autoscaling_set_desired_capacity/} for full documentation.
}
\keyword{internal}
