% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscalingplans_operations.R
\name{autoscalingplans_delete_scaling_plan}
\alias{autoscalingplans_delete_scaling_plan}
\title{Deletes the specified scaling plan}
\usage{
autoscalingplans_delete_scaling_plan(ScalingPlanName, ScalingPlanVersion)
}
\arguments{
\item{ScalingPlanName}{[required] The name of the scaling plan.}

\item{ScalingPlanVersion}{[required] The version number of the scaling plan. Currently, the only valid value
is \code{1}.}
}
\description{
Deletes the specified scaling plan.

See \url{https://paws-r.github.io/docs/autoscalingplans/delete_scaling_plan.html} for full documentation.
}
\keyword{internal}
