% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscalingplans_operations.R
\name{autoscalingplans_delete_scaling_plan}
\alias{autoscalingplans_delete_scaling_plan}
\title{Deletes the specified scaling plan}
\usage{
autoscalingplans_delete_scaling_plan(ScalingPlanName,
  ScalingPlanVersion)
}
\arguments{
\item{ScalingPlanName}{[required] The name of the scaling plan.}

\item{ScalingPlanVersion}{[required] The version number of the scaling plan. Currently, the only valid value
is \code{1}.}
}
\value{
An empty list.
}
\description{
Deletes the specified scaling plan.

Deleting a scaling plan deletes the underlying ScalingInstruction for
all of the scalable resources that are covered by the plan.

If the plan has launched resources or has scaling activities in
progress, you must delete those resources separately.
}
\section{Request syntax}{
\preformatted{svc$delete_scaling_plan(
  ScalingPlanName = "string",
  ScalingPlanVersion = 123
)
}
}

\keyword{internal}
