% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscalingplans_operations.R
\name{autoscalingplans_update_scaling_plan}
\alias{autoscalingplans_update_scaling_plan}
\title{Updates the specified scaling plan}
\usage{
autoscalingplans_update_scaling_plan(
  ScalingPlanName,
  ScalingPlanVersion,
  ApplicationSource = NULL,
  ScalingInstructions = NULL
)
}
\arguments{
\item{ScalingPlanName}{[required] The name of the scaling plan.}

\item{ScalingPlanVersion}{[required] The version number of the scaling plan. The only valid value is \code{1}.
Currently, you cannot have multiple scaling plan versions.}

\item{ApplicationSource}{A CloudFormation stack or set of tags.

For more information, see
\href{https://docs.aws.amazon.com/autoscaling/plans/APIReference/API_ApplicationSource.html}{ApplicationSource}
in the \emph{AWS Auto Scaling API Reference}.}

\item{ScalingInstructions}{The scaling instructions.

For more information, see
\href{https://docs.aws.amazon.com/autoscaling/plans/APIReference/API_ScalingInstruction.html}{ScalingInstruction}
in the \emph{AWS Auto Scaling API Reference}.}
}
\description{
Updates the specified scaling plan.

See \url{https://www.paws-r-sdk.com/docs/autoscalingplans_update_scaling_plan/} for full documentation.
}
\keyword{internal}
