% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_account_limits}
\alias{cloudformation_describe_account_limits}
\title{Retrieves your account's CloudFormation limits, such as the maximum
number of stacks that you can create in your account}
\usage{
cloudformation_describe_account_limits(NextToken = NULL)
}
\arguments{
\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Retrieves your account's CloudFormation limits, such as the maximum number of stacks that you can create in your account. For more information about account limits, see \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cloudformation-limits.html}{Understand CloudFormation quotas} in the \emph{CloudFormation User Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_account_limits/} for full documentation.
}
\keyword{internal}
