% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_events}
\alias{cloudformation_describe_events}
\title{Returns CloudFormation events based on flexible query criteria}
\usage{
cloudformation_describe_events(
  StackName = NULL,
  ChangeSetName = NULL,
  OperationId = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StackName}{The name or unique stack ID for which you want to retrieve events.}

\item{ChangeSetName}{The name or Amazon Resource Name (ARN) of the change set for which you
want to retrieve events.}

\item{OperationId}{The unique identifier of the operation for which you want to retrieve
events.}

\item{Filters}{Filters to apply when retrieving events.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Returns CloudFormation events based on flexible query criteria. Groups events by operation ID, enabling you to focus on individual stack operations during deployment.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_events/} for full documentation.
}
\keyword{internal}
