% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_events}
\alias{cloudformation_describe_stack_events}
\title{Returns all stack related events for a specified stack in reverse
chronological order}
\usage{
cloudformation_describe_stack_events(StackName = NULL, NextToken = NULL)
}
\arguments{
\item{StackName}{The name or the unique stack ID that's associated with the stack, which
aren't always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack's name or its unique
stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}

Default: There is no default value.}

\item{NextToken}{A string that identifies the next page of events that you want to
retrieve.}
}
\description{
Returns all stack related events for a specified stack in reverse chronological order. For more information about a stack's event history, see \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/stack-resource-configuration-complete.html}{CloudFormation stack creation events} in the \emph{CloudFormation User Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_stack_events/} for full documentation.
}
\keyword{internal}
