% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_resource}
\alias{cloudformation_describe_stack_resource}
\title{Returns a description of the specified resource in the specified stack}
\usage{
cloudformation_describe_stack_resource(StackName, LogicalResourceId)
}
\arguments{
\item{StackName}{[required] The name or the unique stack ID that's associated with the stack, which
aren't always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack's name or its
unique stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}}

\item{LogicalResourceId}{[required] The logical name of the resource as specified in the template.}
}
\description{
Returns a description of the specified resource in the specified stack.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_describe_stack_resource/} for full documentation.
}
\keyword{internal}
