% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_type_registration}
\alias{cloudformation_describe_type_registration}
\title{Returns information about a type's registration, including its current
status and type and version identifiers}
\usage{
cloudformation_describe_type_registration(RegistrationToken)
}
\arguments{
\item{RegistrationToken}{[required] The identifier for this registration request.

This registration token is generated by CloudFormation when you initiate
a registration request using \verb{<a>RegisterType</a>}.}
}
\description{
Returns information about a type's registration, including its current
status and type and version identifiers.

When you initiate a registration request using \verb{<a>RegisterType</a>},
you can then use \verb{<a>DescribeTypeRegistration</a>} to monitor the
progress of that registration request.

Once the registration request has completed, use \verb{<a>DescribeType</a>}
to return detailed informaiton about a type.
}
\section{Request syntax}{
\preformatted{svc$describe_type_registration(
  RegistrationToken = "string"
)
}
}

\keyword{internal}
