% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_change_sets}
\alias{cloudformation_list_change_sets}
\title{Returns the ID and status of each active change set for a stack}
\usage{
cloudformation_list_change_sets(StackName, NextToken = NULL)
}
\arguments{
\item{StackName}{[required] The name or the Amazon Resource Name (ARN) of the stack for which you
want to list change sets.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Returns the ID and status of each active change set for a stack. For example, CloudFormation lists change sets that are in the \code{CREATE_IN_PROGRESS} or \code{CREATE_PENDING} state.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_change_sets/} for full documentation.
}
\keyword{internal}
