% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_imports}
\alias{cloudformation_list_imports}
\title{Lists all stacks that are importing an exported output value}
\usage{
cloudformation_list_imports(ExportName, NextToken = NULL)
}
\arguments{
\item{ExportName}{[required] The name of the exported output value. CloudFormation returns the stack
names that are importing this value.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Lists all stacks that are importing an exported output value. To modify or remove an exported output value, first use this action to see which stacks are using it. To see the exported output values in your account, see \code{\link[=cloudformation_list_exports]{list_exports}}.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_imports/} for full documentation.
}
\keyword{internal}
