% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_list_stack_resources}
\alias{cloudformation_list_stack_resources}
\title{Returns descriptions of all resources of the specified stack}
\usage{
cloudformation_list_stack_resources(StackName, NextToken = NULL)
}
\arguments{
\item{StackName}{[required] The name or the unique stack ID that is associated with the stack, which
aren't always interchangeable:
\itemize{
\item Running stacks: You can specify either the stack's name or its
unique stack ID.
\item Deleted stacks: You must specify the unique stack ID.
}}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Returns descriptions of all resources of the specified stack.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_list_stack_resources/} for full documentation.
}
\keyword{internal}
