% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_set_type_default_version}
\alias{cloudformation_set_type_default_version}
\title{Specify the default version of an extension}
\usage{
cloudformation_set_type_default_version(
  Arn = NULL,
  Type = NULL,
  TypeName = NULL,
  VersionId = NULL
)
}
\arguments{
\item{Arn}{The Amazon Resource Name (ARN) of the extension for which you want
version summary information.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{Type}{The kind of extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{TypeName}{The name of the extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{VersionId}{The ID of a specific version of the extension. The version ID is the
value at the end of the Amazon Resource Name (ARN) assigned to the
extension version when it is registered.}
}
\description{
Specify the default version of an extension. The default version of an extension will be used in CloudFormation operations.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_set_type_default_version/} for full documentation.
}
\keyword{internal}
