% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_validate_template}
\alias{cloudformation_validate_template}
\title{Validates a specified template}
\usage{
cloudformation_validate_template(TemplateBody = NULL, TemplateURL = NULL)
}
\arguments{
\item{TemplateBody}{Structure that contains the template body with a minimum length of 1
byte and a maximum length of 51,200 bytes.

Conditional: You must pass \code{TemplateURL} or \code{TemplateBody}. If both are
passed, only \code{TemplateBody} is used.}

\item{TemplateURL}{The URL of a file that contains the template body. The URL must point to
a template (max size: 1 MB) that is located in an Amazon S3 bucket or a
Systems Manager document. The location for an Amazon S3 bucket must
start with \verb{https://}.

Conditional: You must pass \code{TemplateURL} or \code{TemplateBody}. If both are
passed, only \code{TemplateBody} is used.}
}
\description{
Validates a specified template. CloudFormation first checks if the template is valid JSON. If it isn't, CloudFormation checks if the template is valid YAML. If both these checks fail, CloudFormation returns a template validation error.

See \url{https://www.paws-r-sdk.com/docs/cloudformation_validate_template/} for full documentation.
}
\keyword{internal}
