% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_delete_channel}
\alias{cloudtrail_delete_channel}
\title{Deletes a channel}
\usage{
cloudtrail_delete_channel(Channel)
}
\arguments{
\item{Channel}{[required] The ARN or the \code{UUID} value of the channel that you want to delete.}
}
\description{
Deletes a channel.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_delete_channel/} for full documentation.
}
\keyword{internal}
