% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_delete_dashboard}
\alias{cloudtrail_delete_dashboard}
\title{Deletes the specified dashboard}
\usage{
cloudtrail_delete_dashboard(DashboardId)
}
\arguments{
\item{DashboardId}{[required] The name or ARN for the dashboard.}
}
\description{
Deletes the specified dashboard. You cannot delete a dashboard that has termination protection enabled.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_delete_dashboard/} for full documentation.
}
\keyword{internal}
