% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_delete_trail}
\alias{cloudtrail_delete_trail}
\title{Deletes a trail}
\usage{
cloudtrail_delete_trail(Name)
}
\arguments{
\item{Name}{[required] Specifies the name or the CloudTrail ARN of the trail to be deleted. The
following is the format of a trail ARN.
\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}
}
\description{
Deletes a trail. This operation must be called from the region in which the trail was created. \code{\link[=cloudtrail_delete_trail]{delete_trail}} cannot be called on the shadow trails (replicated trails in other regions) of a trail that is enabled in all regions.

See \url{https://paws-r.github.io/docs/cloudtrail/delete_trail.html} for full documentation.
}
\keyword{internal}
