% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_describe_query}
\alias{cloudtrail_describe_query}
\title{Returns metadata about a query, including query run time in
milliseconds, number of events scanned and matched, and query status}
\usage{
cloudtrail_describe_query(EventDataStore = NULL, QueryId)
}
\arguments{
\item{EventDataStore}{The ARN (or the ID suffix of the ARN) of an event data store on which
the specified query was run.}

\item{QueryId}{[required] The query ID.}
}
\description{
Returns metadata about a query, including query run time in milliseconds, number of events scanned and matched, and query status. You must specify an ARN for \code{EventDataStore}, and a value for \code{QueryID}.

See \url{https://paws-r.github.io/docs/cloudtrail/describe_query.html} for full documentation.
}
\keyword{internal}
