% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_channel}
\alias{cloudtrail_get_channel}
\title{Returns information about a specific channel}
\usage{
cloudtrail_get_channel(Channel)
}
\arguments{
\item{Channel}{[required] The ARN or \code{UUID} of a channel.}
}
\description{
Returns information about a specific channel.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_get_channel/} for full documentation.
}
\keyword{internal}
