% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_dashboard}
\alias{cloudtrail_get_dashboard}
\title{Returns the specified dashboard}
\usage{
cloudtrail_get_dashboard(DashboardId)
}
\arguments{
\item{DashboardId}{[required] The name or ARN for the dashboard.}
}
\description{
Returns the specified dashboard.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_get_dashboard/} for full documentation.
}
\keyword{internal}
