% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_event_data_store}
\alias{cloudtrail_get_event_data_store}
\title{Returns information about an event data store specified as either an ARN
or the ID portion of the ARN}
\usage{
cloudtrail_get_event_data_store(EventDataStore)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or ID suffix of the ARN) of the event data store about which
you want information.}
}
\description{
Returns information about an event data store specified as either an ARN or the ID portion of the ARN.

See \url{https://paws-r.github.io/docs/cloudtrail/get_event_data_store.html} for full documentation.
}
\keyword{internal}
