% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_event_selectors}
\alias{cloudtrail_get_event_selectors}
\title{Describes the settings for the event selectors that you configured for
your trail}
\usage{
cloudtrail_get_event_selectors(TrailName)
}
\arguments{
\item{TrailName}{[required] Specifies the name of the trail or trail ARN. If you specify a trail
name, the string must meet the following requirements:
\itemize{
\item Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.),
underscores (\_), or dashes (-)
\item Start with a letter or number, and end with a letter or number
\item Be between 3 and 128 characters
\item Have no adjacent periods, underscores or dashes. Names like
\code{my-_namespace} and \code{my--namespace} are not valid.
\item Not be in IP address format (for example, 192.168.5.4)
}

If you specify a trail ARN, it must be in the format:

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}
}
\description{
Describes the settings for the event selectors that you configured for
your trail. The information returned for your event selectors includes
the following:
}
\details{
\itemize{
\item If your event selector includes read-only events, write-only events,
or all events. This applies to both management events and data
events.
\item If your event selector includes management events.
\item If your event selector includes data events, the Amazon S3 objects
or AWS Lambda functions that you are logging for data events.
}

For more information, see \href{http://docs.aws.amazon.com/awscloudtrail/latest/userguide/logging-management-and-data-events-with-cloudtrail.html}{Logging Data and Management Events for Trails}
in the \emph{AWS CloudTrail User Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_event_selectors(
  TrailName = "string"
)
}
}

\keyword{internal}
