% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_import}
\alias{cloudtrail_get_import}
\title{Returns information about a specific import}
\usage{
cloudtrail_get_import(ImportId)
}
\arguments{
\item{ImportId}{[required] The ID for the import.}
}
\description{
Returns information about a specific import.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_get_import/} for full documentation.
}
\keyword{internal}
