% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_get_trail}
\alias{cloudtrail_get_trail}
\title{Returns settings information for a specified trail}
\usage{
cloudtrail_get_trail(Name)
}
\arguments{
\item{Name}{[required] The name or the Amazon Resource Name (ARN) of the trail for which you
want to retrieve settings information.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Trail = list(
    Name = "string",
    S3BucketName = "string",
    S3KeyPrefix = "string",
    SnsTopicName = "string",
    SnsTopicARN = "string",
    IncludeGlobalServiceEvents = TRUE|FALSE,
    IsMultiRegionTrail = TRUE|FALSE,
    HomeRegion = "string",
    TrailARN = "string",
    LogFileValidationEnabled = TRUE|FALSE,
    CloudWatchLogsLogGroupArn = "string",
    CloudWatchLogsRoleArn = "string",
    KmsKeyId = "string",
    HasCustomEventSelectors = TRUE|FALSE,
    HasInsightSelectors = TRUE|FALSE,
    IsOrganizationTrail = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns settings information for a specified trail.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_trail(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
