% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_channels}
\alias{cloudtrail_list_channels}
\title{Lists the channels in the current account, and their source names}
\usage{
cloudtrail_list_channels(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of CloudTrail channels to display on a single page.}

\item{NextToken}{The token to use to get the next page of results after a previous API
call. This token must be passed in with the same parameters that were
specified in the original call. For example, if the original call
specified an AttributeKey of 'Username' with a value of 'root', the call
with NextToken should include those same parameters.}
}
\description{
Lists the channels in the current account, and their source names.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_channels/} for full documentation.
}
\keyword{internal}
