% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_event_data_stores}
\alias{cloudtrail_list_event_data_stores}
\title{Returns information about all event data stores in the account, in the
current Region}
\usage{
cloudtrail_list_event_data_stores(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token you can use to get the next page of event data store results.}

\item{MaxResults}{The maximum number of event data stores to display on a single page.}
}
\description{
Returns information about all event data stores in the account, in the current Region.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_event_data_stores/} for full documentation.
}
\keyword{internal}
