% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_import_failures}
\alias{cloudtrail_list_import_failures}
\title{Returns a list of failures for the specified import}
\usage{
cloudtrail_list_import_failures(ImportId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ImportId}{[required] The ID of the import.}

\item{MaxResults}{The maximum number of failures to display on a single page.}

\item{NextToken}{A token you can use to get the next page of import failures.}
}
\description{
Returns a list of failures for the specified import.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_import_failures/} for full documentation.
}
\keyword{internal}
