% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_imports}
\alias{cloudtrail_list_imports}
\title{Returns information on all imports, or a select set of imports by
ImportStatus or Destination}
\usage{
cloudtrail_list_imports(
  MaxResults = NULL,
  Destination = NULL,
  ImportStatus = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of imports to display on a single page.}

\item{Destination}{The ARN of the destination event data store.}

\item{ImportStatus}{The status of the import.}

\item{NextToken}{A token you can use to get the next page of import results.}
}
\description{
Returns information on all imports, or a select set of imports by \code{ImportStatus} or \code{Destination}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_imports/} for full documentation.
}
\keyword{internal}
