% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_tags}
\alias{cloudtrail_list_tags}
\title{Lists the tags for the trail in the current region}
\usage{
cloudtrail_list_tags(ResourceIdList, NextToken)
}
\arguments{
\item{ResourceIdList}{[required] Specifies a list of trail ARNs whose tags will be listed. The list has a
limit of 20 ARNs. The format of a trail ARN is:

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}

\item{NextToken}{Reserved for future use.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceTagList = list(
    list(
      ResourceId = "string",
      TagsList = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags for the trail in the current region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags(
  ResourceIdList = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
