% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_list_trails}
\alias{cloudtrail_list_trails}
\title{Lists trails that are in the current account}
\usage{
cloudtrail_list_trails(NextToken = NULL)
}
\arguments{
\item{NextToken}{The token to use to get the next page of results after a previous API
call. This token must be passed in with the same parameters that were
specified in the original call. For example, if the original call
specified an AttributeKey of 'Username' with a value of 'root', the call
with NextToken should include those same parameters.}
}
\description{
Lists trails that are in the current account.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_list_trails/} for full documentation.
}
\keyword{internal}
