% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_restore_event_data_store}
\alias{cloudtrail_restore_event_data_store}
\title{Restores a deleted event data store specified by EventDataStore, which
accepts an event data store ARN}
\usage{
cloudtrail_restore_event_data_store(EventDataStore)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or the ID suffix of the ARN) of the event data store that you
want to restore.}
}
\description{
Restores a deleted event data store specified by \code{EventDataStore}, which accepts an event data store ARN. You can only restore a deleted event data store within the seven-day wait period after deletion. Restoring an event data store can take several minutes, depending on the size of the event data store.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_restore_event_data_store/} for full documentation.
}
\keyword{internal}
