% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_start_event_data_store_ingestion}
\alias{cloudtrail_start_event_data_store_ingestion}
\title{Starts the ingestion of live events on an event data store specified as
either an ARN or the ID portion of the ARN}
\usage{
cloudtrail_start_event_data_store_ingestion(EventDataStore)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or ID suffix of the ARN) of the event data store for which you
want to start ingestion.}
}
\description{
Starts the ingestion of live events on an event data store specified as either an ARN or the ID portion of the ARN. To start ingestion, the event data store \code{Status} must be \code{STOPPED_INGESTION} and the \code{eventCategory} must be \code{Management}, \code{Data}, \code{NetworkActivity}, or \code{ConfigurationItem}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_start_event_data_store_ingestion/} for full documentation.
}
\keyword{internal}
