% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_stop_event_data_store_ingestion}
\alias{cloudtrail_stop_event_data_store_ingestion}
\title{Stops the ingestion of live events on an event data store specified as
either an ARN or the ID portion of the ARN}
\usage{
cloudtrail_stop_event_data_store_ingestion(EventDataStore)
}
\arguments{
\item{EventDataStore}{[required] The ARN (or ID suffix of the ARN) of the event data store for which you
want to stop ingestion.}
}
\description{
Stops the ingestion of live events on an event data store specified as either an ARN or the ID portion of the ARN. To stop ingestion, the event data store \code{Status} must be \code{ENABLED} and the \code{eventCategory} must be \code{Management}, \code{Data}, \code{NetworkActivity}, or \code{ConfigurationItem}.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_stop_event_data_store_ingestion/} for full documentation.
}
\keyword{internal}
