% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_stop_import}
\alias{cloudtrail_stop_import}
\title{Stops a specified import}
\usage{
cloudtrail_stop_import(ImportId)
}
\arguments{
\item{ImportId}{[required] The ID of the import.}
}
\description{
Stops a specified import.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_stop_import/} for full documentation.
}
\keyword{internal}
