% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_update_channel}
\alias{cloudtrail_update_channel}
\title{Updates a channel specified by a required channel ARN or UUID}
\usage{
cloudtrail_update_channel(Channel, Destinations = NULL, Name = NULL)
}
\arguments{
\item{Channel}{[required] The ARN or ID (the ARN suffix) of the channel that you want to update.}

\item{Destinations}{The ARNs of event data stores that you want to log events arriving
through the channel.}

\item{Name}{Changes the name of the channel.}
}
\description{
Updates a channel specified by a required channel ARN or UUID.

See \url{https://paws-r.github.io/docs/cloudtrail/update_channel.html} for full documentation.
}
\keyword{internal}
