% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail_operations.R
\name{cloudtrail_update_dashboard}
\alias{cloudtrail_update_dashboard}
\title{Updates the specified dashboard}
\usage{
cloudtrail_update_dashboard(
  DashboardId,
  Widgets = NULL,
  RefreshSchedule = NULL,
  TerminationProtectionEnabled = NULL
)
}
\arguments{
\item{DashboardId}{[required] The name or ARN of the dashboard.}

\item{Widgets}{An array of widgets for the dashboard. A custom dashboard can have a
maximum of 10 widgets.

To add new widgets, pass in an array that includes the existing widgets
along with any new widgets. Run the
\code{\link[=cloudtrail_get_dashboard]{get_dashboard}} operation to get the list of
widgets for the dashboard.

To remove widgets, pass in an array that includes the existing widgets
minus the widgets you want removed.}

\item{RefreshSchedule}{The refresh schedule configuration for the dashboard.}

\item{TerminationProtectionEnabled}{Specifies whether termination protection is enabled for the dashboard.
If termination protection is enabled, you cannot delete the dashboard
until termination protection is disabled.}
}
\description{
Updates the specified dashboard.

See \url{https://www.paws-r-sdk.com/docs/cloudtrail_update_dashboard/} for full documentation.
}
\keyword{internal}
