% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_anomaly_detector}
\alias{cloudwatch_delete_anomaly_detector}
\title{Deletes the specified anomaly detection model from your account}
\usage{
cloudwatch_delete_anomaly_detector(Namespace, MetricName, Dimensions,
  Stat)
}
\arguments{
\item{Namespace}{[required] The namespace associated with the anomaly detection model to delete.}

\item{MetricName}{[required] The metric name associated with the anomaly detection model to delete.}

\item{Dimensions}{The metric dimensions associated with the anomaly detection model to
delete.}

\item{Stat}{[required] The statistic associated with the anomaly detection model to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified anomaly detection model from your account.
}
\section{Request syntax}{
\preformatted{svc$delete_anomaly_detector(
  Namespace = "string",
  MetricName = "string",
  Dimensions = list(
    list(
      Name = "string",
      Value = "string"
    )
  ),
  Stat = "string"
)
}
}

\keyword{internal}
