% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_alarm_history}
\alias{cloudwatch_describe_alarm_history}
\title{Retrieves the history for the specified alarm}
\usage{
cloudwatch_describe_alarm_history(AlarmName, AlarmTypes,
  HistoryItemType, StartDate, EndDate, MaxRecords, NextToken, ScanBy)
}
\arguments{
\item{AlarmName}{The name of the alarm.}

\item{AlarmTypes}{Use this parameter to specify whether you want the operation to return
metric alarms or composite alarms. If you omit this parameter, only
metric alarms are returned.}

\item{HistoryItemType}{The type of alarm histories to retrieve.}

\item{StartDate}{The starting date to retrieve alarm history.}

\item{EndDate}{The ending date to retrieve alarm history.}

\item{MaxRecords}{The maximum number of alarm history records to retrieve.}

\item{NextToken}{The token returned by a previous call to indicate that there is more
data available.}

\item{ScanBy}{Specified whether to return the newest or oldest alarm history first.
Specify \code{TimestampDescending} to have the newest event history returned
first, and specify \code{TimestampAscending} to have the oldest history
returned first.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AlarmHistoryItems = list(
    list(
      AlarmName = "string",
      AlarmType = "CompositeAlarm"|"MetricAlarm",
      Timestamp = as.POSIXct(
        "2015-01-01"
      ),
      HistoryItemType = "ConfigurationUpdate"|"StateUpdate"|"Action",
      HistorySummary = "string",
      HistoryData = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the history for the specified alarm. You can filter the
results by date range or item type. If an alarm name is not specified,
the histories for either all metric alarms or all composite alarms are
returned.

CloudWatch retains the history of an alarm even if you delete the alarm.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_alarm_history(
  AlarmName = "string",
  AlarmTypes = list(
    "CompositeAlarm"|"MetricAlarm"
  ),
  HistoryItemType = "ConfigurationUpdate"|"StateUpdate"|"Action",
  StartDate = as.POSIXct(
    "2015-01-01"
  ),
  EndDate = as.POSIXct(
    "2015-01-01"
  ),
  MaxRecords = 123,
  NextToken = "string",
  ScanBy = "TimestampDescending"|"TimestampAscending"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
