% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_alarms_for_metric}
\alias{cloudwatch_describe_alarms_for_metric}
\title{Retrieves the alarms for the specified metric}
\usage{
cloudwatch_describe_alarms_for_metric(MetricName, Namespace, Statistic,
  ExtendedStatistic, Dimensions, Period, Unit)
}
\arguments{
\item{MetricName}{[required] The name of the metric.}

\item{Namespace}{[required] The namespace of the metric.}

\item{Statistic}{The statistic for the metric, other than percentiles. For percentile
statistics, use \code{ExtendedStatistics}.}

\item{ExtendedStatistic}{The percentile statistic for the metric. Specify a value between p0.0
and p100.}

\item{Dimensions}{The dimensions associated with the metric. If the metric has any
associated dimensions, you must specify them in order for the call to
succeed.}

\item{Period}{The period, in seconds, over which the statistic is applied.}

\item{Unit}{The unit for the metric.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MetricAlarms = list(
    list(
      AlarmName = "string",
      AlarmArn = "string",
      AlarmDescription = "string",
      AlarmConfigurationUpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      ActionsEnabled = TRUE|FALSE,
      OKActions = list(
        "string"
      ),
      AlarmActions = list(
        "string"
      ),
      InsufficientDataActions = list(
        "string"
      ),
      StateValue = "OK"|"ALARM"|"INSUFFICIENT_DATA",
      StateReason = "string",
      StateReasonData = "string",
      StateUpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      MetricName = "string",
      Namespace = "string",
      Statistic = "SampleCount"|"Average"|"Sum"|"Minimum"|"Maximum",
      ExtendedStatistic = "string",
      Dimensions = list(
        list(
          Name = "string",
          Value = "string"
        )
      ),
      Period = 123,
      Unit = "Seconds"|"Microseconds"|"Milliseconds"|"Bytes"|"Kilobytes"|"Megabytes"|"Gigabytes"|"Terabytes"|"Bits"|"Kilobits"|"Megabits"|"Gigabits"|"Terabits"|"Percent"|"Count"|"Bytes/Second"|"Kilobytes/Second"|"Megabytes/Second"|"Gigabytes/Second"|"Terabytes/Second"|"Bits/Second"|"Kilobits/Second"|"Megabits/Second"|"Gigabits/Second"|"Terabits/Second"|"Count/Second"|"None",
      EvaluationPeriods = 123,
      DatapointsToAlarm = 123,
      Threshold = 123.0,
      ComparisonOperator = "GreaterThanOrEqualToThreshold"|"GreaterThanThreshold"|"LessThanThreshold"|"LessThanOrEqualToThreshold"|"LessThanLowerOrGreaterThanUpperThreshold"|"LessThanLowerThreshold"|"GreaterThanUpperThreshold",
      TreatMissingData = "string",
      EvaluateLowSampleCountPercentile = "string",
      Metrics = list(
        list(
          Id = "string",
          MetricStat = list(
            Metric = list(
              Namespace = "string",
              MetricName = "string",
              Dimensions = list(
                list(
                  Name = "string",
                  Value = "string"
                )
              )
            ),
            Period = 123,
            Stat = "string",
            Unit = "Seconds"|"Microseconds"|"Milliseconds"|"Bytes"|"Kilobytes"|"Megabytes"|"Gigabytes"|"Terabytes"|"Bits"|"Kilobits"|"Megabits"|"Gigabits"|"Terabits"|"Percent"|"Count"|"Bytes/Second"|"Kilobytes/Second"|"Megabytes/Second"|"Gigabytes/Second"|"Terabytes/Second"|"Bits/Second"|"Kilobits/Second"|"Megabits/Second"|"Gigabits/Second"|"Terabits/Second"|"Count/Second"|"None"
          ),
          Expression = "string",
          Label = "string",
          ReturnData = TRUE|FALSE,
          Period = 123
        )
      ),
      ThresholdMetricId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the alarms for the specified metric. To filter the results,
specify a statistic, period, or unit.

This operation retrieves only standard alarms that are based on the
specified metric. It does not return alarms based on math expressions
that use the specified metric, or composite alarms that use the
specified metric.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_alarms_for_metric(
  MetricName = "string",
  Namespace = "string",
  Statistic = "SampleCount"|"Average"|"Sum"|"Minimum"|"Maximum",
  ExtendedStatistic = "string",
  Dimensions = list(
    list(
      Name = "string",
      Value = "string"
    )
  ),
  Period = 123,
  Unit = "Seconds"|"Microseconds"|"Milliseconds"|"Bytes"|"Kilobytes"|"Megabytes"|"Gigabytes"|"Terabytes"|"Bits"|"Kilobits"|"Megabits"|"Gigabits"|"Terabits"|"Percent"|"Count"|"Bytes/Second"|"Kilobytes/Second"|"Megabytes/Second"|"Gigabytes/Second"|"Terabytes/Second"|"Bits/Second"|"Kilobits/Second"|"Megabits/Second"|"Gigabits/Second"|"Terabits/Second"|"Count/Second"|"None"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
