% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_disable_alarm_actions}
\alias{cloudwatch_disable_alarm_actions}
\title{Disables the actions for the specified alarms}
\usage{
cloudwatch_disable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\value{
An empty list.
}
\description{
Disables the actions for the specified alarms. When an alarm's actions
are disabled, the alarm actions do not execute when the alarm state
changes.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_alarm_actions(
  AlarmNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
