% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_enable_alarm_actions}
\alias{cloudwatch_enable_alarm_actions}
\title{Enables the actions for the specified alarms}
\usage{
cloudwatch_enable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\description{
Enables the actions for the specified alarms.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_enable_alarm_actions/} for full documentation.
}
\keyword{internal}
