% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_enable_alarm_actions}
\alias{cloudwatch_enable_alarm_actions}
\title{Enables the actions for the specified alarms}
\usage{
cloudwatch_enable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\value{
An empty list.
}
\description{
Enables the actions for the specified alarms.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$enable_alarm_actions(
  AlarmNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
