% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_list_metric_streams}
\alias{cloudwatch_list_metric_streams}
\title{Returns a list of metric streams in this account}
\usage{
cloudwatch_list_metric_streams(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Include this value, if it was returned by the previous call, to get the
next set of metric streams.}

\item{MaxResults}{The maximum number of results to return in one operation.}
}
\description{
Returns a list of metric streams in this account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_list_metric_streams/} for full documentation.
}
\keyword{internal}
