% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_put_metric_stream}
\alias{cloudwatch_put_metric_stream}
\title{Creates or updates a metric stream}
\usage{
cloudwatch_put_metric_stream(Name, IncludeFilters, ExcludeFilters,
  FirehoseArn, RoleArn, OutputFormat, Tags)
}
\arguments{
\item{Name}{[required] If you are creating a new metric stream, this is the name for the new
stream. The name must be different than the names of other metric
streams in this account and Region.

If you are updating a metric stream, specify the name of that stream
here.

Valid characters are A-Z, a-z, 0-9, "-" and "_".}

\item{IncludeFilters}{If you specify this parameter, the stream sends only the metrics from
the metric namespaces that you specify here.

You cannot include \code{IncludeFilters} and \code{ExcludeFilters} in the same
operation.}

\item{ExcludeFilters}{If you specify this parameter, the stream sends metrics from all metric
namespaces except for the namespaces that you specify here.

You cannot include \code{ExcludeFilters} and \code{IncludeFilters} in the same
operation.}

\item{FirehoseArn}{[required] The ARN of the Amazon Kinesis Firehose delivery stream to use for this
metric stream. This Amazon Kinesis Firehose delivery stream must already
exist and must be in the same account as the metric stream.}

\item{RoleArn}{[required] The ARN of an IAM role that this metric stream will use to access Amazon
Kinesis Firehose resources. This IAM role must already exist and must be
in the same account as the metric stream. This IAM role must include the
following permissions:
\itemize{
\item firehose:PutRecord
\item firehose:PutRecordBatch
}}

\item{OutputFormat}{[required] The output format for the stream. Valid values are \code{json} and
\code{opentelemetry0.7}. For more information about metric stream output
formats, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-metric-streams-formats.html}{Metric streams output formats}.}

\item{Tags}{A list of key-value pairs to associate with the metric stream. You can
associate as many as 50 tags with a metric stream.

Tags can help you organize and categorize your resources. You can also
use them to scope user permissions by granting a user permission to
access or change only resources with certain tag values.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string"
)
}
}
\description{
Creates or updates a metric stream. Metric streams can automatically
stream CloudWatch metrics to AWS destinations including Amazon S3 and to
many third-party solutions.

For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/}{Using Metric Streams}.

To create a metric stream, you must be logged on to an account that has
the \code{iam:PassRole} permission and either the \code{CloudWatchFullAccess}
policy or the \code{cloudwatch:PutMetricStream} permission.

When you create or update a metric stream, you choose one of the
following:
\itemize{
\item Stream metrics from all metric namespaces in the account.
\item Stream metrics from all metric namespaces in the account, except for
the namespaces that you list in \code{ExcludeFilters}.
\item Stream metrics from only the metric namespaces that you list in
\code{IncludeFilters}.
}

When you use \code{\link[=cloudwatch_put_metric_stream]{put_metric_stream}} to
create a new metric stream, the stream is created in the \code{running}
state. If you use it to update an existing stream, the state of the
stream is not changed.
}
\section{Request syntax}{
\preformatted{svc$put_metric_stream(
  Name = "string",
  IncludeFilters = list(
    list(
      Namespace = "string"
    )
  ),
  ExcludeFilters = list(
    list(
      Namespace = "string"
    )
  ),
  FirehoseArn = "string",
  RoleArn = "string",
  OutputFormat = "json"|"opentelemetry0.7",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
