% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_start_o_tel_enrichment}
\alias{cloudwatch_start_o_tel_enrichment}
\title{Enables enrichment and PromQL access for CloudWatch vended metrics for
supported Amazon Web Services resources in the account}
\usage{
cloudwatch_start_o_tel_enrichment()
}
\description{
Enables enrichment and PromQL access for CloudWatch vended metrics for \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/UsingResourceTagsForTelemetry.html}{supported Amazon Web Services resources} in the account. Once enabled, metrics that contain a resource identifier dimension (for example, EC2 \code{CPUUtilization} with an \code{InstanceId} dimension) are enriched with resource ARN and resource tag labels and become queryable using PromQL.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_start_o_tel_enrichment/} for full documentation.
}
\keyword{internal}
