% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_untag_resource}
\alias{cloudwatch_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
cloudwatch_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the CloudWatch resource that you're removing tags from.

The ARN format of an alarm is
\code{arn:aws:cloudwatch:Region:account-id:alarm:alarm-name }

The ARN format of a Contributor Insights rule is
\code{arn:aws:cloudwatch:Region:account-id:insight-rule:insight-rule-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazoncloudwatch.html#amazoncloudwatch-resources-for-iam-policies}{Resource Types Defined by Amazon CloudWatch}
in the \emph{Amazon Web Services General Reference}.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatch_untag_resource/} for full documentation.
}
\keyword{internal}
