% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_service.R
\name{cloudwatchapplicationsignals}
\alias{cloudwatchapplicationsignals}
\title{Amazon CloudWatch Application Signals}
\usage{
cloudwatchapplicationsignals(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Use CloudWatch Application Signals for comprehensive observability of your cloud-based applications. It enables real-time service health dashboards and helps you track long-term performance trends against your business goals. The application-centric view provides you with unified visibility across your applications, services, and dependencies, so you can proactively monitor and efficiently triage any issues that may arise, ensuring optimal customer experience.

Application Signals provides the following benefits:
\itemize{
\item Automatically collect metrics and traces from your applications, and display key metrics such as call volume, availability, latency, faults, and errors.
\item Create and monitor service level objectives (SLOs).
\item See a map of your application topology that Application Signals automatically discovers, that gives you a visual representation of your applications, dependencies, and their connectivity.
}

Application Signals works with CloudWatch RUM, CloudWatch Synthetics canaries, and Amazon Web Services Service Catalog AppRegistry, to display your client pages, Synthetics canaries, and application names within dashboards and maps.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- cloudwatchapplicationsignals(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=cloudwatchapplicationsignals_batch_delete_instrumentation_configurations]{batch_delete_instrumentation_configurations} \tab Deletes multiple instrumentation configurations in a single request\cr
\link[=cloudwatchapplicationsignals_batch_get_service_level_objective_budget_report]{batch_get_service_level_objective_budget_report} \tab Use this operation to retrieve one or more service level objective (SLO) budget reports\cr
\link[=cloudwatchapplicationsignals_batch_update_exclusion_windows]{batch_update_exclusion_windows} \tab Add or remove time window exclusions for one or more Service Level Objectives (SLOs)\cr
\link[=cloudwatchapplicationsignals_create_instrumentation_configuration]{create_instrumentation_configuration} \tab Creates a dynamic instrumentation configuration for a specific code or endpoint location within a service and environment\cr
\link[=cloudwatchapplicationsignals_create_service_level_objective]{create_service_level_objective} \tab Creates a service level objective (SLO), which can help you ensure that your critical business operations are meeting customer expectations\cr
\link[=cloudwatchapplicationsignals_delete_grouping_configuration]{delete_grouping_configuration} \tab Deletes the grouping configuration for this account\cr
\link[=cloudwatchapplicationsignals_delete_instrumentation_configuration]{delete_instrumentation_configuration} \tab Deletes the specified instrumentation configuration\cr
\link[=cloudwatchapplicationsignals_delete_service_level_objective]{delete_service_level_objective} \tab Deletes the specified service level objective\cr
\link[=cloudwatchapplicationsignals_get_instrumentation_configuration]{get_instrumentation_configuration} \tab Returns the details of a single instrumentation configuration identified by service, environment, signal type, and location\cr
\link[=cloudwatchapplicationsignals_get_instrumentation_configuration_status]{get_instrumentation_configuration_status} \tab Retrieves the status history for a single instrumentation configuration during a specified time range\cr
\link[=cloudwatchapplicationsignals_get_service]{get_service} \tab Returns information about a service discovered by Application Signals\cr
\link[=cloudwatchapplicationsignals_get_service_level_objective]{get_service_level_objective} \tab Returns information about one SLO created in the account\cr
\link[=cloudwatchapplicationsignals_list_audit_findings]{list_audit_findings} \tab Returns a list of audit findings that provide automated analysis of service behavior and root cause analysis\cr
\link[=cloudwatchapplicationsignals_list_entity_events]{list_entity_events} \tab Returns a list of change events for a specific entity, such as deployments, configuration changes, or other state-changing activities\cr
\link[=cloudwatchapplicationsignals_list_grouping_attribute_definitions]{list_grouping_attribute_definitions} \tab Returns the current grouping configuration for this account, including all custom grouping attribute definitions that have been configured\cr
\link[=cloudwatchapplicationsignals_list_instrumentation_configurations]{list_instrumentation_configurations} \tab Returns all active instrumentation configurations for a service and environment\cr
\link[=cloudwatchapplicationsignals_list_service_dependencies]{list_service_dependencies} \tab Returns a list of service dependencies of the service that you specify\cr
\link[=cloudwatchapplicationsignals_list_service_dependents]{list_service_dependents} \tab Returns the list of dependents that invoked the specified service during the provided time range\cr
\link[=cloudwatchapplicationsignals_list_service_level_objective_exclusion_windows]{list_service_level_objective_exclusion_windows} \tab Retrieves all exclusion windows configured for a specific SLO\cr
\link[=cloudwatchapplicationsignals_list_service_level_objectives]{list_service_level_objectives} \tab Returns a list of SLOs created in this account\cr
\link[=cloudwatchapplicationsignals_list_service_operations]{list_service_operations} \tab Returns a list of the operations of this service that have been discovered by Application Signals\cr
\link[=cloudwatchapplicationsignals_list_services]{list_services} \tab Returns a list of services that have been discovered by Application Signals\cr
\link[=cloudwatchapplicationsignals_list_service_states]{list_service_states} \tab Returns information about the last deployment and other change states of services\cr
\link[=cloudwatchapplicationsignals_list_tags_for_resource]{list_tags_for_resource} \tab Displays the tags associated with a CloudWatch resource\cr
\link[=cloudwatchapplicationsignals_put_grouping_configuration]{put_grouping_configuration} \tab Creates or updates the grouping configuration for this account\cr
\link[=cloudwatchapplicationsignals_report_instrumentation_configuration_status]{report_instrumentation_configuration_status} \tab Reports the status of one or more instrumentation configurations from SDK instances\cr
\link[=cloudwatchapplicationsignals_start_discovery]{start_discovery} \tab Enables this Amazon Web Services account to be able to use CloudWatch Application Signals by creating the AWSServiceRoleForCloudWatchApplicationSignals service-linked role\cr
\link[=cloudwatchapplicationsignals_tag_resource]{tag_resource} \tab Assigns one or more tags (key-value pairs) to the specified CloudWatch resource, such as a service level objective\cr
\link[=cloudwatchapplicationsignals_untag_resource]{untag_resource} \tab Removes one or more tags from the specified resource\cr
\link[=cloudwatchapplicationsignals_update_service_level_objective]{update_service_level_objective} \tab Updates an existing service level objective (SLO)
}
}

\examples{
\dontrun{
svc <- cloudwatchapplicationsignals()
svc$batch_delete_instrumentation_configurations(
  Foo = 123
)
}

}
