% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchapplicationsignals_operations.R
\name{cloudwatchapplicationsignals_tag_resource}
\alias{cloudwatchapplicationsignals_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified CloudWatch
resource, such as a service level objective}
\usage{
cloudwatchapplicationsignals_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudWatch resource that you want
to set tags for.

The ARN format of an Application Signals SLO is
\code{arn:aws:cloudwatch:Region:account-id:slo:slo-name }

For more information about ARN format, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazoncloudwatch.html#amazoncloudwatch-resources-for-iam-policies}{Resource Types Defined by Amazon CloudWatch}
in the \emph{Amazon Web Services General Reference}.}

\item{Tags}{[required] The list of key-value pairs to associate with the alarm.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified CloudWatch resource, such as a service level objective.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchapplicationsignals_tag_resource/} for full documentation.
}
\keyword{internal}
