% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_cancel_replay}
\alias{cloudwatchevents_cancel_replay}
\title{Cancels the specified replay}
\usage{
cloudwatchevents_cancel_replay(ReplayName)
}
\arguments{
\item{ReplayName}{[required] The name of the replay to cancel.}
}
\description{
Cancels the specified replay.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_cancel_replay/} for full documentation.
}
\keyword{internal}
