% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_deauthorize_connection}
\alias{cloudwatchevents_deauthorize_connection}
\title{Removes all authorization parameters from the connection}
\usage{
cloudwatchevents_deauthorize_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to remove authorization from.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConnectionArn = "string",
  ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastAuthorizedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Removes all authorization parameters from the connection. This lets you
remove the secret from the connection so you can reuse it without having
to create a new connection.
}
\section{Request syntax}{
\preformatted{svc$deauthorize_connection(
  Name = "string"
)
}
}

\keyword{internal}
