% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_connection}
\alias{cloudwatchevents_delete_connection}
\title{Deletes a connection}
\usage{
cloudwatchevents_delete_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConnectionArn = "string",
  ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastAuthorizedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Deletes a connection.
}
\section{Request syntax}{
\preformatted{svc$delete_connection(
  Name = "string"
)
}
}

\keyword{internal}
