% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_delete_event_bus}
\alias{cloudwatchevents_delete_event_bus}
\title{Deletes the specified custom event bus or partner event bus}
\usage{
cloudwatchevents_delete_event_bus(Name)
}
\arguments{
\item{Name}{[required] The name of the event bus to delete.}
}
\description{
Deletes the specified custom event bus or partner event bus. All rules
associated with this event bus need to be deleted. You can't delete your
account's default event bus.
}
\section{Request syntax}{
\preformatted{svc$delete_event_bus(
  Name = "string"
)
}
}

\keyword{internal}
