% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_archive}
\alias{cloudwatchevents_describe_archive}
\title{Retrieves details about an archive}
\usage{
cloudwatchevents_describe_archive(ArchiveName)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to retrieve.}
}
\description{
Retrieves details about an archive.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_describe_archive/} for full documentation.
}
\keyword{internal}
