% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_rule}
\alias{cloudwatchevents_describe_rule}
\title{Describes the specified rule}
\usage{
cloudwatchevents_describe_rule(Name, EventBusName)
}
\arguments{
\item{Name}{[required] The name of the rule.}

\item{EventBusName}{The event bus associated with the rule. If you omit this, the default
event bus is used.}
}
\description{
Describes the specified rule.
}
\details{
\code{DescribeRule} doesn't list the targets of a rule. To see the targets
associated with a rule, use ListTargetsByRule.
}
\section{Request syntax}{
\preformatted{svc$describe_rule(
  Name = "string",
  EventBusName = "string"
)
}
}

\keyword{internal}
