% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_list_tags_for_resource}
\alias{cloudwatchevents_list_tags_for_resource}
\title{Displays the tags associated with an EventBridge resource}
\usage{
cloudwatchevents_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the EventBridge resource for which you want to view tags.}
}
\description{
Displays the tags associated with an EventBridge resource. In EventBridge, rules and event buses can be tagged.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevents_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
