% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_update_api_destination}
\alias{cloudwatchevents_update_api_destination}
\title{Updates an API destination}
\usage{
cloudwatchevents_update_api_destination(
  Name,
  Description = NULL,
  ConnectionArn = NULL,
  InvocationEndpoint = NULL,
  HttpMethod = NULL,
  InvocationRateLimitPerSecond = NULL
)
}
\arguments{
\item{Name}{[required] The name of the API destination to update.}

\item{Description}{The name of the API destination to update.}

\item{ConnectionArn}{The ARN of the connection to use for the API destination.}

\item{InvocationEndpoint}{The URL to the endpoint to use for the API destination.}

\item{HttpMethod}{The method to use for the API destination.}

\item{InvocationRateLimitPerSecond}{The maximum number of invocations per second to send to the API
destination.}
}
\description{
Updates an API destination.

See \url{https://paws-r.github.io/docs/cloudwatchevents/update_api_destination.html} for full documentation.
}
\keyword{internal}
