% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_tags_for_resource}
\alias{cloudwatchevidently_list_tags_for_resource}
\title{Displays the tags associated with an Evidently resource}
\usage{
cloudwatchevidently_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that you want to see the tags of.}
}
\description{
Displays the tags associated with an Evidently resource.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
