% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_start_launch}
\alias{cloudwatchevidently_start_launch}
\title{Starts an existing launch}
\usage{
cloudwatchevidently_start_launch(launch, project)
}
\arguments{
\item{launch}{[required] The name of the launch to start.}

\item{project}{[required] The name or ARN of the project that contains the launch to start.}
}
\description{
Starts an existing launch. To create a launch, use \code{\link[=cloudwatchevidently_create_launch]{create_launch}}.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchevidently_start_launch/} for full documentation.
}
\keyword{internal}
