% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_update_project}
\alias{cloudwatchevidently_update_project}
\title{Updates the description of an existing project}
\usage{
cloudwatchevidently_update_project(description = NULL, project)
}
\arguments{
\item{description}{An optional description of the project.}

\item{project}{[required] The name or ARN of the project to update.}
}
\description{
Updates the description of an existing project.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/update_project.html} for full documentation.
}
\keyword{internal}
