% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_get_health_event}
\alias{cloudwatchinternetmonitor_get_health_event}
\title{Gets information the Amazon CloudWatch Internet Monitor has created and
stored about a health event for a specified monitor}
\usage{
cloudwatchinternetmonitor_get_health_event(MonitorName, EventId)
}
\arguments{
\item{MonitorName}{[required] The name of the monitor.}

\item{EventId}{[required] The internally-generated identifier of a health event. Because \code{EventID}
contains the forward slash (“/”) character, you must URL-encode the
\code{EventID} field in the request URL.}
}
\description{
Gets information the Amazon CloudWatch Internet Monitor has created and stored about a health event for a specified monitor. This information includes the impacted locations, and all the information related to the event, by location.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_get_health_event/} for full documentation.
}
\keyword{internal}
