% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchinternetmonitor_operations.R
\name{cloudwatchinternetmonitor_get_monitor}
\alias{cloudwatchinternetmonitor_get_monitor}
\title{Gets information about a monitor in Amazon CloudWatch Internet Monitor
based on a monitor name}
\usage{
cloudwatchinternetmonitor_get_monitor(MonitorName)
}
\arguments{
\item{MonitorName}{[required] The name of the monitor.}
}
\description{
Gets information about a monitor in Amazon CloudWatch Internet Monitor based on a monitor name. The information returned includes the Amazon Resource Name (ARN), create time, modified time, resources included in the monitor, and status information.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchinternetmonitor_get_monitor/} for full documentation.
}
\keyword{internal}
