% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_cancel_export_task}
\alias{cloudwatchlogs_cancel_export_task}
\title{Cancels the specified export task}
\usage{
cloudwatchlogs_cancel_export_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the export task.}
}
\description{
Cancels the specified export task.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/cancel_export_task.html} for full documentation.
}
\keyword{internal}
