% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_cancel_export_task}
\alias{cloudwatchlogs_cancel_export_task}
\title{Cancels the specified export task}
\usage{
cloudwatchlogs_cancel_export_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the export task.}
}
\value{
An empty list.
}
\description{
Cancels the specified export task.

The task must be in the \code{PENDING} or \code{RUNNING} state.
}
\section{Request syntax}{
\preformatted{svc$cancel_export_task(
  taskId = "string"
)
}
}

\keyword{internal}
