% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_create_lookup_table}
\alias{cloudwatchlogs_create_lookup_table}
\title{Creates a lookup table by uploading CSV data}
\usage{
cloudwatchlogs_create_lookup_table(
  lookupTableName,
  description = NULL,
  tableBody,
  kmsKeyId = NULL,
  tags = NULL
)
}
\arguments{
\item{lookupTableName}{[required] The name of the lookup table. The name must be unique within your
account and Region. The name can contain only alphanumeric characters
and underscores, and can be up to 256 characters long.}

\item{description}{A description of the lookup table. The description can be up to 1024
characters long.}

\item{tableBody}{[required] The CSV content of the lookup table. The first row must be a header row
with column names. The content must use UTF-8 encoding and not exceed 10
MB.}

\item{kmsKeyId}{The ARN of the KMS key to use to encrypt the lookup table data. If you
don't specify a key, the data is encrypted with an Amazon Web
Services-owned key.}

\item{tags}{A list of key-value pairs to associate with the lookup table. You can
associate as many as 50 tags with a lookup table. Tags can help you
organize and categorize your resources.}
}
\description{
Creates a lookup table by uploading CSV data. You can use lookup tables to enrich log data in CloudWatch Logs Insights queries with reference data such as user details, application names, or error descriptions.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_create_lookup_table/} for full documentation.
}
\keyword{internal}
