% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_log_group}
\alias{cloudwatchlogs_delete_log_group}
\title{Deletes the specified log group and permanently deletes all the archived
log events associated with the log group}
\usage{
cloudwatchlogs_delete_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\value{
An empty list.
}
\description{
Deletes the specified log group and permanently deletes all the archived
log events associated with the log group.
}
\section{Request syntax}{
\preformatted{svc$delete_log_group(
  logGroupName = "string"
)
}
}

\keyword{internal}
