% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_query_definition}
\alias{cloudwatchlogs_delete_query_definition}
\title{Deletes a saved CloudWatch Logs Insights query definition}
\usage{
cloudwatchlogs_delete_query_definition(queryDefinitionId)
}
\arguments{
\item{queryDefinitionId}{[required] The ID of the query definition that you want to delete. You can use
\href{https://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DescribeQueryDefinitions.html}{DescribeQueryDefinitions}
to retrieve the IDs of your saved query definitions.}
}
\description{
Deletes a saved CloudWatch Logs Insights query definition. A query
definition contains details about a saved CloudWatch Logs Insights
query.

Each \code{DeleteQueryDefinition} operation can delete one query definition.

You must have the \code{logs:DeleteQueryDefinition} permission to be able to
perform this operation.
}
\section{Request syntax}{
\preformatted{svc$delete_query_definition(
  queryDefinitionId = "string"
)
}
}

\keyword{internal}
