% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_query_definition}
\alias{cloudwatchlogs_delete_query_definition}
\title{Deletes a saved CloudWatch Logs Insights query definition}
\usage{
cloudwatchlogs_delete_query_definition(queryDefinitionId)
}
\arguments{
\item{queryDefinitionId}{[required] The ID of the query definition that you want to delete. You can use
\code{\link[=cloudwatchlogs_describe_query_definitions]{describe_query_definitions}}
to retrieve the IDs of your saved query definitions.}
}
\description{
Deletes a saved CloudWatch Logs Insights query definition. A query definition contains details about a saved CloudWatch Logs Insights query.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/delete_query_definition.html} for full documentation.
}
\keyword{internal}
