% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_query_definition}
\alias{cloudwatchlogs_delete_query_definition}
\title{Delete query definition}
\usage{
cloudwatchlogs_delete_query_definition(queryDefinitionId)
}
\arguments{
\item{queryDefinitionId}{[required]}
}
\description{
Delete query definition
}
\section{Request syntax}{
\preformatted{svc$delete_query_definition(
  queryDefinitionId = "string"
)
}
}

\keyword{internal}
