% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_delete_retention_policy}
\alias{cloudwatchlogs_delete_retention_policy}
\title{Deletes the specified retention policy}
\usage{
cloudwatchlogs_delete_retention_policy(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Deletes the specified retention policy.

Log events do not expire if they belong to log groups without a
retention policy.
}
\section{Request syntax}{
\preformatted{svc$delete_retention_policy(
  logGroupName = "string"
)
}
}

\keyword{internal}
