% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_delivery_sources}
\alias{cloudwatchlogs_describe_delivery_sources}
\title{Retrieves a list of the delivery sources that have been created in the
account}
\usage{
cloudwatchlogs_describe_delivery_sources(nextToken = NULL, limit = NULL)
}
\arguments{
\item{nextToken}{}

\item{limit}{Optionally specify the maximum number of delivery sources to return in
the response.}
}
\description{
Retrieves a list of the delivery sources that have been created in the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_describe_delivery_sources/} for full documentation.
}
\keyword{internal}
