% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_describe_resource_policies}
\alias{cloudwatchlogs_describe_resource_policies}
\title{Lists the resource policies in this account}
\usage{
cloudwatchlogs_describe_resource_policies(nextToken, limit)
}
\arguments{
\item{nextToken}{}

\item{limit}{The maximum number of resource policies to be displayed with one call of
this API.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  resourcePolicies = list(
    list(
      policyName = "string",
      policyDocument = "string",
      lastUpdatedTime = 123
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the resource policies in this account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_resource_policies(
  nextToken = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
