% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_delivery}
\alias{cloudwatchlogs_get_delivery}
\title{Returns complete information about one delivery}
\usage{
cloudwatchlogs_get_delivery(id)
}
\arguments{
\item{id}{[required] The ID of the delivery that you want to retrieve.}
}
\description{
Returns complete information about one \emph{delivery}. A delivery is a connection between a logical \emph{delivery source} and a logical \emph{delivery destination}

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_delivery/} for full documentation.
}
\keyword{internal}
