% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_query_results}
\alias{cloudwatchlogs_get_query_results}
\title{Returns the results from the specified query}
\usage{
cloudwatchlogs_get_query_results(queryId)
}
\arguments{
\item{queryId}{[required] The ID number of the query.}
}
\description{
Returns the results from the specified query. If the query is in
progress, partial results of that current execution are returned.
}
\details{
Only the fields requested in the query are returned, along with a \code{@ptr}
field which is the identifier for the log record. You can use the value
of \code{@ptr} in a operation to get the full log record.

\code{GetQueryResults} does not start a query execution. To run a query, use
.
}
\section{Request syntax}{
\preformatted{svc$get_query_results(
  queryId = "string"
)
}
}

\keyword{internal}
