% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_get_query_results}
\alias{cloudwatchlogs_get_query_results}
\title{Returns the results from the specified query}
\usage{
cloudwatchlogs_get_query_results(queryId)
}
\arguments{
\item{queryId}{[required] The ID number of the query.}
}
\description{
Returns the results from the specified query.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_get_query_results/} for full documentation.
}
\keyword{internal}
