% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_scheduled_queries}
\alias{cloudwatchlogs_list_scheduled_queries}
\title{Lists all scheduled queries in your account and region}
\usage{
cloudwatchlogs_list_scheduled_queries(
  maxResults = NULL,
  nextToken = NULL,
  state = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of scheduled queries to return. Valid range is 1 to
1000.}

\item{nextToken}{}

\item{state}{Filter scheduled queries by state. Valid values are \code{ENABLED} and
\code{DISABLED}. If not specified, all scheduled queries are returned.}
}
\description{
Lists all scheduled queries in your account and region. You can filter results by state to show only enabled or disabled queries.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_scheduled_queries/} for full documentation.
}
\keyword{internal}
