% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_tags_log_group}
\alias{cloudwatchlogs_list_tags_log_group}
\title{Lists the tags for the specified log group}
\usage{
cloudwatchlogs_list_tags_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Lists the tags for the specified log group.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/list_tags_log_group.html} for full documentation.
}
\keyword{internal}
