% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_tags_log_group}
\alias{cloudwatchlogs_list_tags_log_group}
\title{The ListTagsLogGroup operation is on the path to deprecation}
\usage{
cloudwatchlogs_list_tags_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
The ListTagsLogGroup operation is on the path to deprecation. We recommend that you use \code{\link[=cloudwatchlogs_list_tags_for_resource]{list_tags_for_resource}} instead.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_list_tags_log_group/} for full documentation.
}
\keyword{internal}
