% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_list_tags_log_group}
\alias{cloudwatchlogs_list_tags_log_group}
\title{Lists the tags for the specified log group}
\usage{
cloudwatchlogs_list_tags_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags for the specified log group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_log_group(
  logGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
