% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_tag_log_group}
\alias{cloudwatchlogs_tag_log_group}
\title{The TagLogGroup operation is on the path to deprecation}
\usage{
cloudwatchlogs_tag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{tags}{[required] The key-value pairs to use for the tags.}
}
\description{
The TagLogGroup operation is on the path to deprecation. We recommend that you use \code{\link[=cloudwatchlogs_tag_resource]{tag_resource}} instead.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_tag_log_group/} for full documentation.
}
\keyword{internal}
