% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_test_metric_filter}
\alias{cloudwatchlogs_test_metric_filter}
\title{Tests the filter pattern of a metric filter against a sample of log
event messages}
\usage{
cloudwatchlogs_test_metric_filter(filterPattern, logEventMessages)
}
\arguments{
\item{filterPattern}{[required]}

\item{logEventMessages}{[required] The log event messages to test.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  matches = list(
    list(
      eventNumber = 123,
      eventMessage = "string",
      extractedValues = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Tests the filter pattern of a metric filter against a sample of log
event messages. You can use this operation to validate the correctness
of a metric filter pattern.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$test_metric_filter(
  filterPattern = "string",
  logEventMessages = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
