% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_test_metric_filter}
\alias{cloudwatchlogs_test_metric_filter}
\title{Tests the filter pattern of a metric filter against a sample of log
event messages}
\usage{
cloudwatchlogs_test_metric_filter(filterPattern, logEventMessages)
}
\arguments{
\item{filterPattern}{[required]}

\item{logEventMessages}{[required] The log event messages to test.}
}
\description{
Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_test_metric_filter/} for full documentation.
}
\keyword{internal}
