% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_untag_log_group}
\alias{cloudwatchlogs_untag_log_group}
\title{Removes the specified tags from the specified log group}
\usage{
cloudwatchlogs_untag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{tags}{[required] The tag keys. The corresponding tags are removed from the log group.}
}
\value{
An empty list.
}
\description{
Removes the specified tags from the specified log group.

To list the tags for a log group, use
\code{\link[=cloudwatchlogs_list_tags_log_group]{list_tags_log_group}}. To add
tags, use \code{\link[=cloudwatchlogs_tag_log_group]{tag_log_group}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_log_group(
  logGroupName = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
