% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_untag_log_group}
\alias{cloudwatchlogs_untag_log_group}
\title{Removes the specified tags from the specified log group}
\usage{
cloudwatchlogs_untag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{tags}{[required] The tag keys. The corresponding tags are removed from the log group.}
}
\description{
Removes the specified tags from the specified log group.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/untag_log_group.html} for full documentation.
}
\keyword{internal}
