% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_update_lookup_table}
\alias{cloudwatchlogs_update_lookup_table}
\title{Updates an existing lookup table by replacing all of its CSV content}
\usage{
cloudwatchlogs_update_lookup_table(
  lookupTableArn,
  description = NULL,
  tableBody,
  kmsKeyId = NULL
)
}
\arguments{
\item{lookupTableArn}{[required] The ARN of the lookup table to update.}

\item{description}{An updated description of the lookup table.}

\item{tableBody}{[required] The new CSV content to replace the existing data. The first row must be
a header row with column names. The content must use UTF-8 encoding and
not exceed 10 MB.}

\item{kmsKeyId}{The ARN of the KMS key to use to encrypt the lookup table data. You can
use this parameter to add, update, or remove the KMS key. To remove the
KMS key and use an Amazon Web Services-owned key instead, specify an
empty string.}
}
\description{
Updates an existing lookup table by replacing all of its CSV content. After the update completes, queries that use this table will use the new data.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchlogs_update_lookup_table/} for full documentation.
}
\keyword{internal}
