% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_get_sink}
\alias{cloudwatchobservabilityaccessmanager_get_sink}
\title{Returns complete information about one monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_get_sink(Identifier, IncludeTags = NULL)
}
\arguments{
\item{Identifier}{[required] The ARN of the sink to retrieve information for.}

\item{IncludeTags}{Specifies whether to include the tags associated with the sink in the
response. When \code{IncludeTags} is set to \code{true} and the caller has the
required permission, \code{oam:ListTagsForResource}, the API will return the
tags for the specified resource. If the caller doesn't have the required
permission, \code{oam:ListTagsForResource}, the API will raise an exception.

The default value is \code{false}.}
}
\description{
Returns complete information about one monitoring account sink.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_get_sink/} for full documentation.
}
\keyword{internal}
