% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cloudwatchobservabilityaccessmanager_operations.R
\name{cloudwatchobservabilityaccessmanager_update_link}
\alias{cloudwatchobservabilityaccessmanager_update_link}
\title{Use this operation to change what types of data are shared from a source
account to its linked monitoring account sink}
\usage{
cloudwatchobservabilityaccessmanager_update_link(
  Identifier,
  IncludeTags = NULL,
  LinkConfiguration = NULL,
  ResourceTypes
)
}
\arguments{
\item{Identifier}{[required] The ARN of the link that you want to update.}

\item{IncludeTags}{Specifies whether to include the tags associated with the link in the
response after the update operation. When \code{IncludeTags} is set to \code{true}
and the caller has the required permission, \code{oam:ListTagsForResource},
the API will return the tags for the specified resource. If the caller
doesn't have the required permission, \code{oam:ListTagsForResource}, the API
will raise an exception.

The default value is \code{false}.}

\item{LinkConfiguration}{Use this structure to filter which metric namespaces and which log
groups are to be shared from the source account to the monitoring
account.}

\item{ResourceTypes}{[required] An array of strings that define which types of data that the source
account will send to the monitoring account.

Your input here replaces the current set of data types that are shared.}
}
\description{
Use this operation to change what types of data are shared from a source account to its linked monitoring account sink. You can't change the sink or change the monitoring account with this operation.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchobservabilityaccessmanager_update_link/} for full documentation.
}
\keyword{internal}
