% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_get_app_monitor}
\alias{cloudwatchrum_get_app_monitor}
\title{Retrieves the complete configuration information for one app monitor}
\usage{
cloudwatchrum_get_app_monitor(Name)
}
\arguments{
\item{Name}{[required] The app monitor to retrieve information for.}
}
\description{
Retrieves the complete configuration information for one app monitor.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_get_app_monitor/} for full documentation.
}
\keyword{internal}
