% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_list_app_monitors}
\alias{cloudwatchrum_list_app_monitors}
\title{Returns a list of the Amazon CloudWatch RUM app monitors in the account}
\usage{
cloudwatchrum_list_app_monitors(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in one operation. The default is
50. The maximum that you can specify is 100.}

\item{NextToken}{Use the token returned by the previous operation to request the next
page of results.}
}
\description{
Returns a list of the Amazon CloudWatch RUM app monitors in the account.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_list_app_monitors/} for full documentation.
}
\keyword{internal}
