% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_put_rum_events}
\alias{cloudwatchrum_put_rum_events}
\title{Sends telemetry events about your application performance and user
behavior to CloudWatch RUM}
\usage{
cloudwatchrum_put_rum_events(
  Id,
  BatchId,
  AppMonitorDetails,
  UserDetails,
  RumEvents,
  Alias = NULL
)
}
\arguments{
\item{Id}{[required] The ID of the app monitor that is sending this data.}

\item{BatchId}{[required] A unique identifier for this batch of RUM event data.}

\item{AppMonitorDetails}{[required] A structure that contains information about the app monitor that
collected this telemetry information.}

\item{UserDetails}{[required] A structure that contains information about the user session that this
batch of events was collected from.}

\item{RumEvents}{[required] An array of structures that contain the telemetry event data.}

\item{Alias}{If the app monitor uses a resource-based policy that requires
\code{\link[=cloudwatchrum_put_rum_events]{put_rum_events}} requests to specify a
certain alias, specify that alias here. This alias will be compared to
the \code{rum:alias} context key in the resource-based policy. For more
information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-RUM-resource-policies.html}{Using resource-based policies with CloudWatch RUM}.}
}
\description{
Sends telemetry events about your application performance and user behavior to CloudWatch RUM. The code snippet that RUM generates for you to add to your application includes \code{\link[=cloudwatchrum_put_rum_events]{put_rum_events}} operations to send this data to RUM.

See \url{https://www.paws-r-sdk.com/docs/cloudwatchrum_put_rum_events/} for full documentation.
}
\keyword{internal}
