% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_aggregation_authorization}
\alias{configservice_delete_aggregation_authorization}
\title{Deletes the authorization granted to the specified configuration
aggregator account in a specified region}
\usage{
configservice_delete_aggregation_authorization(AuthorizedAccountId,
  AuthorizedAwsRegion)
}
\arguments{
\item{AuthorizedAccountId}{[required] The 12-digit account ID of the account authorized to aggregate data.}

\item{AuthorizedAwsRegion}{[required] The region authorized to collect aggregated data.}
}
\value{
An empty list.
}
\description{
Deletes the authorization granted to the specified configuration
aggregator account in a specified region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_aggregation_authorization(
  AuthorizedAccountId = "string",
  AuthorizedAwsRegion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
