% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_conformance_pack}
\alias{configservice_delete_conformance_pack}
\title{Deletes the specified conformance pack and all the AWS Config rules,
remediation actions, and all evaluation results within that conformance
pack}
\usage{
configservice_delete_conformance_pack(ConformancePackName)
}
\arguments{
\item{ConformancePackName}{[required] Name of the conformance pack you want to delete.}
}
\description{
Deletes the specified conformance pack and all the AWS Config rules,
remediation actions, and all evaluation results within that conformance
pack.

AWS Config sets the conformance pack to \code{DELETE_IN_PROGRESS} until the
deletion is complete. You cannot update a conformance pack while it is
in this state.
}
\section{Request syntax}{
\preformatted{svc$delete_conformance_pack(
  ConformancePackName = "string"
)
}
}

\keyword{internal}
