% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_remediation_configuration}
\alias{configservice_delete_remediation_configuration}
\title{Deletes the remediation configuration}
\usage{
configservice_delete_remediation_configuration(
  ConfigRuleName,
  ResourceType = NULL
)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule for which you want to delete remediation
configuration.}

\item{ResourceType}{The type of a resource.}
}
\description{
Deletes the remediation configuration.

See \url{https://paws-r.github.io/docs/configservice/delete_remediation_configuration.html} for full documentation.
}
\keyword{internal}
