% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_service_linked_configuration_recorder}
\alias{configservice_delete_service_linked_configuration_recorder}
\title{Deletes an existing service-linked configuration recorder}
\usage{
configservice_delete_service_linked_configuration_recorder(ServicePrincipal)
}
\arguments{
\item{ServicePrincipal}{[required] The service principal of the Amazon Web Services service for the
service-linked configuration recorder that you want to delete.}
}
\description{
Deletes an existing service-linked configuration recorder.

See \url{https://www.paws-r-sdk.com/docs/configservice_delete_service_linked_configuration_recorder/} for full documentation.
}
\keyword{internal}
