% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_stored_query}
\alias{configservice_delete_stored_query}
\title{Deletes the stored query for a single Amazon Web Services account and a
single Amazon Web Services Region}
\usage{
configservice_delete_stored_query(QueryName)
}
\arguments{
\item{QueryName}{[required] The name of the query that you want to delete.}
}
\description{
Deletes the stored query for a single Amazon Web Services account and a single Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/configservice_delete_stored_query/} for full documentation.
}
\keyword{internal}
