% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_config_rules}
\alias{configservice_describe_config_rules}
\title{Returns details about your Config rules}
\usage{
configservice_describe_config_rules(ConfigRuleNames = NULL, NextToken = NULL)
}
\arguments{
\item{ConfigRuleNames}{The names of the Config rules for which you want details. If you do not
specify any names, Config returns details for all your rules.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns details about your Config rules.

See \url{https://paws-r.github.io/docs/configservice/describe_config_rules.html} for full documentation.
}
\keyword{internal}
