% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_recorder_status}
\alias{configservice_describe_configuration_recorder_status}
\title{Returns the current status of the specified configuration recorder as
well as the status of the last recording event for the recorder}
\usage{
configservice_describe_configuration_recorder_status(
  ConfigurationRecorderNames = NULL
)
}
\arguments{
\item{ConfigurationRecorderNames}{The name(s) of the configuration recorder. If the name is not specified,
the action returns the current status of all the configuration recorders
associated with the account.}
}
\description{
Returns the current status of the specified configuration recorder as well as the status of the last recording event for the recorder. If a configuration recorder is not specified, this action returns the status of all configuration recorders associated with the account.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_configuration_recorder_status/} for full documentation.
}
\keyword{internal}
