% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_organization_config_rule_statuses}
\alias{configservice_describe_organization_config_rule_statuses}
\title{Describe organization config rule statuses}
\usage{
configservice_describe_organization_config_rule_statuses(
  OrganizationConfigRuleNames, Limit, NextToken)
}
\arguments{
\item{OrganizationConfigRuleNames}{}

\item{Limit}{}

\item{NextToken}{}
}
\description{
Describe organization config rule statuses
}
\section{Request syntax}{
\preformatted{svc$describe_organization_config_rule_statuses(
  OrganizationConfigRuleNames = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
