% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_organization_config_rules}
\alias{configservice_describe_organization_config_rules}
\title{Returns a list of organization Config rules}
\usage{
configservice_describe_organization_config_rules(
  OrganizationConfigRuleNames = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OrganizationConfigRuleNames}{The names of organization Config rules for which you want details. If
you do not specify any names, Config returns details for all your
organization Config rules.}

\item{Limit}{The maximum number of organization Config rules returned on each page.
If you do no specify a number, Config uses the default. The default is
100.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\description{
Returns a list of organization Config rules.

See \url{https://www.paws-r-sdk.com/docs/configservice_describe_organization_config_rules/} for full documentation.
}
\keyword{internal}
