% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_remediation_exceptions}
\alias{configservice_describe_remediation_exceptions}
\title{Returns the details of one or more remediation exceptions}
\usage{
configservice_describe_remediation_exceptions(ConfigRuleName,
  ResourceKeys, Limit, NextToken)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the AWS Config rule.}

\item{ResourceKeys}{An exception list of resource exception keys to be processed with the
current request. AWS Config adds exception for each resource key. For
example, AWS Config adds 3 exceptions for 3 resource keys.}

\item{Limit}{The maximum number of RemediationExceptionResourceKey returned on each
page. The default is 25. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned in a previous request that you use to
request the next page of results in a paginated response.}
}
\value{
A list with the following syntax:\preformatted{list(
  RemediationExceptions = list(
    list(
      ConfigRuleName = "string",
      ResourceType = "string",
      ResourceId = "string",
      Message = "string",
      ExpirationTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns the details of one or more remediation exceptions. A detailed
view of a remediation exception for a set of resources that includes an
explanation of an exception and the time when the exception will be
deleted. When you specify the limit and the next token, you receive a
paginated response.

AWS Config generates a remediation exception when a problem occurs
executing a remediation action to a specific resource. Remediation
exceptions blocks auto-remediation until the exception is cleared.

When you specify the limit and the next token, you receive a paginated
response.

Limit and next token are not applicable if you request resources in
batch. It is only applicable, when you request all resources.
}
\section{Request syntax}{
\preformatted{svc$describe_remediation_exceptions(
  ConfigRuleName = "string",
  ResourceKeys = list(
    list(
      ResourceType = "string",
      ResourceId = "string"
    )
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
