% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_aggregate_conformance_pack_compliance_summary}
\alias{configservice_get_aggregate_conformance_pack_compliance_summary}
\title{Returns the count of compliant and noncompliant conformance packs across
all AWS Accounts and AWS Regions in an aggregator}
\usage{
configservice_get_aggregate_conformance_pack_compliance_summary(
  ConfigurationAggregatorName, Filters, GroupByKey, Limit, NextToken)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{Filters}{Filters the results based on the
\code{AggregateConformancePackComplianceSummaryFilters} object.}

\item{GroupByKey}{Groups the result based on AWS Account ID or AWS Region.}

\item{Limit}{The maximum number of results returned on each page. The default is
maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}
}
\value{
A list with the following syntax:\preformatted{list(
  AggregateConformancePackComplianceSummaries = list(
    list(
      ComplianceSummary = list(
        CompliantConformancePackCount = 123,
        NonCompliantConformancePackCount = 123
      ),
      GroupName = "string"
    )
  ),
  GroupByKey = "string",
  NextToken = "string"
)
}
}
\description{
Returns the count of compliant and noncompliant conformance packs across
all AWS Accounts and AWS Regions in an aggregator. You can filter based
on AWS Account ID or AWS Region.

The results can return an empty result page, but if you have a
nextToken, the results are displayed on the next page.
}
\section{Request syntax}{
\preformatted{svc$get_aggregate_conformance_pack_compliance_summary(
  ConfigurationAggregatorName = "string",
  Filters = list(
    AccountId = "string",
    AwsRegion = "string"
  ),
  GroupByKey = "ACCOUNT_ID"|"AWS_REGION",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
