% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_compliance_summary_by_config_rule}
\alias{configservice_get_compliance_summary_by_config_rule}
\title{Returns the number of Config rules that are compliant and noncompliant,
up to a maximum of 25 for each}
\usage{
configservice_get_compliance_summary_by_config_rule()
}
\description{
Returns the number of Config rules that are compliant and noncompliant, up to a maximum of 25 for each.

See \url{https://paws-r.github.io/docs/configservice/get_compliance_summary_by_config_rule.html} for full documentation.
}
\keyword{internal}
