% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_compliance_summary_by_config_rule}
\alias{configservice_get_compliance_summary_by_config_rule}
\title{Returns the number of AWS Config rules that are compliant and
noncompliant, up to a maximum of 25 for each}
\usage{
configservice_get_compliance_summary_by_config_rule()
}
\value{
A list with the following syntax:\preformatted{list(
  ComplianceSummary = list(
    CompliantResourceCount = list(
      CappedCount = 123,
      CapExceeded = TRUE|FALSE
    ),
    NonCompliantResourceCount = list(
      CappedCount = 123,
      CapExceeded = TRUE|FALSE
    ),
    ComplianceSummaryTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Returns the number of AWS Config rules that are compliant and
noncompliant, up to a maximum of 25 for each.
}
\keyword{internal}
