% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_get_stored_query}
\alias{configservice_get_stored_query}
\title{Returns the details of a specific stored query}
\usage{
configservice_get_stored_query(QueryName)
}
\arguments{
\item{QueryName}{[required] The name of the query.}
}
\value{
A list with the following syntax:\preformatted{list(
  StoredQuery = list(
    QueryId = "string",
    QueryArn = "string",
    QueryName = "string",
    Description = "string",
    Expression = "string"
  )
)
}
}
\description{
Returns the details of a specific stored query.
}
\section{Request syntax}{
\preformatted{svc$get_stored_query(
  QueryName = "string"
)
}
}

\keyword{internal}
