% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_aggregation_authorization}
\alias{configservice_put_aggregation_authorization}
\title{Authorizes the aggregator account and region to collect data from the
source account and region}
\usage{
configservice_put_aggregation_authorization(AuthorizedAccountId,
  AuthorizedAwsRegion)
}
\arguments{
\item{AuthorizedAccountId}{[required] The 12-digit account ID of the account authorized to aggregate data.}

\item{AuthorizedAwsRegion}{[required] The region authorized to collect aggregated data.}
}
\description{
Authorizes the aggregator account and region to collect data from the
source account and region.
}
\section{Request syntax}{
\preformatted{svc$put_aggregation_authorization(
  AuthorizedAccountId = "string",
  AuthorizedAwsRegion = "string"
)
}
}

\keyword{internal}
