% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_configuration_aggregator}
\alias{configservice_put_configuration_aggregator}
\title{Creates and updates the configuration aggregator with the selected
source accounts and regions}
\usage{
configservice_put_configuration_aggregator(
  ConfigurationAggregatorName,
  AccountAggregationSources = NULL,
  OrganizationAggregationSource = NULL,
  Tags = NULL,
  AggregatorFilters = NULL
)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{AccountAggregationSources}{A list of AccountAggregationSource object.}

\item{OrganizationAggregationSource}{An OrganizationAggregationSource object.}

\item{Tags}{An array of tag object.}

\item{AggregatorFilters}{An object to filter configuration recorders in an aggregator. Either
\code{ResourceType} or \code{ServicePrincipal} is required.}
}
\description{
Creates and updates the configuration aggregator with the selected source accounts and regions. The source account can be individual account(s) or an organization.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_configuration_aggregator/} for full documentation.
}
\keyword{internal}
