% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_configuration_aggregator}
\alias{configservice_put_configuration_aggregator}
\title{Creates and updates the configuration aggregator with the selected
source accounts and regions}
\usage{
configservice_put_configuration_aggregator(ConfigurationAggregatorName,
  AccountAggregationSources, OrganizationAggregationSource, Tags)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{AccountAggregationSources}{A list of AccountAggregationSource object.}

\item{OrganizationAggregationSource}{An OrganizationAggregationSource object.}

\item{Tags}{An array of tag object.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConfigurationAggregator = list(
    ConfigurationAggregatorName = "string",
    ConfigurationAggregatorArn = "string",
    AccountAggregationSources = list(
      list(
        AccountIds = list(
          "string"
        ),
        AllAwsRegions = TRUE|FALSE,
        AwsRegions = list(
          "string"
        )
      )
    ),
    OrganizationAggregationSource = list(
      RoleArn = "string",
      AwsRegions = list(
        "string"
      ),
      AllAwsRegions = TRUE|FALSE
    ),
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTime = as.POSIXct(
      "2015-01-01"
    ),
    CreatedBy = "string"
  )
)
}
}
\description{
Creates and updates the configuration aggregator with the selected
source accounts and regions. The source account can be individual
account(s) or an organization.

\code{accountIds} that are passed will be replaced with existing accounts. If
you want to add additional accounts into the aggregator, call
\code{DescribeAggregator} to get the previous accounts and then append new
ones.

AWS Config should be enabled in source accounts and regions you want to
aggregate.

If your source type is an organization, you must be signed in to the
management account or a registered delegated administrator and all the
features must be enabled in your organization. If the caller is a
management account, AWS Config calls \code{EnableAwsServiceAccess} API to
enable integration between AWS Config and AWS Organizations. If the
caller is a registered delegated administrator, AWS Config calls
\code{ListDelegatedAdministrators} API to verify whether the caller is a
valid delegated administrator.

To register a delegated administrator, see \href{https://docs.aws.amazon.com/config/latest/developerguide/set-up-aggregator-cli.html#register-a-delegated-administrator-cli}{Register a Delegated Administrator}
in the AWS Config developer guide.
}
\section{Request syntax}{
\preformatted{svc$put_configuration_aggregator(
  ConfigurationAggregatorName = "string",
  AccountAggregationSources = list(
    list(
      AccountIds = list(
        "string"
      ),
      AllAwsRegions = TRUE|FALSE,
      AwsRegions = list(
        "string"
      )
    )
  ),
  OrganizationAggregationSource = list(
    RoleArn = "string",
    AwsRegions = list(
      "string"
    ),
    AllAwsRegions = TRUE|FALSE
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
