% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_configuration_recorder}
\alias{configservice_put_configuration_recorder}
\title{Creates a new configuration recorder to record the selected resource
configurations}
\usage{
configservice_put_configuration_recorder(ConfigurationRecorder)
}
\arguments{
\item{ConfigurationRecorder}{[required] The configuration recorder object that records each configuration change
made to the resources.}
}
\description{
Creates a new configuration recorder to record the selected resource configurations.

See \url{https://paws-r.github.io/docs/configservice/put_configuration_recorder.html} for full documentation.
}
\keyword{internal}
