% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_conformance_pack}
\alias{configservice_put_conformance_pack}
\title{Creates or updates a conformance pack}
\usage{
configservice_put_conformance_pack(
  ConformancePackName,
  TemplateS3Uri = NULL,
  TemplateBody = NULL,
  DeliveryS3Bucket = NULL,
  DeliveryS3KeyPrefix = NULL,
  ConformancePackInputParameters = NULL,
  TemplateSSMDocumentDetails = NULL
)
}
\arguments{
\item{ConformancePackName}{[required] The unique name of the conformance pack you want to deploy.}

\item{TemplateS3Uri}{The location of the file containing the template body
(\verb{s3://bucketname/prefix}). The uri must point to a conformance pack
template (max size: 300 KB) that is located in an Amazon S3 bucket in
the same Region as the conformance pack.

You must have access to read Amazon S3 bucket. In addition, in order to
ensure a successful deployment, the template object must not be in an
\href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage-class-intro.html}{archived storage class}
if this parameter is passed.}

\item{TemplateBody}{A string that contains the full conformance pack template body. The
structure containing the template body has a minimum length of 1 byte
and a maximum length of 51,200 bytes.

You can use a YAML template with two resource types: Config rule
(\verb{AWS::Config::ConfigRule}) and remediation action
(\verb{AWS::Config::RemediationConfiguration}).}

\item{DeliveryS3Bucket}{The name of the Amazon S3 bucket where Config stores conformance pack
templates.

This field is optional.}

\item{DeliveryS3KeyPrefix}{The prefix for the Amazon S3 bucket.

This field is optional.}

\item{ConformancePackInputParameters}{A list of \code{ConformancePackInputParameter} objects.}

\item{TemplateSSMDocumentDetails}{An object of type \code{TemplateSSMDocumentDetails}, which contains the name
or the Amazon Resource Name (ARN) of the Amazon Web Services Systems
Manager document (SSM document) and the version of the SSM document that
is used to create a conformance pack.}
}
\description{
Creates or updates a conformance pack. A conformance pack is a collection of Config rules that can be easily deployed in an account and a region and across an organization. For information on how many conformance packs you can have per account, see \href{https://docs.aws.amazon.com/config/latest/developerguide/configlimits.html}{\strong{Service Limits}} in the \emph{Config Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_conformance_pack/} for full documentation.
}
\keyword{internal}
