% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_delivery_channel}
\alias{configservice_put_delivery_channel}
\title{Creates or updates a delivery channel to deliver configuration
information and other compliance information}
\usage{
configservice_put_delivery_channel(DeliveryChannel)
}
\arguments{
\item{DeliveryChannel}{[required] An object for the delivery channel. A delivery channel sends
notifications and updated configuration states.}
}
\description{
Creates or updates a delivery channel to deliver configuration information and other compliance information.

See \url{https://www.paws-r-sdk.com/docs/configservice_put_delivery_channel/} for full documentation.
}
\keyword{internal}
