% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_external_evaluation}
\alias{configservice_put_external_evaluation}
\title{Add or updates the evaluations for process checks}
\usage{
configservice_put_external_evaluation(ConfigRuleName,
  ExternalEvaluation)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the AWS Config rule.}

\item{ExternalEvaluation}{[required] An \code{ExternalEvaluation} object that provides details about compliance.}
}
\value{
An empty list.
}
\description{
Add or updates the evaluations for process checks. This API checks if
the rule is a process check when the name of the AWS Config rule is
provided.
}
\section{Request syntax}{
\preformatted{svc$put_external_evaluation(
  ConfigRuleName = "string",
  ExternalEvaluation = list(
    ComplianceResourceType = "string",
    ComplianceResourceId = "string",
    ComplianceType = "COMPLIANT"|"NON_COMPLIANT"|"NOT_APPLICABLE"|"INSUFFICIENT_DATA",
    Annotation = "string",
    OrderingTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}

\keyword{internal}
