% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_organization_config_rule}
\alias{configservice_put_organization_config_rule}
\title{Put organization config rule}
\usage{
configservice_put_organization_config_rule(OrganizationConfigRuleName,
  OrganizationManagedRuleMetadata, OrganizationCustomRuleMetadata,
  ExcludedAccounts)
}
\arguments{
\item{OrganizationConfigRuleName}{[required]}

\item{OrganizationManagedRuleMetadata}{}

\item{OrganizationCustomRuleMetadata}{}

\item{ExcludedAccounts}{}
}
\description{
Put organization config rule
}
\section{Request syntax}{
\preformatted{svc$put_organization_config_rule(
  OrganizationConfigRuleName = "string",
  OrganizationManagedRuleMetadata = list(
    Description = "string",
    RuleIdentifier = "string",
    InputParameters = "string",
    MaximumExecutionFrequency = "One_Hour"|"Three_Hours"|"Six_Hours"|"Twelve_Hours"|"TwentyFour_Hours",
    ResourceTypesScope = list(
      "string"
    ),
    ResourceIdScope = "string",
    TagKeyScope = "string",
    TagValueScope = "string"
  ),
  OrganizationCustomRuleMetadata = list(
    Description = "string",
    LambdaFunctionArn = "string",
    OrganizationConfigRuleTriggerTypes = list(
      "ConfigurationItemChangeNotification"|"OversizedConfigurationItemChangeNotification"|"ScheduledNotification"
    ),
    InputParameters = "string",
    MaximumExecutionFrequency = "One_Hour"|"Three_Hours"|"Six_Hours"|"Twelve_Hours"|"TwentyFour_Hours",
    ResourceTypesScope = list(
      "string"
    ),
    ResourceIdScope = "string",
    TagKeyScope = "string",
    TagValueScope = "string"
  ),
  ExcludedAccounts = list(
    "string"
  )
)
}
}

\keyword{internal}
