% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_start_configuration_recorder}
\alias{configservice_start_configuration_recorder}
\title{Starts recording configurations of the Amazon Web Services resources you
have selected to record in your Amazon Web Services account}
\usage{
configservice_start_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{[required] The name of the recorder object that records each configuration change
made to the resources.}
}
\description{
Starts recording configurations of the Amazon Web Services resources you have selected to record in your Amazon Web Services account.

See \url{https://paws-r.github.io/docs/configservice/start_configuration_recorder.html} for full documentation.
}
\keyword{internal}
