% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_stop_configuration_recorder}
\alias{configservice_stop_configuration_recorder}
\title{Stops recording configurations of the AWS resources you have selected to
record in your AWS account}
\usage{
configservice_stop_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{[required] The name of the recorder object that records each configuration change
made to the resources.}
}
\description{
Stops recording configurations of the AWS resources you have selected to
record in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$stop_configuration_recorder(
  ConfigurationRecorderName = "string"
)
}
}

\keyword{internal}
