% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_untag_resource}
\alias{configservice_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
configservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to
list the tags. The following resources are supported:
\itemize{
\item \code{ConfigurationRecorder}
\item \code{ConfigRule}
\item \code{OrganizationConfigRule}
\item \code{ConformancePack}
\item \code{OrganizationConformancePack}
\item \code{ConfigurationAggregator}
\item \code{AggregationAuthorization}
\item \code{StoredQuery}
}}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/configservice_untag_resource/} for full documentation.
}
\keyword{internal}
