% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_baselines}
\alias{controltower_list_baselines}
\title{Returns a summary list of all available baselines}
\usage{
controltower_list_baselines(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{A pagination token.}

\item{maxResults}{The maximum number of results to be shown.}
}
\description{
Returns a summary list of all available baselines. For usage examples, see \href{https://docs.aws.amazon.com/controltower/latest/userguide/baseline-api-examples.html}{\emph{the Amazon Web Services Control Tower User Guide}} .

See \url{https://www.paws-r-sdk.com/docs/controltower_list_baselines/} for full documentation.
}
\keyword{internal}
