% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_landing_zone_operations}
\alias{controltower_list_landing_zone_operations}
\title{Lists all landing zone operations from the past 90 days}
\usage{
controltower_list_landing_zone_operations(
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filter}{An input filter for the
\code{\link[=controltower_list_landing_zone_operations]{list_landing_zone_operations}}
API that lets you select the types of landing zone operations to view.}

\item{nextToken}{The token to continue the list from a previous API call with the same
parameters.}

\item{maxResults}{How many results to return per API call.}
}
\description{
Lists all landing zone operations from the past 90 days. Results are sorted by time, with the most recent operation first.

See \url{https://www.paws-r-sdk.com/docs/controltower_list_landing_zone_operations/} for full documentation.
}
\keyword{internal}
