% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_list_landing_zones}
\alias{controltower_list_landing_zones}
\title{Returns the landing zone ARN for the landing zone deployed in your
managed account}
\usage{
controltower_list_landing_zones(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token to continue the list from a previous API call with the same
parameters.}

\item{maxResults}{The maximum number of returned landing zone ARNs, which is one.}
}
\description{
Returns the landing zone ARN for the landing zone deployed in your managed account. This API also creates an ARN for existing accounts that do not yet have a landing zone ARN.

See \url{https://www.paws-r-sdk.com/docs/controltower_list_landing_zones/} for full documentation.
}
\keyword{internal}
