% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controltower_operations.R
\name{controltower_reset_landing_zone}
\alias{controltower_reset_landing_zone}
\title{This API call resets a landing zone}
\usage{
controltower_reset_landing_zone(landingZoneIdentifier)
}
\arguments{
\item{landingZoneIdentifier}{[required] The unique identifier of the landing zone.}
}
\description{
This API call resets a landing zone. It starts an asynchronous operation that resets the landing zone to the parameters specified in the original configuration, which you specified in the manifest file. Nothing in the manifest file's original landing zone configuration is changed during the reset process, by default. This API is not the same as a rollback of a landing zone version, which is not a supported operation.

See \url{https://www.paws-r-sdk.com/docs/controltower_reset_landing_zone/} for full documentation.
}
\keyword{internal}
