% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_delete_kx_user}
\alias{finspace_delete_kx_user}
\title{Deletes a user in the specified kdb environment}
\usage{
finspace_delete_kx_user(userName, environmentId)
}
\arguments{
\item{userName}{[required] A unique identifier for the user that you want to delete.}

\item{environmentId}{[required] A unique identifier for the kdb environment.}
}
\description{
Deletes a user in the specified kdb environment.

See \url{https://www.paws-r-sdk.com/docs/finspace_delete_kx_user/} for full documentation.
}
\keyword{internal}
