% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_list_tags_for_resource}
\alias{finspace_list_tags_for_resource}
\title{A list of all tags for a resource}
\usage{
finspace_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name of the resource.}
}
\description{
A list of all tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/finspace_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
