% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_describe_event_details_for_organization}
\alias{health_describe_event_details_for_organization}
\title{Returns detailed information about one or more specified events for one
or more AWS accounts in your organization}
\usage{
health_describe_event_details_for_organization(
  organizationEventDetailFilters, locale)
}
\arguments{
\item{organizationEventDetailFilters}{[required] A set of JSON elements that includes the \code{awsAccountId} and the
\code{eventArn}.}

\item{locale}{The locale (language) to return information in. English (en) is the
default and the only supported value at this time.}
}
\value{
A list with the following syntax:\preformatted{list(
  successfulSet = list(
    list(
      awsAccountId = "string",
      event = list(
        arn = "string",
        service = "string",
        eventTypeCode = "string",
        eventTypeCategory = "issue"|"accountNotification"|"scheduledChange"|"investigation",
        region = "string",
        availabilityZone = "string",
        startTime = as.POSIXct(
          "2015-01-01"
        ),
        endTime = as.POSIXct(
          "2015-01-01"
        ),
        lastUpdatedTime = as.POSIXct(
          "2015-01-01"
        ),
        statusCode = "open"|"closed"|"upcoming",
        eventScopeCode = "PUBLIC"|"ACCOUNT_SPECIFIC"|"NONE"
      ),
      eventDescription = list(
        latestDescription = "string"
      ),
      eventMetadata = list(
        "string"
      )
    )
  ),
  failedSet = list(
    list(
      awsAccountId = "string",
      eventArn = "string",
      errorName = "string",
      errorMessage = "string"
    )
  )
)
}
}
\description{
Returns detailed information about one or more specified events for one
or more AWS accounts in your organization. This information includes
standard event data (such as the AWS Region and service), an event
description, and (depending on the event) possible metadata. This
operation doesn't return affected entities, such as the resources
related to the event. To return affected entities, use the
\code{\link[=health_describe_affected_entities_for_organization]{describe_affected_entities_for_organization}}
operation.

Before you can call this operation, you must first enable AWS Health to
work with AWS Organizations. To do this, call the
\code{\link[=health_enable_health_service_access_for_organization]{enable_health_service_access_for_organization}}
operation from your organization's management account.

When you call the
\code{\link[=health_describe_event_details_for_organization]{describe_event_details_for_organization}}
operation, specify the \code{organizationEventDetailFilters} object in the
request. Depending on the AWS Health event type, note the following
differences:
\itemize{
\item To return event details for a public event, you must specify a null
value for the \code{awsAccountId} parameter. If you specify an account ID
for a public event, AWS Health returns an error message because
public events aren't specific to an account.
\item To return event details for an event that is specific to an account
in your organization, you must specify the \code{awsAccountId} parameter
in the request. If you don't specify an account ID, AWS Health
returns an error message because the event is specific to an account
in your organization.
}

For more information, see
\href{https://docs.aws.amazon.com/health/latest/APIReference/API_Event.html}{Event}.

This operation doesn't support resource-level permissions. You can't use
this operation to allow or deny access to specific AWS Health events.
For more information, see \href{https://docs.aws.amazon.com/health/latest/ug/security_iam_id-based-policy-examples.html#resource-action-based-conditions}{Resource- and action-based conditions}
in the \emph{AWS Health User Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_event_details_for_organization(
  organizationEventDetailFilters = list(
    list(
      eventArn = "string",
      awsAccountId = "string"
    )
  ),
  locale = "string"
)
}
}

\keyword{internal}
