% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_enable_health_service_access_for_organization}
\alias{health_enable_health_service_access_for_organization}
\title{Enables AWS Health to work with AWS Organizations}
\usage{
health_enable_health_service_access_for_organization()
}
\value{
An empty list.
}
\description{
Enables AWS Health to work with AWS Organizations. You can use the
organizational view feature to aggregate events from all AWS accounts in
your organization in a centralized location.

This operation also creates a service-linked role for the management
account in the organization.

To call this operation, you must meet the following requirements:
\itemize{
\item You must have a Business or Enterprise Support plan from \href{https://aws.amazon.com/premiumsupport/}{AWS Support} to use the AWS
Health API. If you call the AWS Health API from an AWS account that
doesn't have a Business or Enterprise Support plan, you receive a
\code{SubscriptionRequiredException} error.
\item You must have permission to call this operation from the
organization's management account. For example IAM policies, see
\href{https://docs.aws.amazon.com/health/latest/ug/security_iam_id-based-policy-examples.html}{AWS Health identity-based policy examples}.
}

If you don't have the required support plan, you can instead use the AWS
Health console to enable the organizational view feature. For more
information, see \href{https://docs.aws.amazon.com/health/latest/ug/aggregate-events.html}{Aggregating AWS Health events}
in the \emph{AWS Health User Guide}.
}
\keyword{internal}
