% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_enable_health_service_access_for_organization}
\alias{health_enable_health_service_access_for_organization}
\title{Enables Health to work with Organizations}
\usage{
health_enable_health_service_access_for_organization()
}
\description{
Enables Health to work with Organizations. You can use the organizational view feature to aggregate events from all Amazon Web Services accounts in your organization in a centralized location.

See \url{https://www.paws-r-sdk.com/docs/health_enable_health_service_access_for_organization/} for full documentation.
}
\keyword{internal}
