% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_accept_grant}
\alias{licensemanager_accept_grant}
\title{Accepts the specified grant}
\usage{
licensemanager_accept_grant(GrantArn)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}
}
\description{
Accepts the specified grant.

See \url{https://paws-r.github.io/docs/licensemanager/accept_grant.html} for full documentation.
}
\keyword{internal}
