% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_create_license_version}
\alias{licensemanager_create_license_version}
\title{Creates a new version of the specified license}
\usage{
licensemanager_create_license_version(
  LicenseArn,
  LicenseName,
  ProductName,
  Issuer,
  HomeRegion,
  Validity,
  LicenseMetadata = NULL,
  Entitlements,
  ConsumptionConfiguration,
  Status,
  ClientToken,
  SourceVersion = NULL
)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}

\item{LicenseName}{[required] License name.}

\item{ProductName}{[required] Product name.}

\item{Issuer}{[required] License issuer.}

\item{HomeRegion}{[required] Home Region of the license.}

\item{Validity}{[required] Date and time range during which the license is valid, in ISO8601-UTC
format.}

\item{LicenseMetadata}{Information about the license.}

\item{Entitlements}{[required] License entitlements.}

\item{ConsumptionConfiguration}{[required] Configuration for consumption of the license. Choose a provisional
configuration for workloads running with continuous connectivity. Choose
a borrow configuration for workloads with offline usage.}

\item{Status}{[required] License status.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{SourceVersion}{Current version of the license.}
}
\description{
Creates a new version of the specified license.

See \url{https://paws-r.github.io/docs/licensemanager/create_license_version.html} for full documentation.
}
\keyword{internal}
