% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_token}
\alias{licensemanager_delete_token}
\title{Deletes the specified token}
\usage{
licensemanager_delete_token(TokenId)
}
\arguments{
\item{TokenId}{[required] Token ID.}
}
\description{
Deletes the specified token. Must be called in the license home Region.

See \url{https://paws-r.github.io/docs/licensemanager/delete_token.html} for full documentation.
}
\keyword{internal}
