% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_extend_license_consumption}
\alias{licensemanager_extend_license_consumption}
\title{Extends the expiration date for license consumption}
\usage{
licensemanager_extend_license_consumption(
  LicenseConsumptionToken,
  DryRun = NULL
)
}
\arguments{
\item{LicenseConsumptionToken}{[required] License consumption token.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request. Provides an error response if you do not
have the required permissions.}
}
\description{
Extends the expiration date for license consumption.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_extend_license_consumption/} for full documentation.
}
\keyword{internal}
