% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_access_token}
\alias{licensemanager_get_access_token}
\title{Gets a temporary access token to use with AssumeRoleWithWebIdentity}
\usage{
licensemanager_get_access_token(Token, TokenProperties)
}
\arguments{
\item{Token}{[required] Refresh token, encoded as a JWT token.}

\item{TokenProperties}{Token properties to validate against those present in the JWT token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccessToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a temporary access token to use with AssumeRoleWithWebIdentity.
Access tokens are valid for one hour.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_access_token(
  Token = "string",
  TokenProperties = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
