% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_grant}
\alias{licensemanager_get_grant}
\title{Gets detailed information about the specified grant}
\usage{
licensemanager_get_grant(GrantArn, Version = NULL)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}

\item{Version}{Grant version.}
}
\description{
Gets detailed information about the specified grant.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_get_grant/} for full documentation.
}
\keyword{internal}
