% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_configuration}
\alias{licensemanager_get_license_configuration}
\title{Gets detailed information about the specified license configuration}
\usage{
licensemanager_get_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] Amazon Resource Name (ARN) of the license configuration.}
}
\description{
Gets detailed information about the specified license configuration.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_get_license_configuration/} for full documentation.
}
\keyword{internal}
