% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_conversion_task}
\alias{licensemanager_get_license_conversion_task}
\title{Gets information about the specified license type conversion task}
\usage{
licensemanager_get_license_conversion_task(LicenseConversionTaskId)
}
\arguments{
\item{LicenseConversionTaskId}{[required] ID of the license type conversion task to retrieve information on.}
}
\description{
Gets information about the specified license type conversion task.

See \url{https://paws-r.github.io/docs/licensemanager/get_license_conversion_task.html} for full documentation.
}
\keyword{internal}
