% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_usage}
\alias{licensemanager_get_license_usage}
\title{Gets detailed information about the usage of the specified license}
\usage{
licensemanager_get_license_usage(LicenseArn)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}
}
\description{
Gets detailed information about the usage of the specified license.
}
\section{Request syntax}{
\preformatted{svc$get_license_usage(
  LicenseArn = "string"
)
}
}

\keyword{internal}
