% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_associations_for_license_configuration}
\alias{licensemanager_list_associations_for_license_configuration}
\title{Lists the resource associations for the specified license configuration}
\usage{
licensemanager_list_associations_for_license_configuration(
  LicenseConfigurationArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{LicenseConfigurationArn}{[required] Amazon Resource Name (ARN) of a license configuration.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists the resource associations for the specified license configuration.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_associations_for_license_configuration/} for full documentation.
}
\keyword{internal}
