% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_distributed_grants}
\alias{licensemanager_list_distributed_grants}
\title{Lists the grants distributed for the specified license}
\usage{
licensemanager_list_distributed_grants(
  GrantArns = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{GrantArns}{Amazon Resource Names (ARNs) of the grants.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{LicenseArn}
\item \code{GrantStatus}
\item \code{GranteePrincipalARN}
\item \code{ProductSKU}
\item \code{LicenseIssuerName}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists the grants distributed for the specified license.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_distributed_grants/} for full documentation.
}
\keyword{internal}
