% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_license_versions}
\alias{licensemanager_list_license_versions}
\title{Lists all versions of the specified license}
\usage{
licensemanager_list_license_versions(LicenseArn, NextToken, MaxResults)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists all versions of the specified license.
}
\section{Request syntax}{
\preformatted{svc$list_license_versions(
  LicenseArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
