% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_licenses}
\alias{licensemanager_list_licenses}
\title{Lists the licenses for your account}
\usage{
licensemanager_list_licenses(
  LicenseArns = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LicenseArns}{Amazon Resource Names (ARNs) of the licenses.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{Beneficiary}
\item \code{ProductSKU}
\item \code{Fingerprint}
\item \code{Status}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists the licenses for your account.

See \url{https://paws-r.github.io/docs/licensemanager/list_licenses.html} for full documentation.
}
\keyword{internal}
