% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_received_licenses}
\alias{licensemanager_list_received_licenses}
\title{Lists received licenses}
\usage{
licensemanager_list_received_licenses(
  LicenseArns = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LicenseArns}{Amazon Resource Names (ARNs) of the licenses.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{ProductSKU}
\item \code{Status}
\item \code{Fingerprint}
\item \code{IssuerName}
\item \code{Beneficiary}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists received licenses.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_received_licenses/} for full documentation.
}
\keyword{internal}
