% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_tags_for_resource}
\alias{licensemanager_list_tags_for_resource}
\title{Lists the tags for the specified license configuration}
\usage{
licensemanager_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the license configuration.}
}
\description{
Lists the tags for the specified license configuration.

See \url{https://www.paws-r-sdk.com/docs/licensemanager_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
