% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_tokens}
\alias{licensemanager_list_tokens}
\title{Lists your tokens}
\usage{
licensemanager_list_tokens(
  TokenIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TokenIds}{Token IDs.}

\item{Filters}{Filters to scope the results. The following filter is supported:
\itemize{
\item \code{LicenseArns}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists your tokens.

See \url{https://paws-r.github.io/docs/licensemanager/list_tokens.html} for full documentation.
}
\keyword{internal}
