% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerlinuxsubscriptions_operations.R
\name{licensemanagerlinuxsubscriptions_list_linux_subscription_instances}
\alias{licensemanagerlinuxsubscriptions_list_linux_subscription_instances}
\title{Lists the running Amazon EC2 instances that were discovered with
commercial Linux subscriptions}
\usage{
licensemanagerlinuxsubscriptions_list_linux_subscription_instances(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results to those
that match one or more sets of key-value pairs that you specify. For
example, you can filter by the name of \code{AmiID} with an optional operator
to see subscriptions that match, partially match, or don't match a
certain Amazon Machine Image (AMI) ID.

The valid names for this filter are:
\itemize{
\item \code{AmiID}
\item \code{InstanceID}
\item \code{AccountID}
\item \code{Status}
\item \code{Region}
\item \code{UsageOperation}
\item \code{ProductCode}
\item \code{InstanceType}
}

The valid Operators for this filter are:
\itemize{
\item \code{contains}
\item \code{equals}
\item \code{Notequal}
}}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists the running Amazon EC2 instances that were discovered with commercial Linux subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerlinuxsubscriptions_list_linux_subscription_instances/} for full documentation.
}
\keyword{internal}
