% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_associate_user}
\alias{licensemanagerusersubscriptions_associate_user}
\title{Associates the user to an EC2 instance to utilize user-based
subscriptions}
\usage{
licensemanagerusersubscriptions_associate_user(
  Username,
  InstanceId,
  IdentityProvider,
  Domain = NULL,
  Tags = NULL
)
}
\arguments{
\item{Username}{[required] The user name from the identity provider.}

\item{InstanceId}{[required] The ID of the EC2 instance that provides the user-based subscription.}

\item{IdentityProvider}{[required] The identity provider for the user.}

\item{Domain}{The domain name of the Active Directory that contains information for
the user to associate.}

\item{Tags}{The tags that apply for the user association.}
}
\description{
Associates the user to an EC2 instance to utilize user-based subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_associate_user/} for full documentation.
}
\keyword{internal}
