% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_deregister_identity_provider}
\alias{licensemanagerusersubscriptions_deregister_identity_provider}
\title{Deregisters the identity provider from providing user-based
subscriptions}
\usage{
licensemanagerusersubscriptions_deregister_identity_provider(
  IdentityProvider,
  Product
)
}
\arguments{
\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{Product}{[required] The name of the user-based subscription product.}
}
\description{
Deregisters the identity provider from providing user-based subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_deregister_identity_provider/} for full documentation.
}
\keyword{internal}
