% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_disassociate_user}
\alias{licensemanagerusersubscriptions_disassociate_user}
\title{Disassociates the user from an EC2 instance providing user-based
subscriptions}
\usage{
licensemanagerusersubscriptions_disassociate_user(
  Username = NULL,
  InstanceId = NULL,
  IdentityProvider = NULL,
  InstanceUserArn = NULL,
  Domain = NULL
)
}
\arguments{
\item{Username}{The user name from the Active Directory identity provider for the user.}

\item{InstanceId}{The ID of the EC2 instance which provides user-based subscriptions.}

\item{IdentityProvider}{An object that specifies details for the Active Directory identity
provider.}

\item{InstanceUserArn}{The Amazon Resource Name (ARN) of the user to disassociate from the EC2
instance.}

\item{Domain}{The domain name of the Active Directory that contains information for
the user to disassociate.}
}
\description{
Disassociates the user from an EC2 instance providing user-based subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_disassociate_user/} for full documentation.
}
\keyword{internal}
