% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_identity_providers}
\alias{licensemanagerusersubscriptions_list_identity_providers}
\title{Lists the identity providers for user-based subscriptions}
\usage{
licensemanagerusersubscriptions_list_identity_providers(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists the identity providers for user-based subscriptions.

See \url{https://paws-r.github.io/docs/licensemanagerusersubscriptions/list_identity_providers.html} for full documentation.
}
\keyword{internal}
