% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_product_subscriptions}
\alias{licensemanagerusersubscriptions_list_product_subscriptions}
\title{Lists the user-based subscription products available from an identity
provider}
\usage{
licensemanagerusersubscriptions_list_product_subscriptions(
  Filters = NULL,
  IdentityProvider,
  MaxResults = NULL,
  NextToken = NULL,
  Product
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results to those
that match one or more sets of key-value pairs that you specify.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}

\item{Product}{[required] The name of the user-based subscription product.}
}
\description{
Lists the user-based subscription products available from an identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_product_subscriptions/} for full documentation.
}
\keyword{internal}
