% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_product_subscriptions}
\alias{licensemanagerusersubscriptions_list_product_subscriptions}
\title{Lists the user-based subscription products available from an identity
provider}
\usage{
licensemanagerusersubscriptions_list_product_subscriptions(
  Product = NULL,
  IdentityProvider,
  MaxResults = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Product}{The name of the user-based subscription product.

Valid values: \code{VISUAL_STUDIO_ENTERPRISE} | \code{VISUAL_STUDIO_PROFESSIONAL}
| \code{OFFICE_PROFESSIONAL_PLUS} | \code{REMOTE_DESKTOP_SERVICES}}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{MaxResults}{The maximum number of results to return from a single request.}

\item{Filters}{You can use the following filters to streamline results:
\itemize{
\item Status
\item Username
\item Domain
}}

\item{NextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Lists the user-based subscription products available from an identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_product_subscriptions/} for full documentation.
}
\keyword{internal}
