% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_user_associations}
\alias{licensemanagerusersubscriptions_list_user_associations}
\title{Lists user associations for an identity provider}
\usage{
licensemanagerusersubscriptions_list_user_associations(
  InstanceId,
  IdentityProvider,
  MaxResults = NULL,
  Filters = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The ID of the EC2 instance, which provides user-based subscriptions.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{MaxResults}{The maximum number of results to return from a single request.}

\item{Filters}{You can use the following filters to streamline results:
\itemize{
\item Status
\item Username
\item Domain
}}

\item{NextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Lists user associations for an identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_list_user_associations/} for full documentation.
}
\keyword{internal}
