% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_register_identity_provider}
\alias{licensemanagerusersubscriptions_register_identity_provider}
\title{Registers an identity provider for user-based subscriptions}
\usage{
licensemanagerusersubscriptions_register_identity_provider(
  IdentityProvider,
  Product,
  Settings = NULL,
  Tags = NULL
)
}
\arguments{
\item{IdentityProvider}{[required] An object that specifies details for the identity provider to register.}

\item{Product}{[required] The name of the user-based subscription product.

Valid values: \code{VISUAL_STUDIO_ENTERPRISE} | \code{VISUAL_STUDIO_PROFESSIONAL}
| \code{OFFICE_PROFESSIONAL_PLUS}}

\item{Settings}{The registered identity provider’s product related configuration
settings such as the subnets to provision VPC endpoints.}

\item{Tags}{The tags that apply to the identity provider's registration.}
}
\description{
Registers an identity provider for user-based subscriptions.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_register_identity_provider/} for full documentation.
}
\keyword{internal}
