% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_start_product_subscription}
\alias{licensemanagerusersubscriptions_start_product_subscription}
\title{Starts a product subscription for a user with the specified identity
provider}
\usage{
licensemanagerusersubscriptions_start_product_subscription(
  Domain = NULL,
  IdentityProvider,
  Product,
  Tags = NULL,
  Username
)
}
\arguments{
\item{Domain}{The domain name of the Active Directory that contains the user for whom
to start the product subscription.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{Product}{[required] The name of the user-based subscription product.

Valid values: \code{VISUAL_STUDIO_ENTERPRISE} | \code{VISUAL_STUDIO_PROFESSIONAL}
| \code{OFFICE_PROFESSIONAL_PLUS} | \code{REMOTE_DESKTOP_SERVICES}}

\item{Tags}{The tags that apply to the product subscription.}

\item{Username}{[required] The user name from the identity provider of the user.}
}
\description{
Starts a product subscription for a user with the specified identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_start_product_subscription/} for full documentation.
}
\keyword{internal}
