% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_stop_product_subscription}
\alias{licensemanagerusersubscriptions_stop_product_subscription}
\title{Stops a product subscription for a user with the specified identity
provider}
\usage{
licensemanagerusersubscriptions_stop_product_subscription(
  Domain = NULL,
  IdentityProvider,
  Product,
  Username
)
}
\arguments{
\item{Domain}{The domain name of the user.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{Product}{[required] The name of the user-based subscription product.}

\item{Username}{[required] The user name from the identity provider for the user.}
}
\description{
Stops a product subscription for a user with the specified identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_stop_product_subscription/} for full documentation.
}
\keyword{internal}
