% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_stop_product_subscription}
\alias{licensemanagerusersubscriptions_stop_product_subscription}
\title{Stops a product subscription for a user with the specified identity
provider}
\usage{
licensemanagerusersubscriptions_stop_product_subscription(
  Username = NULL,
  IdentityProvider = NULL,
  Product = NULL,
  ProductUserArn = NULL,
  Domain = NULL
)
}
\arguments{
\item{Username}{The user name from the identity provider for the user.}

\item{IdentityProvider}{An object that specifies details for the identity provider.}

\item{Product}{The name of the user-based subscription product.

Valid values: \code{VISUAL_STUDIO_ENTERPRISE} | \code{VISUAL_STUDIO_PROFESSIONAL}
| \code{OFFICE_PROFESSIONAL_PLUS} | \code{REMOTE_DESKTOP_SERVICES}}

\item{ProductUserArn}{The Amazon Resource Name (ARN) of the product user.}

\item{Domain}{The domain name of the Active Directory that contains the user for whom
to stop the product subscription.}
}
\description{
Stops a product subscription for a user with the specified identity provider.

See \url{https://www.paws-r-sdk.com/docs/licensemanagerusersubscriptions_stop_product_subscription/} for full documentation.
}
\keyword{internal}
