% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_update_identity_provider_settings}
\alias{licensemanagerusersubscriptions_update_identity_provider_settings}
\title{Updates additional product configuration settings for the registered
identity provider}
\usage{
licensemanagerusersubscriptions_update_identity_provider_settings(
  IdentityProvider,
  Product,
  UpdateSettings
)
}
\arguments{
\item{IdentityProvider}{[required]}

\item{Product}{[required] The name of the user-based subscription product.}

\item{UpdateSettings}{[required] Updates the registered identity provider’s product related configuration
settings. You can update any combination of settings in a single
operation such as the:
\itemize{
\item Subnets which you want to add to provision VPC endpoints.
\item Subnets which you want to remove the VPC endpoints from.
\item Security group ID which permits traffic to the VPC endpoints.
}}
}
\description{
Updates additional product configuration settings for the registered identity provider.

See \url{https://paws-r.github.io/docs/licensemanagerusersubscriptions/update_identity_provider_settings.html} for full documentation.
}
\keyword{internal}
