% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_delete_workspace}
\alias{managedgrafana_delete_workspace}
\title{Deletes an Amazon Managed Grafana workspace}
\usage{
managedgrafana_delete_workspace(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to delete.}
}
\description{
Deletes an Amazon Managed Grafana workspace.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_delete_workspace/} for full documentation.
}
\keyword{internal}
