% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_describe_workspace_configuration}
\alias{managedgrafana_describe_workspace_configuration}
\title{Gets the current configuration string for the given workspace}
\usage{
managedgrafana_describe_workspace_configuration(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to get configuration information for.}
}
\description{
Gets the current configuration string for the given workspace.

See \url{https://paws-r.github.io/docs/managedgrafana/describe_workspace_configuration.html} for full documentation.
}
\keyword{internal}
