% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_list_tags_for_resource}
\alias{managedgrafana_list_tags_for_resource}
\title{The ListTagsForResource operation returns the tags that are associated
with the Amazon Managed Service for Grafana resource specified by the
resourceArn}
\usage{
managedgrafana_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource the list of tags are associated with.}
}
\description{
The \code{\link[=managedgrafana_list_tags_for_resource]{list_tags_for_resource}} operation returns the tags that are associated with the Amazon Managed Service for Grafana resource specified by the \code{resourceArn}. Currently, the only resource that can be tagged is a workspace.

See \url{https://www.paws-r-sdk.com/docs/managedgrafana_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
