% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_app}
\alias{opsworks_delete_app}
\title{Deletes a specified app}
\usage{
opsworks_delete_app(AppId)
}
\arguments{
\item{AppId}{[required] The app ID.}
}
\value{
An empty list.
}
\description{
Deletes a specified app.

\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack, or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_app(
  AppId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
