% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_user_profile}
\alias{opsworks_delete_user_profile}
\title{Deletes a user profile}
\usage{
opsworks_delete_user_profile(IamUserArn)
}
\arguments{
\item{IamUserArn}{[required] The user's IAM ARN. This can also be a federated user's ARN.}
}
\description{
Deletes a user profile.

See \url{https://www.paws-r-sdk.com/docs/opsworks_delete_user_profile/} for full documentation.
}
\keyword{internal}
