% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_ecs_cluster}
\alias{opsworks_deregister_ecs_cluster}
\title{Deregisters a specified Amazon ECS cluster from a stack}
\usage{
opsworks_deregister_ecs_cluster(EcsClusterArn)
}
\arguments{
\item{EcsClusterArn}{[required] The cluster's Amazon Resource Number (ARN).}
}
\description{
Deregisters a specified Amazon ECS cluster from a stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workinglayers-ecscluster.html#workinglayers-ecscluster-delete}{Resource Management}.

See \url{https://paws-r.github.io/docs/opsworks/deregister_ecs_cluster.html} for full documentation.
}
\keyword{internal}
