% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_deregister_volume}
\alias{opsworks_deregister_volume}
\title{Deregisters an Amazon EBS volume}
\usage{
opsworks_deregister_volume(VolumeId)
}
\arguments{
\item{VolumeId}{[required] The OpsWorks Stacks volume ID, which is the GUID that OpsWorks Stacks
assigned to the instance when you registered the volume with the stack,
not the Amazon EC2 volume ID.}
}
\description{
Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_deregister_volume/} for full documentation.
}
\keyword{internal}
