% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_apps}
\alias{opsworks_describe_apps}
\title{Requests a description of a specified set of apps}
\usage{
opsworks_describe_apps(StackId = NULL, AppIds = NULL)
}
\arguments{
\item{StackId}{The app stack ID. If you use this parameter,
\code{\link[=opsworks_describe_apps]{describe_apps}} returns a description of the
apps in the specified stack.}

\item{AppIds}{An array of app IDs for the apps to be described. If you use this
parameter, \code{\link[=opsworks_describe_apps]{describe_apps}} returns a
description of the specified apps. Otherwise, it returns a description
of every app.}
}
\description{
Requests a description of a specified set of apps.

See \url{https://paws-r.github.io/docs/opsworks/describe_apps.html} for full documentation.
}
\keyword{internal}
