% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_load_based_auto_scaling}
\alias{opsworks_describe_load_based_auto_scaling}
\title{Describes load-based auto scaling configurations for specified layers}
\usage{
opsworks_describe_load_based_auto_scaling(LayerIds)
}
\arguments{
\item{LayerIds}{[required] An array of layer IDs.}
}
\description{
Describes load-based auto scaling configurations for specified layers.
}
\details{
You must specify at least one of the parameters.

\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack, or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{svc$describe_load_based_auto_scaling(
  LayerIds = list(
    "string"
  )
)
}
}

\keyword{internal}
