% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_operating_systems}
\alias{opsworks_describe_operating_systems}
\title{Describes the operating systems that are supported by AWS OpsWorks
Stacks}
\usage{
opsworks_describe_operating_systems()
}
\description{
Describes the operating systems that are supported by AWS OpsWorks Stacks.

See \url{https://paws-r.github.io/docs/opsworks/describe_operating_systems.html} for full documentation.
}
\keyword{internal}
