% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stack_provisioning_parameters}
\alias{opsworks_describe_stack_provisioning_parameters}
\title{Requests a description of a stack's provisioning parameters}
\usage{
opsworks_describe_stack_provisioning_parameters(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AgentInstallerUrl = "string",
  Parameters = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Requests a description of a stack's provisioning parameters.

\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_stack_provisioning_parameters(
  StackId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
