% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stack_summary}
\alias{opsworks_describe_stack_summary}
\title{Describes the number of layers and apps in a specified stack, and the
number of instances in each state, such as running_setup or online}
\usage{
opsworks_describe_stack_summary(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\description{
Describes the number of layers and apps in a specified stack, and the number of instances in each state, such as \code{running_setup} or \code{online}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_stack_summary/} for full documentation.
}
\keyword{internal}
