% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_time_based_auto_scaling}
\alias{opsworks_describe_time_based_auto_scaling}
\title{Describes time-based auto scaling configurations for specified instances}
\usage{
opsworks_describe_time_based_auto_scaling(InstanceIds)
}
\arguments{
\item{InstanceIds}{[required] An array of instance IDs.}
}
\description{
Describes time-based auto scaling configurations for specified instances.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_time_based_auto_scaling/} for full documentation.
}
\keyword{internal}
