% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_time_based_auto_scaling}
\alias{opsworks_describe_time_based_auto_scaling}
\title{Describes time-based auto scaling configurations for specified instances}
\usage{
opsworks_describe_time_based_auto_scaling(InstanceIds)
}
\arguments{
\item{InstanceIds}{[required] An array of instance IDs.}
}
\description{
Describes time-based auto scaling configurations for specified
instances.
}
\details{
You must specify at least one of the parameters.

\strong{Required Permissions}: To use this action, an IAM user must have a
Show, Deploy, or Manage permissions level for the stack, or an attached
policy that explicitly grants permissions. For more information about
user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{svc$describe_time_based_auto_scaling(
  InstanceIds = list(
    "string"
  )
)
}
}

\keyword{internal}
