% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_user_profiles}
\alias{opsworks_describe_user_profiles}
\title{Describe specified users}
\usage{
opsworks_describe_user_profiles(IamUserArns = NULL)
}
\arguments{
\item{IamUserArns}{An array of IAM or federated user ARNs that identify the users to be
described.}
}
\description{
Describe specified users.

See \url{https://www.paws-r-sdk.com/docs/opsworks_describe_user_profiles/} for full documentation.
}
\keyword{internal}
