% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_get_hostname_suggestion}
\alias{opsworks_get_hostname_suggestion}
\title{Gets a generated host name for the specified layer, based on the current
host name theme}
\usage{
opsworks_get_hostname_suggestion(LayerId)
}
\arguments{
\item{LayerId}{[required] The layer ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LayerId = "string",
  Hostname = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a generated host name for the specified layer, based on the current
host name theme.

\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack, or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_hostname_suggestion(
  LayerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
