% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_register_elastic_ip}
\alias{opsworks_register_elastic_ip}
\title{Registers an Elastic IP address with a specified stack}
\usage{
opsworks_register_elastic_ip(ElasticIp, StackId)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}

\item{StackId}{[required] The stack ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ElasticIp = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Registers an Elastic IP address with a specified stack. An address can
be registered with only one stack at a time. If the address is already
registered, you must first deregister it by calling
\code{\link[=opsworks_deregister_elastic_ip]{deregister_elastic_ip}}. For more
information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

\strong{Required Permissions}: To use this action, an IAM user must have a
Manage permissions level for the stack, or an attached policy that
explicitly grants permissions. For more information on user permissions,
see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_elastic_ip(
  ElasticIp = "string",
  StackId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
