% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_start_instance}
\alias{opsworks_start_instance}
\title{Starts a specified instance}
\usage{
opsworks_start_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}
}
\description{
Starts a specified instance. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-starting.html}{Starting, Stopping, and Rebooting Instances}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_start_instance/} for full documentation.
}
\keyword{internal}
