% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_unassign_volume}
\alias{opsworks_unassign_volume}
\title{Unassigns an assigned Amazon EBS volume}
\usage{
opsworks_unassign_volume(VolumeId)
}
\arguments{
\item{VolumeId}{[required] The volume ID.}
}
\description{
Unassigns an assigned Amazon EBS volume. The volume remains registered with the stack. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_unassign_volume/} for full documentation.
}
\keyword{internal}
