% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_untag_resource}
\alias{opsworks_untag_resource}
\title{Removes tags from a specified stack or layer}
\usage{
opsworks_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The stack or layer's Amazon Resource Number (ARN).}

\item{TagKeys}{[required] A list of the keys of tags to be removed from a stack or layer.}
}
\value{
An empty list.
}
\description{
Removes tags from a specified stack or layer.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
