% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_update_elastic_ip}
\alias{opsworks_update_elastic_ip}
\title{Updates a registered Elastic IP address's name}
\usage{
opsworks_update_elastic_ip(ElasticIp, Name = NULL)
}
\arguments{
\item{ElasticIp}{[required] The IP address for which you want to update the name.}

\item{Name}{The new name, which can be a maximum of 32 characters.}
}
\description{
Updates a registered Elastic IP address's name. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://www.paws-r-sdk.com/docs/opsworks_update_elastic_ip/} for full documentation.
}
\keyword{internal}
