% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_update_my_user_profile}
\alias{opsworks_update_my_user_profile}
\title{Updates a user's SSH public key}
\usage{
opsworks_update_my_user_profile(SshPublicKey)
}
\arguments{
\item{SshPublicKey}{The user's SSH public key.}
}
\value{
An empty list.
}
\description{
Updates a user's SSH public key.

\strong{Required Permissions}: To use this action, an IAM user must have
self-management enabled or an attached policy that explicitly grants
permissions. For more information about user permissions, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\section{Request syntax}{
\preformatted{svc$update_my_user_profile(
  SshPublicKey = "string"
)
}
}

\keyword{internal}
