% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_delete_backup}
\alias{opsworkscm_delete_backup}
\title{Deletes a backup}
\usage{
opsworkscm_delete_backup(BackupId)
}
\arguments{
\item{BackupId}{[required] The ID of the backup to delete. Run the DescribeBackups command to get a
list of backup IDs. Backup IDs are in the format
\code{ServerName-yyyyMMddHHmmssSSS}.}
}
\description{
Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous.

See \url{https://www.paws-r-sdk.com/docs/opsworkscm_delete_backup/} for full documentation.
}
\keyword{internal}
