% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_describe_account_attributes}
\alias{opsworkscm_describe_account_attributes}
\title{Describes your OpsWorks-CM account attributes}
\usage{
opsworkscm_describe_account_attributes()
}
\description{
Describes your OpsWorks-CM account attributes.

This operation is synchronous.
}
\section{Request syntax}{
\preformatted{svc$describe_account_attributes()
}
}

\keyword{internal}
