% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_describe_backups}
\alias{opsworkscm_describe_backups}
\title{Describes backups}
\usage{
opsworkscm_describe_backups(
  BackupId = NULL,
  ServerName = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{BackupId}{Describes a single backup.}

\item{ServerName}{Returns backups for the server with the specified ServerName.}

\item{NextToken}{This is not currently implemented for
\code{\link[=opsworkscm_describe_backups]{describe_backups}} requests.}

\item{MaxResults}{This is not currently implemented for
\code{\link[=opsworkscm_describe_backups]{describe_backups}} requests.}
}
\description{
Describes backups. The results are ordered by time, with newest backups first. If you do not specify a BackupId or ServerName, the command returns all backups.

See \url{https://www.paws-r-sdk.com/docs/opsworkscm_describe_backups/} for full documentation.
}
\keyword{internal}
