% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_accept_handshake}
\alias{organizations_accept_handshake}
\title{Accepts a handshake by sending an ACCEPTED response to the sender}
\usage{
organizations_accept_handshake(HandshakeId)
}
\arguments{
\item{HandshakeId}{[required] ID for the handshake that you want to accept.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for handshake
ID string requires "h-" followed by from 8 to 32 lowercase letters or
digits.}
}
\description{
Accepts a handshake by sending an \code{ACCEPTED} response to the sender. You can view accepted handshakes in API responses for 30 days before they are deleted.

See \url{https://www.paws-r-sdk.com/docs/organizations_accept_handshake/} for full documentation.
}
\keyword{internal}
