% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_delete_policy}
\alias{organizations_delete_policy}
\title{Deletes the specified policy from your organization}
\usage{
organizations_delete_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] ID for the policy that you want to delete. You can get the ID from the
\code{\link[=organizations_list_policies]{list_policies}} or
\code{\link[=organizations_list_policies_for_target]{list_policies_for_target}}
operations.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for a policy ID
string requires "p-" followed by from 8 to 128 lowercase or uppercase
letters, digits, or the underscore character (_).}
}
\description{
Deletes the specified policy from your organization. Before you perform this operation, you must first detach the policy from all organizational units (OUs), roots, and accounts.

See \url{https://www.paws-r-sdk.com/docs/organizations_delete_policy/} for full documentation.
}
\keyword{internal}
