% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_effective_policy}
\alias{organizations_describe_effective_policy}
\title{Returns the contents of the effective policy for specified policy type
and account}
\usage{
organizations_describe_effective_policy(PolicyType, TargetId = NULL)
}
\arguments{
\item{PolicyType}{[required] The type of policy that you want information about. You can specify one
of the following values:
\itemize{
\item \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_ai-opt-out.html}{AISERVICES_OPT_OUT_POLICY}
\item \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_backup.html}{BACKUP_POLICY}
\item \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_tag-policies.html}{TAG_POLICY}
}}

\item{TargetId}{When you're signed in as the management account, specify the ID of the
account that you want details about. Specifying an organization root or
organizational unit (OU) as the target is not supported.}
}
\description{
Returns the contents of the effective policy for specified policy type and account. The effective policy is the aggregation of any policies of the specified type that the account inherits, plus any policy of that type that is directly attached to the account.

See \url{https://www.paws-r-sdk.com/docs/organizations_describe_effective_policy/} for full documentation.
}
\keyword{internal}
