% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_policy}
\alias{organizations_describe_policy}
\title{Retrieves information about a policy}
\usage{
organizations_describe_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] ID for the policy that you want details about. You can get the ID from
the \code{\link[=organizations_list_policies]{list_policies}} or
\code{\link[=organizations_list_policies_for_target]{list_policies_for_target}}
operations.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for a policy ID
string requires "p-" followed by from 8 to 128 lowercase or uppercase
letters, digits, or the underscore character (_).}
}
\description{
Retrieves information about a policy.

See \url{https://www.paws-r-sdk.com/docs/organizations_describe_policy/} for full documentation.
}
\keyword{internal}
