% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_accounts}
\alias{organizations_list_accounts}
\title{Lists all the accounts in the organization}
\usage{
organizations_list_accounts(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists all the accounts in the organization. To request only the accounts in a specified root or organizational unit (OU), use the \code{\link[=organizations_list_accounts_for_parent]{list_accounts_for_parent}} operation instead.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_accounts/} for full documentation.
}
\keyword{internal}
