% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_aws_service_access_for_organization}
\alias{organizations_list_aws_service_access_for_organization}
\title{Returns a list of the Amazon Web Services services that you enabled to
integrate with your organization}
\usage{
organizations_list_aws_service_access_for_organization(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Returns a list of the Amazon Web Services services that you enabled to integrate with your organization. After a service on this list creates the resources that it requires for the integration, it can perform operations on your organization and its accounts.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_aws_service_access_for_organization/} for full documentation.
}
\keyword{internal}
