% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_children}
\alias{organizations_list_children}
\title{Lists all of the organizational units (OUs) or accounts that are
contained in the specified parent OU or root}
\usage{
organizations_list_children(
  ParentId,
  ChildType,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ParentId}{[required] The unique identifier (ID) for the parent root or OU whose children you
want to list.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for a parent ID
string requires one of the following:
\itemize{
\item \strong{Root} - A string that begins with "r-" followed by from 4 to 32
lowercase letters or digits.
\item \strong{Organizational unit (OU)} - A string that begins with "ou-"
followed by from 4 to 32 lowercase letters or digits (the ID of the
root that the OU is in). This string is followed by a second "-"
dash and from 8 to 32 additional lowercase letters or digits.
}}

\item{ChildType}{[required] Filters the output to include only the specified child type.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists all of the organizational units (OUs) or accounts that are contained in the specified parent OU or root. This operation, along with \code{\link[=organizations_list_parents]{list_parents}} enables you to traverse the tree structure that makes up this root.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_children/} for full documentation.
}
\keyword{internal}
