% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_create_account_status}
\alias{organizations_list_create_account_status}
\title{Lists the account creation requests that match the specified status that
is currently being tracked for the organization}
\usage{
organizations_list_create_account_status(
  States = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{States}{A list of one or more states that you want included in the response. If
this parameter isn't present, all requests are included in the response.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists the account creation requests that match the specified status that is currently being tracked for the organization.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_create_account_status/} for full documentation.
}
\keyword{internal}
