% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_handshakes_for_account}
\alias{organizations_list_handshakes_for_account}
\title{Lists the recent handshakes that you have received}
\usage{
organizations_list_handshakes_for_account(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{A \code{HandshakeFilter} object. Contains the filer used to select the
handshakes for an operation.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists the recent handshakes that you have received.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_handshakes_for_account/} for full documentation.
}
\keyword{internal}
