% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_organizational_units_for_parent}
\alias{organizations_list_organizational_units_for_parent}
\title{Lists the organizational units (OUs) in a parent organizational unit or
root}
\usage{
organizations_list_organizational_units_for_parent(
  ParentId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ParentId}{[required] ID for the root or OU whose child OUs you want to list.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} for a parent ID
string requires one of the following:
\itemize{
\item \strong{Root} - A string that begins with "r-" followed by from 4 to 32
lowercase letters or digits.
\item \strong{Organizational unit (OU)} - A string that begins with "ou-"
followed by from 4 to 32 lowercase letters or digits (the ID of the
root that the OU is in). This string is followed by a second "-"
dash and from 8 to 32 additional lowercase letters or digits.
}}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}

\item{MaxResults}{The maximum number of items to return in the response. If more results
exist than the specified \code{MaxResults} value, a token is included in the
response so that you can retrieve the remaining results.}
}
\description{
Lists the organizational units (OUs) in a parent organizational unit or root.

See \url{https://www.paws-r-sdk.com/docs/organizations_list_organizational_units_for_parent/} for full documentation.
}
\keyword{internal}
