% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_list_tags_for_resource}
\alias{organizations_list_tags_for_resource}
\title{Lists tags that are attached to the specified resource}
\usage{
organizations_list_tags_for_resource(ResourceId, NextToken = NULL)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource with the tags to list.

You can specify any of the following taggable resources.
\itemize{
\item Amazon Web Services account – specify the account ID number.
\item Organizational unit – specify the OU ID that begins with \verb{ou-} and
looks similar to: \verb{ou-1a2b-34uvwxyz }
\item Root – specify the root ID that begins with \verb{r-} and looks similar
to: \verb{r-1a2b }
\item Policy – specify the policy ID that begins with \verb{p-} andlooks
similar to: \verb{p-12abcdefg3 }
}}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from.}
}
\description{
Lists tags that are attached to the specified resource.

See \url{https://paws-r.github.io/docs/organizations/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
