% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_register_delegated_administrator}
\alias{organizations_register_delegated_administrator}
\title{Enables the specified member account to administer the Organizations
features of the specified Amazon Web Services service}
\usage{
organizations_register_delegated_administrator(AccountId, ServicePrincipal)
}
\arguments{
\item{AccountId}{[required] The account ID number of the member account in the organization to
register as a delegated administrator.}

\item{ServicePrincipal}{[required] The service principal of the Amazon Web Services service for which you
want to make the member account a delegated administrator.}
}
\description{
Enables the specified member account to administer the Organizations features of the specified Amazon Web Services service. It grants read-only access to Organizations service data. The account still requires IAM permissions to access and administer the Amazon Web Services service.

See \url{https://www.paws-r-sdk.com/docs/organizations_register_delegated_administrator/} for full documentation.
}
\keyword{internal}
