% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_create_logging_configuration}
\alias{prometheusservice_create_logging_configuration}
\title{The CreateLoggingConfiguration operation creates a logging configuration
for the workspace}
\usage{
prometheusservice_create_logging_configuration(
  clientToken = NULL,
  logGroupArn,
  workspaceId
)
}
\arguments{
\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}

\item{logGroupArn}{[required] The ARN of the CloudWatch log group to which the vended log data will be
published. This log group must exist prior to calling this operation.}

\item{workspaceId}{[required] The ID of the workspace to create the logging configuration for.}
}
\description{
The \code{\link[=prometheusservice_create_logging_configuration]{create_logging_configuration}} operation creates a logging configuration for the workspace. Use this operation to set the CloudWatch log group to which the logs will be published to.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_create_logging_configuration/} for full documentation.
}
\keyword{internal}
