% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_create_scraper}
\alias{prometheusservice_create_scraper}
\title{Create a scraper}
\usage{
prometheusservice_create_scraper(
  alias = NULL,
  scrapeConfiguration,
  source,
  destination,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{alias}{An optional user-assigned alias for this scraper. This alias is for user
reference and does not need to be unique.}

\item{scrapeConfiguration}{[required] The configuration used to create the scraper.}

\item{source}{[required] The source that the scraper will be discovering and collecting metrics
from.}

\item{destination}{[required] The destination that the scraper will be producing metrics to.}

\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{tags}{Optional, user-provided tags for this scraper.}
}
\description{
Create a scraper.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_create_scraper/} for full documentation.
}
\keyword{internal}
