% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_alert_manager_definition}
\alias{prometheusservice_delete_alert_manager_definition}
\title{Deletes an alert manager definition}
\usage{
prometheusservice_delete_alert_manager_definition(
  clientToken = NULL,
  workspaceId
)
}
\arguments{
\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{workspaceId}{[required] The ID of the workspace in which to delete the alert manager definition.}
}
\description{
Deletes an alert manager definition.

See \url{https://paws-r.github.io/docs/prometheusservice/delete_alert_manager_definition.html} for full documentation.
}
\keyword{internal}
