% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_alert_manager_definition}
\alias{prometheusservice_delete_alert_manager_definition}
\title{Deletes the alert manager definition from a workspace}
\usage{
prometheusservice_delete_alert_manager_definition(
  workspaceId,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to delete the alert manager definition from.}

\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}
}
\description{
Deletes the alert manager definition from a workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_alert_manager_definition/} for full documentation.
}
\keyword{internal}
