% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_logging_configuration}
\alias{prometheusservice_delete_logging_configuration}
\title{Delete logging configuration}
\usage{
prometheusservice_delete_logging_configuration(clientToken = NULL, workspaceId)
}
\arguments{
\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{workspaceId}{[required] The ID of the workspace to vend logs to.}
}
\description{
Delete logging configuration.

See \url{https://paws-r.github.io/docs/prometheusservice/delete_logging_configuration.html} for full documentation.
}
\keyword{internal}
