% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_rule_groups_namespace}
\alias{prometheusservice_delete_rule_groups_namespace}
\title{Delete a rule groups namespace}
\usage{
prometheusservice_delete_rule_groups_namespace(
  clientToken = NULL,
  name,
  workspaceId
)
}
\arguments{
\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{name}{[required] The rule groups namespace name.}

\item{workspaceId}{[required] The ID of the workspace to delete rule group definition.}
}
\description{
Delete a rule groups namespace.

See \url{https://paws-r.github.io/docs/prometheusservice/delete_rule_groups_namespace.html} for full documentation.
}
\keyword{internal}
