% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_rule_groups_namespace}
\alias{prometheusservice_delete_rule_groups_namespace}
\title{Deletes one rule groups namespace and its associated rule groups
definition}
\usage{
prometheusservice_delete_rule_groups_namespace(
  workspaceId,
  name,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace containing the rule groups namespace and
definition to delete.}

\item{name}{[required] The name of the rule groups namespace to delete.}

\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}
}
\description{
Deletes one rule groups namespace and its associated rule groups definition.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_rule_groups_namespace/} for full documentation.
}
\keyword{internal}
