% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_workspace}
\alias{prometheusservice_delete_workspace}
\title{Deletes an AMP workspace}
\usage{
prometheusservice_delete_workspace(clientToken = NULL, workspaceId)
}
\arguments{
\item{clientToken}{Optional, unique, case-sensitive, user-provided identifier to ensure the
idempotency of the request.}

\item{workspaceId}{[required] The ID of the workspace to delete.}
}
\description{
Deletes an AMP workspace.

See \url{https://paws-r.github.io/docs/prometheusservice/delete_workspace.html} for full documentation.
}
\keyword{internal}
