% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_delete_workspace}
\alias{prometheusservice_delete_workspace}
\title{Deletes an existing workspace}
\usage{
prometheusservice_delete_workspace(workspaceId, clientToken = NULL)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to delete.}

\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}
}
\description{
Deletes an existing workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_delete_workspace/} for full documentation.
}
\keyword{internal}
