% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_alert_manager_definition}
\alias{prometheusservice_describe_alert_manager_definition}
\title{Describes an alert manager definition}
\usage{
prometheusservice_describe_alert_manager_definition(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to describe.}
}
\description{
Describes an alert manager definition.

See \url{https://paws-r.github.io/docs/prometheusservice/describe_alert_manager_definition.html} for full documentation.
}
\keyword{internal}
