% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_describe_logging_configuration}
\alias{prometheusservice_describe_logging_configuration}
\title{Returns complete information about the current rules and alerting
logging configuration of the workspace}
\usage{
prometheusservice_describe_logging_configuration(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to describe the logging configuration for.}
}
\description{
Returns complete information about the current rules and alerting logging configuration of the workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_describe_logging_configuration/} for full documentation.
}
\keyword{internal}
