% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_get_default_scraper_configuration}
\alias{prometheusservice_get_default_scraper_configuration}
\title{Gets a default configuration}
\usage{
prometheusservice_get_default_scraper_configuration()
}
\description{
Gets a default configuration.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_get_default_scraper_configuration/} for full documentation.
}
\keyword{internal}
