% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_list_tags_for_resource}
\alias{prometheusservice_list_tags_for_resource}
\title{Lists the tags you have assigned to the resource}
\usage{
prometheusservice_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}
}
\description{
Lists the tags you have assigned to the resource.

See \url{https://paws-r.github.io/docs/prometheusservice/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
