% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_tag_resource}
\alias{prometheusservice_tag_resource}
\title{Creates tags for the specified resource}
\usage{
prometheusservice_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required]}
}
\description{
Creates tags for the specified resource.

See \url{https://paws-r.github.io/docs/prometheusservice/tag_resource.html} for full documentation.
}
\keyword{internal}
