% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_untag_resource}
\alias{prometheusservice_untag_resource}
\title{Deletes tags from the specified resource}
\usage{
prometheusservice_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] One or more tag keys}
}
\description{
Deletes tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_untag_resource/} for full documentation.
}
\keyword{internal}
