% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_logging_configuration}
\alias{prometheusservice_update_logging_configuration}
\title{Updates the log group ARN or the workspace ID of the current logging
configuration}
\usage{
prometheusservice_update_logging_configuration(
  clientToken = NULL,
  logGroupArn,
  workspaceId
)
}
\arguments{
\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}

\item{logGroupArn}{[required] The ARN of the CloudWatch log group to which the vended log data will be
published.}

\item{workspaceId}{[required] The ID of the workspace to update the logging configuration for.}
}
\description{
Updates the log group ARN or the workspace ID of the current logging configuration.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_logging_configuration/} for full documentation.
}
\keyword{internal}
