% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusservice_operations.R
\name{prometheusservice_update_workspace_alias}
\alias{prometheusservice_update_workspace_alias}
\title{Updates the alias of an existing workspace}
\usage{
prometheusservice_update_workspace_alias(
  workspaceId,
  alias = NULL,
  clientToken = NULL
)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to update.}

\item{alias}{The new alias for the workspace. It does not need to be unique.

Amazon Managed Service for Prometheus will automatically strip any blank
spaces from the beginning and end of the alias that you specify.}

\item{clientToken}{A unique identifier that you can provide to ensure the idempotency of
the request. Case-sensitive.}
}
\description{
Updates the alias of an existing workspace.

See \url{https://www.paws-r-sdk.com/docs/prometheusservice_update_workspace_alias/} for full documentation.
}
\keyword{internal}
