% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_batch_update_recommendation_status}
\alias{resiliencehub_batch_update_recommendation_status}
\title{Enables you to include or exclude one or more operational
recommendations}
\usage{
resiliencehub_batch_update_recommendation_status(appArn, requestEntries)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}

\item{requestEntries}{[required] Defines the list of operational recommendations that need to be included
or excluded.}
}
\description{
Enables you to include or exclude one or more operational recommendations.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_batch_update_recommendation_status/} for full documentation.
}
\keyword{internal}
