% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_draft_app_version_resources_import_status}
\alias{resiliencehub_describe_draft_app_version_resources_import_status}
\title{Describes the status of importing resources to an application version}
\usage{
resiliencehub_describe_draft_app_version_resources_import_status(appArn)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}
}
\description{
Describes the status of importing resources to an application version.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_describe_draft_app_version_resources_import_status/} for full documentation.
}
\keyword{internal}
