% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_app_assessment}
\alias{resiliencehub_describe_app_assessment}
\title{Describes an assessment for an Resilience Hub application}
\usage{
resiliencehub_describe_app_assessment(assessmentArn)
}
\arguments{
\item{assessmentArn}{[required] Amazon Resource Name (ARN) of the assessment. The format for this ARN
is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app-assessment/\code{app-id}.
For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}
}
\description{
Describes an assessment for an Resilience Hub application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_describe_app_assessment/} for full documentation.
}
\keyword{internal}
