% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_describe_app_version_resource}
\alias{resiliencehub_describe_app_version_resource}
\title{Describes a resource of the Resilience Hub application}
\usage{
resiliencehub_describe_app_version_resource(
  appArn,
  appVersion,
  awsAccountId = NULL,
  awsRegion = NULL,
  logicalResourceId = NULL,
  physicalResourceId = NULL,
  resourceName = NULL
)
}
\arguments{
\item{appArn}{[required] The Amazon Resource Name (ARN) of the Resilience Hub application. The
format for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}

\item{appVersion}{[required] The Resilience Hub application version.}

\item{awsAccountId}{The Amazon Web Services account that owns the physical resource.}

\item{awsRegion}{The Amazon Web Services region that owns the physical resource.}

\item{logicalResourceId}{The logical identifier of the resource.}

\item{physicalResourceId}{The physical identifier of the resource.}

\item{resourceName}{The name of the resource.}
}
\description{
Describes a resource of the Resilience Hub application.

See \url{https://paws-r.github.io/docs/resiliencehub/describe_app_version_resource.html} for full documentation.
}
\keyword{internal}
