% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_import_resources_to_draft_app_version}
\alias{resiliencehub_import_resources_to_draft_app_version}
\title{Imports resources from sources such as a CloudFormation stack,
resource-groups, or application registry app to a draft application
version}
\usage{
resiliencehub_import_resources_to_draft_app_version(
  appArn,
  sourceArns = NULL,
  terraformSources = NULL
)
}
\arguments{
\item{appArn}{[required] The Amazon Resource Name (ARN) of the application. The format for this
ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}.
For more information about ARNs, see \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference}.}

\item{sourceArns}{The Amazon Resource Names (ARNs) for the resources that you want to
import.}

\item{terraformSources}{A list of terraform file s3 URLs you need to import.}
}
\description{
Imports resources from sources such as a CloudFormation stack, resource-groups, or application registry app to a draft application version.

See \url{https://paws-r.github.io/docs/resiliencehub/import_resources_to_draft_app_version.html} for full documentation.
}
\keyword{internal}
