% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_list_apps}
\alias{resiliencehub_list_apps}
\title{Lists your Resilience Hub applications}
\usage{
resiliencehub_list_apps(
  appArn = NULL,
  maxResults = NULL,
  name = NULL,
  nextToken = NULL
)
}
\arguments{
\item{appArn}{The Amazon Resource Name (ARN) of the application. The format for this
ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}.
For more information about ARNs, see \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference}.}

\item{maxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.}

\item{name}{The name for the one of the listed applications.}

\item{nextToken}{Null, or the token from a previous call to get the next set of results.}
}
\description{
Lists your Resilience Hub applications.

See \url{https://paws-r.github.io/docs/resiliencehub/list_apps.html} for full documentation.
}
\keyword{internal}
