% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_list_resource_grouping_recommendations}
\alias{resiliencehub_list_resource_grouping_recommendations}
\title{Lists the resource grouping recommendations suggested by Resilience Hub
for your application}
\usage{
resiliencehub_list_resource_grouping_recommendations(
  appArn = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{appArn}{Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference} guide.}

\item{maxResults}{Maximum number of grouping recommendations to be displayed per
Resilience Hub application.}

\item{nextToken}{Null, or the token from a previous call to get the next set of results.}
}
\description{
Lists the resource grouping recommendations suggested by Resilience Hub for your application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_list_resource_grouping_recommendations/} for full documentation.
}
\keyword{internal}
