% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_publish_app_version}
\alias{resiliencehub_publish_app_version}
\title{Publishes a new version of a specific Resilience Hub application}
\usage{
resiliencehub_publish_app_version(appArn)
}
\arguments{
\item{appArn}{[required] The Amazon Resource Name (ARN) of the application. The format for this
ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}.
For more information about ARNs, see \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference}.}
}
\description{
Publishes a new version of a specific Resilience Hub application.

See \url{https://paws-r.github.io/docs/resiliencehub/publish_app_version.html} for full documentation.
}
\keyword{internal}
