% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_reject_resource_grouping_recommendations}
\alias{resiliencehub_reject_resource_grouping_recommendations}
\title{Rejects resource grouping recommendations}
\usage{
resiliencehub_reject_resource_grouping_recommendations(appArn, entries)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference} guide.}

\item{entries}{[required] List of resource grouping recommendations you have selected to exclude
from your application.}
}
\description{
Rejects resource grouping recommendations.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_reject_resource_grouping_recommendations/} for full documentation.
}
\keyword{internal}
