% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_remove_draft_app_version_resource_mappings}
\alias{resiliencehub_remove_draft_app_version_resource_mappings}
\title{Removes resource mappings from a draft application version}
\usage{
resiliencehub_remove_draft_app_version_resource_mappings(
  appArn,
  appRegistryAppNames = NULL,
  logicalStackNames = NULL,
  resourceGroupNames = NULL,
  resourceNames = NULL,
  terraformSourceNames = NULL
)
}
\arguments{
\item{appArn}{[required] The Amazon Resource Name (ARN) of the application. The format for this
ARN is: arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}.
For more information about ARNs, see \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference}.}

\item{appRegistryAppNames}{The names of the registered applications to remove from the resource
mappings.}

\item{logicalStackNames}{The names of the CloudFormation stacks to remove from the resource
mappings.}

\item{resourceGroupNames}{The names of the resource groups to remove from the resource mappings.}

\item{resourceNames}{The names of the resources to remove from the resource mappings.}

\item{terraformSourceNames}{\if{html}{\out{
</p>
}}}
}
\description{
Removes resource mappings from a draft application version.

See \url{https://paws-r.github.io/docs/resiliencehub/remove_draft_app_version_resource_mappings.html} for full documentation.
}
\keyword{internal}
