% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_resolve_app_version_resources}
\alias{resiliencehub_resolve_app_version_resources}
\title{Resolves the resources for an application version}
\usage{
resiliencehub_resolve_app_version_resources(appArn, appVersion)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}

\item{appVersion}{[required] The version of the application.}
}
\description{
Resolves the resources for an application version.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_resolve_app_version_resources/} for full documentation.
}
\keyword{internal}
