% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_start_resource_grouping_recommendation_task}
\alias{resiliencehub_start_resource_grouping_recommendation_task}
\title{Starts grouping recommendation task}
\usage{
resiliencehub_start_resource_grouping_recommendation_task(appArn)
}
\arguments{
\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference} guide.}
}
\description{
Starts grouping recommendation task.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_start_resource_grouping_recommendation_task/} for full documentation.
}
\keyword{internal}
