% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_update_app_version_app_component}
\alias{resiliencehub_update_app_version_app_component}
\title{Updates an existing Application Component in the Resilience Hub
application}
\usage{
resiliencehub_update_app_version_app_component(
  additionalInfo = NULL,
  appArn,
  id,
  name = NULL,
  type = NULL
)
}
\arguments{
\item{additionalInfo}{Currently, there is no supported additional information for Application
Components.}

\item{appArn}{[required] Amazon Resource Name (ARN) of the Resilience Hub application. The format
for this ARN is:
arn:\code{partition}:resiliencehub:\code{region}:\code{account}:app/\code{app-id}. For more
information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{AWS General Reference} guide.}

\item{id}{[required] Identifier of the Application Component.}

\item{name}{Name of the Application Component.}

\item{type}{Type of Application Component. For more information about the types of
Application Component, see \href{https://docs.aws.amazon.com/resilience-hub/latest/userguide/AppComponent.grouping.html}{Grouping resources in an AppComponent}.}
}
\description{
Updates an existing Application Component in the Resilience Hub application.

See \url{https://www.paws-r-sdk.com/docs/resiliencehub_update_app_version_app_component/} for full documentation.
}
\keyword{internal}
