% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_delete_group}
\alias{resourcegroups_delete_group}
\title{Deletes the specified resource group}
\usage{
resourcegroups_delete_group(GroupName, Group)
}
\arguments{
\item{GroupName}{Deprecated - don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to delete.}
}
\description{
Deletes the specified resource group. Deleting a resource group does not
delete any resources that are members of the group; it only deletes the
group structure.
}
\details{
\strong{Minimum permissions}

To run this command, you must have the following permissions:
\itemize{
\item \code{resource-groups:DeleteGroup}
}
}
\section{Request syntax}{
\preformatted{svc$delete_group(
  GroupName = "string",
  Group = "string"
)
}
}

\keyword{internal}
