% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_delete_group}
\alias{resourcegroups_delete_group}
\title{Deletes the specified resource group}
\usage{
resourcegroups_delete_group(GroupName = NULL, Group = NULL)
}
\arguments{
\item{GroupName}{Deprecated - don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to delete.}
}
\description{
Deletes the specified resource group. Deleting a resource group does not delete any resources that are members of the group; it only deletes the group structure.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_delete_group/} for full documentation.
}
\keyword{internal}
