% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_account_settings}
\alias{resourcegroups_get_account_settings}
\title{Retrieves the current status of optional features in Resource Groups}
\usage{
resourcegroups_get_account_settings()
}
\description{
Retrieves the current status of optional features in Resource Groups.

See \url{https://paws-r.github.io/docs/resourcegroups/get_account_settings.html} for full documentation.
}
\keyword{internal}
