% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group}
\alias{resourcegroups_get_group}
\title{Returns information about a specified resource group}
\usage{
resourcegroups_get_group(GroupName, Group)
}
\arguments{
\item{GroupName}{Deprecated - don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to retrieve.}
}
\description{
Returns information about a specified resource group.
}
\details{
\strong{Minimum permissions}

To run this command, you must have the following permissions:
\itemize{
\item \code{resource-groups:GetGroup}
}
}
\section{Request syntax}{
\preformatted{svc$get_group(
  GroupName = "string",
  Group = "string"
)
}
}

\keyword{internal}
