% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group_configuration}
\alias{resourcegroups_get_group_configuration}
\title{Returns the service configuration associated with the specified resource
group}
\usage{
resourcegroups_get_group_configuration(Group = NULL)
}
\arguments{
\item{Group}{The name or the ARN of the resource group.}
}
\description{
Returns the service configuration associated with the specified resource group. For details about the service configuration syntax, see \href{https://docs.aws.amazon.com/ARG/latest/userguide/about-slg.html}{Service configurations for resource groups}.

See \url{https://paws-r.github.io/docs/resourcegroups/get_group_configuration.html} for full documentation.
}
\keyword{internal}
