% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_tags}
\alias{resourcegroups_get_tags}
\title{Returns a list of tags that are associated with a resource group,
specified by an ARN}
\usage{
resourcegroups_get_tags(Arn)
}
\arguments{
\item{Arn}{[required] The ARN of the resource group whose tags you want to retrieve.}
}
\description{
Returns a list of tags that are associated with a resource group, specified by an ARN.

See \url{https://paws-r.github.io/docs/resourcegroups/get_tags.html} for full documentation.
}
\keyword{internal}
