% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_group_resources}
\alias{resourcegroups_group_resources}
\title{Adds the specified resources to the specified group}
\usage{
resourcegroups_group_resources(Group, ResourceArns)
}
\arguments{
\item{Group}{[required] The name or the ARN of the resource group to add resources to.}

\item{ResourceArns}{[required] The list of ARNs of the resources to be added to the group.}
}
\description{
Adds the specified resources to the specified group.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_group_resources/} for full documentation.
}
\keyword{internal}
