% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_start_tag_sync_task}
\alias{resourcegroups_start_tag_sync_task}
\title{Creates a new tag-sync task to onboard and sync resources tagged with a
specific tag key-value pair to an application}
\usage{
resourcegroups_start_tag_sync_task(Group, TagKey, TagValue, RoleArn)
}
\arguments{
\item{Group}{[required] The Amazon resource name (ARN) or name of the application group for
which you want to create a tag-sync task.}

\item{TagKey}{[required] The tag key. Resources tagged with this tag key-value pair will be added
to the application. If a resource with this tag is later untagged, the
tag-sync task removes the resource from the application.}

\item{TagValue}{[required] The tag value. Resources tagged with this tag key-value pair will be
added to the application. If a resource with this tag is later untagged,
the tag-sync task removes the resource from the application.}

\item{RoleArn}{[required] The Amazon resource name (ARN) of the role assumed by the service to tag
and untag resources on your behalf.}
}
\description{
Creates a new tag-sync task to onboard and sync resources tagged with a specific tag key-value pair to an application.

See \url{https://www.paws-r-sdk.com/docs/resourcegroups_start_tag_sync_task/} for full documentation.
}
\keyword{internal}
