% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_update_group}
\alias{resourcegroups_update_group}
\title{Updates the description for an existing group}
\usage{
resourcegroups_update_group(GroupName, Group, Description)
}
\arguments{
\item{GroupName}{Don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to modify.}

\item{Description}{The new description that you want to update the resource group with.
Descriptions can contain letters, numbers, hyphens, underscores,
periods, and spaces.}
}
\description{
Updates the description for an existing group. You cannot update the
name of a resource group.

\strong{Minimum permissions}

To run this command, you must have the following permissions:
\itemize{
\item \code{resource-groups:UpdateGroup}
}
}
\section{Request syntax}{
\preformatted{svc$update_group(
  GroupName = "string",
  Group = "string",
  Description = "string"
)
}
}

\keyword{internal}
