% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_update_group_query}
\alias{resourcegroups_update_group_query}
\title{Updates the resource query of a group}
\usage{
resourcegroups_update_group_query(
  GroupName = NULL,
  Group = NULL,
  ResourceQuery
)
}
\arguments{
\item{GroupName}{Don't use this parameter. Use \code{Group} instead.}

\item{Group}{The name or the ARN of the resource group to query.}

\item{ResourceQuery}{[required] The resource query to determine which Amazon Web Services resources are
members of this resource group.

A resource group can contain either a \code{Configuration} or a
\code{ResourceQuery}, but not both.}
}
\description{
Updates the resource query of a group. For more information about resource queries, see \href{https://docs.aws.amazon.com/ARG/latest/userguide/gettingstarted-query.html#gettingstarted-query-cli-tag}{Create a tag-based group in Resource Groups}.

See \url{https://paws-r.github.io/docs/resourcegroups/update_group_query.html} for full documentation.
}
\keyword{internal}
