% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_describe_report_creation}
\alias{resourcegroupstaggingapi_describe_report_creation}
\title{Describes the status of the StartReportCreation operation}
\usage{
resourcegroupstaggingapi_describe_report_creation()
}
\description{
Describes the status of the
\code{\link[=resourcegroupstaggingapi_start_report_creation]{start_report_creation}}
operation.

You can call this operation only from the organization's master account
and from the us-east-1 Region.
}
\section{Request syntax}{
\preformatted{svc$describe_report_creation()
}
}

\keyword{internal}
