% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_get_tag_keys}
\alias{resourcegroupstaggingapi_get_tag_keys}
\title{Returns all tag keys in the specified region for the AWS account}
\usage{
resourcegroupstaggingapi_get_tag_keys(PaginationToken)
}
\arguments{
\item{PaginationToken}{A string that indicates that additional data is available. Leave this
value empty for your initial request. If the response includes a
PaginationToken, use that string for this value to request an additional
page of data.}
}
\description{
Returns all tag keys in the specified region for the AWS account.
}
\section{Request syntax}{
\preformatted{svc$get_tag_keys(
  PaginationToken = "string"
)
}
}

\keyword{internal}
