% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_get_tag_keys}
\alias{resourcegroupstaggingapi_get_tag_keys}
\title{Returns all tag keys currently in use in the specified Amazon Web
Services Region for the calling account}
\usage{
resourcegroupstaggingapi_get_tag_keys(PaginationToken = NULL)
}
\arguments{
\item{PaginationToken}{Specifies a \code{PaginationToken} response value from a previous request to
indicate that you want the next page of results. Leave this parameter
empty in your initial request.}
}
\description{
Returns all tag keys currently in use in the specified Amazon Web Services Region for the calling account.

See \url{https://www.paws-r-sdk.com/docs/resourcegroupstaggingapi_get_tag_keys/} for full documentation.
}
\keyword{internal}
