% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroupstaggingapi_operations.R
\name{resourcegroupstaggingapi_get_tag_keys}
\alias{resourcegroupstaggingapi_get_tag_keys}
\title{Returns all tag keys currently in use in the specified Region for the
calling AWS account}
\usage{
resourcegroupstaggingapi_get_tag_keys(PaginationToken)
}
\arguments{
\item{PaginationToken}{Specifies a \code{PaginationToken} response value from a previous request to
indicate that you want the next page of results. Leave this parameter
empty in your initial request.}
}
\value{
A list with the following syntax:\preformatted{list(
  PaginationToken = "string",
  TagKeys = list(
    "string"
  )
)
}
}
\description{
Returns all tag keys currently in use in the specified Region for the
calling AWS account.

This operation supports pagination, where the response can be sent in
multiple pages. You should check the \code{PaginationToken} response
parameter to determine if there are additional results available to
return. Repeat the query, passing the \code{PaginationToken} response
parameter value as an input to the next request until you recieve a
\code{null} value. A null value for \code{PaginationToken} indicates that there
are no more results waiting to be returned.
}
\section{Request syntax}{
\preformatted{svc$get_tag_keys(
  PaginationToken = "string"
)
}
}

\keyword{internal}
