% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_budget_with_resource}
\alias{servicecatalog_associate_budget_with_resource}
\title{Associates the specified budget with the specified resource}
\usage{
servicecatalog_associate_budget_with_resource(BudgetName, ResourceId)
}
\arguments{
\item{BudgetName}{[required] The name of the budget you want to associate.}

\item{ResourceId}{[required] The resource identifier. Either a portfolio-id or a product-id.}
}
\value{
An empty list.
}
\description{
Associates the specified budget with the specified resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_budget_with_resource(
  BudgetName = "string",
  ResourceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
