% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_tag_option_with_resource}
\alias{servicecatalog_associate_tag_option_with_resource}
\title{Associate the specified TagOption with the specified portfolio or
product}
\usage{
servicecatalog_associate_tag_option_with_resource(ResourceId, TagOptionId)
}
\arguments{
\item{ResourceId}{[required] The resource identifier.}

\item{TagOptionId}{[required] The TagOption identifier.}
}
\description{
Associate the specified TagOption with the specified portfolio or product.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_associate_tag_option_with_resource/} for full documentation.
}
\keyword{internal}
