% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_portfolio}
\alias{servicecatalog_create_portfolio}
\title{Creates a portfolio}
\usage{
servicecatalog_create_portfolio(
  AcceptLanguage = NULL,
  DisplayName,
  Description = NULL,
  ProviderName,
  Tags = NULL,
  IdempotencyToken
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{DisplayName}{[required] The name to use for display purposes.}

\item{Description}{The description of the portfolio.}

\item{ProviderName}{[required] The name of the portfolio provider.}

\item{Tags}{One or more tags.}

\item{IdempotencyToken}{[required] A unique identifier that you provide to ensure idempotency. If multiple
requests differ only by the idempotency token, the same response is
returned for each repeated request.}
}
\description{
Creates a portfolio.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_create_portfolio/} for full documentation.
}
\keyword{internal}
