% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_provisioning_artifact}
\alias{servicecatalog_create_provisioning_artifact}
\title{Creates a provisioning artifact (also known as a version) for the
specified product}
\usage{
servicecatalog_create_provisioning_artifact(
  AcceptLanguage = NULL,
  ProductId,
  Parameters,
  IdempotencyToken
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{Parameters}{[required] The configuration for the provisioning artifact.}

\item{IdempotencyToken}{[required] A unique identifier that you provide to ensure idempotency. If multiple
requests differ only by the idempotency token, the same response is
returned for each repeated request.}
}
\description{
Creates a provisioning artifact (also known as a version) for the specified product.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_create_provisioning_artifact/} for full documentation.
}
\keyword{internal}
