% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_tag_option}
\alias{servicecatalog_create_tag_option}
\title{Creates a TagOption}
\usage{
servicecatalog_create_tag_option(Key, Value)
}
\arguments{
\item{Key}{[required] The TagOption key.}

\item{Value}{[required] The TagOption value.}
}
\description{
Creates a TagOption.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_create_tag_option/} for full documentation.
}
\keyword{internal}
