% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_tag_option}
\alias{servicecatalog_create_tag_option}
\title{Creates a TagOption}
\usage{
servicecatalog_create_tag_option(Key, Value)
}
\arguments{
\item{Key}{[required] The TagOption key.}

\item{Value}{[required] The TagOption value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagOptionDetail = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE,
    Id = "string",
    Owner = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a TagOption.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_tag_option(
  Key = "string",
  Value = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
