% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_portfolio}
\alias{servicecatalog_delete_portfolio}
\title{Deletes the specified portfolio}
\usage{
servicecatalog_delete_portfolio(AcceptLanguage, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}
}
\description{
Deletes the specified portfolio.
}
\details{
You cannot delete a portfolio if it was shared with you or if it has
associated products, users, constraints, or shared accounts.
}
\section{Request syntax}{
\preformatted{svc$delete_portfolio(
  AcceptLanguage = "string",
  Id = "string"
)
}
}

\keyword{internal}
