% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_service_action}
\alias{servicecatalog_delete_service_action}
\title{Deletes a self-service action}
\usage{
servicecatalog_delete_service_action(Id, AcceptLanguage)
}
\arguments{
\item{Id}{[required] The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a self-service action.
}
\section{Request syntax}{
\preformatted{svc$delete_service_action(
  Id = "string",
  AcceptLanguage = "string"
)
}
}

\keyword{internal}
