% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_constraint}
\alias{servicecatalog_describe_constraint}
\title{Gets information about the specified constraint}
\usage{
servicecatalog_describe_constraint(AcceptLanguage = NULL, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The identifier of the constraint.}
}
\description{
Gets information about the specified constraint.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_describe_constraint/} for full documentation.
}
\keyword{internal}
