% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_portfolio_share_status}
\alias{servicecatalog_describe_portfolio_share_status}
\title{Gets the status of the specified portfolio share operation}
\usage{
servicecatalog_describe_portfolio_share_status(PortfolioShareToken)
}
\arguments{
\item{PortfolioShareToken}{[required] The token for the portfolio share operation. This token is returned
either by CreatePortfolioShare or by DeletePortfolioShare.}
}
\description{
Gets the status of the specified portfolio share operation. This API can only be called by the management account in the organization or by a delegated admin.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_portfolio_share_status.html} for full documentation.
}
\keyword{internal}
