% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_product}
\alias{servicecatalog_describe_product}
\title{Gets information about the specified product}
\usage{
servicecatalog_describe_product(AcceptLanguage = NULL, Id = NULL, Name = NULL)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{The product identifier.}

\item{Name}{The product name.}
}
\description{
Gets information about the specified product.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_product.html} for full documentation.
}
\keyword{internal}
