% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_provisioning_artifact}
\alias{servicecatalog_describe_provisioning_artifact}
\title{Gets information about the specified provisioning artifact (also known
as a version) for the specified product}
\usage{
servicecatalog_describe_provisioning_artifact(
  AcceptLanguage = NULL,
  ProvisioningArtifactId = NULL,
  ProductId = NULL,
  ProvisioningArtifactName = NULL,
  ProductName = NULL,
  Verbose = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisioningArtifactId}{The identifier of the provisioning artifact.}

\item{ProductId}{The product identifier.}

\item{ProvisioningArtifactName}{The provisioning artifact name.}

\item{ProductName}{The product name.}

\item{Verbose}{Indicates whether a verbose level of detail is enabled.}
}
\description{
Gets information about the specified provisioning artifact (also known as a version) for the specified product.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_provisioning_artifact.html} for full documentation.
}
\keyword{internal}
