% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_record}
\alias{servicecatalog_describe_record}
\title{Gets information about the specified request operation}
\usage{
servicecatalog_describe_record(
  AcceptLanguage = NULL,
  Id,
  PageToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The record identifier of the provisioned product. This identifier is
returned by the request operation.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Gets information about the specified request operation.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_record.html} for full documentation.
}
\keyword{internal}
