% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_service_action}
\alias{servicecatalog_describe_service_action}
\title{Describes a self-service action}
\usage{
servicecatalog_describe_service_action(Id, AcceptLanguage)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ServiceActionDetail = list(
    ServiceActionSummary = list(
      Id = "string",
      Name = "string",
      Description = "string",
      DefinitionType = "SSM_AUTOMATION"
    ),
    Definition = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a self-service action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_service_action(
  Id = "string",
  AcceptLanguage = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
