% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_service_action_execution_parameters}
\alias{servicecatalog_describe_service_action_execution_parameters}
\title{Finds the default parameters for a specific self-service action on a
specific provisioned product and returns a map of the results to the
user}
\usage{
servicecatalog_describe_service_action_execution_parameters(
  ProvisionedProductId,
  ServiceActionId,
  AcceptLanguage = NULL
)
}
\arguments{
\item{ProvisionedProductId}{[required] The identifier of the provisioned product.}

\item{ServiceActionId}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Finds the default parameters for a specific self-service action on a specific provisioned product and returns a map of the results to the user.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_service_action_execution_parameters.html} for full documentation.
}
\keyword{internal}
