% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_tag_option}
\alias{servicecatalog_describe_tag_option}
\title{Gets information about the specified TagOption}
\usage{
servicecatalog_describe_tag_option(Id)
}
\arguments{
\item{Id}{[required] The TagOption identifier.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagOptionDetail = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE,
    Id = "string",
    Owner = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the specified TagOption.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_tag_option(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
