% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_disable_aws_organizations_access}
\alias{servicecatalog_disable_aws_organizations_access}
\title{Disable portfolio sharing through AWS Organizations feature}
\usage{
servicecatalog_disable_aws_organizations_access()
}
\description{
Disable portfolio sharing through AWS Organizations feature. This
feature will not delete your current shares but it will prevent you from
creating new shares throughout your organization. Current shares will
not be in sync with your organization structure if it changes after
calling this API. This API can only be called by the master account in
the organization.
}
\details{
This API can't be invoked if there are active delegated administrators
in the organization.

Note that a delegated administrator is not authorized to invoke
\code{DisableAWSOrganizationsAccess}.
}
\section{Request syntax}{
\preformatted{svc$disable_aws_organizations_access()
}
}

\keyword{internal}
