% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_disassociate_tag_option_from_resource}
\alias{servicecatalog_disassociate_tag_option_from_resource}
\title{Disassociates the specified TagOption from the specified resource}
\usage{
servicecatalog_disassociate_tag_option_from_resource(ResourceId,
  TagOptionId)
}
\arguments{
\item{ResourceId}{[required] The resource identifier.}

\item{TagOptionId}{[required] The TagOption identifier.}
}
\value{
An empty list.
}
\description{
Disassociates the specified TagOption from the specified resource.
}
\section{Request syntax}{
\preformatted{svc$disassociate_tag_option_from_resource(
  ResourceId = "string",
  TagOptionId = "string"
)
}
}

\keyword{internal}
