% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_enable_aws_organizations_access}
\alias{servicecatalog_enable_aws_organizations_access}
\title{Enable portfolio sharing feature through Organizations}
\usage{
servicecatalog_enable_aws_organizations_access()
}
\description{
Enable portfolio sharing feature through Organizations. This API will allow Service Catalog to receive updates on your organization in order to sync your shares with the current structure. This API can only be called by the management account in the organization.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_enable_aws_organizations_access/} for full documentation.
}
\keyword{internal}
