% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_get_aws_organizations_access_status}
\alias{servicecatalog_get_aws_organizations_access_status}
\title{Get the Access Status for AWS Organization portfolio share feature}
\usage{
servicecatalog_get_aws_organizations_access_status()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccessStatus = "ENABLED"|"UNDER_CHANGE"|"DISABLED"
)
}\if{html}{\out{</div>}}
}
\description{
Get the Access Status for AWS Organization portfolio share feature. This
API can only be called by the management account in the organization or
by a delegated admin.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_aws_organizations_access_status()
}\if{html}{\out{</div>}}
}

\keyword{internal}
