% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_get_provisioned_product_outputs}
\alias{servicecatalog_get_provisioned_product_outputs}
\title{This API takes either a ProvisonedProductId or a ProvisionedProductName,
along with a list of one or more output keys, and responds with the
key/value pairs of those outputs}
\usage{
servicecatalog_get_provisioned_product_outputs(AcceptLanguage,
  ProvisionedProductId, ProvisionedProductName, OutputKeys, PageSize,
  PageToken)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisionedProductId}{The identifier of the provisioned product that you want the outputs
from.}

\item{ProvisionedProductName}{The name of the provisioned product that you want the outputs from.}

\item{OutputKeys}{The list of keys that the API should return with their values. If none
are provided, the API will return all outputs of the provisioned
product.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Outputs = list(
    list(
      OutputKey = "string",
      OutputValue = "string",
      Description = "string"
    )
  ),
  NextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
This API takes either a \code{ProvisonedProductId} or a
\code{ProvisionedProductName}, along with a list of one or more output keys,
and responds with the key/value pairs of those outputs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_provisioned_product_outputs(
  AcceptLanguage = "string",
  ProvisionedProductId = "string",
  ProvisionedProductName = "string",
  OutputKeys = list(
    "string"
  ),
  PageSize = 123,
  PageToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
