% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_constraints_for_portfolio}
\alias{servicecatalog_list_constraints_for_portfolio}
\title{Lists the constraints for the specified portfolio and product}
\usage{
servicecatalog_list_constraints_for_portfolio(
  AcceptLanguage = NULL,
  PortfolioId,
  ProductId = NULL,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{ProductId}{The product identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists the constraints for the specified portfolio and product.

See \url{https://www.paws-r-sdk.com/docs/servicecatalog_list_constraints_for_portfolio/} for full documentation.
}
\keyword{internal}
