% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_provisioning_artifacts_for_service_action}
\alias{servicecatalog_list_provisioning_artifacts_for_service_action}
\title{Lists all provisioning artifacts (also known as versions) for the
specified self-service action}
\usage{
servicecatalog_list_provisioning_artifacts_for_service_action(
  ServiceActionId,
  PageSize = NULL,
  PageToken = NULL,
  AcceptLanguage = NULL
)
}
\arguments{
\item{ServiceActionId}{[required] The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Lists all provisioning artifacts (also known as versions) for the specified self-service action.

See \url{https://paws-r.github.io/docs/servicecatalog/list_provisioning_artifacts_for_service_action.html} for full documentation.
}
\keyword{internal}
