% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_record_history}
\alias{servicecatalog_list_record_history}
\title{Lists the specified requests or all performed requests}
\usage{
servicecatalog_list_record_history(
  AcceptLanguage = NULL,
  AccessLevelFilter = NULL,
  SearchFilter = NULL,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{AccessLevelFilter}{The access level to use to obtain results. The default is \code{User}.}

\item{SearchFilter}{The search filter to scope the results.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists the specified requests or all performed requests.

See \url{https://paws-r.github.io/docs/servicecatalog/list_record_history.html} for full documentation.
}
\keyword{internal}
