% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_scan_provisioned_products}
\alias{servicecatalog_scan_provisioned_products}
\title{Lists the provisioned products that are available (not terminated)}
\usage{
servicecatalog_scan_provisioned_products(AcceptLanguage,
  AccessLevelFilter, PageSize, PageToken)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{AccessLevelFilter}{The access level to use to obtain results. The default is \code{User}.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProvisionedProducts = list(
    list(
      Name = "string",
      Arn = "string",
      Type = "string",
      Id = "string",
      Status = "AVAILABLE"|"UNDER_CHANGE"|"TAINTED"|"ERROR"|"PLAN_IN_PROGRESS",
      StatusMessage = "string",
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      IdempotencyToken = "string",
      LastRecordId = "string",
      LastProvisioningRecordId = "string",
      LastSuccessfulProvisioningRecordId = "string",
      ProductId = "string",
      ProvisioningArtifactId = "string",
      LaunchRoleArn = "string"
    )
  ),
  NextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the provisioned products that are available (not terminated).

To use additional filtering, see
\code{\link[=servicecatalog_search_provisioned_products]{search_provisioned_products}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$scan_provisioned_products(
  AcceptLanguage = "string",
  AccessLevelFilter = list(
    Key = "Account"|"Role"|"User",
    Value = "string"
  ),
  PageSize = 123,
  PageToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
