% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_search_products_as_admin}
\alias{servicecatalog_search_products_as_admin}
\title{Gets information about the products for the specified portfolio or all
products}
\usage{
servicecatalog_search_products_as_admin(
  AcceptLanguage = NULL,
  PortfolioId = NULL,
  Filters = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  PageToken = NULL,
  PageSize = NULL,
  ProductSource = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{The portfolio identifier.}

\item{Filters}{The search filters. If no search filters are specified, the output
includes all products to which the administrator has access.}

\item{SortBy}{The sort field. If no value is specified, the results are not sorted.}

\item{SortOrder}{The sort order. If no value is specified, the results are not sorted.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{ProductSource}{Access level of the source of the product.}
}
\description{
Gets information about the products for the specified portfolio or all products.

See \url{https://paws-r.github.io/docs/servicecatalog/search_products_as_admin.html} for full documentation.
}
\keyword{internal}
