% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_terminate_provisioned_product}
\alias{servicecatalog_terminate_provisioned_product}
\title{Terminates the specified provisioned product}
\usage{
servicecatalog_terminate_provisioned_product(ProvisionedProductName,
  ProvisionedProductId, TerminateToken, IgnoreErrors, AcceptLanguage,
  RetainPhysicalResources)
}
\arguments{
\item{ProvisionedProductName}{The name of the provisioned product. You cannot specify both
\code{ProvisionedProductName} and \code{ProvisionedProductId}.}

\item{ProvisionedProductId}{The identifier of the provisioned product. You cannot specify both
\code{ProvisionedProductName} and \code{ProvisionedProductId}.}

\item{TerminateToken}{[required] An idempotency token that uniquely identifies the termination request.
This token is only valid during the termination process. After the
provisioned product is terminated, subsequent requests to terminate the
same provisioned product always return \strong{ResourceNotFound}.}

\item{IgnoreErrors}{If set to true, AWS Service Catalog stops managing the specified
provisioned product even if it cannot delete the underlying resources.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{RetainPhysicalResources}{When this boolean parameter is set to true, the
\code{\link[=servicecatalog_terminate_provisioned_product]{terminate_provisioned_product}}
API deletes the Service Catalog provisioned product. However, it does
not remove the CloudFormation stack, stack set, or the underlying
resources of the deleted provisioned product. The default value is
false.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RecordDetail = list(
    RecordId = "string",
    ProvisionedProductName = "string",
    Status = "CREATED"|"IN_PROGRESS"|"IN_PROGRESS_IN_ERROR"|"SUCCEEDED"|"FAILED",
    CreatedTime = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTime = as.POSIXct(
      "2015-01-01"
    ),
    ProvisionedProductType = "string",
    RecordType = "string",
    ProvisionedProductId = "string",
    ProductId = "string",
    ProvisioningArtifactId = "string",
    PathId = "string",
    RecordErrors = list(
      list(
        Code = "string",
        Description = "string"
      )
    ),
    RecordTags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    LaunchRoleArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Terminates the specified provisioned product.

This operation does not delete any records associated with the
provisioned product.

You can check the status of this request using
\code{\link[=servicecatalog_describe_record]{describe_record}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$terminate_provisioned_product(
  ProvisionedProductName = "string",
  ProvisionedProductId = "string",
  TerminateToken = "string",
  IgnoreErrors = TRUE|FALSE,
  AcceptLanguage = "string",
  RetainPhysicalResources = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
