% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_update_service_action}
\alias{servicecatalog_update_service_action}
\title{Updates a self-service action}
\usage{
servicecatalog_update_service_action(
  Id,
  Name = NULL,
  Definition = NULL,
  Description = NULL,
  AcceptLanguage = NULL
)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{Name}{The self-service action name.}

\item{Definition}{A map that defines the self-service action.}

\item{Description}{The self-service action description.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Updates a self-service action.

See \url{https://paws-r.github.io/docs/servicecatalog/update_service_action.html} for full documentation.
}
\keyword{internal}
