% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_requested_service_quota_change}
\alias{servicequotas_get_requested_service_quota_change}
\title{Retrieves information about the specified quota increase request}
\usage{
servicequotas_get_requested_service_quota_change(RequestId)
}
\arguments{
\item{RequestId}{[required] The ID of the quota increase request.}
}
\description{
Retrieves information about the specified quota increase request.

See \url{https://paws-r.github.io/docs/servicequotas/get_requested_service_quota_change.html} for full documentation.
}
\keyword{internal}
