% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_get_service_quota}
\alias{servicequotas_get_service_quota}
\title{Retrieves the applied quota value for the specified quota}
\usage{
servicequotas_get_service_quota(ServiceCode, QuotaCode)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{QuotaCode}{[required] The quota identifier.}
}
\description{
Retrieves the applied quota value for the specified quota. For some quotas, only the default values are available. If the applied quota value is not available for a quota, the quota is not retrieved.

See \url{https://paws-r.github.io/docs/servicequotas/get_service_quota.html} for full documentation.
}
\keyword{internal}
