% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_service_quotas}
\alias{servicequotas_list_service_quotas}
\title{Lists the applied quota values for the specified AWS service}
\usage{
servicequotas_list_service_quotas(ServiceCode, NextToken, MaxResults)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, if any, make another call with the token returned
from this call.}
}
\description{
Lists the applied quota values for the specified AWS service. For some
quotas, only the default values are available. If the applied quota
value is not available for a quota, the quota is not retrieved.
}
\section{Request syntax}{
\preformatted{svc$list_service_quotas(
  ServiceCode = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
