% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_services}
\alias{servicequotas_list_services}
\title{Lists the names and codes for the services integrated with Service
Quotas}
\usage{
servicequotas_list_services(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, if any, make another call with the token returned
from this call.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Services = list(
    list(
      ServiceCode = "string",
      ServiceName = "string"
    )
  )
)
}
}
\description{
Lists the names and codes for the services integrated with Service
Quotas.
}
\section{Request syntax}{
\preformatted{svc$list_services(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
