% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_request_service_quota_increase}
\alias{servicequotas_request_service_quota_increase}
\title{Submits a quota increase request for the specified quota}
\usage{
servicequotas_request_service_quota_increase(
  ServiceCode,
  QuotaCode,
  DesiredValue
)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{QuotaCode}{[required] The quota identifier.}

\item{DesiredValue}{[required] The new, increased value for the quota.}
}
\description{
Submits a quota increase request for the specified quota.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_request_service_quota_increase/} for full documentation.
}
\keyword{internal}
