% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_request_service_quota_increase}
\alias{servicequotas_request_service_quota_increase}
\title{Submits a quota increase request for the specified quota}
\usage{
servicequotas_request_service_quota_increase(
  ServiceCode,
  QuotaCode,
  DesiredValue,
  ContextId = NULL
)
}
\arguments{
\item{ServiceCode}{[required] Specifies the service identifier. To find the service code value for an
Amazon Web Services service, use the
\code{\link[=servicequotas_list_services]{list_services}} operation.}

\item{QuotaCode}{[required] Specifies the quota identifier. To find the quota code for a specific
quota, use the
\code{\link[=servicequotas_list_service_quotas]{list_service_quotas}} operation,
and look for the \code{QuotaCode} response in the output for the quota you
want.}

\item{DesiredValue}{[required] Specifies the new, increased value for the quota.}

\item{ContextId}{Specifies the Amazon Web Services account or resource to which the quota
applies. The value in this field depends on the context scope associated
with the specified service quota.}
}
\description{
Submits a quota increase request for the specified quota.

See \url{https://www.paws-r-sdk.com/docs/servicequotas_request_service_quota_increase/} for full documentation.
}
\keyword{internal}
