% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_associate_ops_item_related_item}
\alias{ssm_associate_ops_item_related_item}
\title{Associates a related resource to a Systems Manager OpsCenter OpsItem}
\usage{
ssm_associate_ops_item_related_item(OpsItemId, AssociationType,
  ResourceType, ResourceUri)
}
\arguments{
\item{OpsItemId}{[required] The ID of the OpsItem to which you want to associate a resource as a
related item.}

\item{AssociationType}{[required] The type of association that you want to create between an OpsItem and a
resource. OpsCenter supports \code{IsParentOf} and \code{RelatesTo} association
types.}

\item{ResourceType}{[required] The type of resource that you want to associate with an OpsItem.
OpsCenter supports the following types:

\verb{AWS::SSMIncidents::IncidentRecord}: an Incident Manager incident.
Incident Manager is a capability of AWS Systems Manager.

\verb{AWS::SSM::Document}: a Systems Manager (SSM) document.}

\item{ResourceUri}{[required] The Amazon Resource Name (ARN) of the AWS resource that you want to
associate with the OpsItem.}
}
\value{
A list with the following syntax:\preformatted{list(
  AssociationId = "string"
)
}
}
\description{
Associates a related resource to a Systems Manager OpsCenter OpsItem.
For example, you can associate an Incident Manager incident or analysis
with an OpsItem. Incident Manager is a capability of AWS Systems
Manager.
}
\section{Request syntax}{
\preformatted{svc$associate_ops_item_related_item(
  OpsItemId = "string",
  AssociationType = "string",
  ResourceType = "string",
  ResourceUri = "string"
)
}
}

\keyword{internal}
