% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_maintenance_window}
\alias{ssm_delete_maintenance_window}
\title{Deletes a maintenance window}
\usage{
ssm_delete_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window to delete.}
}
\description{
Deletes a maintenance window.

See \url{https://paws-r.github.io/docs/ssm/delete_maintenance_window.html} for full documentation.
}
\keyword{internal}
