% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_maintenance_window}
\alias{ssm_delete_maintenance_window}
\title{Deletes a maintenance window}
\usage{
ssm_delete_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  WindowId = "string"
)
}
}
\description{
Deletes a maintenance window.
}
\section{Request syntax}{
\preformatted{svc$delete_maintenance_window(
  WindowId = "string"
)
}
}

\keyword{internal}
