% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_parameters}
\alias{ssm_delete_parameters}
\title{Delete a list of parameters}
\usage{
ssm_delete_parameters(Names)
}
\arguments{
\item{Names}{[required] The names of the parameters to delete.}
}
\description{
Delete a list of parameters.
}
\section{Request syntax}{
\preformatted{svc$delete_parameters(
  Names = list(
    "string"
  )
)
}
}

\keyword{internal}
