% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_parameters}
\alias{ssm_delete_parameters}
\title{Delete a list of parameters}
\usage{
ssm_delete_parameters(Names)
}
\arguments{
\item{Names}{[required] The names of the parameters to delete. After deleting a parameter, wait
for at least 30 seconds to create a parameter with the same name.}
}
\description{
Delete a list of parameters. After deleting a parameter, wait for at least 30 seconds to create a parameter with the same name.

See \url{https://www.paws-r-sdk.com/docs/ssm_delete_parameters/} for full documentation.
}
\keyword{internal}
