% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_patch_baseline}
\alias{ssm_delete_patch_baseline}
\title{Deletes a patch baseline}
\usage{
ssm_delete_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to delete.}
}
\description{
Deletes a patch baseline.

See \url{https://www.paws-r-sdk.com/docs/ssm_delete_patch_baseline/} for full documentation.
}
\keyword{internal}
