% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_patch_baseline}
\alias{ssm_delete_patch_baseline}
\title{Deletes a patch baseline}
\usage{
ssm_delete_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BaselineId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a patch baseline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_patch_baseline(
  BaselineId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
