% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_resource_data_sync}
\alias{ssm_delete_resource_data_sync}
\title{Deletes a Resource Data Sync configuration}
\usage{
ssm_delete_resource_data_sync(SyncName)
}
\arguments{
\item{SyncName}{[required] The name of the configuration to delete.}
}
\description{
Deletes a Resource Data Sync configuration. After the configuration is
deleted, changes to inventory data on managed instances are no longer
synced with the target Amazon S3 bucket. Deleting a sync configuration
does not delete data in the target Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{svc$delete_resource_data_sync(
  SyncName = "string"
)
}
}

\keyword{internal}
