% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_resource_policy}
\alias{ssm_delete_resource_policy}
\title{Deletes a Systems Manager resource policy}
\usage{
ssm_delete_resource_policy(ResourceArn, PolicyId, PolicyHash)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the resource to which the policies are
attached.}

\item{PolicyId}{[required] The policy ID.}

\item{PolicyHash}{[required] ID of the current policy version. The hash helps to prevent multiple
calls from attempting to overwrite a policy.}
}
\description{
Deletes a Systems Manager resource policy. A resource policy helps you to define the IAM entity (for example, an Amazon Web Services account) that can manage your Systems Manager resources. Currently, \code{OpsItemGroup} is the only resource that supports Systems Manager resource policies. The resource policy for \code{OpsItemGroup} enables Amazon Web Services accounts to view and interact with OpsCenter operational work items (OpsItems).

See \url{https://www.paws-r-sdk.com/docs/ssm_delete_resource_policy/} for full documentation.
}
\keyword{internal}
