% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_deregister_task_from_maintenance_window}
\alias{ssm_deregister_task_from_maintenance_window}
\title{Removes a task from a maintenance window}
\usage{
ssm_deregister_task_from_maintenance_window(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window the task should be removed from.}

\item{WindowTaskId}{[required] The ID of the task to remove from the maintenance window.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  WindowId = "string",
  WindowTaskId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Removes a task from a maintenance window.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$deregister_task_from_maintenance_window(
  WindowId = "string",
  WindowTaskId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
