% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_activations}
\alias{ssm_describe_activations}
\title{Describes details about the activation, such as the date and time the
activation was created, its expiration date, the Identity and Access
Management (IAM) role assigned to the managed nodes in the activation,
and the number of nodes registered by using this activation}
\usage{
ssm_describe_activations(Filters = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Filters}{A filter to view information about your activations.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Describes details about the activation, such as the date and time the activation was created, its expiration date, the Identity and Access Management (IAM) role assigned to the managed nodes in the activation, and the number of nodes registered by using this activation.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_activations/} for full documentation.
}
\keyword{internal}
