% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_association_executions}
\alias{ssm_describe_association_executions}
\title{Views all executions for a specific association ID}
\usage{
ssm_describe_association_executions(
  AssociationId,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssociationId}{[required] The association ID for which you want to view execution history details.}

\item{Filters}{Filters for the request. You can specify the following filters and
values.

ExecutionId (EQUAL)

Status (EQUAL)

CreatedTime (EQUAL, GREATER_THAN, LESS_THAN)}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Views all executions for a specific association ID.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_association_executions/} for full documentation.
}
\keyword{internal}
