% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_document}
\alias{ssm_describe_document}
\title{Describes the specified Amazon Web Services Systems Manager document
(SSM document)}
\usage{
ssm_describe_document(Name, DocumentVersion = NULL, VersionName = NULL)
}
\arguments{
\item{Name}{[required] The name of the SSM document.

If you're calling a shared SSM document from a different Amazon Web
Services account, \code{Name} is the full Amazon Resource Name (ARN) of the
document.}

\item{DocumentVersion}{The document version for which you want information. Can be a specific
version or the default version.}

\item{VersionName}{An optional field specifying the version of the artifact associated with
the document. For example, 12.6. This value is unique across all
versions of a document, and can't be changed.}
}
\description{
Describes the specified Amazon Web Services Systems Manager document (SSM document).

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_document/} for full documentation.
}
\keyword{internal}
