% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_instance_associations_status}
\alias{ssm_describe_instance_associations_status}
\title{The status of the associations for the managed node(s)}
\usage{
ssm_describe_instance_associations_status(
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The managed node IDs for which you want association status information.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
The status of the associations for the managed node(s).

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_instance_associations_status/} for full documentation.
}
\keyword{internal}
