% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_instance_information}
\alias{ssm_describe_instance_information}
\title{Describes one or more of your instances, including information about the
operating system platform, the version of SSM Agent installed on the
instance, instance status, and so on}
\usage{
ssm_describe_instance_information(InstanceInformationFilterList,
  Filters, MaxResults, NextToken)
}
\arguments{
\item{InstanceInformationFilterList}{This is a legacy method. We recommend that you don't use this method.
Instead, use the \code{Filters} data type. \code{Filters} enables you to return
instance information by filtering based on tags applied to managed
instances.

Attempting to use \code{InstanceInformationFilterList} and \code{Filters} leads to
an exception error.}

\item{Filters}{One or more filters. Use a filter to return a more specific list of
instances. You can filter based on tags applied to EC2 instances. Use
this \code{Filters} data type instead of \code{InstanceInformationFilterList},
which is deprecated.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\value{
A list with the following syntax:\preformatted{list(
  InstanceInformationList = list(
    list(
      InstanceId = "string",
      PingStatus = "Online"|"ConnectionLost"|"Inactive",
      LastPingDateTime = as.POSIXct(
        "2015-01-01"
      ),
      AgentVersion = "string",
      IsLatestVersion = TRUE|FALSE,
      PlatformType = "Windows"|"Linux",
      PlatformName = "string",
      PlatformVersion = "string",
      ActivationId = "string",
      IamRole = "string",
      RegistrationDate = as.POSIXct(
        "2015-01-01"
      ),
      ResourceType = "ManagedInstance"|"Document"|"EC2Instance",
      Name = "string",
      IPAddress = "string",
      ComputerName = "string",
      AssociationStatus = "string",
      LastAssociationExecutionDate = as.POSIXct(
        "2015-01-01"
      ),
      LastSuccessfulAssociationExecutionDate = as.POSIXct(
        "2015-01-01"
      ),
      AssociationOverview = list(
        DetailedStatus = "string",
        InstanceAssociationStatusAggregatedCount = list(
          123
        )
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Describes one or more of your instances, including information about the
operating system platform, the version of SSM Agent installed on the
instance, instance status, and so on.

If you specify one or more instance IDs, it returns information for
those instances. If you do not specify instance IDs, it returns
information for all your instances. If you specify an instance ID that
is not valid or an instance that you do not own, you receive an error.

The IamRole field for this API action is the Amazon Identity and Access
Management (IAM) role assigned to on-premises instances. This call does
not return the IAM role for EC2 instances.
}
\section{Request syntax}{
\preformatted{svc$describe_instance_information(
  InstanceInformationFilterList = list(
    list(
      key = "InstanceIds"|"AgentVersion"|"PingStatus"|"PlatformTypes"|"ActivationIds"|"IamRole"|"ResourceType"|"AssociationStatus",
      valueSet = list(
        "string"
      )
    )
  ),
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
