% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_instance_patch_states}
\alias{ssm_describe_instance_patch_states}
\title{Retrieves the high-level patch state of one or more managed nodes}
\usage{
ssm_describe_instance_patch_states(
  InstanceIds,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceIds}{[required] The ID of the managed node for which patch state information should be
retrieved.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of managed nodes to return (per page).}
}
\description{
Retrieves the high-level patch state of one or more managed nodes.

See \url{https://www.paws-r-sdk.com/docs/ssm_describe_instance_patch_states/} for full documentation.
}
\keyword{internal}
