% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_maintenance_window_tasks}
\alias{ssm_describe_maintenance_window_tasks}
\title{Lists the tasks in a maintenance window}
\usage{
ssm_describe_maintenance_window_tasks(WindowId, Filters, MaxResults,
  NextToken)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window whose tasks should be retrieved.}

\item{Filters}{Optional filters used to narrow down the scope of the returned tasks.
The supported filter keys are WindowTaskId, TaskArn, Priority, and
TaskType.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Lists the tasks in a maintenance window.
}
\details{
For maintenance window tasks without a specified target, you cannot
supply values for \code{--max-errors} and \code{--max-concurrency}. Instead, the
system inserts a placeholder value of \code{1}, which may be reported in the
response to this command. These values do not affect the running of your
task and can be ignored.
}
\section{Request syntax}{
\preformatted{svc$describe_maintenance_window_tasks(
  WindowId = "string",
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
