% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_parameters}
\alias{ssm_describe_parameters}
\title{Get information about a parameter}
\usage{
ssm_describe_parameters(
  Filters = NULL,
  ParameterFilters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{This data type is deprecated. Instead, use \code{ParameterFilters}.}

\item{ParameterFilters}{Filters to limit the request results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Get information about a parameter.

See \url{https://paws-r.github.io/docs/ssm/describe_parameters.html} for full documentation.
}
\keyword{internal}
