% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_baselines}
\alias{ssm_describe_patch_baselines}
\title{Lists the patch baselines in your AWS account}
\usage{
ssm_describe_patch_baselines(Filters, MaxResults, NextToken)
}
\arguments{
\item{Filters}{Each element in the array is a structure containing:

Key: (string, "NAME_PREFIX" or "OWNER")

Value: (array of strings, exactly 1 entry, between 1 and 255 characters)}

\item{MaxResults}{The maximum number of patch baselines to return (per page).}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\value{
A list with the following syntax:\preformatted{list(
  BaselineIdentities = list(
    list(
      BaselineId = "string",
      BaselineName = "string",
      OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"|"MACOS",
      BaselineDescription = "string",
      DefaultBaseline = TRUE|FALSE
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the patch baselines in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_patch_baselines(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
