% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_group_state}
\alias{ssm_describe_patch_group_state}
\title{Returns high-level aggregated patch compliance state information for a
patch group}
\usage{
ssm_describe_patch_group_state(PatchGroup)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch snapshot should be retrieved.}
}
\description{
Returns high-level aggregated patch compliance state information for a patch group.

See \url{https://paws-r.github.io/docs/ssm/describe_patch_group_state.html} for full documentation.
}
\keyword{internal}
