% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_groups}
\alias{ssm_describe_patch_groups}
\title{Lists all patch groups that have been registered with patch baselines}
\usage{
ssm_describe_patch_groups(MaxResults, Filters, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of patch groups to return (per page).}

\item{Filters}{One or more filters. Use a filter to return a more specific list of
results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Lists all patch groups that have been registered with patch baselines.
}
\section{Request syntax}{
\preformatted{svc$describe_patch_groups(
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}

\keyword{internal}
