% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_connection_status}
\alias{ssm_get_connection_status}
\title{Retrieves the Session Manager connection status for a managed node to
determine whether it is running and ready to receive Session Manager
connections}
\usage{
ssm_get_connection_status(Target)
}
\arguments{
\item{Target}{[required] The managed node ID.}
}
\description{
Retrieves the Session Manager connection status for a managed node to determine whether it is running and ready to receive Session Manager connections.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_connection_status/} for full documentation.
}
\keyword{internal}
