% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_default_patch_baseline}
\alias{ssm_get_default_patch_baseline}
\title{Retrieves the default patch baseline}
\usage{
ssm_get_default_patch_baseline(OperatingSystem = NULL)
}
\arguments{
\item{OperatingSystem}{Returns the default patch baseline for the specified operating system.}
}
\description{
Retrieves the default patch baseline. Amazon Web Services Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_default_patch_baseline/} for full documentation.
}
\keyword{internal}
