% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window}
\alias{ssm_get_maintenance_window}
\title{Retrieves a maintenance window}
\usage{
ssm_get_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the maintenance window for which you want to retrieve
information.}
}
\description{
Retrieves a maintenance window.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_maintenance_window/} for full documentation.
}
\keyword{internal}
