% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_parameters}
\alias{ssm_get_parameters}
\title{Get information about one or more parameters by specifying multiple
parameter names}
\usage{
ssm_get_parameters(Names, WithDecryption = NULL)
}
\arguments{
\item{Names}{[required] Names of the parameters for which you want to query information.

To query by parameter label, use \code{"Name": "name:label"}. To query by
parameter version, use \code{"Name": "name:version"}.}

\item{WithDecryption}{Return decrypted secure string value. Return decrypted values for secure
string parameters. This flag is ignored for \code{String} and \code{StringList}
parameter types.}
}
\description{
Get information about one or more parameters by specifying multiple parameter names.

See \url{https://www.paws-r-sdk.com/docs/ssm_get_parameters/} for full documentation.
}
\keyword{internal}
