% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_parameters}
\alias{ssm_get_parameters}
\title{Get details of a parameter}
\usage{
ssm_get_parameters(Names, WithDecryption)
}
\arguments{
\item{Names}{[required] Names of the parameters for which you want to query information.}

\item{WithDecryption}{Return decrypted secure string value. Return decrypted values for secure
string parameters. This flag is ignored for String and StringList
parameter types.}
}
\description{
Get details of a parameter. Don\'t confuse this API action with the
GetParameter API action.
}
\section{Request syntax}{
\preformatted{svc$get_parameters(
  Names = list(
    "string"
  ),
  WithDecryption = TRUE|FALSE
)
}
}

\keyword{internal}
