% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_patch_baseline_for_patch_group}
\alias{ssm_get_patch_baseline_for_patch_group}
\title{Retrieves the patch baseline that should be used for the specified patch
group}
\usage{
ssm_get_patch_baseline_for_patch_group(PatchGroup, OperatingSystem)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch baseline should be retrieved.}

\item{OperatingSystem}{Returns he operating system rule specified for patch groups using the
patch baseline.}
}
\value{
A list with the following syntax:\preformatted{list(
  BaselineId = "string",
  PatchGroup = "string",
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"|"MACOS"
)
}
}
\description{
Retrieves the patch baseline that should be used for the specified patch
group.
}
\section{Request syntax}{
\preformatted{svc$get_patch_baseline_for_patch_group(
  PatchGroup = "string",
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"|"ORACLE_LINUX"|"DEBIAN"|"MACOS"
)
}
}

\keyword{internal}
