% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_commands}
\alias{ssm_list_commands}
\title{Lists the commands requested by users of the AWS account}
\usage{
ssm_list_commands(CommandId, InstanceId, MaxResults, NextToken, Filters)
}
\arguments{
\item{CommandId}{(Optional) If provided, lists only the specified command.}

\item{InstanceId}{(Optional) Lists commands issued against this instance ID.}

\item{MaxResults}{(Optional) The maximum number of items to return for this call. The call
also returns a token that you can specify in a subsequent call to get
the next set of results.}

\item{NextToken}{(Optional) The token for the next set of items to return. (You received
this token from a previous call.)}

\item{Filters}{(Optional) One or more filters. Use a filter to return a more specific
list of results.}
}
\description{
Lists the commands requested by users of the AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_commands(
  CommandId = "string",
  InstanceId = "string",
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      key = "InvokedAfter"|"InvokedBefore"|"Status"|"ExecutionStage"|"DocumentName",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
