% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_document_metadata_history}
\alias{ssm_list_document_metadata_history}
\title{Information about approval reviews for a version of a change template in
Change Manager}
\usage{
ssm_list_document_metadata_history(
  Name,
  DocumentVersion = NULL,
  Metadata,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Name}{[required] The name of the change template.}

\item{DocumentVersion}{The version of the change template.}

\item{Metadata}{[required] The type of data for which details are being requested. Currently, the
only supported value is \code{DocumentReviews}.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}
}
\description{
Information about approval reviews for a version of a change template in Change Manager.

See \url{https://paws-r.github.io/docs/ssm/list_document_metadata_history.html} for full documentation.
}
\keyword{internal}
