% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_ops_item_related_items}
\alias{ssm_list_ops_item_related_items}
\title{Lists all related-item resources associated with an OpsItem}
\usage{
ssm_list_ops_item_related_items(OpsItemId, Filters, MaxResults,
  NextToken)
}
\arguments{
\item{OpsItemId}{The ID of the OpsItem for which you want to list all related-item
resources.}

\item{Filters}{One or more OpsItem filters. Use a filter to return a more specific list
of results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Summaries = list(
    list(
      OpsItemId = "string",
      AssociationId = "string",
      ResourceType = "string",
      AssociationType = "string",
      ResourceUri = "string",
      CreatedBy = list(
        Arn = "string"
      ),
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedBy = list(
        Arn = "string"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Lists all related-item resources associated with an OpsItem.
}
\section{Request syntax}{
\preformatted{svc$list_ops_item_related_items(
  OpsItemId = "string",
  Filters = list(
    list(
      Key = "ResourceType"|"AssociationId"|"ResourceUri",
      Values = list(
        "string"
      ),
      Operator = "Equal"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
