% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_list_ops_metadata}
\alias{ssm_list_ops_metadata}
\title{Systems Manager calls this API action when displaying all Application
Manager OpsMetadata objects or blobs}
\usage{
ssm_list_ops_metadata(Filters, MaxResults, NextToken)
}
\arguments{
\item{Filters}{One or more filters to limit the number of OpsMetadata objects returned
by the call.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\value{
A list with the following syntax:\preformatted{list(
  OpsMetadataList = list(
    list(
      ResourceId = "string",
      OpsMetadataArn = "string",
      LastModifiedDate = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedUser = "string",
      CreationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Systems Manager calls this API action when displaying all Application
Manager OpsMetadata objects or blobs.
}
\section{Request syntax}{
\preformatted{svc$list_ops_metadata(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
