% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_put_inventory}
\alias{ssm_put_inventory}
\title{Bulk update custom inventory items on one more instance}
\usage{
ssm_put_inventory(InstanceId, Items)
}
\arguments{
\item{InstanceId}{[required] An instance ID where you want to add or update inventory items.}

\item{Items}{[required] The inventory items that you want to add or update on instances.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Message = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Bulk update custom inventory items on one more instance. The request
adds an inventory item, if it doesn't already exist, or updates an
inventory item, if it does exist.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_inventory(
  InstanceId = "string",
  Items = list(
    list(
      TypeName = "string",
      SchemaVersion = "string",
      CaptureTime = "string",
      ContentHash = "string",
      Content = list(
        list(
          "string"
        )
      ),
      Context = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
