% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_register_default_patch_baseline}
\alias{ssm_register_default_patch_baseline}
\title{Defines the default patch baseline for the relevant operating system}
\usage{
ssm_register_default_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline that should be the default patch baseline.}
}
\description{
Defines the default patch baseline for the relevant operating system.

See \url{https://paws-r.github.io/docs/ssm/register_default_patch_baseline.html} for full documentation.
}
\keyword{internal}
