% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_register_patch_baseline_for_patch_group}
\alias{ssm_register_patch_baseline_for_patch_group}
\title{Registers a patch baseline for a patch group}
\usage{
ssm_register_patch_baseline_for_patch_group(BaselineId, PatchGroup)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to register the patch group with.}

\item{PatchGroup}{[required] The name of the patch group that should be registered with the patch
baseline.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BaselineId = "string",
  PatchGroup = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Registers a patch baseline for a patch group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_patch_baseline_for_patch_group(
  BaselineId = "string",
  PatchGroup = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
