% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_register_patch_baseline_for_patch_group}
\alias{ssm_register_patch_baseline_for_patch_group}
\title{Registers a patch baseline for a patch group}
\usage{
ssm_register_patch_baseline_for_patch_group(BaselineId, PatchGroup)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to register with the patch group.}

\item{PatchGroup}{[required] The name of the patch group to be registered with the patch baseline.}
}
\description{
Registers a patch baseline for a patch group.

See \url{https://paws-r.github.io/docs/ssm/register_patch_baseline_for_patch_group.html} for full documentation.
}
\keyword{internal}
