% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_start_session}
\alias{ssm_start_session}
\title{Initiates a connection to a target (for example, an instance) for a
Session Manager session}
\usage{
ssm_start_session(Target, DocumentName, Parameters)
}
\arguments{
\item{Target}{[required] The instance to connect to for the session.}

\item{DocumentName}{The name of the SSM document to define the parameters and plugin
settings for the session. For example, \code{SSM-SessionManagerRunShell}. You
can call the \code{\link[=ssm_get_document]{get_document}} API to verify the
document exists before attempting to start a session. If no document
name is provided, a shell to the instance is launched by default.}

\item{Parameters}{Reserved for future use.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SessionId = "string",
  TokenValue = "string",
  StreamUrl = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Initiates a connection to a target (for example, an instance) for a
Session Manager session. Returns a URL and token that can be used to
open a WebSocket connection for sending input and receiving outputs.

AWS CLI usage: \code{start-session} is an interactive command that requires
the Session Manager plugin to be installed on the client machine making
the call. For information, see \href{https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html}{Install the Session Manager plugin for the AWS CLI}
in the \emph{AWS Systems Manager User Guide}.

AWS Tools for PowerShell usage: Start-SSMSession is not currently
supported by AWS Tools for PowerShell on Windows local machines.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_session(
  Target = "string",
  DocumentName = "string",
  Parameters = list(
    list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
