% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_unlabel_parameter_version}
\alias{ssm_unlabel_parameter_version}
\title{Remove a label or labels from a parameter}
\usage{
ssm_unlabel_parameter_version(Name, ParameterVersion, Labels)
}
\arguments{
\item{Name}{[required] The parameter name of which you want to delete one or more labels.}

\item{ParameterVersion}{[required] The specific version of the parameter which you want to delete one or
more labels from. If it is not present, the call will fail.}

\item{Labels}{[required] One or more labels to delete from the specified parameter version.}
}
\value{
A list with the following syntax:\preformatted{list(
  RemovedLabels = list(
    "string"
  ),
  InvalidLabels = list(
    "string"
  )
)
}
}
\description{
Remove a label or labels from a parameter.
}
\section{Request syntax}{
\preformatted{svc$unlabel_parameter_version(
  Name = "string",
  ParameterVersion = 123,
  Labels = list(
    "string"
  )
)
}
}

\keyword{internal}
