% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_document_metadata}
\alias{ssm_update_document_metadata}
\title{Updates information related to approval reviews for a specific version
of a change template in Change Manager}
\usage{
ssm_update_document_metadata(Name, DocumentVersion = NULL, DocumentReviews)
}
\arguments{
\item{Name}{[required] The name of the change template for which a version's metadata is to be
updated.}

\item{DocumentVersion}{The version of a change template in which to update approval metadata.}

\item{DocumentReviews}{[required] The change template review details to update.}
}
\description{
Updates information related to approval reviews for a specific version of a change template in Change Manager.

See \url{https://paws-r.github.io/docs/ssm/update_document_metadata.html} for full documentation.
}
\keyword{internal}
