% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_ops_metadata}
\alias{ssm_update_ops_metadata}
\title{Systems Manager calls this API action when you edit OpsMetadata in
Application Manager}
\usage{
ssm_update_ops_metadata(OpsMetadataArn, MetadataToUpdate, KeysToDelete)
}
\arguments{
\item{OpsMetadataArn}{[required] The Amazon Resoure Name (ARN) of the OpsMetadata Object to update.}

\item{MetadataToUpdate}{Metadata to add to an OpsMetadata object.}

\item{KeysToDelete}{The metadata keys to delete from the OpsMetadata object.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OpsMetadataArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Systems Manager calls this API action when you edit OpsMetadata in
Application Manager.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_ops_metadata(
  OpsMetadataArn = "string",
  MetadataToUpdate = list(
    list(
      Value = "string"
    )
  ),
  KeysToDelete = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
