% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_contact}
\alias{ssmcontacts_delete_contact}
\title{To remove a contact from Incident Manager, you can delete the contact}
\usage{
ssmcontacts_delete_contact(ContactId)
}
\arguments{
\item{ContactId}{[required] The Amazon Resource Name (ARN) of the contact that you're deleting.}
}
\description{
To remove a contact from Incident Manager, you can delete the contact. However, deleting a contact does not remove it from escalation plans and related response plans. Deleting an escalation plan also does not remove it from all related response plans. To modify an escalation plan, we recommend using the \code{\link[=ssmcontacts_update_contact]{update_contact}} action to specify a different existing contact.

See \url{https://www.paws-r-sdk.com/docs/ssmcontacts_delete_contact/} for full documentation.
}
\keyword{internal}
