% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_describe_page}
\alias{ssmcontacts_describe_page}
\title{Lists details of the engagement to a contact channel}
\usage{
ssmcontacts_describe_page(PageId)
}
\arguments{
\item{PageId}{[required] The ID of the engagement to a contact channel.}
}
\description{
Lists details of the engagement to a contact channel.

See \url{https://paws-r.github.io/docs/ssmcontacts/describe_page.html} for full documentation.
}
\keyword{internal}
