% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_get_rotation}
\alias{ssmcontacts_get_rotation}
\title{Retrieves information about an on-call rotation}
\usage{
ssmcontacts_get_rotation(RotationId)
}
\arguments{
\item{RotationId}{[required] The Amazon Resource Name (ARN) of the on-call rotation to retrieve
information about.}
}
\description{
Retrieves information about an on-call rotation.

See \url{https://paws-r.github.io/docs/ssmcontacts/get_rotation.html} for full documentation.
}
\keyword{internal}
