% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_delete_resource_policy}
\alias{ssmincidents_delete_resource_policy}
\title{Deletes the resource policy that Resource Access Manager uses to share
your Incident Manager resource}
\usage{
ssmincidents_delete_resource_policy(policyId, resourceArn)
}
\arguments{
\item{policyId}{[required] The ID of the resource policy you're deleting.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource you're deleting the
policy from.}
}
\description{
Deletes the resource policy that Resource Access Manager uses to share your Incident Manager resource.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_delete_resource_policy/} for full documentation.
}
\keyword{internal}
