% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_resource_policies}
\alias{ssmincidents_get_resource_policies}
\title{Retrieves the resource policies attached to the specified response plan}
\usage{
ssmincidents_get_resource_policies(
  maxResults = NULL,
  nextToken = NULL,
  resourceArn
)
}
\arguments{
\item{maxResults}{The maximum number of resource policies to display for each page of
results.}

\item{nextToken}{The pagination token for the next set of items to return. (You received
this token from a previous call.)}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the response plan with the attached
resource policy.}
}
\description{
Retrieves the resource policies attached to the specified response plan.

See \url{https://www.paws-r-sdk.com/docs/ssmincidents_get_resource_policies/} for full documentation.
}
\keyword{internal}
