% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_list_replication_sets}
\alias{ssmincidents_list_replication_sets}
\title{Lists details about the replication set configured in your account}
\usage{
ssmincidents_list_replication_sets(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results per page.}

\item{nextToken}{The pagination token to continue to the next page of results.}
}
\description{
Lists details about the replication set configured in your account.

See \url{https://paws-r.github.io/docs/ssmincidents/list_replication_sets.html} for full documentation.
}
\keyword{internal}
