% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_update_related_items}
\alias{ssmincidents_update_related_items}
\title{Add or remove related items from the related items tab of an incident
record}
\usage{
ssmincidents_update_related_items(
  clientToken = NULL,
  incidentRecordArn,
  relatedItemsUpdate
)
}
\arguments{
\item{clientToken}{A token that ensures that a client calls the operation only once with
the specified details.}

\item{incidentRecordArn}{[required] The Amazon Resource Name (ARN) of the incident record that contains the
related items that you update.}

\item{relatedItemsUpdate}{[required] Details about the item that you are add to, or delete from, an incident.}
}
\description{
Add or remove related items from the related items tab of an incident record.

See \url{https://paws-r.github.io/docs/ssmincidents/update_related_items.html} for full documentation.
}
\keyword{internal}
