% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_attachment}
\alias{support_describe_attachment}
\title{Returns the attachment that has the specified ID}
\usage{
support_describe_attachment(attachmentId)
}
\arguments{
\item{attachmentId}{[required] The ID of the attachment to return. Attachment IDs are returned by the
\code{\link[=support_describe_communications]{describe_communications}} operation.}
}
\value{
A list with the following syntax:\preformatted{list(
  attachment = list(
    fileName = "string",
    data = raw
  )
)
}
}
\description{
Returns the attachment that has the specified ID. Attachments can
include screenshots, error logs, or other files that describe your
issue. Attachment IDs are generated by the case management system when
you add an attachment to a case or case communication. Attachment IDs
are returned in the AttachmentDetails objects that are returned by the
\code{\link[=support_describe_communications]{describe_communications}} operation.
\itemize{
\item You must have a Business or Enterprise support plan to use the AWS
Support API.
\item If you call the AWS Support API from an account that does not have a
Business or Enterprise support plan, the
\code{SubscriptionRequiredException} error message appears. For
information about changing your support plan, see \href{https://aws.amazon.com/premiumsupport/}{AWS Support}.
}
}
\section{Request syntax}{
\preformatted{svc$describe_attachment(
  attachmentId = "string"
)
}
}

\keyword{internal}
