% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_services}
\alias{support_describe_services}
\title{Returns the current list of Amazon Web Services services and a list of
service categories for each service}
\usage{
support_describe_services(serviceCodeList = NULL, language = NULL)
}
\arguments{
\item{serviceCodeList}{A JSON-formatted list of service codes available for Amazon Web Services
services.}

\item{language}{The ISO 639-1 code for the language in which Amazon Web Services
provides support. Amazon Web Services Support currently supports English
("en") and Japanese ("ja"). Language parameters must be passed
explicitly for operations that take them.}
}
\description{
Returns the current list of Amazon Web Services services and a list of service categories for each service. You then use service names and categories in your \code{\link[=support_create_case]{create_case}} requests. Each Amazon Web Services service has its own set of categories.

See \url{https://paws-r.github.io/docs/support/describe_services.html} for full documentation.
}
\keyword{internal}
