% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_check_result}
\alias{support_describe_trusted_advisor_check_result}
\title{Returns the results of the Trusted Advisor check that has the specified
check ID}
\usage{
support_describe_trusted_advisor_check_result(checkId, language = NULL)
}
\arguments{
\item{checkId}{[required] The unique identifier for the Trusted Advisor check.}

\item{language}{The ISO 639-1 code for the language that you want your check results to
appear in.

The Amazon Web Services Support API currently supports the following
languages for Trusted Advisor:
\itemize{
\item Chinese, Simplified - \code{zh}
\item Chinese, Traditional - \code{zh_TW}
\item English - \code{en}
\item French - \code{fr}
\item German - \code{de}
\item Indonesian - \code{id}
\item Italian - \code{it}
\item Japanese - \code{ja}
\item Korean - \code{ko}
\item Portuguese, Brazilian - \code{pt_BR}
\item Spanish - \code{es}
}}
}
\description{
Returns the results of the Trusted Advisor check that has the specified check ID. You can get the check IDs by calling the \code{\link[=support_describe_trusted_advisor_checks]{describe_trusted_advisor_checks}} operation.

See \url{https://www.paws-r-sdk.com/docs/support_describe_trusted_advisor_check_result/} for full documentation.
}
\keyword{internal}
