% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_check_result}
\alias{support_describe_trusted_advisor_check_result}
\title{Returns the results of the Trusted Advisor check that has the specified
check ID}
\usage{
support_describe_trusted_advisor_check_result(checkId, language = NULL)
}
\arguments{
\item{checkId}{[required] The unique identifier for the Trusted Advisor check.}

\item{language}{The ISO 639-1 code for the language in which Amazon Web Services
provides support. Amazon Web Services Support currently supports English
("en") and Japanese ("ja"). Language parameters must be passed
explicitly for operations that take them.}
}
\description{
Returns the results of the Trusted Advisor check that has the specified check ID. You can get the check IDs by calling the \code{\link[=support_describe_trusted_advisor_checks]{describe_trusted_advisor_checks}} operation.

See \url{https://paws-r.github.io/docs/support/describe_trusted_advisor_check_result.html} for full documentation.
}
\keyword{internal}
