% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_refresh_trusted_advisor_check}
\alias{support_refresh_trusted_advisor_check}
\title{Refreshes the Trusted Advisor check that you specify using the check ID}
\usage{
support_refresh_trusted_advisor_check(checkId)
}
\arguments{
\item{checkId}{[required] The unique identifier for the Trusted Advisor check to refresh.

Specifying the check ID of a check that is automatically refreshed
causes an \code{InvalidParameterValue} error.}
}
\description{
Refreshes the Trusted Advisor check that you specify using the check ID. You can get the check IDs by calling the \code{\link[=support_describe_trusted_advisor_checks]{describe_trusted_advisor_checks}} operation.

See \url{https://www.paws-r-sdk.com/docs/support_refresh_trusted_advisor_check/} for full documentation.
}
\keyword{internal}
