% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_resolve_case}
\alias{support_resolve_case}
\title{Resolves a support case}
\usage{
support_resolve_case(caseId = NULL)
}
\arguments{
\item{caseId}{The support case ID requested or returned in the call. The case ID is an
alphanumeric string formatted as shown in this example:
case-\emph{12345678910-2013-c4c1d2bf33c5cf47}}
}
\description{
Resolves a support case. This operation takes a \code{caseId} and returns the initial and final state of the case.

See \url{https://paws-r.github.io/docs/support/resolve_case.html} for full documentation.
}
\keyword{internal}
