% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_get_canary_runs}
\alias{synthetics_get_canary_runs}
\title{Retrieves a list of runs for a specified canary}
\usage{
synthetics_get_canary_runs(Name, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to see runs for.}

\item{NextToken}{A token that indicates that there is more data available. You can use
this token in a subsequent
\code{\link[=synthetics_get_canary_runs]{get_canary_runs}} operation to retrieve
the next set of results.}

\item{MaxResults}{Specify this parameter to limit how many runs are returned each time you
use the \code{\link[=synthetics_get_canary_runs]{get_canary_runs}} operation. If
you omit this parameter, the default of 100 is used.}
}
\description{
Retrieves a list of runs for a specified canary.

See \url{https://paws-r.github.io/docs/synthetics/get_canary_runs.html} for full documentation.
}
\keyword{internal}
