% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_start_canary_dry_run}
\alias{synthetics_start_canary_dry_run}
\title{Use this operation to start a dry run for a canary that has already been
created}
\usage{
synthetics_start_canary_dry_run(
  Name,
  Code = NULL,
  RuntimeVersion = NULL,
  RunConfig = NULL,
  VpcConfig = NULL,
  ExecutionRoleArn = NULL,
  SuccessRetentionPeriodInDays = NULL,
  FailureRetentionPeriodInDays = NULL,
  VisualReference = NULL,
  ArtifactS3Location = NULL,
  ArtifactConfig = NULL,
  ProvisionedResourceCleanup = NULL
)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to dry run. To find canary names,
use \code{\link[=synthetics_describe_canaries]{describe_canaries}}.}

\item{Code}{}

\item{RuntimeVersion}{Specifies the runtime version to use for the canary. For a list of valid
runtime versions and for more information about runtime versions, see
\href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Synthetics_Canaries_Library.html}{Canary Runtime Versions}.}

\item{RunConfig}{}

\item{VpcConfig}{}

\item{ExecutionRoleArn}{The ARN of the IAM role to be used to run the canary. This role must
already exist, and must include \code{lambda.amazonaws.com} as a principal in
the trust policy. The role must also have the following permissions:}

\item{SuccessRetentionPeriodInDays}{The number of days to retain data on the failed runs for this canary.
The valid range is 1 to 455 days.

This setting affects the range of information returned by
\code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of
information displayed in the Synthetics console.}

\item{FailureRetentionPeriodInDays}{The number of days to retain data on the failed runs for this canary.
The valid range is 1 to 455 days.

This setting affects the range of information returned by
\code{\link[=synthetics_get_canary_runs]{get_canary_runs}}, as well as the range of
information displayed in the Synthetics console.}

\item{VisualReference}{}

\item{ArtifactS3Location}{The location in Amazon S3 where Synthetics stores artifacts from the
test runs of this canary. Artifacts include the log file, screenshots,
and HAR files. The name of the Amazon S3 bucket can't include a period
(.).}

\item{ArtifactConfig}{}

\item{ProvisionedResourceCleanup}{Specifies whether to also delete the Lambda functions and layers used by
this canary when the canary is deleted. If the value of this parameter
is \code{AUTOMATIC}, it means that the Lambda functions and layers will be
deleted when the canary is deleted.

If the value of this parameter is \code{OFF}, then the value of the
\code{DeleteLambda} parameter of the
\code{\link[=synthetics_delete_canary]{delete_canary}} operation determines whether
the Lambda functions and layers will be deleted.}
}
\description{
Use this operation to start a dry run for a canary that has already been created

See \url{https://www.paws-r-sdk.com/docs/synthetics_start_canary_dry_run/} for full documentation.
}
\keyword{internal}
